% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_multinomial.R
\name{di_mpmf}
\alias{di_mpmf}
\title{Multinomial Probability Mass Function}
\usage{
di_mpmf(F, P, method = "loggamma")
}
\arguments{
\item{F}{list with the observed counts}

\item{P}{list with the probabilities for each category}

\item{method}{optional the calculation method to use. Either "loggamma" (default), "factorial", "gamma", "mprob".}
}
\value{
A float with the requested probability
}
\description{
This is a function for the multinomial probability. It returns the probability of a distribution as given in F for a sample size of sum of F, where the probability for each category is given as in P. It is a generalization of the binomial distribution.
}
\details{
If \emph{method=factorial} the following formula is used:
\deqn{mpmf\left(F, P\right) = \frac{n!}{\prod_{i=1}^{k} \left(F_i!\right)} \times \prod_{i=1}^{k} P_i^{F_i}}
This formula was most likely already used by for example Edgeworth (1905), but can for example also be found in Berry and Mielke (1995, p. 769)

If \emph{method=gamma}:
\deqn{mpmf\left(F, P\right) = \frac{\Gamma\left(1+n\right)}{\prod_{i=1}^n \Gamma\left(1+F_i\right)} \times \prod_{i=1}^{k} P_i^{F_i}}
If \emph{method=loggamma}:
\deqn{mpmf\left(F, P\right) = e^{\ln\left(mpmf\left(F, P\right)\right)}}
\deqn{\ln\left(mpmf\left(F, P\right)\right) = \ln\left(\Gamma\left(n + 1\right)\right) + \sum_{i=1}^k F_i\times\ln\left(P_i\right) - \ln\left(\Gamma\left(F_i + 1\right)\right)}
This formula can for example be found in Arnold (2018).

If \emph{method=mprob} the algorithm from García-Pérez (1999) is used:
\enumerate{
\item Determine \eqn{F^*}, the counts in descending order, and move the elements in \eqn{P} accordingly creating \eqn{P^*}.
\item Set \eqn{pmf = 1}, \eqn{t=P_1^*}, \eqn{i=2}, \eqn{x=0}, and \eqn{m=F_1^*}
\item Set \eqn{l = F_i^*}. For \eqn{r=1} to \eqn{l} do:
\itemize{
\item update \eqn{x = x + 1}
\item if \eqn{x > F_1^*} then set \eqn{t = 1} (else nothing)
\item update \eqn{pmf = pmf\times t\times P_i^*\times \frac{r + m}{r}}
}
\item If \eqn{i=k}, then go to step 5, otherwise update \eqn{i = i + 1}, \eqn{m = m + F_i^*} and go to step 3
\item If \eqn{x < F_1^*} then for \eqn{r=x + 1} to \eqn{F_1^*} update \eqn{pmf = pmf\times P_1^*}
}
}
\references{
Arnold, J. (2018, December 3). Maximum Likelihood for the multinomial distribution (bag of words) \link{Blog}. Jakuba. https://blog.jakuba.net/maximum-likelihood-for-multinomial-distribution/

Berry, K. J., & Mielke, P. W. (1995). Exact cumulative probabilities for the multinomial distribution. \emph{Educational and Psychological Measurement, 55}(5), 769–772. doi:10.1177/0013164495055005008

Edgeworth, F. Y. (1905). The law of error. \emph{Transactions of the Cambridge Philosophical Society, 20}, 36–66.

García-Pérez, M. A. (1999). MPROB: Computation of multinomial probabilities. \emph{Behavior Research Methods, Instruments, & Computers, 31}(4), 701–705. doi:10.3758/BF03200749
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
