% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mann_whitney_wilcoxon.R
\name{di_mwwcdf}
\alias{di_mwwcdf}
\title{Mann-Whitney-Wilcoxon Cumulative Distribution Function}
\usage{
di_mwwcdf(u, n1, n2)
}
\arguments{
\item{u}{int, the U test statistic}

\item{n1}{int, the sample size of the first category}

\item{n2}{int the sample size of the second category}
}
\value{
p : the cumulative probability
}
\description{
This function returns the cumulative probability for the specified U statistic, given n1 and n2 cases in each category.

It first uses the di_mwwd function to determine the distribution up to the specified u value, sums these results and divides it by the total number of possible arrangements.
}
\details{
ee the details in di_mwwd() on how the frequency distribution is determined. The sum of these is then divided by the total number of possibilities, which is the number of ways we can choose $n_1$ items out of $n$, without replacement. This is the binomial coefficient, or number of combinations:
\deqn{C(n, n_1) = nCr(n, n_1) = \binom{n}{n_1} = \frac{n!}{n_1!\times\left(n - n_1\right)!}}

To convert a W statistic to a U statistic use:
\deqn{U = W - \frac{n_1\times\left(n_1 + 1\right)}{2}}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
