% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mann_whitney_wilcoxon.R
\name{di_mwwd}
\alias{di_mwwd}
\title{Mann-Whitney-Wilcoxon Distribution}
\usage{
di_mwwd(u, n1, n2)
}
\arguments{
\item{u}{int, the U test statistic}

\item{n1}{int, the sample size of the first category}

\item{n2}{int the sample size of the second category}
}
\value{
result : a list with the counts starting with the count for U=0
}
\description{
This distribution is also referred to as a permutation distribution.

It is used in the Mann-Whitney U and Wilcoxon Rank Sum test.

In this version the U-statistic is used as input, and the sample sizes of each of the two categories. This function will return the counts (frequency) of each possible U value from 0 till the provided u value. If all possible values need to be shown, simply set u = i*j.
}
\details{
A recursive formula is used:
\deqn{f_{n_1, n_2}(U) = \begin{cases} 0 & \text{ if } U < 0 \text{ or } U > n_1\times n_2 \\ 1 & \text{ if } (n_1=1 \text{ or } n_2=1) \text{ and } 0 \leq U \leq n_1\times n_2 \\ f_{n_1, n_2-1}(U) + f_{n_1-1, n_2}(U - n_2) & \text{ else } \end{cases}}

This formula is found in Mann and Whitney (1947, p. 51), Dinneen and Blakesley 1973, p. 269) and described also in Festinger (1946).

To convert a W statistic to a U statistic use:
\deqn{U = W - \frac{n_1\times\left(n_1 + 1\right)}{2}}
}
\references{
Dinneen, L. C., & Blakesley, B. C. (1973). Algorithm AS 62: A generator for the sampling distribution of the Mann- Whitney U statistic. \emph{Journal of the Royal Statistical Society. Series C (Applied Statistics), 22}(2), 269–273. doi:10.2307/2346934

Festinger, L. (1946). The significance of difference between means without reference to the frequency distribution function. \emph{Psychometrika, 11}(2), 97–105. doi:10.1007/BF02288926

Mann, H. B., & Whitney, D. R. (1947). On a Test of Whether one of Two Random Variables is Stochastically Larger than the Other. \emph{The Annals of Mathematical Statistics, 18}(1), 50–60. doi:10.1214/aoms/1177730491
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
