% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mann_whitney_wilcoxon.R
\name{di_mwwpmf}
\alias{di_mwwpmf}
\title{Mann-Whitney-Wilcoxon Probability Mass Function}
\usage{
di_mwwpmf(u, n1, n2)
}
\arguments{
\item{u}{int, the U test statistic}

\item{n1}{int, the sample size of the first category}

\item{n2}{int the sample size of the second category}
}
\value{
result : a list with the counts starting with the count for U=0
}
\description{
This function returns the probability for the specified U statistic, given n1 and n2 cases in each category.

It first uses the di_mwwf function to determine the count for the u value, and divides it by the total number of possible arrangements.

The dwilcox() function from R's stats library does the same, and is probably more optimized than this function.
}
\details{
See the details in di_mwwf() on how the frequency is determined. This is then divided by the total number of possibilities, which is the number of ways we can choose $n_1$ items out of $n$, without replacement. This is the binomial coefficient, or number of combinations:
\deqn{C(n, n_1) = nCr(n, n_1) = \binom{n}{n_1} = \frac{n!}{n_1!\times\left(n - n_1\right)!}}

To convert a W statistic to a U statistic use:
\deqn{U = W - \frac{n_1\times\left(n_1 + 1\right)}{2}}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
