% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_wilcoxon.R
\name{di_wcdf}
\alias{di_wcdf}
\title{Wilcoxon Cumulative Distribution Function}
\usage{
di_wcdf(T, n, method = "shift")
}
\arguments{
\item{T}{int with the sum of ranks}

\item{n}{int with the sample size}

\item{method}{optional the calculation method to use. Either "shift" (default), "enumerate", "recursive".}
}
\value{
A float with the requested probability
}
\description{
This function will give the cumulative probability of a sum of ranks of T, given a sample size of n.
}
\details{
The enumeration method will create all possible combinations of ranks 1 to n, sum each of these, and then determines the count of each unique sum of ranks. It then uses this to determine the probability and cumulative probabilities.

The recursive method uses the formula from McCornack (1965, p. 864):
\deqn{srf\left(x,y\right)=\begin{cases} 0 & x < 0 \\ 0 & x > \binom{y+1}{2} \\ 1 & y=1 \wedge \left( x=0\vee x=1 \right) \\ srf^*\left(x,y\right) & y\geq 0 \end{cases}}

with:
\deqn{srf^*\left(x,y\right) = srf\left(x-y,y-1\right) + srf\left(x,y-1\right)}

The shift-algorithm from Streitberg and Röhmel (1987), and can also be found in Munzel and Brunner (2002). This works as follows.
\itemize{
\item Start with listing all values from 0 to the maximum possible sum of ranks, so 0 to (n×(n+1))/2
\item Create a vector with the value 1 followed by n times a 0.
\item Create a shifted vector by moving all values by 1.
\item Add the two results (the original and the shifted version)
\item This will be the updated vector
\item Shift the vector now by 2
\item Add the two results (the updated vector with and the two shifted version)
\item Repeat these steps each time shifting by one more than the previous. Stop when n-times shifting has been done.
}
}
\references{
McCornack, R. L. (1965). Extended tables of the Wilcoxon matched pair signed rank statistic. \emph{Journal of the American Statistical Association, 60}(311), 864–871. doi:10.2307/2283253

Munzel, U., & Brunner, E. (2002). An exact paired rank test. \emph{Biometrical Journal, 44}(5), 584. doi:10.1002/1521-4036(200207)44:5<584::AID-BIMJ584>3.0.CO;2-9

Streitberg, B., & Röhmel, J. (1987). Exakte Verteilungen für Rang-und Randomisierungstests im allgemeinen c-Stichprobenproblem. \emph{EDV in Medizin und Biologie, 18}(1), 12–19.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
