% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_alroy_f.R
\name{es_alroy_f}
\alias{es_alroy_f}
\title{Alroy's Forbes Adjustment}
\usage{
es_alroy_f(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Alroy F
}
\description{
A measure of association between two binary variables.

An adjustment to the Forbes coefficient.
}
\details{
The formula used (Alroy, 2015):
\deqn{F' = \frac{a\times\left(n' + \sqrt{n'}\right)}{a\times\left(n' + \sqrt{n'}\right) + \frac{3}{2}\times b\times c}}
With:
\deqn{n' = a + b + c}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
}
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_alroy_f(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Alroy, J. (2015). A new twist on a very old binary similarity coefficient. \emph{Ecology, 96}(2), 575–586. https://doi.org/10.1890/14-0471.1
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
