% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_bag_s.R
\name{es_bag_s}
\alias{es_bag_s}
\title{Bennett-Alpert-Goldstein S}
\usage{
es_bag_s(field1, field2, categories = NULL)
}
\arguments{
\item{field1}{vector, the first categorical field}

\item{field2}{vector, the first categorical field}

\item{categories}{vector, optional, order and/or selection for categories of field1 and field2}
}
\value{
S, the Bennett-Alpert-Goldstein value
}
\description{
An effect size meaure, that measures the how strongly two raters or variables, agree with each other.

It takes the proportions of cases that both agree, and adjusts for the number of categories. Scott's pi (see es_scott_pi()) does this as well, and improves on this measure.
}
\details{
The formula used (Bennett et al., 1954, p. 307):
\deqn{S = \frac{k}{k-1}\times\left(p_0 - \frac{1}{k}\right)}

With:
\deqn{P = \sum_{i=1}^r F_{i,i}}
\deqn{p_0 = \frac{P}{n}}

\emph{Symbols used}
\itemize{
\item \eqn{F_{i,j}}, the observed count in row i and column j.
\item \eqn{r}, is the number of rows (categories in the first variable)
\item \eqn{n}, is the total number of scores
}
}
\references{
Bennett, E. M., Alpert, R., & Goldstein, A. C. (1954). Communications through limited response questioning. \emph{Public Opinion Quarterly, 18}(3), 303. doi:10.1086/266520
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
