% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_becker_clogg_r.R
\name{es_becker_clogg_r}
\alias{es_becker_clogg_r}
\title{Becker and Clogg rho}
\usage{
es_becker_clogg_r(
  field1,
  field2,
  categories1 = NULL,
  categories2 = NULL,
  version = 1
)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}

\item{version}{c(1, 2) optional, which version to calculate (see details), default ver=1}
}
\value{
Becker and Clogg r
}
\description{
An approximation for the tetrachoric correlation coefficient.
}
\details{
Version 1 will calculate:
\deqn{\rho^* = \frac{g-1}{g+1}}

Version 2 will calculate:
\deqn{\rho^{**} = \frac{OR^{13.3/\Delta} - 1}{OR^{13.3/\Delta} + 1}}

With:
\deqn{g=e^{12.4\times\phi - 24.6\times\phi^3}}
\deqn{\phi = \frac{\ln\left(OR\right)}{\Delta}}
\deqn{OR=\frac{\left(\frac{a}{c}\right)}{\left(\frac{b}{d}\right)} = \frac{a\times d}{b\times c}}
\deqn{\Delta = \left(\mu_{R1} - \mu_{R2}\right) \times \left(v_{C1} - v_{C2}\right)}
\deqn{\mu_{R1} = \frac{-e^{-\frac{t_r^2}{2}}}{p_{R1}}, \mu_{R2} = \frac{e^{-\frac{t_r^2}{2}}}{p_{R2}}}
\deqn{v_{C1} = \frac{-e^{-\frac{t_c^2}{2}}}{p_{C1}}, v_{C2} = \frac{e^{-\frac{t_c^2}{2}}}{p_{C2}}}
\deqn{t_r = \Phi^{-1}\left(p_{R1}\right), t_c = \Phi^{-1}\left(p_{C1}\right)}
\deqn{p_{x} = \frac{x}{n}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
\item \eqn{R_i} the sum of counts in the i-th row
\item \eqn{C_i} the sum of counts in the i-th column
\item \eqn{n} the sum of all counts
\item \eqn{\Phi^{-1}\left(x\right)} for the the inverse standard normal cumulative distribution function
}

These formulas can be found in Becker and Clogg (1988, pp. 410-412)
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_becker_clogg_r(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Becker, M. P., & Clogg, C. C. (1988). A note on approximating correlations from Odds Ratios. \emph{Sociological Methods & Research, 16}(3), 407–424. https://doi.org/10.1177/0049124188016003003
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
