% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_bin_bin.R
\name{es_bin_bin}
\alias{es_bin_bin}
\title{Effect Sizes for Binary vs. Binary}
\usage{
es_bin_bin(
  field1,
  field2,
  categories1 = NULL,
  categories2 = NULL,
  method = "odds-ratio"
)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}

\item{method}{: the method to use. Default is odds-ratio}
}
\value{
the effect size measure value
}
\description{
Various measures of association/agreement/similarity for binary vs. binary cases.
}
\details{
The method can be set to any of the following:\tabular{llll}{
    \tab  \tab  \tab  \cr
   alroy \tab ample \tab anderberg \tab austin-colwell \cr
   baroni-urbani-buser-1 \tab baroni-urbani-buser-2 \tab becker-clogg-1 \tab becker-clogg-2 \cr
   bonett-price-1 \tab bonett-price-2 \tab bonett-price-3 \tab braun-blanquet \cr
   camp-1 \tab camp-2 \tab camp-3 \tab chen-popovich \cr
   clement \tab cohen-kappa \tab cohen-w \tab cole-c1 \cr
   cole-c2 \tab cole-c3 \tab cole-c4 \tab cole-c5 \cr
   cole-c6 \tab cole-c7 \tab cole-c8 \tab colligation \cr
   contingency \tab czekanowski \tab dennis \tab dice-1 \cr
   dice-2 \tab dice-3 \tab digby \tab doolittle \cr
   driver-kroeber-1 \tab driver-kroeber-2 \tab edward \tab eyraud \cr
   fager-mcgowan-1 \tab fager-mcgowan-2 \tab faith \tab fleiss \cr
   forbes-1 \tab forbes-2 \tab fossum-kaskey \tab gilbert \cr
   gilbert-wells \tab gk-lambda-1 \tab gk-lambda-2 \tab gleason \cr
   gower \tab gower-legendre \tab hamann \tab harris-lahey \cr
   hawkins-dotson \tab hurlbert \tab jaccard \tab johnson \cr
   kent-foster-1 \tab kent-foster-2 \tab kuder-richardson \tab kulczynski-1 \cr
   kulczynski-2 \tab loevinger \tab matching \tab maxwell-pilliner \cr
   mcconnaughey \tab mcewen-michael \tab mountford \tab nei-li \cr
   ochiai-1 \tab ochiai-2 \tab odds-ratio \tab otsuka \cr
   pearson \tab pearson-heron \tab pearson-q1 \tab pearson-q1 \cr
   pearson-q3 \tab pearson-q4 \tab pearson-q5 \tab peirce-1 \cr
   peirce-2 \tab peirce-3 \tab phi \tab rogers-tanimoto \cr
   rogot-goldberg \tab russell-rao \tab scott \tab simpson \cr
   sokal-michener \tab sokal-sneath-1 \tab sokal-sneath-2 \tab sokal-sneath-3 \cr
   sokal-sneath-4 \tab sokal-sneath-5 \tab sorgenfrei \tab stiles \cr
   tanimoto \tab tarantula \tab tarwid \tab tulloss \cr
   yule-q \tab yule-r \tab yule-y \tab  \cr
}


If we have a 2x2 table with the following values:\tabular{lccc}{
    \tab Column 1 \tab Column 2 \tab total \cr
   row 1 \tab \eqn{a} \tab \eqn{b} \tab \eqn{R_1} \cr
   row 2 \tab \eqn{c} \tab \eqn{d} \tab \eqn{R_2} \cr
   total \tab \eqn{C_1} \tab \eqn{C_2} \tab \eqn{n} \cr
}


then the following formula are used:\tabular{lll}{
   Nr \tab Label \tab Formula \cr
   1 \tab Russell-Rao \tab \eqn{\frac{a}{n}} \cr
   2 \tab Dice-1 \tab \eqn{\frac{a}{R_1}} \cr
   3 \tab Dice-2 \tab \eqn{\frac{a}{C_1}} \cr
   4 \tab Braun-Blanquet \tab \eqn{\frac{a}{\max\left(R_1, C_1\right)}} \cr
   5 \tab Simpson Similarity \tab \eqn{\frac{a}{\min\left(R_1, C_1\right)}} \cr
   6 \tab Kulczynski-1 \tab \eqn{\frac{a}{b+c}} \cr
   7 \tab Jaccard \tab \eqn{\frac{a}{a+b+c}} \cr
   8 \tab Sokal-Sneath-1 \tab \eqn{\frac{a}{a+2b+2c}} \cr
   9 \tab Gleasson \tab \eqn{\frac{2a}{2a+b+c}} \cr
   10 \tab Mountford \tab \eqn{\frac{2a}{a\left(b+c\right)+2bc}} \cr
   11 \tab Driver-Kroeber \tab \eqn{\frac{a}{\sqrt{R_1 C_1}}} \cr
   12 \tab Sorgenfrei \tab \eqn{\frac{a^2}{R_1 C_1}} \cr
   13 \tab Johnson \tab \eqn{\frac{a}{R_1}+\frac{a}{C_1}} \cr
   14 \tab Kulczynski-2 \tab \eqn{\frac{1}{2}\left(\frac{a}{R_1}+\frac{a}{C_1}\right)} \cr
   15 \tab Fager-McGowan-1 \tab \eqn{\frac{a}{\sqrt{R_1 C_1}}-\frac{1}{2\sqrt{\max\left(R_1, C_1\right)}}} \cr
   16 \tab Fager-McGowan-2 \tab \eqn{\frac{a}{\sqrt{R_1 C_1}}-\frac{\sqrt{\max\left(R_1, C_1\right)}}{2}} \cr
   17 \tab tarantula \tab \eqn{\frac{aR_2}{cR_1}} \cr
   18 \tab Ample \tab \eqn{\left| \frac{aR_2}{cR_1}\right|} \cr
   19 \tab Gilbert \tab \eqn{\frac{an-R_1 C_1}{C_1 n + R_1 n - an - R_1 C_1}} \cr
   20 \tab Fossum-Kaskey \tab \eqn{\frac{n\left(a - \frac{1}{2}\right)^2}{R_1 C_1}} \cr
   21 \tab Forbes - 1 \tab \eqn{\frac{na}{R_1 C_1}} \cr
   22 \tab Eyraud \tab \eqn{\frac{a-R_1 C_1}{R_1 R_2 C_1 C_2}} \cr
   23 \tab Sokal-Michener \tab \eqn{\frac{a+d}{n}} \cr
   24 \tab Faith \tab \eqn{\frac{a+\frac{1}{2}}{n}} \cr
   25 \tab Sokal-Sneath-5 \tab \eqn{\frac{a+d}{b+c}} \cr
   26 \tab Rogers-Tanimoto \tab \eqn{\frac{a+d}{a+2\left(b+c\right)+d}} \cr
   27 \tab Sokal-Sneath-2 \tab \eqn{\frac{2a+2d}{2a+b+c+2d}} \cr
   28 \tab Gower \tab \eqn{\frac{a+d}{\sqrt{R_1 R_2 C_1 C_2}}} \cr
   29 \tab Sokal-Sneath-4 \tab \eqn{\frac{ad}{\sqrt{R_1 R_2 C_1 C_2}}} \cr
   30 \tab Rogot-Goldberg \tab \eqn{\frac{a}{R_1 + C_1}+\frac{d}{R_2 + C_2}} \cr
   31 \tab Sokal-Sneath-3 \tab \eqn{\frac{1}{4}\left(\frac{a}{R_1}+\frac{a}{C_1}+\frac{d}{R_2}+\frac{d}{C_2}\right)} \cr
   32 \tab Hawkin-Dotson \tab \eqn{\frac{1}{2}\left(\frac{a}{a+b+c}+\frac{d}{b+c+d}\right)} \cr
   33 \tab Clement \tab \eqn{\frac{aR_2}{nR_1}+\frac{dR_1}{nR_2}} \cr
   34 \tab Harris-Lahey \tab \eqn{\frac{a\left(R_2+C_2\right)}{2n\left(a+b+c\right)}+\frac{d\left(R_1+C_1\right)}{2n\left(b+c+d\right)}} \cr
   35 \tab Austin-Colwell \tab \eqn{\frac{2}{\pi}\text{arcsin}\sqrt{\frac{a+d}{n}}} \cr
   36 \tab Baroni-Urbani-Buser-1 \tab \eqn{\frac{\sqrt{ad}+a}{\sqrt{ad}+a+b+c}} \cr
   37 \tab Peirce-1 \tab \eqn{\frac{ad-bc}{R_1 R_2}} \cr
   38 \tab Peirce-2 \tab \eqn{\frac{ad-bc}{C_1 C_2}} \cr
   39 \tab Cole C1 \tab \eqn{\frac{ad-bc}{R_1 C_1}} \cr
   40 \tab Loevinger \tab \eqn{\frac{ad-bc}{\min\left(R_1 C_2, R_2 C_1\right)}} \cr
   41 \tab Cole C7 \tab \eqn{\begin{cases} \frac{ad-bc}{R_1 C_2} & \text{ if } ad \geq bc \\ \frac{ad-bc}{R_1 C_1} & \text{ if } ad < bc \text{ and } a \leq d \\ \frac{ad-bc}{R_2 C_2} & \text{ if } ad < bc \text{ and } a > d\end{cases}} \cr
   42 \tab Dennis \tab \eqn{\frac{ad-bc}{\sqrt{nR_1 C_1}}} \cr
   43 \tab Phi \tab \eqn{\frac{ad-bc}{\sqrt{R_1 R_2 C_1 C_2}}} \cr
   44 \tab Doolittle \tab \eqn{\frac{\left(ad-bc\right)^2}{R_1 R_2 C_1 C_2}} \cr
   45 \tab Peirce-3 \tab \eqn{\frac{ad-bc}{ab+2bc+cd}} \cr
   46 \tab Cohen-kappa \tab \eqn{\frac{2\left(ad-bc\right)}{R_1 C_2 + R_2 C_1}} \cr
   47 \tab McEwen-Michael \tab \eqn{\frac{4\left(ad-bc\right)}{\left(a+d\right)^2 + \left(b+c\right)^2}} \cr
   48 \tab Kuder-Richardson \tab \eqn{\frac{4\left(ad-bc\right)}{R_1 R_2 + C_1 C_2 + 2ad-2bc}} \cr
   49 \tab Scott \tab \eqn{\frac{4ad-\left(b+c\right)^2}{\left(R_1 + C_1\right)\left(R_2 + C_2\right)}} \cr
   50 \tab Maxwell-Pilliner \tab \eqn{\frac{2\left(ad-bc\right)}{R_1 R_2 + C_1 C_2}} \cr
   51 \tab Cole C5 \tab \eqn{\frac{\sqrt{2}\left(ad-bc\right)}{\sqrt{\left(ad-bc\right)^2+R_1 R_2 C_1 C_2}}} \cr
   52 \tab Hamann \tab \eqn{\frac{\left(a+d\right)-\left(b+c\right)}{n}} \cr
   53 \tab Fleiss \tab \eqn{\frac{\left(ad-bc\right)\left(R_1 C_2 + R_2 C_1\right)}{2R_1 R_2 C_1 C_2}} \cr
   54 \tab Yule Q \tab \eqn{\frac{ad-bc}{ad+bc}} \cr
   55 \tab Yule Y \tab \eqn{\frac{\sqrt{ad}-\sqrt{bc}}{\sqrt{ad}+\sqrt{bc}}} \cr
   56 \tab Digby H \tab \eqn{\frac{\left(ad\right)^{3/4}-\left(bc\right)^{3/4}}{\left(ad\right)^{3/4}+\left(bc\right)^{3/4}}} \cr
   57 \tab Edward Q \tab \eqn{\frac{OR^{\pi/4}-1}{OR^{\pi/4}+1}} \cr
   58 \tab Tarwid \tab \eqn{\frac{na-R_1 C_1}{na+R_1 C_1}} \cr
   59 \tab Bonett-Price-1 \tab \eqn{\frac{\hat{w}^x-1}{\hat{w}^x+1}} \cr
   60 \tab Contingency coefficient \tab \eqn{\sqrt{\frac{\chi^2}{n+\chi^2}}} \cr
   61 \tab Cohen w \tab \eqn{\sqrt{\frac{\chi^2}{\chi^2}}} \cr
   62 \tab Pearson \tab \eqn{\sqrt{\frac{\phi^2}{n+\phi^2}}} \cr
   63 \tab Hurlbert \tab \eqn{\frac{ad-bc}{\left\lvert ad-bc\right\rvert}\sqrt{\frac{\chi^2-\chi_{min}^2}{\chi_{max}^2-\chi_{min}^2}}} \cr
   64 \tab Stiles \tab \eqn{\log_{10}\left(\frac{n\left(\left\lvert ad-bc \right\rvert -\frac{n}{2}\right)^2}{R_1 R_2 C_1 C_2}\right)} \cr
   65 \tab McConnaughey \tab \eqn{\frac{a^2-bc}{R_1 C_1}} \cr
   66 \tab Baroni-Urbani-Buser-2 \tab \eqn{\frac{a-b-c+\sqrt{ad}}{a+b+c+\sqrt{ad}}} \cr
   67 \tab Kent-Foster-1 \tab \eqn{\frac{-bc}{bR_1 + cC_1 + bc}} \cr
   68 \tab Kent-Foster-2 \tab \eqn{\frac{-bc}{bR_2 + cC_2 + bc}} \cr
   69 \tab Tulloss \tab \eqn{\sqrt{U\times S\times R}} \cr
   70 \tab Gilbert-Wells \tab \eqn{\ln\left(a\right)-\ln\left(n\right)-\ln\left(\frac{R_1}{n}\right)-\ln\left(\frac{C_1}{n}\right)} \cr
   71 \tab Yule r \tab \eqn{\cos\left(\frac{\pi\sqrt{bc}}{\sqrt{ad}+\sqrt{bc}}\right)} \cr
   72 \tab Anderberg \tab \eqn{\frac{\sigma-\sigma'}{2n}} \cr
   73 \tab Alroy F \tab \eqn{\frac{a\left(n' + \sqrt{n'}\right)}{a\left(n' + \sqrt{n'}\right)+\frac{3}{2}bc}} \cr
   74 \tab Pearson Q1 \tab \eqn{\sin\left(\frac{\pi}{2}\times\frac{R_2 C_1}{R_1 C_2}\right)} \cr
   75 \tab Goodman-Kruskal Lambda-1 \tab \eqn{\frac{\sigma-\sigma'}{2n-\sigma'}} \cr
   76 \tab Goodman-Kruskal Lambda-2 \tab \eqn{\frac{2\min\left(a,d\right)-b-c}{2\min\left(a,d\right)+b+c}} \cr
   77 \tab Odds Ratio \tab \eqn{\frac{ad}{bc}} \cr
   78 \tab Pearson Q4 \tab \eqn{\sin\frac{\pi}{2}\times\frac{1}{1+\frac{2bcn}{\left(ad-bc\right)\left(b+c\right)}}} \cr
   79 \tab Pearson Q5 \tab \eqn{\sin\frac{\pi}{2}\times\frac{1}{\sqrt{1+\frac{4abcdn^2}{\left(ad-bc\right)^2\left(a+d\right)\left(b+c\right)}}}} \cr
   80 \tab Camp (3 ver.) \tab \eqn{\frac{m}{\sqrt{1+\Theta\times m^2}}} \cr
   81 \tab Becker-Clogg-1 \tab \eqn{\frac{g-1}{g+1}} \cr
   82 \tab Becker-Clogg-2 \tab \eqn{\frac{OR^{13.3/ \delta}-1}{OR^{13.3/ \delta}+1}} \cr
   83 \tab Bonett-Price-r \tab \eqn{\cos\left(\frac{\pi}{1+\omega^c}\right)} \cr
   84 \tab Bonett-Price-rhat \tab \eqn{\cos\left(\frac{\pi}{1+\hat{\omega}^{\hat{c}}}\right)} \cr
   85 \tab Chen-Popovich \tab \eqn{\frac{ad-bc}{\lambda_x \lambda_y n^2}} \cr
}


\strong{Equation 57}

\deqn{OR = \frac{ad}{bc}}

\strong{Equation 59}

\deqn{x = \frac{1}{2}-\left(\frac{1}{2}-p_{min}\right)^2}
\deqn{p_{min} = \frac{\min\left(R_1, R_2, C_1, C_2\right)}{n}}
\deqn{\hat{\omega} = \frac{\left(a+0.1\right)\times\left(d+0.1\right)}{\left(b+0.1\right)\times\left(c+0.1\right)}}

\strong{Equations 60, 61, and 63}

\deqn{\chi^2 = \frac{n\left(ad-bc\right)^2}{R_1 R_2 C_1 C_2}}

\strong{Equation 62}

\deqn{\frac{ad-bc}{\sqrt{R_1 R_2 C_1 C_2}}}

\strong{Equation 63:}

\deqn{\chi_{max}^2 = \begin{cases} \frac{nR_1 C_2}{R_2 C_1} & \text{ if } ad \geq bc \\  \frac{nR_1 C_1}{R_2 C_2} & \text{ if } ad < bc \text{ and } a \leq d \\  \frac{nR_2 C_2}{R_1 C_1} & \text{ if } ad < bc \text{ and } a > d \end{cases}}
\deqn{\chi_{min}^2 = \frac{n^3\left(\hat{a} - g\left(\hat{a}\right)\right)^2}{R_1 R_2 C_1 C_2}}
\deqn{\hat{a}=\frac{R_1 C_1}{n}}
\deqn{g\left(\hat{a}\right) = \begin{cases} \lfloor \hat{a} \rfloor & \text{ if } ad < bc \\  \lceil \hat{a}\rceil & \text{ if } ad \geq bc \end{cases}}

\strong{Equation 69}

\deqn{U = \log_2\left(1+\frac{\min\left(b,c\right)+a}{\max\left(b,c\right)+a}\right)}
\deqn{S = \frac{1}{\sqrt{\log_2\left(2+\frac{\min\left(b,c\right)}{a+1}\right)}}}
\deqn{R = \log_2\left(1+\frac{a}{R_1}\right)\log_2\left(1+\frac{a}{RC1}\right)}

\strong{Equation 71 and 75}

\deqn{\sigma = \max\left(a,b\right)+\max\left(c,d\right)+\max\left(a,c\right)+\max\left(b,d\right)}
\deqn{\sigma' = \max\left(R_1, R_2\right)+\max\left(C_1, C_2\right)}

\strong{Equation 73}

\deqn{n' = a+b+c}

\strong{Equation 80}

Camp (1934, pp. 309) describes the following steps for the calculation:
Step 1: If total of column 1 (C1) is less than column 2 (C2), swop the two columns

Step 2: Calculate \eqn{p = \frac{C1}{n}}, \eqn{p_1 = \frac{a}{n}}, and \eqn{p_2 = \frac{c}{C2}}

Step 3: Determine \eqn{z_1}, \eqn{z_2} as the normal deviate
corresponding to the area \eqn{p_1}, \eqn{p_2} resp. (inverse standard normal cumulative distribution)

Step 4: Determine y the normal ordinate corresponding to \eqn{p} (the height of the normal distribution)

Step 5: Calculate \eqn{m = \frac{p\times\left(1-p\right)\times\left(z_1 + z_2\right)}{y}}

Step 6: Find phi in a table of phi values

Camp suggested for a very basic approximation to simply use \eqn{\phi=1}.

For a better approximation Camp made the following table:\tabular{llllll}{
   p \tab 0.5 \tab 0.6 \tab 0.7 \tab 0.8 \tab 0.9 \cr
   phi \tab 0.637 \tab 0.63 \tab 0.62 \tab 0.60 \tab 0.56 \cr
}


Cureton (1968, p. 241) expanded on this table and produced:\tabular{llllllllllll}{
   p \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab 10 \cr
   0.5 \tab 0.637 \tab 0.636 \tab 0.636 \tab 0.635 \tab 0.635 \tab 0.634 \tab 0.634 \tab 0.633 \tab 0.633 \tab 0.632 \tab 0.631 \cr
   0.6 \tab 0.631 \tab 0.631 \tab 0.630 \tab 0.629 \tab 0.628 \tab 0.627 \tab 0.626 \tab 0.625 \tab 0.624 \tab 0.622 \tab 0.621 \cr
   0.7 \tab 0.621 \tab 0.620 \tab 0.618 \tab 0.616 \tab 0.614 \tab 0.612 \tab 0.610 \tab 0.608 \tab 0.606 \tab 0.603 \tab 0.600 \cr
   0.8 \tab 0.600 \tab 0.597 \tab 0.594 \tab 0.591 \tab 0.587 \tab 0.583 \tab 0.579 \tab 0.574 \tab 0.569 \tab 0.564 \tab 0.559 \cr
}


Step 7: Calculate \eqn{r_t = \frac{m}{\sqrt{1+\phi\times m^2}}}

Cureton (1968) describes quite a few shortcomings with this approximation, and circumstances when it might be appropriate.

\strong{Equation 81 and 82}

Version 81 will calculate:
\deqn{\rho^* = \frac{g-1}{g+1}}

Version 82 will calculate:
\deqn{\rho^{**} = \frac{OR^{13.3/\Delta} - 1}{OR^{13.3/\Delta} + 1}}

With:
\deqn{g=e^{12.4\times\phi - 24.6\times\phi^3}}
\deqn{\phi = \frac{\ln\left(OR\right)}{\Delta}}
\deqn{OR=\frac{\left(\frac{a}{c}\right)}{\left(\frac{b}{d}\right)} = \frac{a\times d}{b\times c}}
\deqn{\Delta = \left(\mu_{R1} - \mu_{R2}\right) \times \left(v_{C1} - v_{C2}\right)}
\deqn{\mu_{R1} = \frac{-e^{-\frac{t_r^2}{2}}}{p_{R1}}, \mu_{R2} = \frac{e^{-\frac{t_r^2}{2}}}{p_{R2}}}
\deqn{v_{C1} = \frac{-e^{-\frac{t_c^2}{2}}}{p_{C1}}, v_{C2} = \frac{e^{-\frac{t_c^2}{2}}}{p_{C2}}}
\deqn{t_r = \Phi^{-1}\left(p_{R1}\right), t_c = \Phi^{-1}\left(p_{C1}\right)}
\deqn{p_{x} = \frac{x}{n}}

\strong{Equations 83 and 84}

Formula for version 1 is (Bonett & Price, 2005, p. 216):
\deqn{\rho^* = \cos\left(\frac{\pi}{1+\omega^c}\right)}
With:
\deqn{\omega = OR = \frac{a\times d}{b\times c}}
\deqn{c = \frac{1-\frac{\left|R_1-C_1\right|}{5\times n} - \left(\frac{1}{2}-p_{min}\right)^2}{2}}
\deqn{p_{min} = \frac{\text{MIN}\left(R_1, R_2, C_1, C_2\right)}{n}}

Formula for version 2 is  (Bonett & Price, 2005, p. 216):
\deqn{\hat{\rho}^* = \cos\left(\frac{\pi}{1+\hat{\omega}^{\hat{c}}}\right)}
with:
\deqn{\hat{\omega} = \frac{\left(a+\frac{1}{2}\right)\times \left(d+\frac{1}{2}\right)}{\left(b+\frac{1}{2}\right)\times \left(c+\frac{1}{2}\right)}}
\deqn{\hat{c} = \frac{1-\frac{\left|R_1-C_1\right|}{5\times \left(n+2\right)} - \left(\frac{1}{2}-\hat{p}_{min}\right)^2}{2}}
\deqn{\hat{p}_{min} = \frac{\text{MIN}\left(R_1, R_2, C_1, C_2\right)+1}{n+2}}

\strong{Equation 85}

\deqn{\lambda_x = \Phi^{-1}\left(\frac{R_1}{n}\right)}
\deqn{\lambda_y = \Phi^{-1}\left(\frac{C_1}{n}\right)}

with \eqn{\Phi^{-1}\left(\dots\right)} being the inverse standard normal cumulative distribution

\strong{Sources for formulas}

The formulas were obtained from the following sources.
The columns W-C-H show which equation corresponds to my label in:
\itemize{
\item W: Warrens (2008, pp. 219–222). Equation 4 from Warrens is the chi-square value and not added.
\item C: Choi et al. (2010, pp. 44–45). Equations not added from this source are: Eq. 4 is a ‘3w Jaccard’, could not find a source for this and not added. Equation 12 is just the intersection (a), eq. 13 the innerproduct (a+d), and 66 the dispersion. Equation 51 is the chi-square value and measures 15 to 30 and 62 are just distance measures.
\item H: Hubálek (Hubálek, 1982, pp. 671–673)
}

If no page number is listed in the original source, the formula was taken from Warrens, Choi et al. and/or Hubálek.\tabular{llllll}{
   nr \tab Label \tab Original \tab W \tab C \tab H \cr
   1 \tab Russell-Rao \tab (Russell & Rao, 1940) \tab 15 \tab 14 \tab 14 \cr
   2 \tab Dice-1 \tab (Dice, 1945, p. 302) \tab 17a \tab  \tab  \cr
   3 \tab Dice-2 \tab (Dice, 1945, p. 302) \tab 17b \tab  \tab  \cr
   4 \tab Braun-Blanquet \tab (Braun-Blanquet, 1932) \tab 12 \tab 46 \tab 1 \cr
   5 \tab Simpson Similarity \tab (Simpson, 1943, p. 20, 1960, p. 301) \tab 16 \tab 45 \tab 2 \cr
   6 \tab Kulczynski-1 \tab (Kulczynski, 1927) \tab 11b \tab 64 \tab 3 \cr
   7 \tab Jaccard = \tab (Jaccard, 1901, 1912, p. 39) \tab 6 \tab 1 \tab 4 \cr
    \tab Tanimoto \tab (Tanimoto, 1958, p. 5) \tab  \tab 65 \tab  \cr
   8 \tab Sokal-Sneath-1 \tab (Sokal & Sneath, 1963, p. 129) \tab 30a \tab 6 \tab 6 \cr
   9 \tab Gleasson = \tab (Gleason, 1920, p. 31) \tab 9 \tab  \tab 5 \cr
    \tab Dice-3 = \tab (Dice, 1945, p. 302) \tab  \tab 2 \tab  \cr
    \tab Nei-Li = \tab (Nei & Li, 1979, p. 5270) \tab  \tab 5 \tab  \cr
    \tab Czekanowski \tab  \tab  \tab 3 \tab  \cr
   10 \tab Mountford \tab (Mountford, 1962, p. 45) \tab 28 \tab 37 \tab 15 \cr
   11 \tab Driver-Kroeber = \tab (Driver & Kroeber, 1932, p. 219) \tab 13 \tab 31 \tab 11 \cr
    \tab Ochiai-1  = \tab (Ochiai, 1957) \tab  \tab 33 \tab  \cr
    \tab Otsuka \tab (Otsuka, 1936) \tab  \tab 38 \tab  \cr
   12 \tab Sorgenfrei \tab (Sorgenfrei, 1958) \tab 23 \tab 36 \tab 12 \cr
   13 \tab Johnson \tab (Johnson, 1967) \tab 33 \tab 43 \tab 9 \cr
   14 \tab Driver-Kroeber-2 = \tab (Driver & Kroeber, 1932, p. 219) \tab 11a \tab 42 \tab 8 \cr
    \tab Kulczynski-2 \tab (Kulczynski, 1927) \tab  \tab 41 \tab 7 \cr
   15 \tab Fager-McGowan-1 \tab (Fager & McGowan, 1963, p. 454) \tab 29 \tab  \tab  \cr
   16 \tab Fager-McGowan-2 \tab (Fager & McGowan, 1963, p. 454) \tab  \tab 47 \tab 13 \cr
   17 \tab tarantula \tab (Jones & Harrold, 2005) \tab  \tab 75 \tab  \cr
   18 \tab ample \tab  \tab  \tab 76 \tab  \cr
   19 \tab Gilbert \tab (Gilbert, 1884, p. 171) \tab  \tab  \tab  \cr
   20 \tab Fossum-Kaskey \tab (Fossum & Kaskey, 1966, p. 65) \tab  \tab 35 \tab  \cr
   21 \tab Forbes - 1 \tab (Forbes, 1907, p. 279) \tab 5 \tab 34 \tab 40 \cr
   22 \tab Eyraud \tab (Eyraud, 1936) \tab  \tab 74 \tab 17 \cr
   23 \tab Sokal-Michener \tab (Sokal & Michener, 1958, p. 1417) \tab 22 \tab 7 \tab 20 \cr
   24 \tab Faith \tab (Faith, 1983, p. 290) \tab  \tab 10 \tab  \cr
   25 \tab Sokal-Sneath-5 \tab (Sokal & Sneath, 1963, p. 129) \tab 30e \tab 56 \tab 19 \cr
   26 \tab Rogers-Tanimoto \tab (Rogers & Tanimoto, 1960) \tab 25 \tab 9 \tab 23 \cr
   27 \tab Sokal-Sneath-2 = \tab (Sokal & Sneath, 1963, p. 129) \tab 30b \tab 8 \tab 22 \cr
    \tab Gower-Legendre \tab (Gower & Legendre, 1986) \tab  \tab 11 \tab  \cr
   28 \tab Gower \tab (Gower, 1971) \tab  \tab 50 \tab  \cr
   29 \tab Sokal-Sneath-4 = \tab (Sokal & Sneath, 1963, p. 130) \tab 30d \tab 57 \tab 25 \cr
    \tab Ochiai-2 \tab (Ochiai, 1957) \tab  \tab 60 \tab  \cr
   30 \tab Rogot-Goldberg \tab (Rogot & Goldberg, 1966, p. 997) \tab 32 \tab  \tab  \cr
   31 \tab Sokal-Sneath-3 \tab (Sokal & Sneath, 1963, p. 130) \tab 30c \tab 49 \tab 18 \cr
   32 \tab Hawkin-Dotson \tab (Hawkins & Dotson, 1975, pp. 372–373) \tab 34 \tab  \tab  \cr
   33 \tab Clement \tab (Clement, 1976, p. 258) \tab 37 \tab  \tab  \cr
   34 \tab Harris-Lahey \tab (Harris & Lahey, 1978, p. 526) \tab 40 \tab  \tab  \cr
   35 \tab Austin-Colwell \tab (Austin & Colwell, 1977, p. 205) \tab  \tab  \tab 21 \cr
   36 \tab Baroni-Urbani-Buser-1 \tab (Baroni-Urbani & Buser, 1976, p. 258) \tab 38a \tab 71 \tab 32 \cr
   37 \tab Peirce-1 \tab (Peirce, 1884, p. 453) \tab 1a \tab  \tab  \cr
   38 \tab Peirce-2 \tab (Peirce, 1884, p. 453) \tab 1b \tab  \tab 26 \cr
   39 \tab Cole C1 \tab (Cole, 1949, p. 415) \tab  \tab  \tab  \cr
   40 \tab Loevinger = \tab (Loevinger, 1947, p. 30) \tab 18 \tab  \tab  \cr
    \tab Forbes 2 \tab (Forbes, 1925) \tab  \tab 48 \tab 42 \cr
   41 \tab Cole C7 \tab (Cole, 1949, p. 420) \tab 19 \tab  \tab 34 \cr
   42 \tab Dennis \tab (Dennis, 1965, p. 69) \tab  \tab 44 \tab  \cr
   43 \tab Pearson Phi = \tab (Pearson, 1900a, p. 12) \tab 7 \tab 54 \tab 30 \cr
    \tab Yule Phi = \tab (Yule, 1912, p. 596) \tab  \tab  \tab  \cr
    \tab Cole C2 \tab (Cole, 1949, p. 415) \tab  \tab  \tab  \cr
   44 \tab Doolittle \tab (Doolittle, 1885, p. 123) \tab 2 \tab  \tab 31 \cr
   45 \tab Peirce-3 \tab (Peirce, 1884) \tab  \tab 73 \tab 16 \cr
   46 \tab Cohen-kappa \tab (Cohen, 1960, p. 40) \tab 24 \tab  \tab  \cr
   47 \tab McEwen-Michael = \tab (Michael, 1920, p. 57) \tab 10 \tab 68 \tab 39 \cr
    \tab Cole C3 \tab (Cole, 1949, p. 415) \tab  \tab  \tab  \cr
   48 \tab Kuder-Richardson \tab (Kuder & Richardson, 1937) \tab 14 \tab  \tab  \cr
   49 \tab Scott \tab (Scott, 1955, p. 324) \tab 21 \tab  \tab  \cr
   50 \tab Maxwell-Pilliner \tab (Maxwell & Pilliner, 1968) \tab 35 \tab  \tab  \cr
   51 \tab Cole C5 \tab (Cole, 1949, p. 416) \tab  \tab 58 \tab 29 \cr
   52 \tab Hamann \tab (Hamann, 1961) \tab 27 \tab 67 \tab 24 \cr
   53 \tab Fleiss \tab (Fleiss, 1975, p. 656) \tab 36 \tab  \tab  \cr
   54 \tab Yule Q = \tab (Yule, 1900, p. 272) \tab 3 \tab 61 \tab 36 \cr
    \tab Cole C4 = \tab (Cole, 1949, p. 415) \tab  \tab  \tab  \cr
    \tab Pearson Q2 \tab (Pearson, 1900, p. 15) \tab  \tab  \tab  \cr
   55 \tab Yule Y \tab (Yule, 1912, p. 592) \tab 8 \tab 63 \tab 37 \cr
   56 \tab Digby H \tab (Digby, 1983, p. 754) \tab 41 \tab  \tab  \cr
   57 \tab Edward Q \tab (Edwards, 1957; Becker & Clogg, 1988, p. 409) \tab  \tab  \tab  \cr
   58 \tab Tarwid \tab (Tarwid, 1960, p. 117) \tab  \tab 40 \tab 43 \cr
   59 \tab Bonett-Price-1 \tab (Bonett & Price, 2007, p. 433) \tab  \tab  \tab  \cr
   60 \tab Contingency \tab (Pearson, 1904, p. 9) \tab  \tab 52 \tab 28 \cr
   61 \tab Cohen w \tab (Cohen, 1988, p. 216) \tab  \tab  \tab  \cr
   62 \tab Pearson \tab (Pearson, 1904) \tab  \tab 53 \tab  \cr
   63 \tab Hurlbert / Cole C8 \tab (Hurlbert, 1969, p. 1) \tab  \tab  \tab 35 \cr
   64 \tab Stiles \tab (Stiles, 1961, p. 272) \tab 26 \tab 59 \tab  \cr
   65 \tab McConnaughey \tab (McConnaughey, 1964) \tab 31 \tab 39 \tab 10 \cr
   66 \tab Baroni-Urbani-Buser-2 \tab (Baroni-Urbani & Buser, 1976, p. 258) \tab 38b \tab 72 \tab 33 \cr
   67 \tab Kent-Foster-1 \tab (Kent & Foster, 1977, p. 311) \tab 39a \tab  \tab  \cr
   68 \tab Kent-Foster-2 \tab (Kent & Foster, 1977, p. 311) \tab 39b \tab  \tab  \cr
   69 \tab Tulloss \tab (Tulloss, 1997, p. 133) \tab  \tab  \tab  \cr
   70 \tab Gilbert-Wells \tab (Gilbert & Wells, 1966) \tab  \tab  \tab  \cr
   71 \tab Yule r \tab (Yule, 1900, p. 276) \tab  \tab  \tab  \cr
    \tab Pearson-Q3 \tab (Pearson, 1900a, p. 16) \tab  \tab  \tab  \cr
    \tab Cole C6 \tab (Cole, 1949, p. 416) \tab  \tab  \tab  \cr
    \tab Pearson-Heron \tab (Pearson & Heron, 1913) \tab  \tab 55 \tab 38 \cr
   72 \tab Anderberg \tab (Anderberg, 1973) \tab  \tab 70 \tab  \cr
   73 \tab Alroy F \tab (Alroy, 2015, eq. 6) \tab  \tab  \tab  \cr
   74 \tab Pearson Q1 \tab (Pearson, 1900a, p. 15) \tab  \tab  \tab  \cr
   75 \tab Goodman-Kruskal Lambda-1 \tab (Goodman & Kruskal, 1954, p. 743) \tab  \tab 69 \tab  \cr
   76 \tab Goodman-Kruskal Lambda-2 \tab (Goodman & Kruskal, 1954) \tab 20 \tab  \tab  \cr
   77 \tab Odds Ratio \tab (Fisher, 1935, p. 50) \tab  \tab  \tab  \cr
   78 \tab Pearson Q4 \tab (Pearson, 1900, p. 16) \tab  \tab  \tab  \cr
   79 \tab Pearson Q5 \tab (Pearson, 1900, p. 16) \tab  \tab  \tab  \cr
   80 \tab Camp \tab (Camp, 1934, p. 309) \tab  \tab  \tab  \cr
   81 \tab Becker-Clogg-1 \tab (Becker & Clogg, 1988, pp. 410–412) \tab  \tab  \tab  \cr
   82 \tab Becker-Clogg-2 \tab (Becker & Clogg, 1988, pp. 410–412) \tab  \tab  \tab  \cr
   83 \tab Bonett-Price-2 \tab (Bonett & Price, 2005, p. 216) \tab  \tab  \tab  \cr
   84 \tab Bonett-Price-3 \tab (Bonett & Price, 2005, p. 216) \tab  \tab  \tab  \cr
   85 \tab Ched-Popovich \tab (Chen & Popovich, 2002, p. 37) \tab  \tab  \tab  \cr
}
}
\references{
Alroy, J. (2015). A new twist on a very old binary similarity coefficient. \emph{Ecology, 96}(2), 575–586. doi:10.1890/14-0471.1

Anderberg, M. R. (1973). \emph{Cluster analysis for applications}. New York, NY, Academic Press.

Austin, B., & Colwell, R. R. (1977). Evaluation of some coefficients for use in numerical taxonomy of microorganisms. \emph{International Journal of Systematic Bacteriology, 27}(3), 204–210. doi:10.1099/00207713-27-3-204

Baroni-Urbani, C., & Buser, M. W. (1976). Similarity of binary data. \emph{Systematic Zoology, 25}(3), 251–259. doi:10.2307/2412493

Becker, M. P., & Clogg, C. C. (1988). A note on approximating correlations from Odds Ratios. \emph{Sociological Methods & Research, 16}(3), 407–424. doi:10.1177/0049124188016003003

Bonett, D. G., & Price, R. M. (2007). Statistical inference for generalized yule coefficients in 2 x 2 contingency tables. \emph{Sociological Methods & Research, 35}(3), 429–446. doi:10.1177/0049124106292358

Braun-Blanquet, J. (1932). \emph{Plant sociology: The study of plant communities}. McGraw Hill.

Choi, S.-S., Cha, S.-H., & Tappert, C. (2010). A survey of binary similarity and distance measures. \emph{Journal on Systemics, Cybernetics and Informatics, 8}(1), 43–48.

Clement, P. W. (1976). A formula for computing inter-observer agreement. \emph{Psychological Reports, 39}(1), 257–258. doi:10.2466/pr0.1976.39.1.257

Cohen, J. (1960). A coefficient of agreement for nominal scales. \emph{Educational and Psychological Measurement, 20}(1), 37–46. doi:10.1177/001316446002000104

Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.

Cole, L. C. (1949). \emph{The measurement of interspecific associaton. Ecology, 30}(4), 411–424. doi:10.2307/1932444

Dennis, S. F. (1965). The construction of a thesaurus automatically from a sample of text. In M. E. Stevens, V. E. Giuliano, & L. B. Heilprin (Eds.), \emph{Statistical Association Methods for Mechanized Documentation} (Vol. 14, pp. 61–148). U.S. Government Printing Office.

Dice, L. R. (1945). Measures of the amount of ecologic association between species. \emph{Ecology, 26}(3), 297–302. doi:10.2307/1932409

Digby, P. G. N. (1983). Approximating the tetrachoric correlation coefficient. \emph{Biometrics, 39}(3), 753–757. doi:10.2307/2531104

Doolittle, M. H. (1885). The verification of predictions. \emph{Bulletin of the Philosophical Society of Washington, 7}, 122–127.

Driver, H. E., & Kroeber, A. L. (1932). Quantitative expression of cultural relationships. \emph{University Of California Publications in American Archeology and Ethnology, 31}(4), 211–256.

Edwards, J. H. (1957). A note on the practical interpretation of 2 x 2 tables. \emph{Journal of Epidemiology & Community Health, 11}(2), 73–78. doi:10.1136/jech.11.2.73

Eyraud, H. (1936). Les principes de la mesure des correlations. \emph{Ann. Univ. Lyon, III. Ser., Sect. A, 1}(30–47), 111.

Fager, E. W., & McGowan, J. A. (1963). Zooplankton species groups in the north pacific: Co-occurrences of species can be used to derive groups whose members react similarly to water-mass types. \emph{Science, 140}(3566), 453–460. doi:10.1126/science.140.3566.453

Faith, D. P. (1983). Asymmetric binary similarity measures. \emph{Oecologia, 57}(3), 287–290. doi:10.1007/BF00377169

Fisher, R. A. (1935). The logic of inductive inference. \emph{Journal of the Royal Statistical Society, 98}(1), 39–82. doi:10.2307/2342435

Fleiss, J. L. (1975). Measuring agreement between two judges on the presence or absence of a trait. \emph{Biometrics, 31}(3), 651–659. doi:10.2307/2529549

Forbes, S. A. (1907). On the local distribution of certain Illinois fishes: An essay in statistical ecology. \emph{Illinois Natural History Survey Bulletin, 7}(8), 273–303.

Forbes, S. A. (1925). Method of determining and measuring the associative relations of species. \emph{Science, 61}, 524.

Fossum, E. G., & Kaskey, G. (1966). Optimization and standardization of information retrieval language and systems (AF49(638)1194; p. 96). Univac Division.

Gilbert, G. K. (1884). Finley’s tornado predictions. \emph{American Meteorological Journal, 1}(5), 166–172.

Gilbert, N., & Wells, T. C. E. (1966). \emph{Analysis of quadrat data. Journal of Ecology, 54}(3), 675–685. doi:10.2307/2257810

Gleason, H. A. (1920). Some applications of the quadrat method. \emph{Bulletin of the Torrey Botanical Club, 47}(1), 21–33. doi:10.2307/2480223

Goodman, L. A., & Kruskal, W. H. (1954). Measures of association for cross classifications. \emph{Journal of the American Statistical Association, 49}(268), 732–764. doi:10.2307/2281536

Gower, J. C. (1971). A general coefficient of similarity and some of its properties. \emph{Biometrics, 27}(4), 857. doi:10.2307/2528823

Gower, J. C., & Legendre, P. (1986). Metric and Euclidean properties of dissimilarity coefficients. \emph{Journal of Classification, 3}(1), 5–48. doi:10.1007/BF01896809

Hamann, U. (1961). Merkmalsbestand und verwandtschaftsbeziehungen der farinosae: Ein beitrag zum system der monokotyledonen. \emph{Willdenowia, 2}(5), 639–768.

Harris, F. C., & Lahey, B. B. (1978). A method for combining occurrence and nonoccurrence interobserver agreement scores. \emph{Journal of Applied Behavior Analysis, 11}(4), 523–527. doi:10.1901/jaba.1978.11-523

Hawkins, R. P., & Dotson, V. A. (1975). Reliability scores that delude: An Alice in wonderland trip through the misleading characteristics of inter-observer agreement scores in interval recording. In E. Ramp & G. Semb (Eds.), *Behavior  analysis: Areas of research and application *(pp. 359–376). Prentice Hall.

Hubálek, Z. (1982). Coefficients of association and similarity, based on binary (presence-absence) data: An evaluation. \emph{Biological Reviews, 57}(4), 669–689. doi:10.1111/j.1469-185X.1982.tb00376.x

Hurlbert, S. H. (1969). A coefficient of interspecific assciation. \emph{Ecology, 50}(1), 1–9. doi:10.2307/1934657

Jaccard, P. (1901). Étude comparative de la distribution florale dans une portion des Alpes et des Jura. \emph{Bulletin Del La Société Vaudoise Des Sciences Naturelles, 37}, 547–579.

Jaccard, P. (1912). The distribution of the flora in the alpine zone. \emph{The New Phytologist, 11}(2), 37–50.

Johnson, S. C. (1967). Hierarchical clustering schemes. \emph{Psychometrika, 32}(3), 241–254. doi:10.1007/BF02289588

Jones, J. A., & Harrold, M. J. (2005). Empirical evaluation of the tarantula automatic fault-localization technique. Proceedings of the 20th IEEE/ACM International Conference on Automated Software Engineering, 273–282. doi:10.1145/1101908.1101949

Kent, R. N., & Foster, S. L. (1977). Direct observational procedures: Methodological issues in naturalistic settings. In A. R. Ciminero, K. S. Calhoun, & H. E. Adams (Eds.), \emph{Handbook of behavioral assessment} (pp. 279–328). New York, NY: Wiley. http://archive.org/details/handbookofbehavi00cimi

Kuder, G. F., & Richardson, M. W. (1937). The theory of the estimation of test reliability. \emph{Psychometrika, 2}(3), 151–160. doi:10.1007/BF02288391

Kulczynski, S. (1927). Die Pflanzenassoziationen der Pieninen. \emph{Bulletin International de l’Academie Polonaise Des Sciences et Des Lettres, Classe Des Sciences Mathematiques et Naturelles, B (Sciences Naturelles)}, II, 57–203.

Ling, M. H. T. (2010). COPADS, I: Distance coefficients between two lists or sets. \emph{The Python Papers Source Codes, 2}(2), 1–31.

Loevinger, J. (1947). A systematic approach to the construction and evaluation of tests of ability. \emph{Psychological Monographs, 61}(4), i–49. doi:10.1037/h0093565

Maxwell, A. E., & Pilliner, A. E. (1968). Deriving coefficients of reliability and agreement for ratings. \emph{The British Journal of Mathematical and Statistical Psychology, 21}(1), 105–116. doi:10.1111/j.2044-8317.1968.tb00401.x

McConnaughey, B. H. (1964). The determination and analysis of plankton communities. \emph{Marine Research, 7}, 1–40.

Michael, E. L. (1920). Marine Ecology and the coefficient of association: A plea in behalf of quantitative biology. \emph{Journal of Ecology, 8}(1), 54–59. doi:10.2307/2255213

Mountford, M. D. (1962). An index of similarity and its application to classification problems. In P. W. Murphy & D. Phil (Eds.), \emph{Progress in soil zoology} (pp. 43–50). Butterworths.

Nei, M., & Li, W. H. (1979). Mathematical model for studying genetic variation in terms of restriction endonucleases. \emph{Proceedings of the National Academy of Sciences, 76}(10), 5269–5273. doi:10.1073/pnas.76.10.5269

Ochiai, A. (1957). Zoogeographical studies on the soleoid fishes found in Japan and its neighbouring regions-I. \emph{Nippon Suisan Gakkaishi, 22}(9), 522–525. doi:10.2331/suisan.22.522

Otsuka, Y. (1936). The faunal character of the Japanese Pleistocene marine Mollusca, as evidence of the climate having become colder during the Pleistocene in Japan. \emph{Bulletin of the Biogeographical Society of Japan, 6}(16), 165–170.

Pearson, K. (1900). Contributions to the mathematical theory of evolution. VII. On the correlation of characters not quantitatively measurable. \emph{Philosophical Transactions of the Royal Society of London, 195}, 1–405. doi:10.1098/rsta.1900.0022

Pearson, K. (1904). \emph{Contributions to the Mathematical Theory of Evolution. XIII. On the theory of contingency and its relation to association and normal correlation}. Dulau and Co.

Pearson, K., & Heron, D. (1913). On theories of association. \emph{Biometrika, 9}(1/2), 159–315. doi:10.2307/2331805

Peirce, C. S. (1884). The numerical measure of the success of predictions. \emph{Science, 4}(93), 453–454. doi:10.1126/science.ns-4.93.453-a

Rogers, D. J., & Tanimoto, T. T. (1960). A computer program for classifying plants. \emph{Science, 132}(3434), 1115–1118. doi:10.1126/science.132.3434.1115

Rogot, E., & Goldberg, I. D. (1966). A proposed index for measuring agreement in test-retest studies. \emph{Journal of Chronic Diseases, 19}(9), 991–1006. doi:10.1016/0021-9681(66)90032-4

Russell, P. F., & Rao, T. R. (1940). On habitat and association of species of anopheline larvae in south-eastern Madras. \emph{Journal of the Malaria Institute of India, 3}(1), 153–178.

Scott, W. A. (1955). Reliability of content analysis: The case of nominal scale coding. \emph{The Public Opinion Quarterly, 19}(3), 321–325.

Simpson, G. G. (1943). Mammals and the nature of continents. \emph{American Journal of Science, 241}(1), 1–31. doi:10.2475/ajs.241.1.1

Simpson, G. G. (1960). Notes on the measurement of faunal resemblance. \emph{American Journal of Science, 258-A}, 300–311.

Sokal, P. H. A., & Sneath, R. R. (1963). Principles of numerical taxonomy. W.H. Freeman and Company.

Sokal, R., & Michener, C. (1958). A statistical method for evaluating systematic relationships. University of Kansas Science Bulletin.

Sorgenfrei, T. (1958). Molluscan assemblages from the marine middle miocene of south Jutland and their Environments. Vol. I. \emph{Danmarks Geologiske Undersøgelse II. Række}, 79, 1–355. doi:10.34194/raekke2.v79.6868

Stiles, H. E. (1961). The association factor in information retrieval. \emph{Journal of the ACM, 8}(2), 271–279. doi:10.1145/321062.321074

Tanimoto, T. T. (1958). An elementary mathematical theory of classification and prediction (PB167360). International Business Machines Corp., New York, NY.

Tarwid, K. (1960). Szacowanie zbieinosci nisz ekologicznych gatunkow droga oceny prawdopodobienstwa spotykania sie ich w polowach. \emph{Ekologia Polska, 6}, 115–130.

Tulloss, R. E. (1997). Assessment of similarity indices for undesirable properties and a new tripartite similarity index based on cost functions. In M. E. Palm & I. H. Chapela (Eds.), \emph{Mycology in sustainable development: Expanding concepts, vanishing borders} (pp. 122–143). Parkway Pub.

Warrens, M. J. (2008). Similarity coefficients for binary data: Properties of coefficients, coefficient matrices, multi-way metrics and multivariate coefficients \link{Doctoral dissertation, Universiteit Leiden}. https://hdl.handle.net/1887/12987

Yule, G. U. (1900). On the association of attributes in statistics: With illustrations from the material of the childhood society, &c. \emph{Philosophical Transactions of the Royal Society of London, 194}, 257–319. doi:10.1098/rsta.1900.0019

Yule, G. U. (1912). On the methods of measuring association between two attributes. \emph{Journal of the Royal Statistical Society, 75}(6), 579–652. doi:10.2307/2340126
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
