% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_bonett_price_r.R
\name{es_bonett_price_r}
\alias{es_bonett_price_r}
\title{Bonett and Price rho}
\usage{
es_bonett_price_r(
  field1,
  field2,
  categories1 = NULL,
  categories2 = NULL,
  version = 2
)
}
\arguments{
\item{version}{c(1, 2) Optional either version 1 for rho, or version 2 (default) for rho-hat}

\item{var1}{A vector with the binary data from the first variable}

\item{var2}{A vector with the binary data from the second variable}
}
\value{
Bonett and Price r
}
\description{
An approximation for the tetrachoric correlation coefficient.
}
\details{
An approximation for the tetrachoric correlation coefficient.

Formula for version 1 is (Bonett & Price, 2005, p. 216):
\deqn{\rho^* = \cos\left(\frac{\pi}{1+\omega^c}\right)}
With:
\deqn{\omega = OR = \frac{a\times d}{b\times c}}
\deqn{c = \frac{1-\frac{\left|R_1-C_1\right|}{5\times n} - \left(\frac{1}{2}-p_{min}\right)^2}{2}}
\deqn{p_{min} = \frac{\text{MIN}\left(R_1, R_2, C_1, C_2\right)}{n}}

Formula for version 2 is  (Bonett & Price, 2005, p. 216):
\deqn{\hat{\rho}^* = \cos\left(\frac{\pi}{1+\hat{\omega}^{\hat{c}}}\right)}
with:
\deqn{\hat{\omega} = \frac{\left(a+\frac{1}{2}\right)\times \left(d+\frac{1}{2}\right)}{\left(b+\frac{1}{2}\right)\times \left(c+\frac{1}{2}\right)}}
\deqn{\hat{c} = \frac{1-\frac{\left|R_1-C_1\right|}{5\times \left(n+2\right)} - \left(\frac{1}{2}-\hat{p}_{min}\right)^2}{2}}
\deqn{\hat{p}_{min} = \frac{\text{MIN}\left(R_1, R_2, C_1, C_2\right)+1}{n+2}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
\item \eqn{R_i} the sum of counts in the i-th row
\item \eqn{C_i} the sum of counts in the i-th column
\item \eqn{n} the sum of all counts
}
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_bonett_price_r(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Bonett, D. G., & Price, R. M. (2005). Inferential methods for the tetrachoric correlation coefficient. \emph{Journal of Educational and Behavioral Statistics, 30}(2), 213–225. https://doi.org/10.3102/10769986030002213
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
