% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_camp_r.R
\name{es_camp_r}
\alias{es_camp_r}
\title{Camp r}
\usage{
es_camp_r(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Camp r
}
\description{
An approximation for the tetrachoric correlation coefficient.
}
\details{
Camp (1934, pp. 309) describes the following steps for the calculation:
Step 1: If total of column 1 (C1) is less than column 2 (C2), swop the two columns

Step 2: Calculate \eqn{p = \frac{C1}{n}}, \eqn{p_1 = \frac{a}{n}}, and \eqn{p_2 = \frac{c}{C2}}

Step 3: Determine \eqn{z_1}, \eqn{z_2}, \eqn{z_3} as the normal deviate
corresponding to the area \eqn{p_1}, \eqn{p_2}, \eqn{p} resp. (inverse standard normal cumulative distribution)

Step 4: Determine y the normal ordinate corresponding to \eqn{z_3} (the height of the normal distribution)

Step 5: Calculate \eqn{m = \frac{p\times\left(1-p\right)\times\left(z_1 + z_2\right)}{y}}

Step 6: Find phi in a table of phi values\tabular{llllllllllll}{
   p \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab 10 \cr
   0.5 \tab 0.637 \tab 0.636 \tab 0.636 \tab 0.635 \tab 0.635 \tab 0.634 \tab 0.634 \tab 0.633 \tab 0.633 \tab 0.632 \tab 0.631 \cr
   0.6 \tab 0.631 \tab 0.631 \tab 0.630 \tab 0.629 \tab 0.628 \tab 0.627 \tab 0.626 \tab 0.625 \tab 0.624 \tab 0.622 \tab 0.621 \cr
   0.7 \tab 0.621 \tab 0.620 \tab 0.618 \tab 0.616 \tab 0.614 \tab 0.612 \tab 0.610 \tab 0.608 \tab 0.606 \tab 0.603 \tab 0.600 \cr
   0.8 \tab 0.600 \tab 0.597 \tab 0.594 \tab 0.591 \tab 0.587 \tab 0.583 \tab 0.579 \tab 0.574 \tab 0.569 \tab 0.564 \tab 0.559 \cr
}


Step 7: Calculate \eqn{r_t = \frac{m}{\sqrt{1+\phi\times m^2}}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
\item \eqn{R_i} the sum of counts in the i-th row
\item \eqn{C_i} the sum of counts in the i-th column
}

Cureton (1968) describes quite a few shortcomings with this approximation, and circumstances when it might be appropriate.
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_camp_r(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Camp, B. H. (1934). \emph{Mathematical part of elementary statistics}. D.C. Heath and Company, London.

Cureton, E. E. (1968). Tetrachoric correlation by the Camp approximation. \emph{Educational and Psychological Measurement, 28}(2), 239–244. https://doi.org/10.1177/001316446802800202
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
