% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_d_os.R
\name{es_cohen_d_os}
\alias{es_cohen_d_os}
\title{Cohen d' (for one-sample)}
\usage{
es_cohen_d_os(data, mu = NULL)
}
\arguments{
\item{data}{pandas series with the numeric scores}

\item{mu}{optional parameter to set the hypothesized mean. If not used the midrange is used}
}
\value{
Cohen d'. mu is also printed if not provided.
}
\description{
This function will calculate Cohen d' (one-sample). An effect size measure that can be used with a test for a single mean (for example a one-sample Student t-test).
}
\details{
The test is also described at \href{https://peterstatistics.com/Terms/EffectSizes/CohenD.html}{PeterStatistics.com}

The formula used (Cohen, 1988, p. 46):
\deqn{d'=\frac{\bar{x}-\mu_{H_{0}}}{s}}
With:
\deqn{s = \sqrt{\frac{\sum_{i=1}^n \left(x_i - \bar{x}\right)^2}{n - 1}}}
\deqn{\bar{x} = \frac{\sum_{i=1}^n x_i}{n}}

\emph{Symbols used:}
\itemize{
\item \eqn{\bar{x}} the sample mean
\item \eqn{\mu_{H_0}} the hypothesized mean in the population
\item \eqn{n} the sample size (i.e. the number of scores)
\item \eqn{s} the unbiased sample standard deviation
\item \eqn{x_i} the i-th score
}

Note to use a rule-of-thumb from Cohen d, first convert this to a regular Cohen d using
\emph{es_convert(d', from="cohendos", to="cohend")}, then use \emph{th_cohen_d(d)}

Or convert it further to an Odds Ratio using, \emph{es_convert(d, from="cohend", to="or", ex1="chinn")} or \emph{es_convert(d, from="cohend", to="or", ex1="borenstein")}. Then use \emph{th_odds_ratio(or)}
}
\section{Before, After and Alternatives}{

Before this you might want to perform a test:
\code{\link{ts_student_t_os}}, for One-Sample Student t-Test.
\code{\link{ts_trimmed_mean_os}}, for One-Sample Trimmed (Yuen or Yuen-Welch) Mean Test.
\code{\link{ts_z_os}}, for One-Sample Z-Test.

After this you might want a rule-of-thumb for the effect size, first convert to regular Cohen d:
\code{\link{es_convert}}, to convert Cohen's d one-sample to Cohen d, use \emph{fr = "cohendos"} and \emph{to = "cohend"}.
\code{\link{th_cohen_d}}, for rules-of-thumb for Cohen d.

Alternative Effect Sizes:
\code{\link{es_hedges_g_os}}, for Hedges g.
\code{\link{es_common_language_os}}, for the Common Language Effect Size.
}

\examples{
#Example 1: Numeric dataframe
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
df2 = read.csv(file2, sep=';', na.strings=c("", "NA"))
ex1 = df2['Gen_Age']
es_cohen_d_os(ex1)

#Example 2: Numeric list
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
es_cohen_d_os(ex2)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
