% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_d_ow.R
\name{es_cohen_d_ow}
\alias{es_cohen_d_ow}
\title{Cohen d (for one-way ANOVA)}
\usage{
es_cohen_d_ow(scores, groups)
}
\arguments{
\item{scores}{the numeric scores variable}

\item{groups}{the groups variable}
}
\value{
the effect size value
}
\description{
Cohen d (for one-way ANOVA)
}
\details{
The formula used is (Cohen, 1988, p. 276):
\deqn{d = \frac{\bar{x}_{max} - \bar{x}_{min}}{\sigma}}
With:
\deqn{\sigma = \sqrt{\frac{SS_w}{n}}}
\deqn{SS_w = \sum_{j=1}^k \sum_{i=1}^{n_j} \left(x_{i,j}-\bar{x}_j\right)^2}
\deqn{\bar{x}_j = \frac{\sum_{i=1}^{n_j} x_{i,j}}{n_j}}

\emph{Symbols}
\itemize{
\item \eqn{x_{i,j}} the i-th score in category j
\item \eqn{n_j} the number of scores in category j
\item \eqn{k} the number of categories
\item \eqn{\bar{x}_j} the mean of the scores in category j
\item \eqn{SS_w} the within sum of squares (sum of squared deviation of the mean)
}
}
\examples{
scores = c(20, 50, 80, 15, 40, 85, 30, 45, 70, 60, NA, 90, 25, 40, 70, 65, NA, 70, 98, 40, 65, 60, 35, NA, 50, 40, 75, NA, 65, 70, NA, 20, 80, 35, NA, 68, 70, 60, 70, NA, 80, 98, 10, 40, 63, 75, 80, 40, 90, 100, 33, 36, 65, 78, 50)
groups = c("Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Haarlem", "Diemen", "Haarlem", "Haarlem", "Haarlem", "Haarlem", "Haarlem")
es_cohen_d_ow(scores, groups)

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
