% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_g.R
\name{es_cohen_g}
\alias{es_cohen_g}
\title{Cohen's g}
\usage{
es_cohen_g(data, codes = NULL)
}
\arguments{
\item{data}{A vector with the data}

\item{codes}{Optional vector with the two codes to use}
}
\value{
Cohen's g as a single value
}
\description{
Cohen's g
}
\details{
The formula used is (Cohen, 1988, p. 147):
\deqn{g=p-0.5}

\emph{Symbols used}:
\itemize{
\item \eqn{p} is the sample proportion
}

For a classification use: \emph{th_cohen_g(g)}

\strong{Alternative}

I'm not aware of any alternative library that has this function.
}
\examples{
data <- c("Female", "Male", "Male", "Female", "Male", "Male", "Female", "Female", "Male", "Male", "Male", "Male", "Male", "Male", "Female", "Male", "Female", "Male", "Male")
es_cohen_g(data)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
