% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_h.R
\name{es_cohen_h}
\alias{es_cohen_h}
\title{Cohen h}
\usage{
es_cohen_h(p1, p2)
}
\arguments{
\item{p1}{the first proportion}

\item{p2}{the second proportion}
}
\value{
Cohen h
}
\description{
Cohen h
}
\details{
Formula used (Cohen, 1988, p. 181):
\deqn{h=\phi_{1}-\phi_{2}}
With:
\deqn{\phi_{i}=2\times\textup{arcsin}\sqrt{p_{i}}}

\emph{Symbols used}:
\itemize{
\item \eqn{p_i} the proportion of cases in category i
}

For classification rule-of-thumb use: \emph{th_cohen_h()}
}
\examples{
es_cohen_h(0.2, 0.4)

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
