% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_h_os.R
\name{es_cohen_h_os}
\alias{es_cohen_h_os}
\title{Cohen's h'}
\usage{
es_cohen_h_os(data, codes = NULL, p0 = 0.5)
}
\arguments{
\item{data}{A vector with the data}

\item{codes}{Optional vector with the two codes to use}

\item{p0}{Optional the hypothesized proportion for the first category (default is 0.5)}
}
\value{
Cohen's h'
}
\description{
Cohen's h'
}
\details{
An adaptation of Cohen h for a one-sample case.

Formula used (Cohen, 1988, p. 202):
\deqn{h'=\phi_{1}-\phi_{h_0}}
With:
\deqn{\phi_{i}=2\times\textup{arcsin}\sqrt{p_{i}}}
\deqn{p_i = \frac{F_i}{n}}
\deqn{n = \sum_{i=1}^k F_i}

\emph{Symbols used}:
\itemize{
\item \eqn{F_i} is the (absolute) frequency (count) of category i
\item \eqn{n} is the sample size, i.e. the sum of all frequencies
\item \eqn{p_i} the proportion of cases in category i
\item \eqn{p_{h_0}} the expected proportion (i.e. the proportion according to the null hypothesis)
}

For classification rule-of-thumb use a conversion to Cohen h: \emph{es_convert(h2, from="cohenhos", to="cohenh")}

Then use: \emph{th_cohen_h()}

\strong{Alternative}

The \emph{'pwr'} library has a similar function: \emph{ES.h()}
}
\examples{
data <- c("Female", "Male", "Male", "Female", "Male", "Male", "Female", "Female", "Male", "Male", "Male", "Male", "Male", "Male", "Female", "Male", "Female", "Male", "Male")
es_cohen_h_os(data)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
