% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_kappa.R
\name{es_cohen_kappa}
\alias{es_cohen_kappa}
\title{Cohen Kappa}
\usage{
es_cohen_kappa(nom1, nom2, ase = "exact")
}
\arguments{
\item{nom1}{the scores on the first variable}

\item{nom2}{the scores on the second variable}

\item{ase}{c("exact", "approximate") optional to indicate which method to use to calculate asymptotic standard errors}
}
\value{
dataframe with the effect size value, the asymptotic standard errors (assuming null and alternative), test statistic, degrees of freedom, and p-value (sig.)
}
\description{
Cohen Kappa
}
\details{
The formula used is (Cohen, 1960, p. 40):
\deqn{\kappa = \frac{p_0 - p_c}{1 - p_c}}

With:
\deqn{p_0 = \frac{P}{n}}
\deqn{p_c = \frac{Q}{n^2}}
\deqn{P = \sum_{i=1}^r F_{i,i}}
\deqn{Q = \sum_{i=1}^r R_i \times C_i}
\deqn{R_i = \sum_{j=1}^c F_{i,j}}
\deqn{C_j = \sum_{i=1}^r F_{i,j}}

\emph{Symbols used}
\itemize{
\item \eqn{r} is the number of rows (categories in the first variable)
\item \eqn{c} is the number of columns (categories in the second variable)
\item \eqn{n} is the total number of scores
\item \eqn{F_{i,j}} is the frequency (count) of scores equal to the i-th category in the first variable, and the j-th category in the second.
}

The approximate asymptotic standard errors (ase="approximate") use (Cohen, 1960, pp. 43-44):
\deqn{ASE_0 \approx \sqrt{\frac{p_c}{n\times \left(1 - p_c\right)}}}
\deqn{ASE_1 \approx \sqrt{\frac{p_0 \times \left(1 - p_0\right)}{n\times \left(1 - p_c\right)^2}}}

The exact asymptotic standard errors (ase="exact") use (Fleiss et al., 1969, p. 325):
\deqn{ASE_0 = \sqrt{\frac{SS_0}{n\times\left(1 - p_c\right)^2}}}
\deqn{ASE_1 = \sqrt{\frac{SS_1}{n\times\left(1 - p_c\right)^4}}}
With:
\deqn{SS_0 = \left(\sum_{i=1}^r p_{i,.} \times p_{.,i} \times \left(1 - \left(p_{.,i} + p_{i,.}\right)\right)^2\right) - p_c^2 + \left(1 - p_0\right)^2 \times \sum_{i=1}^r \underset{j\neq i}{\sum_{j=1}^c} p_{i,.} \times p_{.,j} \times \left(p_{.,i} + p_{j,.}\right)^2}
\deqn{SS_1 = \left(\sum_{i=1}^r p_{i,i} \times \left(\left(1 - p_c\right) - \left(p_{.,i} + p_{i,.}\right)\times\left(1 - p_0\right)\right)^2\right) - \left(p_0\times p_c - 2\times p_c + p_0\right)^2 + \left(1 - p_0\right)^2 \times \sum_{i=1}^r \underset{j\neq i}{\sum_{j=1}^c} p_{i,j}\times\left(p_{.,i} + p_{j,.}\right)^2}
\deqn{p_{i,j} = \frac{F_{i,j}}{n}}
\deqn{p_{i,.} = \frac{R_i}{n}}
\deqn{p_{.,j} = \frac{C_j}{n}}

The test is then performed using (Cohen, 1960, p. 44):
\deqn{z_{\kappa} = \frac{\kappa}{ASE_0}}
\deqn{sig. = 2\times\left(1 - \Phi\left(\left|z_{\kappa}\right|\right)\right)}

Where \eqn{\Phi\left(\dots\right)} is the cumulative density function of the standard normal distribution
}
\references{
Cohen, J. (1960). A coefficient of agreement for nominal scales. \emph{Educational and Psychological Measurement, 20}(1), 37–46. https://doi.org/10.1177/001316446002000104

Fleiss, J. L., Cohen, J., & Everitt, B. S. (1969). Large sample standard errors of kappa and weighted kappa. \emph{Psychological Bulletin, 72}(5), 323–327. https://doi.org/10.1037/h0028106
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
