% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_u.R
\name{es_cohen_u}
\alias{es_cohen_u}
\title{Cohen U}
\usage{
es_cohen_u(d, version = "u3")
}
\arguments{
\item{d}{the Cohen d_s value (a.k.a. Hedges g)}

\item{version}{c("u3", "u2", "u1"), Optional, the version of Cohen U to determine}
}
\value{
The Cohen U value
}
\description{
Cohen (1988, p. 23) provided three measures that relate to Cohen's \eqn{d_s} (a.k.a. Hedges g).
\itemize{
\item  \eqn{U_1}, is (supposedly) the proportion of non-overlap between distributions
\item  \eqn{U_2}, is (supposedly) the proportion of overlap between distributions
\item  \eqn{U_3}, is the proportion of one group's scores below the mean of another group
}

\eqn{U_1} and \eqn{U_2} are probably the least used of these three, since most likely Cohen was wrong in his calculation (Grice & Barrett, 2014). A nice interactive visualisation of the relation between Cohen and the Common Language Effect size, can be found at https://rpsychologist.com/therapist-effects. It actually corrects Cohen U.

By converting each back to Cohen's d, the rule-of-thumb from Cohen d could be used as classification.
}
\details{
The following formulas are used (Cohen, 1988, p. 23):
\deqn{U_3 = \Phi\left(d\right)}
\deqn{U_2 = \Phi\left(\frac{d}{2}\right)}
\deqn{U_1 = \Phi\left(\frac{2\times U_2 - 1}{U_2}\right)}

\emph{Symbols used:}
\itemize{
\item  \eqn{d}, Cohen's d_s value, the Cohen d for independent samples, a.k.a. Hedges g.
\item  \eqn{\Phi\left(\dots\right)} the cumulative density function of the standard normal distribution
}
}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\seealso{
\code{\link{es_convert}}, to convert an U to Cohen d use \verb{fr="cohenu.", to="cohend"}.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
