% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cole_c5.R
\name{es_cole_c5}
\alias{es_cole_c5}
\title{Cole C5}
\usage{
es_cole_c5(var1, var2)
}
\arguments{
\item{var1}{A vector with the binary data from the first variable}

\item{var2}{A vector with the binary data from the second variable}
}
\value{
Cole C5
}
\description{
Cole C5
}
\details{
The formula used (Cole, 1949, p. 416):
\deqn{C_5 = \frac{\sqrt{2}\times\left(a\times d - b\times c\right)}{\sqrt{\left(a\times d - b\times c\right)^2 + R1\times R2\times C1\times C2}}}
\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
\item \eqn{R_i} the sum of counts in the i-th row
\item \eqn{C_i} the sum of counts in the i-th column
}

This is actually an adjustment of the contingency coefficient, by dividing it over \eqn{\sqrt{\frac{1}{2}}}, which is the maximum value in 2x2 tables.

Note that Cole C2 is the phi coefficient, Cole C3 is McEwen-Michael coefficient, Cole C4 is Yule Q, and Cole C6 is Yule r.
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
es_cole_c5(bin1, bin2)

}
\references{
Cole, L. C. (1949). The measurement of interspecific associaton. \emph{Ecology, 30}(4), 411–424. https://doi.org/10.2307/1932444
}
\author{
P. Stikker
Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
