% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cont_coeff.R
\name{es_cont_coeff}
\alias{es_cont_coeff}
\title{(Pearson) Contingency Coefficient}
\usage{
es_cont_coeff(chi2, n, adj = NULL, r = NULL, c = NULL)
}
\arguments{
\item{chi2}{the chi-square test statistic}

\item{n}{the sample size}

\item{adj}{c(NULL, "sakoda") adjustment to use (see details)}

\item{r}{the number of rows, required if adj="sakado"}

\item{c}{the number of columns, required if adj="sakado"}
}
\value{
value of (Pearson) Contingency Coefficient
}
\description{
(Pearson) Contingency Coefficient
}
\details{
The formula used is (Pearson, 1904, p. 9):
\deqn{C = \sqrt\frac{\chi^{2}}{n + \chi^{2}}}

\emph{Symbols used}:
\itemize{
\item \eqn{\chi^{2}} the chi-square test statistic
\item \eqn{n} the sample size, i.e. the sum of all frequencies
}

The maximum value for C would be (Sakoda, 1977, p. 778):
\deqn{C_{max} = \sqrt{\frac{m - 1}{m}}}
Where \eqn{m} is the minimum of the number of rows, or number of columns.

Sakoda propses to divide the contingency coefficient by this maximum, i.e.
\deqn{\frac{C}{C_{max}}}

For a 2x2 table Cole C1 will also divide by the maximum and produce the same result.

\strong{Alternative}

The \emph{'DescTools'} library also has a function for this: \emph{ContCoef()}
}
\examples{
chi2Value <- 3.105263
n <- 19
es_cont_coeff(chi2Value, n)

}
\references{
Pearson, K. (1904). \emph{Contributions to the Mathematical Theory of Evolution. XIII. On the theory of contingency and its relation to association and normal correlation}. Dulau and Co.

Sakoda, J. M. (1977). Measures of Association for Multivariate Contingency Tables. \emph{In Proceedings of the Social Statistics Section of the American Statistical Association: Vol. Part III} (pp. 777–780).
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
