% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_es.R
\name{es_convert}
\alias{es_convert}
\title{Convert Effect Size}
\usage{
es_convert(es, fr, to, ex1 = NULL, ex2 = NULL)
}
\arguments{
\item{es}{the effect size value to convert}

\item{fr}{name of the original effect size (see details)}

\item{to}{name of the effect size to convert to (see details)}

\item{ex1}{extra for some conversions (see details)}

\item{ex2}{extra for some conversions (see details)}
}
\value{
the converted effect size value
}
\description{
Convert Effect Size
}
\details{
\strong{COHEN D}

\strong{Cohen d to Odds Ratio}

fr="cohend", to="or", ex1="chinn"

This uses (Chinn, 2000, p. 3129):

\deqn{OR = e^{d\times 1.81}}

fr="cohend", to="or", ex1="borenstein"

This uses (Borenstein et. al, 2009, p. 3):

\deqn{OR = e^{\frac{d\times\pi}{\sqrt{3}}}}

\strong{Cohen d to Rank Biserial (Cliff delta)}

fr = "cohend", to = "rb"

This uses (Marfo & Okyere, 2019, p. 4):
\deqn{r_b = \frac{2\times\Phi\left(\frac{d}{2}\right)-1}{\Phi\left(\frac{d}{2}\right)}}

\strong{COHEN D'}

\strong{Convert a Cohen d' to Cohen d}

fr="cohendos" to="cohend"

This uses (Cohen , 1988, p. 46):
\deqn{d = d'\times\sqrt{2}}

\strong{COHEN F}

\strong{Cohen f to Eta-squared}

fr="cohenf" to="etasq"

This uses (Cohen, 1988, p. 284):
\deqn{\eta^2 = \frac{f^2}{1 + f^2}}

\strong{COHEN H'}

\strong{Cohen h' to Cohen h}

fr = "cohenhos", to = "cohenh"

This uses (Cohen, 1988, p. 203):
\deqn{h = h'\times\sqrt{2}}

\strong{COHEN U}

\strong{Cohen U to Cohen d}

fr="cohenu1", to="cohend"
fr="cohenu2", to="cohend"
fr="cohenu3", to="cohend"

This uses (Cohen, 1988, p. 23):
\deqn{d = \Phi^{-1}\left(U_3\right)}
\deqn{d = 2\times\Phi^{-1}\left(U_2\right)}
\deqn{d = 2\times\Phi^{-1}\left(\frac{1}{2 - U_1}\right)}

\strong{COHEN w}

\strong{Cohen w to Contingency Coefficient}

fr="cohenw", to="cc"

\emph{Cohen w to Cramér V GoF}

fr="cohenw", to="cramervgof", ex1=k

This uses (Cohen, 1988, p. 223):
\deqn{v = \frac{w}{\sqrt{k - 1}}}

\emph{Cohen w to Cramér V ind.}

fr="cohenw", to="cramervind", ex1=r, ex2=c

This uses:
\deqn{v = \frac{w}{\sqrt{\min\left(r - 1, c - 1\right)}}}

\emph{Cohen w to Fei}

fr="cohenw", to="fei", ex1=minExp/n

This uses:
\deqn{Fei = \frac{w}{\sqrt{\frac{1}{p_E}-1}}}

\strong{CRAMER V GoF}

\strong{Cramer's v for Goodness-of-Fit to Cohen w}

fr="cramervgof", to = "cohenw", ex1 = k

This uses (Cohen, 1988, p. 223):
\deqn{w = v\times\sqrt{df}}

\strong{EPSILON SQUARED}

\strong{Epsilon Squared to Eta Squared}

fr="epsilonsq", to="etasq", ex1 = n, ex2 = k

This uses:
\deqn{\eta^2 = 1 - \frac{\left(1 - \epsilon^2\right)\times\left(n - k\right)}{n - 1}}

\strong{Epsilon Squared to Omega Squared}

fr="epsilonsq", to="omegasq", ex1=MS_w, ex2 = SS_t

This uses:
\deqn{\hat{\omega}^2 = \epsilon^2\times\left(1 - \frac{MS_w}{SS_t + MS_w}\right)}

\strong{ETA SQUARED}

\strong{Eta squared to Cohen f}
fr="etasq", to="cohenf")

This uses:
\deqn{f = \sqrt{\frac{\eta^2}{1 - \eta^2}}}

\strong{Eta squared to Epsilon Squared}
fr="etasq", to="epsilonsq", ex1=n, ex2=k

This uses:
\deqn{\epsilon^2 = \frac{n\times\eta^2 - k + \left(1 - \eta^2\right)}{n - l}}

\strong{FEI}

\emph{Fei to Cohen w}

fr="fei", to="cohenw", ex1=minExp/n

This uses:
\deqn{w = Fei\times\sqrt{\frac{1}{p_E}-1}}

\emph{Fei to Johnston-Berry-Mielke E}
fr="fei", to="jbme"

This uses:
\deqn{E = Fei^2}

\strong{JOHNSTON-BERRY-MIELKE}

\strong{Johnston-Berry-Mielke E to Cohen w}

fr="jbme", to="cohenw", ex1=minExp/n

This uses (Johnston et al., 2006, p. 413):
\deqn{w = \sqrt{\frac{E\times\left(1 - \right)}{q}}}

\emph{Johnston-Berry-Mielke E to Cohen w}

fr="jbme", to="fei"

This uses:
\deqn{Fei = \sqrt(E)}

\strong{ODDS RATIO}

\strong{Odds Ratio to Cohen d}

fr="or", to="cohend", ex1="chinn"

This uses (Chinn, 2000, p. 3129):

\deqn{d = \frac{\ln{\left(OR\right)}}{1.81}}

fr="or", to="cohend", ex1="borenstein"

This uses (Borenstein et. al, 2009, p. 3):

\deqn{d = \ln\left(OR\right)\times\frac{\sqrt{3}}{\pi}}

\strong{Odds Ratio to Yule Q}

fr="or", to="yuleq"

This uses:

\deqn{Q = \frac{OR - 1}{OR + 1}}

\strong{Odds Ratio to Yule Y}

This uses
\deqn{Y = \frac{\sqrt{OR} - 1}{\sqrt{OR} + 1}}

\strong{OMEGA SQUARED}

\strong{Omega Squared to Epsilon Squared}

fr="omegasq", to="epsilonsq", ex1=MS_w, ex2 = SS_t

This uses:
\deqn{\epsilon^2 = \frac{\hat{\omega}^2}{1 - \frac{MS_w}{SS_t + MS_w}}}

\strong{RANK BISERIAL (CLIFF DELTA)}

\strong{Rank Biserial (Cliff delta) to Cohen d}

fr = "rb", to = "cohend"

This uses (Marfo & Okyere, 2019, p. 4):
\deqn{d =\sqrt{2} \times \Phi^{-1}\left(-\frac{1}{r_b-2}\right)}

\strong{Rank Biserial (Cliff delta) to Vargha-Delaney A}

fr = "rb", to = "vda"

This uses:
\deqn{r_b = 2\times A - 1}

\strong{VARGHA-DELANEY A}

\strong{Vargha-Delaney A to Rank Biserial (Cliff delta)}

fr = "vda", to = "rb"

This uses:
\deqn{A = \frac{r_b + 1}{2}}

\strong{YULE Q}

\strong{Yule Q to Odds Ratio}

fr="yuleq", to="or"

This uses:

\deqn{OR = \frac{1 + Q}{1 - Q}}

\strong{Yule Q to Yule Y}

fr="yuleq", to="yuley"

This uses:
\deqn{Y = \frac{1 - sqrt{1 - Q^2}}{Q}}

\strong{YULE Y}

\strong{Yule Y to Yule Q}

fr="yuley", to=="yuleq"

This uses:
\deqn{Q = \frac{2\times Y}{1 + Y^2}}

\strong{Yule Y to Odds Ratio}

fr="yuley", to=="or"

This uses
\deqn{OR = \left(\frac{1 + Y}{1 - Y}\right)^2}
}
\examples{
es_convert(0.3, fr="cohenhos", to = "cohenh")

}
\references{
Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009). Converting Among Effect Sizes. In \emph{Introduction to Meta-Analysis}. John Wiley & Sons, Ltd. https://doi.org/10.1002/9780470743386

Chinn, S. (2000). A simple method for converting an odds ratio to effect size for use in meta-analysis. \emph{Statistics in Medicine, 19}(22), 3127–3131. https://doi.org/10.1002/1097-0258(20001130)19:22<3127::aid-sim784>3.0.co;2-m

Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.

Johnston, J. E., Berry, K. J., & Mielke, P. W. (2006). Measures of effect size for chi-squared and likelihood-ratio goodness-of-fit tests. \emph{Perceptual and Motor Skills, 103}(2), 412–414. https://doi.org/10.2466/pms.103.2.412-414
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
