% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cramer_v_ind.R
\name{es_cramer_v_ind}
\alias{es_cramer_v_ind}
\title{Cramer's V for Independence Test}
\usage{
es_cramer_v_ind(chi2, n, r, c, cc = NULL)
}
\arguments{
\item{chi2}{the chi-square test statistic}

\item{n}{the sample size}

\item{r}{the number of categories in the first variable (i.e. the number of rows)}

\item{c}{the number of categories in the second variable (i.e. the number of columns)}

\item{cc}{c(NULL, "bergsma") optional to indicate correction to use (default is NULL)}
}
\value{
Cramer's V value
}
\description{
Cramer's V for Independence Test
}
\details{
The formula used is:
\deqn{V=\sqrt\frac{\chi^{2}}{n\times \left(\text{min}\left(r, c\right) - 1\right)}}

\emph{Symbols used}:
\itemize{
\item \eqn{r} the number of categories in the first variable (i.e. the number of rows)
\item \eqn{c} the number of categories in the second variable (i.e. the number of columns)
\item \eqn{n} the sample size, i.e. the sum of all frequencies
\item \eqn{\chi^{2}} the chi-square statistic
}

The Bergsma correction uses a different formula (Bergsma, 2013, pp. 324-325):
\deqn{V_B = \sqrt{\frac{\tilde{\varphi}^2}{\text{min}\left(\tilde{r}, \tilde{c}\right) - 1}}}
With:
\deqn{\tilde{\varphi}^2 = \text{max}\left(0,\varphi^2 - \frac{\left(r - 1\right)\times\left(c - 1\right)}{n - 1}\right)}
\deqn{\tilde{r} = r - \frac{\left(r - 1\right)^2}{n - 1}}
\deqn{\tilde{c} = r - \frac{\left(c - 1\right)^2}{n - 1}}
\deqn{\varphi^2 = \frac{\chi^{2}}{n}}
}
\examples{
chi2Val = 16.98975
n = 1941
nRows = 5
nCols = 2
es_cramer_v_ind(chi2Val, n, nRows, nCols)
es_cramer_v_ind(chi2Val, n, nRows, nCols, cc="bergsma")

}
\references{
Bergsma, W. (2013). A bias-correction for Cramér’s and Tschuprow’s. Journal of the Korean Statistical Society, 42(3), 323–328. https://doi.org/10.1016/j.jkss.2012.10.002

Cramér, H. (1946). Mathematical methods of statistics. Princeton University Press.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
