% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_dominance.R
\name{es_dominance}
\alias{es_dominance}
\title{Dominance and a Vargha-Delaney A like effect size measure}
\usage{
es_dominance(data, mu = NULL, out = "dominance")
}
\arguments{
\item{data}{vector with the numeric scores}

\item{mu}{optional parameter to set the hypothesized median. If not used the midrange is used}

\item{out}{c("dominance","vda"). optional to either show the dominance score (default), or a VDA like measure}
}
\value{
dataframe with the hypothesized median (mu) and the effect size value
}
\description{
Dominance and a Vargha-Delaney A like effect size measure
}
\details{
The formula used is (Mangiafico, 2016, p. 223-224):
\deqn{D = p_{pos} - p_{neg}}
Where:
\deqn{p_i = \frac{n_i}{n}}
\emph{Symbols used:}
\itemize{
\item \eqn{p_{pos}} the proportion of cases above the hypothesized median
\item \eqn{p_{neg}} the proportion of cases below the hypothesized median
\item \eqn{n_{pos}} the number of cases above the hypothesized median
\item \eqn{n_{neg}} the number of cases below the hypothesized median
\item \eqn{n} the total number of cases
}

The dominance score will range from -1 to 1.

A Vargha-Delaney A (VDA) style effect size is calculated with (Mangiafico, 2016, p. 223-224):
\deqn{VDA_{like} = \frac{D + 1}{2}}

This will range from 0 to 1, with 0.5 being the same as a dominance score of 0.

\strong{Alternative}
The \emph{rcompanion} library has a similar function: \emph{oneSampleDominance()}
}
\examples{
 
data <- c(1, 2, 5, 1, 1, 5, 3, 1, 5, 1, 1, 5, 1, 1, 3, 3, 3, 4, 2, 4)
es_dominance(data)
es_dominance(data, mu=2)

}
\references{
Mangiafico, S. S. (2016). Summary and analysis of extension program evaluation in R (1.20.01). Rutger Cooperative Extension.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
