% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_fei.R
\name{es_fei}
\alias{es_fei}
\title{Fei}
\usage{
es_fei(chi2, n, minExp)
}
\arguments{
\item{chi2}{the chi-square test statistic}

\item{n}{the sample size}

\item{minExp}{the minimum expected count}
}
\value{
the value of Fei
}
\description{
An effect size measure that could be used with a chi-square test or g-test.
}
\details{
The formula used (Ben-Shachar et al., 2023, p. 6):
\deqn{Fei = \\sqrt{\\frac{\\chi_{GoF}^2}{n\\times\\left(\\frac{1}{\\min\\left(p_E\\right)}-1\\right)}}}

\emph{Symbols used}
\itemize{
\item \eqn{\\chi_{GoF}^2}, the chi-square value of the goodness-of-fit chi-square test
\item \eqn{n}, the sample size
\item \eqn{p_E}, the expected proportions
}

\emph{Classification}
A qualification rule-of-thumb could be obtained by converting this to Cohen's w (use \strong{es_convert(Fei, fr="fei", to="cohenw", ex1=minExp/n)})
}
\references{
Ben-Shachar, M. S., Patil, I., Thériault, R., Wiernik, B. M., & Lüdecke, D. (2023). Phi, fei, fo, fum: Effect sizes for categorical data that use the chi-squared statistic. \emph{Mathematics, 11}(1982), 1–10. doi:10.3390/math11091982
}
\seealso{
\code{\link{es_convert}} to convert Fei to Cohen w, use from="fei", to="cohenw", and ex1=minExp/n
\code{\link{th_cohen_w}} rules-of-thumb for Cohen w
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
