% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_freeman_theta.R
\name{es_freeman_theta}
\alias{es_freeman_theta}
\title{Freeman Theta}
\usage{
es_freeman_theta(catField, ordField, categories = NULL, levels = NULL)
}
\arguments{
\item{catField}{vector with categories}

\item{ordField}{vector with the scores}

\item{categories}{vector, optional. the categories to use from catField}

\item{levels}{vector, optional. the levels or order used in ordField.}
}
\value{
theta, float. The Freeman Theta value
}
\description{
According to Jacobson (1972, p. 42), this is the only measure for nominal-ordinal data, and is a modification of Somers d.

It can range from 0 to 1, with 0 indicating no influence of the catField on the scores of the ordField, and a 1 a perfect relationship.

Alternatives could be eta-squared and epsilon-squared.
}
\details{
The formula used is (Freeman, 1965, p. 116):
\deqn{\theta = \frac{D}{T}}

With:
\deqn{D = \sum D_{x,y}}
\deqn{D_{x,y} = \left|f_a - f_b\right|}
\deqn{f_a = \sum_{i=1}^{n_{lvl} - 1}\left(F_{x,i}\times\sum_{j=i+1}^{n_{lvl}} F_{y,j}\right)}
\deqn{f_b = \sum_{i=2}^{n_{lvl}}\left(F_{x,i}\times\sum_{j=1}^{i-1} F_{y,j}\right)}

\emph{Symbols used:}

\itemize{
\item \eqn{F_{x,i}}, from category x, the number of cases with level i.
\item \eqn{n_{lvl}}, the number of levels.
\item \eqn{n_i}, the total number of cases from category i
}
}
\references{
Freeman, L. C. (1965). \emph{Elementary applied statistics: For students in behavioral science}. Wiley.

Jacobson, P. E. (1972). Applying measures of association to nominal-ordinal data. \emph{The Pacific Sociological Review, 15}(1), 41–60. doi:10.2307/1388286
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
