% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_goodman_kruskal_lambda.R
\name{es_goodman_kruskal_lambda}
\alias{es_goodman_kruskal_lambda}
\title{Goodman-Kruskal Lambda}
\usage{
es_goodman_kruskal_lambda(
  field1,
  field2,
  categories1 = NULL,
  categories2 = NULL,
  ties = "first"
)
}
\arguments{
\item{field1}{the scores on the first variable}

\item{field2}{the scores on the second variable}

\item{categories1}{optional, categories to use for field1}

\item{categories2}{optional, categories to use for field2}

\item{ties}{c("first", "random", "average") optional to indicate what to do in case of multimodal situations.}
}
\value{
dataframe with the effect size value, asymptotic standard error (both assuming null and alternative), the test statistic, and p-value
}
\description{
Goodman-Kruskal Lambda
}
\details{
The formula used is (Goodman & Kruskal, 1954, p. 743):
\deqn{\lambda_{Y|X} = \frac{\left(\sum_{i=1}^r f_{i,max}\right) - C_{max}}{n - C_{max}}}
\deqn{\lambda_{X|Y} = \frac{\left(\sum_{j=1}^c f_{max,j}\right) - R_{max}}{n - R_{max}}}
\deqn{\lambda = \frac{\left(\sum_{i=1}^r f_{i,max}\right) + \left(\sum_{j=1}^c f_{max,j}\right)- C_{max} - R_{max}}{2\times n - C_{max} - R_{max}}}

The asymptotic standard errors are calculated using (Hartwig, 1973, p. 308):
\deqn{ASE\left(\lambda_{Y|X}\right)_0 = \sqrt{\left(\sum_{i,j}\left(F_{i,j}\times\left(\delta_{i,j}^c - \delta_j^c\right)^2\right)\right) - \frac{\left(\left(\sum_{i=1}^r f_{i, max}\right) - C_{max}\right)^2}{n}}{n - C_{max}}}
\deqn{ASE\left(\lambda_{X|Y}\right)_0 = \sqrt{\left(\sum_{i,j}\left(F_{i,j}\times\left(\delta_{i,j}^r - \delta_i^r\right)^2\right)\right) - \frac{\left(\left(\sum_{j=1}^c f_{max, j}\right) - R_{max}\right)^2}{n}}{n - R_{max}}}
\deqn{ASE\left(\lambda\right)_0 = \sqrt{\left(\sum_{i,j}\left(F_{i,j}\times\left(\delta_{i,j}^c - \delta_j^c + \delta_{i,j}^r - \delta_i^r\right)^2\right)\right) - \frac{\left(\left(\sum_{i=1}^r f_{i, max}\right) - C_{max} + \left(\sum_{j=1}^c f_{max, j}\right) - R_{max}\right)^2}{n}}{2\times n - C_{max} - R_{max}}}

\deqn{ASE\left(\lambda_{Y|X}\right)_1 = \sqrt{\frac{\left(n - \sum_{i=1}^r f_{i, max}\right) \times \left(\left(\sum_{i=1}^r f_{i, max}\right) + C_{max} - 2\times\sum_{i,j}\left(F_{i,j}\times\delta_{i,j}^c\times\delta_j^c\right)\right)}{\left(n - C_{max}\right)^3}}}
\deqn{ASE\left(\lambda_{X|Y}\right)_1 = \sqrt{\frac{\left(n - \sum_{j=1}^c f_{max, j}\right) \times \left(\left(\sum_{j=1}^c f_{max, j}\right) + R_{max} - 2\times\sum_{i,j}\left(F_{i,j}\times\delta_{i,j}^r\times\delta_i^r\right)\right)}{\left(n - R_{max}\right)^3}}}
\deqn{ASE\left(\lambda\right)_1 = \frac{\left(\sum_{i,j}\left(F_{i,j}\times\left(\delta_{i,j}^c + \delta_{i,j}^r - \delta_{j}^c - \delta_{i}^r + \lambda\times\left(\delta_{j}^c + \delta_{i}^r\right)\right)^2\right)\right) - 4\times n\times\lambda^2}{2\times n - C_{max} - R_{max}}}

With:
\deqn{\delta_{i,j}^c \begin{cases} 1 & \text{ if } j = \text{column index} f_{i,max} \\ 0 & \text{ else } \end{cases}}
\deqn{\delta_{j}^c \begin{cases} 1 & \text{ if } j = \text{column index} C_{max} \\ 0 & \text{ else } \end{cases}}
\deqn{\delta_{i,j}^r \begin{cases} 1 & \text{ if } i = \text{row index} f_{max,j} \\ 0 & \text{ else } \end{cases}}
\deqn{\delta_{i}^r \begin{cases} 1 & \text{ if } i = \text{column index} R_{max} \\ 0 & \text{ else } \end{cases}}

The test is performed using:
\deqn{z_{i} = \frac{\lambda_{i}}{ASE\left(i\right)_0}}
\deqn{sig. = 2\times\left(1 - \Phi\left(\left|z_i\right|\right)\right)}

\emph{Symbols used}
\itemize{
\item \eqn{F_{i,j}} the absolute frequency (observed count) from row i and column j.
\item \eqn{c} the number of columns
\item \eqn{r} the number of rows
\item \eqn{R_i} row total of row i, it can be calculated using \eqn{R_i=\sum_{j=1}^c F_{i,j}}
\item \eqn{C_j} column total of column j, it can be calculated using \eqn{C_j=\sum_{i=1}^r F_{i,j}}
\item \eqn{n} the total number of cases, it can be calculated in various ways, \eqn{n = \sum_{j=1}^c C_j =\sum_{i=1}^r R_i = \sum_{i=1}^r \sum_{j=1}^c F_{i,j}}
\item \eqn{f_{i,max}} is the maximum count of row i, i.e. \eqn{f_{i,max}=\text{max}\left\{ F_{i,1},F_{i,2},\dots,F_{i,c}\right\}}
\item \eqn{f_{max,j}} is the maximum count of column j, i.e. \eqn{f_{max,j}=\text{max}\left\{ F_{1,j},F_{2,j},\dots,F_{r,j}\right\}}
\item \eqn{R_{max}} is the maximum of the row totals, i.e. \eqn{R_{max}=\text{max}\left\{ R_1,R_2,\dots,R_r\right\}}
\item \eqn{C_{max}} is the maximum of the column totals, i.e. \eqn{C_{max}=\text{max}\left\{ C_1,C_2,\dots,C_c\right\}}
\item \eqn{\Phi\left(\dots\right)} the cumulative density function of the standard normal distribution
}

Unfortunately not much is written about how to deal with situations if more than one row / column / cell
has the highest (i.e. a multimodal situation).
Hartwig proposed three options in case multi-modal situation occurs: choose random,
choose the largest ASE, or average them.
This function can allow you to simply choose the first only (I think SPSS uses this),
average them, or simply choose one at random.
}
\references{
Goodman, L. A., & Kruskal, W. H. (1954). Measures of Association for Cross Classifications. \emph{Journal of the American Statistical Association, 49}(268), 732-764. https://doi.org/10.2307/2281536

Gray, L. N., & Campbell, R. (1975). Statistical significance of the Lambda coefficients: A comment. \emph{Behavioral Science, 20}(4), 258-259. https://doi.org/10.1002/bs.3830200407

Hartwig, F. (1973). Statistical significance of the lambda coefficients. \emph{Behavioral Science, 18}(4), 307-310. https://doi.org/10.1002/bs.3830180409

SPSS. (2006). SPSS 15.0 algorithms.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
