% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_goodman_kruskal_tau.R
\name{es_goodman_kruskal_tau}
\alias{es_goodman_kruskal_tau}
\title{Goodman-Kruskal tau}
\usage{
es_goodman_kruskal_tau(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{the scores on the first variable}

\item{field2}{the scores on the second variable}

\item{categories1}{optional, categories to use for field1}

\item{categories2}{optional, categories to use for field2}
}
\value{
dataframe with the effect size value, the test statistic, degrees of freedom, and p-value
}
\description{
Goodman-Kruskal tau
}
\details{
The formula used is (Goodman & Kruskal, 1954, p. 759):
\deqn{\tau_{Y|X} = \frac{n\times\sum_{i=1}^r\sum_{j=1}^c\frac{F_{i,j}^2}{R_i} - \sum_{j=1}^c C_j^2}{n^2 - \sum_{j=1}^c C_j^2}}
\deqn{sig. = 1 - \chi^2\left(\left(n - 1\right)\times\left(c - 1\right)\times\tau_{Y|X}, df\right)}
\deqn{\tau_{X|Y} = \frac{n\times\sum_{i=1}^r\sum_{j=1}^c\frac{F_{i,j}^2}{C_i} - \sum_{i=1}^r R_i^2}{n^2 - \sum_{i=1}^r R_i^2}}
\deqn{sig. = 1 - \chi^2\left(\left(n - 1\right)\times\left(r - 1\right)\times\tau_{X|Y}, df\right)}

With:
\deqn{df = \left(r - 1\right)\times\left(c - 1\right)}

\emph{Symbols used}
\itemize{
\item \eqn{F_{i,j}} the absolute frequency (observed count) from row i and column j.
\item \eqn{c} the number of columns
\item \eqn{r} the number of rows
\item \eqn{R_i} row total of row i, it can be calculated using \eqn{R_i=\sum_{j=1}^c F_{i,j}}
\item \eqn{C_j} column total of column j, it can be calculated using \eqn{C_j=\sum_{i=1}^r F_{i,j}}
\item \eqn{n} the total number of cases, it can be calculated in various ways, \eqn{n = \sum_{j=1}^c C_j =\sum_{i=1}^r R_i = \sum_{i=1}^r \sum_{j=1}^c F_{i,j}}
\item \eqn{\chi\left(\dots, \dots\right)} the cumulative density function of the chi-square distribution
}

Light and Margolin developed a R2 measure for categorical data,
they proposed a test CATANOVA (Categorical Anova) for this measure.
This was a chi-square test (p. 538). Sarndal (1974, p. 178) concluded that
R2 from Light and Mangolin, was the same as Goodman-Kendal tau, and uses their
test for tau. Margolin and Light (1974) reach the same conclusion and proof the
equivelance.
}
\references{
Goodman, L. A., & Kruskal, W. H. (1954). Measures of association for cross classifications. \emph{Journal of the American Statistical Association, 49}(268), 732–764. https://doi.org/10.2307/2281536

Light, R. J., & Margolin, B. H. (1971). An analysis of variance for categorical data. \emph{Journal of the American Statistical Association, 66}(335), 534–544. https://doi.org/10.1080/01621459.1971.10482297

Margolin, B. H., & Light, R. J. (1974). An analysis of variance for categorical data, II: Small sample comparisons with chi square and other competitors. \emph{Journal of the American Statistical Association, 69}(347), 755–764. https://doi.org/10.1080/01621459.1974.10480201

Särndal, C. E. (1974). A comparative study of association measures. \emph{Psychometrika, 39}(2), 165–187. https://doi.org/10.1007/BF02291467
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
