% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_hedges_g_ps.R
\name{es_hedges_g_ps}
\alias{es_hedges_g_ps}
\title{Hedges g (Paired Samples)}
\usage{
es_hedges_g_ps(
  field1,
  field2,
  dmu = 0,
  appr = c("none", "hedges", "durlak", "xue"),
  within = TRUE
)
}
\arguments{
\item{field1}{the scores on the first variable}

\item{field2}{the scores on the second variable}

\item{dmu}{optional the difference according to null hypothesis (default is 0)}

\item{appr}{approximation to use (see details), default is "none"}

\item{within}{boolean to use a correction for correlated pairs}
}
\value{
A dataframe with:
\item{g}{the Hedges g value}
\item{version}{version that was used}
}
\description{
Effect size measure for paired samples. This is very similar as Hedges g for independent samples.
}
\details{
The formula used is the same as for Cohen d_z

The same corrections can then be applied as for the independent samples version. See es_hedges_g_is() for details.

\strong{Alternatives}

\emph{library(effsize)}

datF = na.omit(data.frame(field1, field2))

cohen.d(datF$field1, datF$field2, paired=TRUE, within=TRUE, hedges.correction=TRUE)

cohen.d(datF$field1, datF$field2, paired=TRUE, within=FALSE, hedges.correction=TRUE)
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
