% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_jbm_r.R
\name{es_jbm_r}
\alias{es_jbm_r}
\title{Berry-Johnston-Mielke R}
\usage{
es_jbm_r(data, success = NULL)
}
\arguments{
\item{data}{dataframe with the scores}

\item{success}{indicator for what is considered a success (default is 1)}
}
\value{
\item{R}{the effect size measure}
}
\description{
A chance-corrected version of eta-squared, as an effect size measure for a Cochran Q test.
}
\details{
The formula used (Berry et al., 2007 pp. 1237, 1239):
\deqn{R = 1 - \frac{\delta}{\mu_{\delta}}}
With:
\deqn{\mu_{\delta} = \frac{2}{n\times\left(n - 1\right)} \times \left(\sum_{i=1}^n p_i\right) \times \left(n - \sum_{i=1}^n p_i\right) - \sum_{i=1}^n p_i\times\left(1 - p_i\right)}
\deqn{\delta = \frac{1}{k\times\binom{n}{k}} \times \sum_{c=1}^k\sum_{i=1}^{n - 1}\sum_{j=i+1}^n \left|x_{i,c} - x_{j,c}\right|}
\deqn{p_i = \frac{\sum_{j=1}^k}{k}}

\emph{Symbols used}
\itemize{
\item \eqn{n} the number of rows
\item \eqn{k} the number of columns
\item \eqn{x_{i,j}} the score in row i and column j
}

The function actually uses for:
\deqn{\sum_{c=1}^k\sum_{i=1}^{n - 1}\sum_{j=i+1}^n \left|x_{i,c} - x_{j,c}\right| = \sum_{j=1}^k C_j \times\left(n - \sum_{j=1}^k C_j\right)}
With:
\deqn{C_j = \sum_{i = 1}^n x_{i,j}}

The original article has in the equation for \eqn{\mu_{\delta}} the first factor written as
\eqn{\frac{2}{k\times\left(k - 1\right)}}. In personal communication with one of the authors
Alexis (2014) indicated this was wrong and \eqn{n} should be used.
}
\references{
Alexis. (2014, September 7). Answer to "Effect size of Cochran's Q." Cross Validated. https://stats.stackexchange.com/a/114649

Berry, K. J., Johnston, J. E., & Mielke, P. W. (2007). An alternative measure of effect size for Cochran's Q test for related proportions. \emph{Perceptual and Motor Skills, 104}(3_suppl), 1236-1242. doi:10.2466/pms.104.4.1236-1242
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
