% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_kendall_w.R
\name{es_kendall_w}
\alias{es_kendall_w}
\title{Eta Squared (Maximum Corrected) for Cochran Q}
\usage{
es_kendall_w(Q, n, k)
}
\arguments{
\item{Q}{the Cochran Q statistic}

\item{n}{the sample size (number of rows)}

\item{k}{the number of variables (number of columns)}
}
\value{
\item{es}{the effect size measure}
}
\description{
Eta Squared (Maximum Corrected) for Cochran Q
}
\details{
The formula used (Serlin et al., 1982 p. 788):
\deqn{\eta_Q^2 = \frac{Q}{n\times\left(k - 1\right)}}

\emph{Symbols used}
\itemize{
\item \eqn{n} the number of rows
\item \eqn{k} the number of columns
\item \eqn{Q} the Cochran Q statistic
}

The Cochran Q statistic can be obtained using \emph{ts_cochran_q()} function.
The number of rows and columns of a dataframe with R's \emph{nrow(dataframe)} and
\emph{ncol(dataframe)} functions.
}
\references{
Serlin, R. C., Carr, J., & Marascuilo, L. A. (1982). A measure of association for selected nonparametric procedures. \emph{Psychological Bulletin, 92}(3), 786–790. https://doi.org/10.1037/0033-2909.92.3.786
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
