% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_mcewen_michael.R
\name{es_mcewen_michael}
\alias{es_mcewen_michael}
\title{McEwen-Michael Coefficient / Cole C3}
\usage{
es_mcewen_michael(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
McEwen-Michael Coefficient
}
\description{
McEwen-Michael Coefficient / Cole C3
}
\details{
The formula used (Michael, 1920, p. 57)):
\deqn{\frac{a\times d - b\times c}{\left(\frac{a + d}{2}\right)^2 + \left(\frac{b + c}{2}\right)^2}}
\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
}

Note that Cole (1949, p. 415) refers to this as C3.
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
es_mcewen_michael(bin1, bin2)

}
\references{
Cole, L. C. (1949). The measurement of interspecific associaton. \emph{Ecology, 30}(4), 411–424. https://doi.org/10.2307/1932444

Michael, E. L. (1920). Marine Ecology and the coefficient of association: A plea in behalf of quantitative biology. \emph{Journal of Ecology, 8}(1), 54–59. https://doi.org/10.2307/2255213
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
