% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_odds_ratio.R
\name{es_odds_ratio}
\alias{es_odds_ratio}
\title{Odds Ratio}
\usage{
es_odds_ratio(var1, var2)
}
\arguments{
\item{var1}{A vector with the binary data from the first variable}

\item{var2}{A vector with the binary data from the second variable}
}
\value{
dataframe with the odds ratio value, z-statistic, and 2-sided p-value
}
\description{
Odds Ratio
}
\details{
The p-value is for the null-hypothesis that the population OR is 1.

\deqn{OR = \frac{a\times d}{b \times c}}
\deqn{sig. = 2\times\left(1 - Z\left(\left|z\right|\right)\right)}

With:
\deqn{SE = \sqrt{\frac{1}{a} + \frac{1}{b} + \frac{1}{c} + \frac{1}{d}}}
\deqn{z = \frac{\ln{\left(OR\right)}}{SE}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell
\item \eqn{b} the count in the top-right cell
\item \eqn{c} the count in the bottom-left cell
\item \eqn{d} the count in the bottom-right cell
\item \eqn{Z\left(\dots\right)} the cumulative density function of the standard normal distribution
}

I could not find an original source for this, but one source is McHugh (2009).
The formula for the standard error can be found on page 123.

For a classification/qualification of the odds ratio use: \emph{th_odds_ratio()}.

Or convert it to Cohen's d using either: \emph{es_convert(or, from="or", to="cohend", ex1="chinn")} or \emph{es_convert(or, from="or", to="cohend", ex1="borenstein")}

Then use classification/qualification of Cohen d: \emph{th_cohen_d()}

Or convert it to Yule Q using: \emph{es_convert(or, from="or", to="yuleq")}

Then use classification/qualification of Yule Q: \emph{th_yule_q()}

Or convert it to Yule Y using: \emph{es_convert(or, from="or", to="yuley")}

\strong{Alternative}

R's \emph{stats} library has a function that also shows an odds ratio: \emph{fisher.test()}
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
es_odds_ratio(bin1, bin2)

}
\references{
McHugh, M. (2009). The odds ratio: Calculation, usage, and interpretation. \emph{Biochemia Medica, 19}(2), 120–126. https://doi.org/10.11613/BM.2009.011
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
