% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_omega_sq.R
\name{es_omega_sq}
\alias{es_omega_sq}
\title{Omega Squared}
\usage{
es_omega_sq(nomField, scaleField, categories = NULL)
}
\arguments{
\item{nomField}{the groups variable}

\item{scaleField}{the numeric scores variable}

\item{categories}{vector, optional. the categories to use from catField}
}
\value{
the omega squared value
}
\description{
An effect size measure for a one-way ANOVA. It indicates the the strength of the categories on the scale field. A 0 would indicate no influence, and 1 a perfect relationship.

Although a popular belief is that \eqn{\omega^2} is preferred over \eqn{\epsilon^2} (Keselman, 1975), a later study actually showed that \eqn{\epsilon^2} might be preferred (Okada, 2013).
}
\details{
The formula used (Kirk, 1996, p. 751):
\deqn{\omega^2 = \frac{\left(F - 1\right)\times df_b}{df_b\times\left(F-1\right)+n}}

There are quite some variations on the formula above, all giving the same final result.

Hays (1973, p. 486) and Albers and Lakens (2018, p. 194):
\deqn{\omega^2 = \frac{F - 1}{\frac{df_w + 1}{df_b} + F}}

Caroll and Nordholm (1975, p. 547)
\deqn{\omega^2 = \frac{F - 1}{\frac{N - k + 1}{k - 1} + F}}

Hays (1973, p. 485):
\deqn{\omega^2 = \frac{SS_b - \left(k - 1\right)\times MS_w}{SS_t + MS_w}}

Olejnik and Algina (2003, p. 435):
\deqn{\omega^2 = \frac{SS_b - df_b\times MS_w}{SS_b + \left(n - df_b\right)\times MS_w}}

\emph{Symbols}
\itemize{
\item \eqn{x_{i,j}} the i-th score in category j
\item \eqn{n} the total sample size
\item \eqn{n_j} the number of scores in category j
\item \eqn{k} the number of categories
\item \eqn{\bar{x}_j} the mean of the scores in category j
\item \eqn{MS_i} the mean square of i
\item \eqn{SS_i} the sum of squares of i (sum of squared deviation of the mean)
\item \eqn{df_i} the degrees of freedom of i
\item \eqn{b} is between = factor = treatment = model
\item \eqn{w} is within = error (the variability within the groups)
\item \eqn{F_F} the test statistic of the Fisher/Classic one-way ANOVA
}

The formula appears in many different formats. Hays (1973, p. 486) shows:
\deqn{\hat{\omega}^2 = \frac{F_F - 1}{\frac{df_w + 1}{df_b}+ F_F}}
Which can also be found in Albers and Lakens (2018, p. 194).

Kirk (1996, p. 751) shows:
\deqn{\hat{\omega}^2 = \frac{df_b\times\left(F_F - 1\right)}{df_b\times\left(F_F - 1\right) + n}}

Instead of using the definitions for \eqn{df_b} and \eqn{df_w}, Caroll and Nordholm (1975, p. 547) show the formula as:
\deqn{\hat{\omega}^2 = \frac{F_F - 1}{\frac{n - k + 1}{k - 1}+ F_F}}
They also show Hays original formula (hays1) on p. 188.

Olejnik and Algina (2003, p. 435) use:
\deqn{\hat{\omega}^2 = \frac{SS_b - df_b\times MS_w}{SS_b + \left(n - df_b\right)\times MS_w}}

\strong{Conversion}

To convert \eqn{\omega^2} to \eqn{\epsilon^2} use  \emph{es_convert(omegasq, from="omegasq", to="epsilonsq", ex1=MS_w, ex2=SS_t)}

\strong{Alternatives}

\emph{library(effectsize)}

anova_stats(aov(scores~groups))

omega_squared(aov(scores~groups))
}
\references{
Albers, C., & Lakens, D. (2018). When power analyses based on pilot data are biased: Inaccurate effect size estimators and follow-up bias. \emph{Journal of Experimental Social Psychology, 74}, 187-195. doi:10.1016/j.jesp.2017.09.004

Carroll, R. M., & Nordholm, L. A. (1975). Sampling characteristics of Kelley's \eqn{\epsilon} and Hays' \eqn{\omega}. \emph{Educational and Psychological Measurement, 35}(3), 541-554. doi:10.1177/001316447503500304

Hays, W. L. (1973). \emph{Statistics for the social sciences} (2nd ed.). Holt, Rinehart and Winston.

Keselman, H. J. (1975). A Monte Carlo investigation of three estimates of treatment magnitude: Epsilon squared, eta squared, and omega squared. \emph{Canadian Psychological Review / Psychologie Canadienne, 16}(1), 44-48. doi:10.1037/h0081789

Kirk, R. E. (1996). Practical significance: A concept whose time has come. \emph{Educational and Psychological Measurement, 56}(5), 746-759. doi:10.1177/0013164496056005002

Okada, K. (2013). Is omega squared less biased? A comparison of three major effect size indices in one-way anova. \emph{Behaviormetrika, 40}(2), 129-147. doi:10.2333/bhmk.40.129

Olejnik, S., & Algina, J. (2003). Generalized eta and omega squared statistics: Measures of effect size for some common research designs. \emph{Psychological Methods, 8}(4), 434-447. doi:10.1037/1082-989X.8.4.434
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
