% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_pairwise_bin.R
\name{es_pairwise_bin}
\alias{es_pairwise_bin}
\title{Binary Effect Size for Pairwise Test}
\usage{
es_pairwise_bin(data, expCounts = NULL, es = "coheng")
}
\arguments{
\item{data, }{list with the data}

\item{expCounts}{dataframe, optional. The categories and expected counts}

\item{es, }{string, optional. effect size to use.Either "coheng" (default), "cohenh", "ar"}
}
\value{
dataframe with
\item{cat1}{label of first category in pair}
\item{cat2}{label of second category in pair}
\item{n1}{number of cases in first category}
\item{n2}{number of cases in second category}
\item{var}{followed by the effect size value}
}
\description{
When using a pairwise post-hoc test for a single nominal variable, the pair has become binary. This function then can determine the effect size for each pair.

Options are to use Cohen g, Cohen h', or the Alternative Ratio.
}
\details{
If expected counts are provided, for Cohen h' and the Alternative Ratio these expected counts are converted to expected proportions

See the separate functions of each effect size for more details.
\itemize{
\item \code{\link{es_cohen_g}} for Cohen g
\item \code{\link{es_cohen_h_os}} for Cohen h'
\item \code{\link{es_alt_ratio}} for the Alternative Ratio
}
}
\seealso{
\code{\link{ph_pairwise_bin}}, performs a pairwise binomial test
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
