% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_tschuprow_t.R
\name{es_tschuprow_t}
\alias{es_tschuprow_t}
\title{Tschuprow T}
\usage{
es_tschuprow_t(chi2, n, r, c, cc = NULL)
}
\arguments{
\item{chi2}{the chi-square test statistic}

\item{n}{the sample size}

\item{r}{the number of categories in the first variable (i.e. the number of rows)}

\item{c}{the number of categories in the second variable (i.e. the number of columns)}

\item{cc}{c(NULL, "bergsma") optional to indicate correction to use (default is NULL)}
}
\value{
Tschuprow T value
}
\description{
Tschuprow T
}
\details{
The formula used is:
\deqn{T=\sqrt\frac{\chi^{2}}{n\times \sqrt{\left(r - 1\right)\times\left(c - 1\right)}}}

\emph{Symbols used}:
\itemize{
\item \eqn{r} the number of categories in the first variable (i.e. the number of rows)
\item \eqn{c} the number of categories in the second variable (i.e. the number of columns)
\item \eqn{n} the sample size, i.e. the sum of all frequencies
\item \eqn{\chi^{2}} the chi-square statistic
}

The formula is taken from Bergsma (2013, p. 324) who refers to Tschuprow (1925, 1939).

The Bergsma correction uses a different formula (Bergsma, 2013, pp. 324-325):
\deqn{V_B = \sqrt{\frac{\tilde{\varphi}^2}{\sqrt{\left(\tilde{r} - 1\right)\times\left(\tilde{c} - 1\right)}}}}
With:
\deqn{\tilde{\varphi}^2 = \text{max}\left(0,\varphi^2 - \frac{\left(r - 1\right)\times\left(c - 1\right)}{n - 1}\right)}
\deqn{\tilde{r} = r - \frac{\left(r - 1\right)^2}{n - 1}}
\deqn{\tilde{c} = r - \frac{\left(c - 1\right)^2}{n - 1}}
\deqn{\varphi^2 = \frac{\chi^{2}}{n}}
}
\examples{
chi2Val = 16.98975
n = 1941
nRows = 5
nCols = 2
es_tschuprow_t(chi2Val, n, nRows, nCols)
es_tschuprow_t(chi2Val, n, nRows, nCols, cc="bergsma")

}
\references{
Bergsma, W. (2013). A bias-correction for Cramer's and Tschuprow's. \emph{Journal of the Korean Statistical Society, 42}(3), 323-328. https://doi.org/10.1016/j.jkss.2012.10.002

Tschuprow, A. A. (1925). \emph{Grundbegriffe und Grundprobleme der Korrelationstheorie}. B.G. Teubner.

Tschuprow, A. A. (1939). \emph{Principles of the mathematical theory of correlation} (M. Kantorowitsch, Trans.). W. Hodge.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
