% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_AS71.R
\name{he_AS71}
\alias{he_AS71}
\title{Helper Function - Algorithm AS 71}
\usage{
he_AS71(S, N)
}
\arguments{
\item{S}{the test statistic}

\item{n}{the sample size}
}
\value{
\item{pValue}{upper tail p-value of Kendall tau Distribution}
}
\description{
Helper Function - Algorithm AS 71
}
\details{
Algorithm AS 71 (Best & Gipps, 1974) uses as test statistic:
\deqn{S = \binom{n}{2}\times\left|\tau\right| = \frac{n\times\left(n - 1\right)}{2}\times\left|\tau\right|}

The Fortran code was translated to R by myself.
}
\references{
Best, D. J., & Gipps, P. G. (1974). Algorithm AS 71: The upper tail probabilities of Kendall’s tau. \emph{Applied Statistics, 23}(1), 98–100. https://doi.org/10.2307/2347062
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
