% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_AS89.R
\name{he_AS89}
\alias{he_AS89}
\title{Helper Function - Algorithm AS 89}
\usage{
he_AS89(n, IS)
}
\arguments{
\item{n}{the number of scores (should be equal in both variables)}

\item{IS}{the test statistic (see details)}
}
\value{
\item{pValue}{the two-sided significance (p-value)}
}
\description{
Algorithm AS 89 (Best & Roberts, 1975) is for upper-tail probabilities
}
\details{
The test statistic \eqn{S} defined as:
\deqn{S = \sum_{i = 1}^n d_i^2 = \sum_{i=1}^n \left(r_{x_i} - r_{y_i}\right)^2}
Which if there are no ties is equal to:
\deqn{S = \frac{\left(n^3 - n\right)\times\left(1 - r_s\right)}{6}}
}
\examples{
n = 12
S = 8
he_AS89(n, S)

}
\references{
Best, D. J., & Roberts, D. E. (1975). Algorithm AS 89: The upper tail probabilities of Spearman's rho. \emph{Applied Statistics, 24}(3), 377-379. https://doi.org/10.2307/2347111
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
