% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_multinomial.R
\name{he_find_combinations}
\alias{he_find_combinations}
\title{Find Combinations}
\usage{
he_find_combinations(n, k)
}
\arguments{
\item{n}{int with the sample size}

\item{k}{int with the number of categories}
}
\value{
A float with the requested probability
}
\description{
Helper function for the multinomial cumulative distribution. Will return all possible combinations to distribute n items over k categories.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
