% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_permutations.R
\name{he_permutations}
\alias{he_permutations}
\title{Helper Function for Permutations}
\usage{
he_permutations(n)
}
\arguments{
\item{n}{the number of scores}
}
\value{
\item{A}{all possible permutations of integers from 1 to n}
}
\description{
This function was posted by Museful (2013). It creates all possible permutations.
}
\examples{
he_permutations(5)

}
\references{
Museful. (2013, November 25). Answer to "Generating all distinct permutations of a list in R." Stack Overflow. https://stackoverflow.com/a/20199902/12149706
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
