% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_spearman_permutation.R
\name{he_spearman_permutation}
\alias{he_spearman_permutation}
\title{Spearman Exact Distribution using Permutations}
\usage{
he_spearman_permutation(ord1, ord2)
}
\arguments{
\item{ord1}{the numeric scores of the first variable}

\item{ord2}{the numeric scores of the second variable}
}
\value{
\item{pValue}{the two-tailed p-value}
}
\description{
This code in this function was posted by cuttlefish44 (2016)
It performs a two-tailed exact test for Spearman rho.
}
\details{
The exact distribution is calculated using the following steps:
\enumerate{
\item Determine all possible permutations of the scores in the first variable
\item Determine for each permutation the Spearman rho with the second variable
\item Count how often the Spearman rho is above the Spearman rho between the original two variables
\item Divide the results by \eqn{n!}
}
}
\examples{
ord1 = c(5, 3, 3, 4, 3, 4, 3)
ord2 = c(5, 3, 3, 3, 3, 3, 5)

he_spearman_permutation(ord1, ord2)

}
\references{
cuttlefish44. (2016, September 16). Answer to "Different methods for finding spearman's coefficient produce diff p-values depending on presence of tied values." Cross Validated. https://stats.stackexchange.com/a/235380/190640
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
