% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_kendall_tau_permutation.R
\name{he_tau_permutation}
\alias{he_tau_permutation}
\title{Helper Function - Kendall Tau Permutation Test}
\usage{
he_tau_permutation(ord1, ord2)
}
\arguments{
\item{ord1}{the numeric scores of the first variable}

\item{ord2}{the numeric scores of the second variable}
}
\value{
\item{pValue}{upper tail p-value of Kendall tau Distribution}
}
\description{
Helper Function - Kendall Tau Permutation Test
}
\details{
Uses a permutation test to calculate the probability. It is an adaption to
the code that was posted by cuttlefish44 (2016) online.

The exact distribution is calculated using the following steps:
\enumerate{
\item Determine all possible permutations of the scores in the first variable
\item Determine for each permutation the Kendall tau with the second variable
\item Count how often the Spearman rho is above the Kendall tau between the original two variables
\item Divide the results by \eqn{n!}
}
}
\examples{
ord1 = c(5, 8, 6, 3, 2, 9)
ord2 = c(2, 1, 4, 5, 8, 7)
he_tau_permutation(ord1, ord2)

}
\references{
cuttlefish44. (2016, September 16). Answer to "Different methods for finding spearman's coefficient produce diff p-values depending on presence of tied values." Cross Validated. https://stats.stackexchange.com/a/235380/190640
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
