% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tetrachoric.R
\name{hEfn1}
\alias{hEfn1}
\title{helper function for Kirk method}
\usage{
hEfn1(xx, w)
}
\arguments{
\item{xx}{H}

\item{w}{K value}
}
\value{
f value
}
\description{
H AND K INTEGRAND EVALUATION FOR GAUSSIAN QUADRATURE
}
