% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tetrachoric.R
\name{hEfn2}
\alias{hEfn2}
\title{helper function for Kirk method}
\usage{
hEfn2(h2k2, hk2, x, xx, v, w)
}
\arguments{
\item{h2k2}{see Kirk method flowchart}

\item{hk2}{see Kirk method flowchart}

\item{x}{see Kirk method flowchart}

\item{xx}{see Kirk method flowchart}

\item{v}{see Kirk method flowchart}

\item{w}{see Kirk method flowchart}
}
\value{
f value
}
\description{
R INTEGRAND EVALUATION
}
