% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_consensus.R
\name{me_consensus}
\alias{me_consensus}
\title{Consensus}
\usage{
me_consensus(data, levels = NULL)
}
\arguments{
\item{data}{a vector with the data}

\item{levels}{optional to indicate the categories in order if data is non-numeric}
}
\value{
cns the consensus score
}
\description{
The Consensus is a measure of agreement or dispersion for ordinal data. If there is no agreement the value is 0, and with full agreement 1.
}
\details{
The formula used (Tastle et al., 2005, p. 98):
\deqn{\text{Cns}\left(X\right) = 1 + \sum_{i=1}^k p_i \log_2\left(1 - \frac{\left|i - \mu_X\right|}{d_X}\right)}
With:
\deqn{\mu_X = \frac{\sum_{i=1}^k i\times F_i}{n}}
\deqn{d_X = k - 1}
\deqn{p_i = \frac{F_i}{n}}

\emph{Symbols used:}
\itemize{
\item \eqn{F_i} the frequency (count) of the i-th category (after they have been sorted)
\item \eqn{n} the sample size
\item \eqn{k} the number of categories.
}
}
\section{Alternatives}{


The \emph{agrmt} library has a function \emph{consensus(table(ordData))}
}

\references{
Tastle, W. J., & Wierman, M. J. (2007). Consensus and dissention: A measure of ordinal dispersion. \emph{International Journal of Approximate Reasoning, 45}(3), 531–545. https://doi.org/10.1016/j.ijar.2006.06.024
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
