% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_hodges_lehmann_is.R
\name{me_hodges_lehmann_is}
\alias{me_hodges_lehmann_is}
\title{Hodges-Lehmann Estimator (Independent Samples)}
\usage{
me_hodges_lehmann_is(catField, scores, categories = NULL, levels = NULL)
}
\arguments{
\item{catField}{A vector with the categorical data}

\item{scores}{A vector with the scores}

\item{categories}{Optional to indicate which two categories of catField to use, otherwise first two found will be used.}

\item{levels}{Optional list with the ordinal text values in order}
}
\value{
HL, the Hodges-Lehmann Estimator
\item{CLE cat. 1}{the effect size for the first category}
\item{CLE cat. 2}{the effect size for  the second category}
}
\description{
The Hodges-Lehmann estimate, is the median of all the possible differences between two sets of data. The authors (Hodges & Lehmann, 1963) describe it as the location shift that is needed to align two distributions (with similar distributions) as much as possible (p. 599).

It is sometimes incorrectly described as the difference between the two medians, but that is incorrect. It is not uncommon to have a different Hodges-Lehmann estimate than simply taking the difference between the two medians.

This measure is sometimes mentioned as an effect size measure for a Mann-Whitney U / Wilcoxon Rank Sum test (van Geloven, 2018), however since it is a median of the possible differences, it is not standardized (i.e. it doesn't range between two fixed values, and depends therefor on the data).
}
\details{
The formula for the Hodges-Lehmann estimator with two samples is (Hodges & Lehmann, 1963, p. 602):

\deqn{HL = \text{median}\left(y_j - x_i | 1 \leq i \leq n_x, 1 \leq j \leq n_y\right)}

\emph{Symbols used:}
\itemize{
\item  \eqn{x_i} the i-th score in category x
\item \eqn{x_j} the j-th score in category y
\item  \eqn{n_i} the number of scores in category i
}

There might be a faster method to actually determine this. Algorithm 616 (Monahan, 1984), but couldn't translate the Fortran to R
}
\references{
Hodges, J. L., & Lehmann, E. L. (1963). Estimates of location based on rank tests. \emph{The Annals of Mathematical Statistics, 34}(2), 598-611. doi:10.1214/aoms/1177704172

Monahan, J. F. (1984). Algorithm 616: Fast computation of the Hodges-Lehmann location estimator. \emph{ACM Transactions on Mathematical Software, 10}(3), 265-270. doi:10.1145/1271.319414

van Geloven, N. (2018, March 13). Mann-Whitney U toets [Wiki]. Wikistatistiek. https://wikistatistiek.amc.nl/Mann-Whitney_U_toets
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
