% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_mean.R
\name{me_mean}
\alias{me_mean}
\title{Mean}
\usage{
me_mean(data, version = "arithmetic", trimProp = 0.1, trimFrac = "down")
}
\arguments{
\item{data, }{vector or dataframe with scores as numbers}

\item{version, }{optional mean to calculate. Either \code{"arithmetic"} (default), \code{"winsorized"}, \code{"trimmed"}, \code{"windsor"}, \code{"truncated"}, \code{"olympic"}, \code{"geometric"}, \code{"harmonic"}, \code{"midrange"}}

\item{trimProp, }{optional to indicate the total proportion to trim. Default at 0.1 i.e. 0.05 from each side.}

\item{trimFrac, }{optional parameter to indicate what to do if trimmed amount is non-integer. Either \code{"down"} (default), \code{"prop"}, \code{"linear"}}
}
\value{
res, the value of the mean
}
\description{
Different types of means can be determined using this function.

The mean is a measure of central tendency, to indicate the center.
}
\details{
\strong{Arithmetic Mean}

One of the three Pythagorean means, and the mean most people would assume if you ask them to calculate the mean.

It is the fulcrum of the distribution (Weinberg & Schumaker, 1962, p.19). One reference can for example be found in Aristotle (384-322 BC) (1850, p. 43).

The formula:
\deqn{\bar{x} = \frac{\sum_{i=1}^n x_i}{n}}

\strong{Harmonic Mean}

The second of the three Pythagorean means:
\deqn{H = \frac{n}{\sum_{i=1}^n \frac{1}{x_i}}}

\strong{Geometric Mean}

The third of the three Pythagorean means:
\deqn{G = e^{\frac{1}{n}\times\sum_{i=1}^n \ln\left(x_i\right)}}

\strong{Olympic Mean}

Simply ignore the maximum and minimum (only once) (Louis et al., 2023, p. 117):
\deqn{OM = \frac{\sum_{i=2}^{n-1} x_i}{n - 2}}

\strong{Mid Range}

The average of the maximum and minimum (Lovitt & Holtzclaw, 1931, p. 91):
\deqn{MR = \frac{\min x + \max x}{2}}

\strong{Trimmed}

With a trimmed (Windsor/Truncated) mean we trim a fixed amount of scores from each side (Tukey, 1962, p. 17). Let \eqn{p_t} be the proportion to trim, we then need to trim \eqn{n_t = \frac{p_t\times n}{2}} from each side.

If this \eqn{n_t} is an integer there isn't a problem, but if it isn't we have options. The  first option is to simply round down, i.e. \eqn{n_l = \lfloor n_t\rfloor}. The trimmed mean is then:
\deqn{\bar{x}_t = \frac{\sum_{i=n_t+1}^{n - n_l + 1} x_i}{n - 2\times n_l}}

This is used if \emph{trimFrac = "down"} is set.

We could also use linear interpolation based on the number of scores to trim. We missed out on: \eqn{f = n_t - n_l} on each side. So the first and last value we do include should only count for \eqn{1 - f} each. The trimmed mean will then be:
\deqn{\bar{x}_t = \frac{\left(x_{n_t + 1} + x_{n - n_l + 1}\right)\times\left(1 - f\right) + \sum_{i=n_l+2}^{n - n_l} x_i}{n - 2\times n_t}}

This is used if \emph{trimFrac = "prop"} is set.

Alternative, we could take the proportion itself and use linear interpolation on that. The found \eqn{n_l} will be \eqn{p_1 = \frac{n_l \times 2}{n}} of the the total sample size. While if we had rounded up, we had used \eqn{p_2 = \frac{\left(n_l + 1\right)\times 2}{n}} of the the total sample size. Using linear interpolation we then get:
\deqn{\bar{x}_t = \frac{p_t - p_1}{p_2 - p_1}\times\left(\bar{x}_{th}-\bar{x}_{tl}\right) + \bar{x}_{tl}}
Where \eqn{\bar{x}_{tl}} is the trimmed mean if \eqn{p_1} would be used as a trim proportion, and \eqn{\bar{x}_{th}} is the trimmed mean if \eqn{p_2} would be used.

This is used if \emph{trimFrac = "linear"} is set.

\strong{Winsorized Mean}

Similar as with a trimmed mean, but now the data is not removed, but replaced by the value equal to the nearest value that is still included (Winsor as cited in Dixon, 1960, p. 385).
\deqn{W = n_l \times \left(x_{n_l + 1} + x_{n - n_l}\right)\times\frac{\sum_{n_l + 1}^{n - n_l} x_i}{n - 2\times n_l}}
}
\examples{
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
df2 = read.csv(file2, sep=';', na.strings=c("", "NA"))
#Example 1: Numeric dataframe
ex1 = df2['Gen_Age']
me_mean(ex1)

#Example 2: Numeric list
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
me_mean(ex2)

}
\references{
Aristotle. (1850). \emph{The nicomachean ethics of Aristotle} (R. W. Browne, Trans.). Henry G. Bohn.

Dixon, W. J. (1960). Simplified estimation from censored normal samples. \emph{The Annals of Mathematical Statistics, 31}(2), 385–391. https://doi.org/10.1214/aoms/1177705900

Louis, P., Núñez, M., & Xefteris, D. (2023). Trimming extreme reports in preference aggregation. \emph{Games and Economic Behavior, 137}, 116–151. https://doi.org/10.1016/j.geb.2022.11.003

Lovitt, W. V., & Holtzclaw, H. F. (1931). \emph{Statistics}. Prentice Hall.

Tukey, J. W. (1962). The future of data analysis. \emph{The Annals of Mathematical Statistics, 33}(1), 1–67. https://doi.org/10.1214/aoms/1177704711

Weinberg, G. H., & Schumaker, J. A. (1962). \emph{Statistics An intuitive approach}. Wadsworth Publishing.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
