% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_median.R
\name{me_median}
\alias{me_median}
\title{Median}
\usage{
me_median(data, levels = NULL, tieBreaker = c("between", "low", "high"))
}
\arguments{
\item{data}{vector with the data}

\item{levels}{optional to indicate what values represent}

\item{tieBreaker}{optional which to return if median falls between two values. Either \code{"between"} (default), \code{"low"}, or \code{"high"}}
}
\value{
the median
}
\description{
Function to determine the median of a set of data. The median can be defined as "the middle value in a distribution, below and above which lie values with equal total frequencies or probabilities" (Porkess, 1991, p. 134). This means that 50\% of the respondents scored equal or higher to the median, and also 50\% of the respondents scored lower or equal.
}
\details{
The formula that is used, assuming the data has been sorted, is:
\deqn{\tilde{x} = \begin{cases} x_{MI} & \text{ if } MI= \left \lfloor MI \right \rfloor\\ \frac{x_{MI-0.5} + x_{MI+0.5}}{2} & \text{ if } MI\neq \left \lfloor MI \right \rfloor \end{cases}}
With:
\deqn{MI = \frac{n + 1}{2}}

\emph{Symbols used:}
\itemize{
\item \eqn{n} the sample size
\item \eqn{x_i} the i-th score of X, assuming X has been sorted.
\item \eqn{MI} the index of the median
\item \eqn{\tilde{x}} the median
}

If the number of scores is an odd number, and the median falls between two categories. With the
\emph{tieBreaker} it can then be set to return the lower value (low), upper (high), or average (between).

Some old references to the median are Pacioli (1523) in Italian, Cournot (1843, p. 120) in French,
and Galton (1881, p. 246) in English.
}
\section{Alternatives}{


R's \emph{stats} library has a function \emph{median}, and also \emph{DescTools}, \emph{missMethods}, and probably
more. Note that many of these do not make it possible to return the "between" score if the data
are ordered factors.
}

\examples{
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
df2 = read.csv(file2, sep=';', na.strings=c("", "NA"))
#Example 1: Text dataframe
ex1 = df2[['Teach_Motivate']]
order = c("Fully Disagree", "Disagree", "Neither disagree nor agree", "Agree", "Fully agree")
me_median(ex1, levels=order)

#Example 2: Numeric data
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
me_median(ex2)

#Example 3: Text data with between median
ex3 = c("a", "b", "f", "d", "e", "c")
order = c("a", "b", "c", "d", "e", "f")
me_median(ex3, levels=order)
me_median(ex3, levels=order, tieBreaker="low")
me_median(ex3, levels=order, tieBreaker="high")

#Example 4: Numeric data with between median
ex4 = c(1, 2, 3, 4, 5, 6)
me_median(ex4)
me_median(ex4, tieBreaker="low")
me_median(ex4, tieBreaker="high")

}
\references{
Cournot, A. A. (1843). \emph{Exposition de la théorie des chances et des probabilités}. L. Hachette.

Galton, F. (1881). Report of the anthropometric committee. \emph{Report of the British Association for the Advancement of Science, 51}, 225–272.

Pacioli, L. (1523). \emph{Summa de arithmetica geometria proportioni: Et proportionalita}. Paganino de Paganini.

Porkess, R. (1991). \emph{The HarperCollins dictionary of statistics}. HarperPerennial.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
