% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_binomial.R
\name{ph_binomial}
\alias{ph_binomial}
\title{Pairwise Binomial Test for Post-Hoc Analysis}
\usage{
ph_binomial(
  data,
  expCount = NULL,
  twoSidedMethod = "eqdist",
  posthoc = "bonferroni"
)
}
\arguments{
\item{data}{dataframe with scores}

\item{expCount}{optional dataframe with categories and expected counts}

\item{twoSidedMethod}{method to use for determining two-sided p-value of binomial test}

\item{posthoc}{the correction to use, currently only "bonferroni" available}
}
\value{
a dataframe with:
\item{category 1}{the label of the first category}
\item{category 2}{the label of the second category}
\item{n1}{the sample size of the first category}
\item{n2}{the sample size of the second category}
\item{obs. prop. cat. 1}{the proportion in the sample of the first category}
\item{exp. prop. cat. 1}{the expected proportion for the first category}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
}
\description{
This function will perform a one-sample binomial test for each possible pair in the
data. It makes use of the ts_binomial_os() function.
}
\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = df1['mar1']
ph_binomial(ex1)

#Example 2: pandas series with various settings
ex2 = df1['mar1']
categories = c("MARRIED", "DIVORCED", "NEVER MARRIED", "SEPARATED")
expcounts = c(5, 5, 5, 5)
eCounts = data.frame('category' = categories, 'count' = ecounts)
ph_binomial(ex2, expCount=eCounts)

#Example 3: a list
ex3 = c("MARRIED", "DIVORCED", "MARRIED", "SEPARATED", "DIVORCED", "NEVER MARRIED", "DIVORCED", "DIVORCED", "NEVER MARRIED", "MARRIED", "MARRIED", "MARRIED", "SEPARATED", "DIVORCED", "NEVER MARRIED", "NEVER MARRIED", "DIVORCED", "DIVORCED", "MARRIED")
ph_binomial(ex3)

}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
