% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_column_proportion.R
\name{ph_column_proportion}
\alias{ph_column_proportion}
\title{Post-Hoc Column Proportion Test}
\usage{
ph_column_proportion(
  field1,
  field2,
  categories1 = NULL,
  categories2 = NULL,
  seMethod = "spss"
)
}
\arguments{
\item{field1}{list or dataframe with the first categorical field}

\item{field2}{list or dataframe with the second categorical field}

\item{categories1}{optional list with order and/or selection for categories of field1}

\item{categories2}{optional list with order and/or selection for categories of field2}

\item{seMethod}{optional methdod for standard error. Either "spss" (default) or "marascuilo".}
}
\description{
Post-Hoc Column Proportion Test
}
