% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_conover_iman.R
\name{ph_conover_iman}
\alias{ph_conover_iman}
\title{Post-Hoc Conover-Iman Test}
\usage{
ph_conover_iman(catField, ordField, categories = NULL, levels = NULL)
}
\arguments{
\item{catField}{vector with categories}

\item{ordField}{vector with the scores}

\item{categories}{vector, optional. the categories to use from catField}

\item{levels}{vector, optional. the levels or order used in ordField.}
}
\value{
A dataframe with:
\item{cat. 1}{one of the two categories being compared}
\item{cat. 2}{second of the two categories being compared}
\item{n1}{number of cat. 1. cases in comparison}
\item{n2}{number of cat. 2 cases in comparison}
\item{mean rank 1}{mean rank of cases in cat. 1, based on all cases (incl. categories not in comparison)}
\item{mean rank 2}{mean rank of cases in cat. 2, based on all cases (incl. categories not in comparison)}
\item{statistic}{the t-value of the test}
\item{df}{the degrees of freedom}
\item{p-value}{the p-value (significance)}
}
\description{
This can be used as a post-hoc test for a Kruskal-Wallis test (see ts_kruskal_wallis()).

The test compares each possible pair of categories from the catField and their mean rank. The null hypothesis is that these are then equal. A simple Bonferroni adjustment is also made for the multiple testing.

Other post-hoc tests that could be considered are Dunn, Nemenyi, Steel-Dwass, a pairwise Mann-Whitney U, or pairwise Mood-Median.
}
\details{
The formula used is (Conover & Iman, 1979, p. 11):
\deqn{t_{1,2} = \frac{\bar{r}_1 - \bar{r}_2}{\sqrt{S^2\times\frac{n-1-T}{n-k}\times\left(\frac{1}{n_1}+\frac{1}{n_2}\right)}}}
\deqn{df = n - k}
\deqn{sig. = 1 - T\left(\left|t_{1,2}\right|, df\right)}

With:
\deqn{S^2=\frac{\sum_{j=1}^k \sum_{i=1}^{n_j} r_{i,j}^2 - \frac{n\times\left(n+1\right)^2}{4}}{n-1}}
\deqn{T = \frac{\sum_{i=1}^k \frac{R_i^2}{n_i} - \frac{n\times\left(n+1\right)^2}{4}}{S^2}}
\deqn{R_i = \sum_{j=1}^{n_i} r_{i,j}}

Note that \eqn{S^2, T, k, n} are all based on all scores, including those not in the selected pair.

The formula can also be found in Conover (1980, pp. 230-231).

\emph{Symbols used}

\itemize{
\item \eqn{k}, the number of categories
\item \eqn{n_i}, the number of scores in category i
\item \eqn{r_{i,j}}, the rank of the j-th score in category i using all original scores (incl. those not in the comparison).
\item \eqn{R_i}, the sum of the ranks in category i
\item \eqn{\bar{r}_i}, the average of the ranks in category i, using all original scores (incl. those not in the comparison).
\item \eqn{T\left(\dots\right)}, the cumulative distribution function of the Student t distribution.
}
}
\references{
Conover, W. J. (1980). \emph{Practical nonparametric statistics} (2nd ed.). Wiley.

Conover, W. J., & Iman, R. L. (1979). \emph{On multiple-comparisons procedures} (LA-7677-MS; pp. 1-14). Los Alamos Scientific Laboratory.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
