% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_dunn.R
\name{ph_dunn}
\alias{ph_dunn}
\title{Post-Hoc Dunn Test}
\usage{
ph_dunn(catField, ordField, categories = NULL, levels = NULL)
}
\arguments{
\item{catField}{vector with categories}

\item{ordField}{vector with the scores}

\item{categories}{vector, optional. the categories to use from catField}

\item{levels}{vector, optional. the levels or order used in ordField.}
}
\value{
A dataframe with:
\item{cat. 1}{one of the two categories being compared}
\item{cat. 2}{second of the two categories being compared}
\item{n1}{number of cat. 1. cases in comparison}
\item{n2}{number of cat. 2 cases in comparison}
\item{mean rank 1}{mean rank of cases in cat. 1, based on all cases (incl. categories not in comparison)}
\item{mean rank 2}{mean rank of cases in cat. 2, based on all cases (incl. categories not in comparison)}
\item{statistic}{the z-value of the test}
\item{p-value}{the p-value (significance)}
\item{adj. p-value}{the Bonferroni adjusted p-value}
}
\description{
This can be used as a post-hoc test for a Kruskal-Wallis test (see ts_kruskal_wallis()).

The test compares each possible pair of categories from the catField and their mean rank. The null hypothesis is that these are then equal. A simple Bonferroni adjustment is also made for the multiple testing.

Dunn (1964) describes two procedures. The first is his own, the second is one from Steel (1960) for comparison. The difference is that in Dunn's procedure the mean rank of each category is based on the scores of all categories, including those that are not being compared, while Steel's procedure re-calculates the mean rank for each category using only the scores from the two categories being compared. This later one wouls make it very similar to a pairwise Mann-Whitney U test (see ph_mann_whitney()).

Other post-hoc tests that could be considered are Nemenyi, Steel-Dwass, Conover, a pairwise Mann-Whitney U, or pairwise Mood-Median.
}
\details{
The formula used (Dunn, 1964, p. 249):
\deqn{z_{1,2} = \frac{\bar{r}_1 - \bar{r}_2}{\sqrt{\sigma_m^2}}}
\deqn{sig. = 2\times\left(1 -\Phi\left(z\right)\right)}

With:
\deqn{\sigma_m^2 = \left(\frac{n\times\left(n+1\right)}{12} - \frac{T}{12\times\left(n - 1\right)}\right)\times\left(\frac{1}{n_1} + \frac{1}{n_2}\right)}
\deqn{T = \sum_{j=1}^k t_j^3 - t_j}
\deqn{\bar{r}_i = \frac{R_i}{n_i}}
\deqn{R_i = \sum_{j=1}^{n_i} r_{i,j}}

\emph{Symbols used}

\itemize{
\item \eqn{k}, the number of categories
\item \eqn{t_j}, the frequency of the j-th unique rank.
\item \eqn{n_i}, the number of scores in category i
\item \eqn{r_{i,j}}, the rank of the j-th score in category i using all original scores (incl. those not in the comparison).
\item \eqn{R_i}, the sum of the ranks in category i
\item \eqn{\bar{r}_i}, the average of the ranks in category i, using all original scores (incl. those not in the comparison).
\item \eqn{\Phi\left(\dots\right)}, the cumulative distribution function of the standard normal distribution.
}
}
\references{
Dunn, O. J. (1964). Multiple comparisons using rank sums. \emph{Technometrics, 6}(3), 241-252. doi:10.1080/00401706.1964.10490181

Steel, R. G. D. (1960). A rank sum test for comparing all pairs of treatments. \emph{Technometrics, 2}(2), 197-207. doi:10.1080/00401706.1960.10489894
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
