% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_dunn_q.R
\name{ph_dunn_q}
\alias{ph_dunn_q}
\title{Post-Hoc Dunn Test (for Cochran Q test)}
\usage{
ph_dunn_q(data, success = NULL)
}
\arguments{
\item{data}{dataframe with the binary scores}

\item{success}{indicator for what is considered a success (default is first value found)}
}
\value{
A dataframe with:
\item{category 1}{label of first variable in comparison}
\item{category 2}{label of second variable in comparison}
\item{n suc. 1}{number of successes in first variable in comparison}
\item{n suc. 2*}{number of successes in second variable in comparison}
\item{statistic}{test statistic}
\item{z-value}{standardized test statistic (z-value)}
\item{p-value}{p-value of the z-value}
\item{adj. p-value}{Bonferroni corrected p-value}
}
\description{
An adaptation from IBM SPSS on the Dunn test, so it can be used as a post-hoc test for a Cochran Q test.
}
\details{
The formula used (IBM, 2021, p. 814):
\deqn{z_{1,2} = \frac{\bar{d}_{1,2}}{SE}}
\deqn{sig. = 2\times\left(1-\Phi\left(\left|z_{1,2}\right|\right)\right)}

With:
\deqn{\bar{d}_{1,2} = \frac{ns_1 - ns_2}{n}}
\deqn{SE = \sqrt{2\times\frac{k\times\sum_{i=1}^n R_i - \sum_{i=1}^n R_i^2}{n^2\times k\times\left(k-1\right)}}}
\deqn{R_i = \sum_{j=1}^k s_{i,j}}
\deqn{ns_j = \sum_{i=1}^n s_{i,j}}
\deqn{s_{i,j} = \begin{cases} 1 & \text{ if } x_{i,j}= \text{success} \\ 0 & \text{ if } x_{i,j} \neq \text{success} \end{cases}}

IBM SPSS mentions this is an adaptation from Dunn (1964), originally for the Kruskal-Wallis test.

The Bonferroni adjustment is done using:
\deqn{sig._{adj} = \min \left(sig. \times n_c, 1\right)}
\deqn{n_c = \frac{k\times\left(k-1\right)}{2}}

\emph{Symbols used}

\itemize{
\item \eqn{x_{i,j}}, the score in row i and column j
\item \eqn{k}, the number of variables
\item \eqn{n}, the total number of cases used
\item \eqn{ns_j}, the total number of successes in column j
\item \eqn{R_i}, the total number of successes in row i
\item \eqn{\Phi\left(\dots\right)}, the standard normal cumulative distribution function.
\item \eqn{n_c}, the number of comparisons (pairs)
}
}
\references{
Dunn, O. J. (1964). Multiple comparisons using rank sums. \emph{Technometrics, 6}(3), 241–252. doi:10.1080/00401706.1964.10490181

IBM. (2021). IBM SPSS Statistics Algorithms. IBM.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
