% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_mann_whitney.R
\name{ph_mann_whitney}
\alias{ph_mann_whitney}
\title{Post-Hoc Pairwise Mann-Whitney U Test}
\usage{
ph_mann_whitney(
  catField,
  ordField,
  categories = NULL,
  levels = NULL,
  method = "appr",
  cc = TRUE
)
}
\arguments{
\item{catField}{vector with categories}

\item{ordField}{vector with the scores}

\item{categories}{vector, optional. the categories to use from catField}

\item{levels}{vector, optional. the levels or order used in ordField.}

\item{method}{string, optional. The use of the exact distribution (only if there are no ties) or the normal. Either "exact", "appr"}

\item{cc}{boolean, optional.Use of continuity correction. Default is True.}
}
\value{
A dataframe with:
\item{category 1}{one of the two categories being compared}
\item{category 2}{second of the two categories being compared}
\item{statistic}{the test statistic (z-value)}
\item{p-value}{the p-value (significance)}
\item{adj. p-value}{the Bonferroni adjusted p-value}
}
\description{
This can be used as a post-hoc test for a Kruskal-Wallis test (see ts_kruskal_wallis()).

The test compares each possible pair of categories from the catField and their mean rank. The null hypothesis is that these are then equal. A simple Bonferroni adjustment is also made for the multiple testing.

Other post-hoc tests that could be considered are Dunn, Nemenyi, Steel-Dwass, Conover-Iman, or pairwise Mood-Median.
}
\details{
This function selects each possible pair of categories and then simply runs a Mann-Whitney U test, using only those two categories.

See ts_mann_whitney() for details of the calculations.

The Bonferroni adjustment is simply:
\deqn{p_{adj} = \min \left(p \times n_{comp}, 1\right)}
\deqn{n_{comp} = \frac{k\times\left(k-1\right)}{2}}

\emph{Symbols used:}

\itemize{
\item \eqn{n_{comp}}, number of comparisons (pairs)
\item \eqn{k}, number of categories
}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
