% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_mcnemar_pw.R
\name{ph_mcnemar_pw}
\alias{ph_mcnemar_pw}
\title{Post-Hoc McNemar Test - Pairwise}
\usage{
ph_mcnemar_pw(field1, field2, categories = NULL, exact = FALSE, cc = FALSE)
}
\arguments{
\item{field1}{vector, the first categorical field}

\item{field2}{vector, the first categorical field}

\item{categories}{vector, optional, order and/or selection for categories of field1 and field2}

\item{exact}{boolean, optional, use of exact binomial distribution (default is False)}

\item{cc}{boolean, optional, use of a continuity correction (default is False)}
}
\value{
Dataframe with:
\item{field1}{the first category compared to the second}
\item{field2}{the second category compared to the first}
\item{n}{the sample size}
\item{statistic}{the chi-squared value (if applicable)}
\item{df}{the degrees of freedom used in the test (if applicable)}
\item{p-value}{the significance (p-value)}
\item{adj. p-value}{the Bonferroni adjusted p-value}
}
\description{
After a (McNemar-)Bowker test a post-hoc test can potentially locate where the changes occured. This can be done use a McNemar test, which is the Bowker test but for 2x2 tables.

There are two variations, one is to simply compare each possible pair of categories (pairwise comparison), or compare each category with all other categories (collapsed comparison). This function is for the pairwise version, see \strong{ph_mcnemar_co()} for the collapsed version.

Instead of using the McNemar test it is also possible to use the binomial test, which will be used if exact is set to True.
}
\details{
The formula used is (McNemar, 1947, p. 156):
\deqn{\chi_{M}^2 = \frac{\left(F_{1,2} - F_{2,1}\right)^2}{F_{1,2} + F_{2,1}}}
\deqn{df = 1}
\deqn{sig. = 1 - \chi^2\left(\chi_M^2, df\right)}
If a continuity correction is applied the formula changes to:
\deqn{\chi_{M*}^2 = \frac{\left(\left|F_{1,2} - F_{2,1}\right| - 1\right)^2}{F_{1,2} + F_{2,1}}}

The formula used for the binomial test is:
\deqn{sig. = 2\times\text{Bin}\left(F_{1,2} + F_{2,1}, \min\left(F_{1,2}, F_{2,1}\right), 0.5\right)}

The formula used for the binomial test with a mid-p correction:
\deqn{sig. = 2\times\text{Bin}\left(F_{1,2} + F_{2,1}, \min\left(F_{1,2}, F_{2,1}\right), 0.5\right) -\text{bin}\left(F_{1,2} + F_{2,1}, \min\left(F_{1,2}, F_{2,1}\right), 0.5\right)}

The number of pairwise tests \eqn{n_{comp}} ) is:
\deqn{n_{comp} = \frac{k\times \left(k - 1\right)}{2}}

The adjusted p-value is then determined using a Bonferroni correction:
\deqn{sig._{adj} = \begin{cases} sig. \times n_{comp} & \text{ if } sig. \times n_{comp} \leq = 1 \\ 1 & \text{ if } sig. \times n_{comp} > 1 \end{cases}}

\emph{Symbols used}
\itemize{
\item \eqn{F_{1,2}}, the observed count of cases that scored category 1 on the first variable, and category 2 on the second.
\item \eqn{F_{2,1}}, the observed count of cases that scored category 2 on the first variable, and category 1 on the second.
\item \eqn{\chi^2\left(\dots\right)}, the cumulative distribution function for the chi-square distribution.
\item \eqn{\text{Bin}\left(\dots\right)}, the cumulative distribution function for the binomial distribution.
\item \eqn{\text{bin}\left(\dots\right)}, the probability mass function for the binomial distribution.
}
}
\references{
McNemar, Q. (1947). Note on the sampling error of the difference between correlated proportions or percentages. \emph{Psychometrika, 12}(2), 153–157. doi:10.1007/BF02295996
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
