% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_mood_median.R
\name{ph_mood_median}
\alias{ph_mood_median}
\title{Post-Hoc Pairwise Mood-Median Test}
\usage{
ph_mood_median(
  catField,
  ordField,
  categories = NULL,
  levels = NULL,
  test = "pearson",
  cc = NULL,
  lambd = 2/3
)
}
\arguments{
\item{catField}{vector with categories}

\item{ordField}{vector with the scores}

\item{categories}{vector, optional. the categories to use from catField}

\item{levels}{vector, optional. the levels or order used in ordField.}

\item{test}{string, optional. The test of independence to use. Options are "pearson" (default), "fisher", "freeman-tukey", "g", "mod-log", "neyman", "power"}

\item{cc}{string, optional. Method for continuity correction. Options are NULL (default), "yates", "pearson", "williams"}

\item{lambd}{float or string, optional. Either name of test or specific value. Default is "cressie-read" i.e. lambda of 2/3. Only applies to Power Divergence test. Options besides a float are float, "cressie-read", "likelihood-ratio", "mod-log", "pearson", "freeman-tukey", "neyman"}
}
\value{
A dataframe with:
\item{category 1}{one of the two categories being compared}
\item{category 2}{second of the two categories being compared}
\item{statistic}{the test statistic}
\item{df}{he degrees of freedom, if applicable}
\item{p-value}{the p-value (significance)}
\item{adj. p-value}{the Bonferroni adjusted p-value}
}
\description{
This can be used as a post-hoc test for a Kruskal-Wallis test (see ts_kruskal_wallis()).

The test compares each possible pair of categories from the catField and their mean rank. The null hypothesis is that these are then equal. A simple Bonferroni adjustment is also made for the multiple testing.

Other post-hoc tests that could be considered are Dunn, Nemenyi, Steel-Dwass, Conover-Iman, or pairwise Mann-Whitney U test.
}
\details{
This function selects each possible pair of categories and then simply runs a Mood-Median test, using only those two categories.

See ts_mood_median() for details of the calculations.

The Bonferroni adjustment is simply:
\deqn{p_{adj} = \min \left(p \times n_{comp}, 1\right)}
\deqn{n_{comp} = \frac{k\times\left(k-1\right)}{2}}

\emph{Symbols used:}

\itemize{
\item \eqn{n_{comp}}, number of comparisons (pairs)
\item \eqn{k}, number of categories
}

The R function pairwiseMedianTest from the rcompanion package, as well as the pairwise.mood.medtest function from the RVAideMemoire package, produce the same result, but will apply the Pearson continuity correction.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
