% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_pairwise_bin.R
\name{ph_pairwise_bin}
\alias{ph_pairwise_bin}
\title{Pairwise Binary Test for Post-Hoc Analysis}
\usage{
ph_pairwise_bin(
  data,
  test = "binomial",
  expCount = NULL,
  mtc = "bonferroni",
  ...
)
}
\arguments{
\item{data}{dataframe with scores}

\item{test}{{"binomial", "score", "wald"}, optional test to use for each pair}

\item{expCount}{optional dataframe with categories and expected counts}

\item{mtc}{optional string. Any of the methods available in p_adjust() to correct for multiple tests}

\item{...}{optional additional arguments for the specific test that are passed along.}
}
\value{
a dataframe with:
\item{category 1}{the label of the first category}
\item{category 2}{the label of the second category}
\item{n1}{the sample size of the first category}
\item{n2}{the sample size of the second category}
\item{n pair}{the sample size of of the pair}
\item{obs. prop. 1}{the proportion in the sample of the first category}
\item{exp. prop. 1}{the expected proportion for the first category}
\item{statistic}{the test statistic}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
\item{test}{description of the test used}
}
\description{
This function will perform a one-sample binary test for each possible pair in the data. This could either be a binomial, Wald or score test.

The unadjusted p-values and Bonferroni adjusted p-values are both determined.
}
\seealso{
\code{\link{ts_binomial_os}}, more info on the one-sample binomial test
\code{\link{ts_wald_os}}, more info on the one-sample Wald test
\code{\link{ts_score_os}}, more info on the one-sample score test
\code{\link{ps_adjust}}, the various methods for correction with multiple testing
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
