% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_pairwise_is.R
\name{ph_pairwise_is}
\alias{ph_pairwise_is}
\title{Post-Hoc Pairwise Independent Samples Test}
\usage{
ph_pairwise_is(
  nomField,
  scaleField,
  categories = NULL,
  isTest = "student",
  trimProp = 0.1
)
}
\arguments{
\item{nomField}{the groups variable}

\item{scaleField}{the numeric scores variable}

\item{categories}{vector, optional. the categories to use from catField}

\item{isTest}{string, optional. The independent samples test to use. Either "student" (default), "welch", "trimmed", "yuen", "z"}

\item{trimProp}{float, optional. The trim proportion to use, if applicable. Default is 0.1.}
}
\value{
A dataframe with:
\item{category 1}{the first category in the pair}
\item{category 2}{the second category in the pair}
\item{n1}{sample size of first category}
\item{n2}{sample size of second category}
\item{mean 1}{arithmetic mean of scores in first category}
\item{mean 2}{arithmetic mean of scores in second category}
\item{sample diff.}{difference between the two arithmetic means}
\item{hyp diff.}{the hypothesized difference}
\item{statistic}{the test-statistic}
\item{df}{the degrees of freedom}
\item{p-value}{the unadjusted p-value (significance)}
\item{adj. p-value}{the Bonferroni adjusted p-values}
\item{test}{description of test used}
}
\description{
This function can perform various pairwise independent samples tests, for use after a one-way ANOVA, to determine which categories significantly differ from each other.

A simple Bonferroni correction is also applied.

The independent samples tests that can be used are:
\itemize{
\item Student t, see ts_student_t_is() for details. An alternative version for this is available by using the ph_pairwise_t() function.
\item Welch t, see ts_welch_t_is() for details
\item Trimmed Mean / Yuen, see ts_trimmed_mean_is() for details
\item Z, see ts_z_is() for details
}
}
\details{
The Bonferroni adjustment is simply:
\deqn{p_{adj} = \min \left(p \times n_{comp}, 1\right)}
\deqn{n_{comp} = \frac{k\times\left(k-1\right)}{2}}

\emph{Symbols used:}

\itemize{
\item \eqn{n_{comp}}, number of comparisons (pairs)
\item \eqn{k}, number of categories
}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
