% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_pairwise_iso.R
\name{ph_pairwise_iso}
\alias{ph_pairwise_iso}
\title{Post-Hoc Pairwise Independent Samples Test for Ordinal}
\usage{
ph_pairwise_iso(
  catField,
  ordField,
  categories = NULL,
  levels = NULL,
  iso_test = "mann-whitney",
  ...
)
}
\arguments{
\item{catField}{vector with categories}

\item{ordField}{vector with the scores}

\item{categories}{vector, optional. the categories to use from catField}

\item{levels}{vector, optional. the levels or order used in ordField.}

\item{iso_test}{string, optional. The test of independence to use. Options are "mann-whitney" (default), "mood", "fligner-policello}

\item{...}{other, optional. other arguments to pass on for the specific test used.}
}
\value{
A dataframe with:
\item{category 1}{one of the two categories being compared}
\item{category 2}{second of the two categories being compared}
\item{statistic}{the test statistic}
\item{df}{he degrees of freedom, if applicable}
\item{p-value}{the p-value (significance)}
\item{adj. p-value}{the Bonferroni adjusted p-value}
}
\description{
The Mann-Whitney U, Mood Median and Fligner-Policello test are designed for two categories and an ordinal variable. These can therefor be used as a post-hoc test for a Kruskal-Wallis test (see ts_kruskal_wallis()).
The test compares each possible pair of categories from the catField and their mean rank. The null hypothesis is that these are then equal. A simple Bonferroni adjustment is also made for the multiple testing.
Other post-hoc tests that could be considered are Dunn, Nemenyi, Steel-Dwass, Conover-Iman.
}
\details{
This function selects each possible pair of categories and then simply runs the requested test, using only those two categories.

See ts_mann_whitney(), ts_mood_median() and/or ts_fligner_policello() for details of the calculations.

The Bonferroni adjustment is simply:
\deqn{p_{adj} = \min \left(p \times n_{comp}, 1\right)}
\deqn{n_{comp} = \frac{k\times\left(k-1\right)}{2}}

\emph{Symbols used:}

\itemize{
\item \eqn{n_{comp}}, number of comparisons (pairs)
\item \eqn{k}, number of categories
}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
