% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_pairwise_ps.R
\name{ph_pairwise_ps}
\alias{ph_pairwise_ps}
\title{Post-Hoc Pairwise Paired Samples Tests}
\usage{
ph_pairwise_ps(
  data,
  levels = NULL,
  test = "sign",
  appr = "wilcoxon",
  noDiff = "wilcoxon",
  ties = TRUE,
  cc = FALSE
)
}
\arguments{
\item{data}{dataframe. A column for each variable}

\item{levels}{vector, optional. Indication of what the levels are in order}

\item{test}{string, optional. Test to use in pairwise comparisons. Either "sign" (details), "wilcoxon", "trinomial".}

\item{appr}{string, optional. Option for sign and wilcoxon test. Default for wilcoxon is wilcoxon, for sign is appr. Either "exact", "appr", "wilcoxon", "imanz", "imant"}

\item{noDiff}{string, optional. Method to deal with scores equal to mu. Either "wilcoxon" (default), "pratt", "zsplit". Only applies if test="wilcoxon"}

\item{ties}{boolean, optional. Apply a ties correction. Default is True}

\item{cc}{boolean, optional. use a continuity correction. Default is False. Only applies if test="wilcoxon"}
}
\value{
res, a dataframe with the test results and:
\item{var 1}{the name of the first variable in the pair}
\item{var 2}{the name of the second variable in the pair}
\item{adj. p-value}{the Bonferroni adjusted p-value}
}
\description{
This function simply performs pairwise paired tests: Sign, Wilcoxon and Trinomial. It then adds a Bonferroni correction.

These could be used with a Friedman test, but other post-hoc tests are also available in the ph_friedman() function (a Dunn, Nemenyi and Conover test).
}
\details{
This function creates each possible pair of the variables (columns) and then uses the requested paired samples test.

See for the calculations:
\itemize{
\item Sign test -> ts_sign_ps()
\item Wilcoxon signed rank test -> ts_wilcoxon_ps()
\item Trinomial test -> ts_trinomial_ps()
}

The Bonferroni adjustment is done using:
\deqn{sig._{adj} = \min \left(sig. \times n_c, 1\right)}
\deqn{n_c = \frac{k\times\left(k-1\right)}{2}}

Where \eqn{n_c} is the number of comparisons (pairs)
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
