% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_residual_gof.R
\name{ph_residual_gof}
\alias{ph_residual_gof}
\title{Post-Hoc Residuals Goodness-of-Fit Test}
\usage{
ph_residual_gof(data, expCount = NULL)
}
\arguments{
\item{data}{dataframe with scores}

\item{expCount}{optional dataframe with categories and expected counts}
}
\value{
a dataframe with:
\item{category}{the label of the first category}
\item{z-statistic}{the standardized residuals}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
}
\description{
This function will perform a standardized residuals post-hoc test for each of the categories in a nominal field.

The unadjusted p-values and Bonferroni adjusted p-values are both determined.
}
\details{
The formula used is:
\deqn{z = \frac{F_i - E_i}{\sqrt{E_i}}}
\deqn{sig = 2\times\left(1 - \Phi\left(\left|z\right|\right)\right)}

With:
\itemize{
\item \eqn{F_i}, the observed count for category $i$
\item \eqn{E_i}, the expected count for category $i$
\item \eqn{\Phi\left(\dots\right)}, the cumulative distribution function of the standard normal distribution
}

If no expected counts are provide it is assumed they are all equal for each category, i.e. \eqn{E_i = \frac{n}{k}}

The Bonferroni adjustment is calculated using:
\deqn{p_{adj} = \min \left(p \times k, 1\right)}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
