% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_residual_gof_gof.R
\name{ph_residual_gof_gof}
\alias{ph_residual_gof_gof}
\title{Post-Hoc Residuals Using GoF for GoF}
\usage{
ph_residual_gof_gof(
  data,
  test = "pearson",
  expCount = NULL,
  mtc = "bonferroni",
  ...
)
}
\arguments{
\item{data}{dataframe with scores}

\item{test}{{"pearson", "freeman-tukey", "freeman-tukey-read", "g", "mod-log-g", "neyman", "powerdivergence", "multinomial"} optional test to use}

\item{expCount}{optional dataframe with categories and expected counts}

\item{mtc}{optional string. Any of the methods available in p_adjust() to correct for multiple tests}

\item{...}{optional additional parameters to be passed to the test}
}
\value{
a dataframe with:
\item{category}{the label of the first category}
\item{obs. count}{the observed count of the category}
\item{exp. count}{the expected count of the category}
\item{statistic}{the chi-square test statistic}
\item{df}{the degrees of freedom}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
\item{minExp}{the minimum expected count}
\item{propBelow5}{the proportion of cells with an expected count below 5}
\item{test}{description of the test used}

In case of a multinomial test, the same columns except there are no \emph{minExp} and \emph{propBelow5} columns and:
\item{p obs}{instead of \emph{statistic}, showing the probability of the observed sample table}
\item{n combs.}{instead of \emph{df}, showing the number of possible tables}
}
\description{
This function will perform a goodness-of-fit test using each category and collapsing the other categories.

The unadjusted p-values and Bonferroni adjusted p-values are both determined.

The test is also described at \href{https://peterstatistics.com/Terms/Tests/PostHocAfterGoF.html}{PeterStatistics.com}
}
\section{Before, After and Alternatives}{

Before this an omnibus test might be helpful, these are also the tests used on each category:
\code{\link{ts_freeman_tukey_gof}}, for Freeman-Tukey Test of Goodness-of-Fit.
\code{\link{ts_freeman_tukey_read}}, for Freeman-Tukey-Read Test of Goodness-of-Fit.
\code{\link{ts_g_gof}}, for G (Likelihood Ratio) Goodness-of-Fit Test.
\code{\link{ts_mod_log_likelihood_gof}}, for Mod-Log Likelihood Test of Goodness-of-Fit.
\code{\link{ts_neyman_gof}}, for Neyman Test of Goodness-of-Fit.
\code{\link{ts_pearson_gof}}, for Pearson Test of Goodness-of-Fit.
\code{\link{ts_powerdivergence_gof}}, for Power Divergence GoF Test.

After this you might want to add an effect size measure:
\code{\link{es_post_hoc_gof}} for various effect sizes.

Alternative post-hoc tests:
\code{\link{ph_pairwise_bin}} for Pairwise Binary Tests.
\code{\link{ph_pairwise_gof}} for Pairwise Goodness-of-Fit Tests.
\code{\link{ph_residual_gof_bin}} for Residuals Tests using Binary Tests

More info on the adjustment for multiple testing:
\code{\link{p_adjust}}, various adjustment methods.
}

\examples{
# Examples: get data
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
gssDf <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = gssDf['mar1']

#Example 1 using default settings
ph_residual_gof_gof(ex1)

#Example 2 using a G test and Holm correction:
ph_residual_gof_gof(ex1, test="g", mtc='holm')

}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
