% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_point_biserial.R
\name{r_point_biserial}
\alias{r_point_biserial}
\title{Point Biserial Correlation Coefficient}
\usage{
r_point_biserial(t, df)
}
\arguments{
\item{t}{the test statistic value}

\item{df}{the degrees of freedom}
}
\value{
Point Biserial Correlation Coefficient
}
\description{
This can be seen as coding a binary variable with the groups into 0 and 1, and then calculates a Pearson correlation coefficient between the those values and the scores.

This gives the same result as the formula used and as input the Student t-test statistic and corresponding degrees of freedom.
}
\details{
The formula used is (Friedman, 1968, p. 245):
\deqn{r_{pb} = \sqrt{\frac{t^2}{t^2 + df}}}

\emph{Symbols used:}
\itemize{
\item \eqn{t} the test statistic of the independent samples Student t-test
\item \eqn{df} the degrees of freedom of the independent samples Student t-test
}
}
\examples{
r_point_biserial(0.9984, 1967)

}
\references{
Friedman, H. (1968). Magnitude of experimental effect and a table for its rapid estimation. \emph{Psychological Bulletin, 70}(4), 245–251. https://doi.org/10.1037/h0026258
}
\seealso{
\code{\link{ts_student_t_is}}, Student t-test
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
