% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_polychoric.R
\name{r_polychoric}
\alias{r_polychoric}
\title{Polychoric Correlation Coefficient}
\usage{
r_polychoric(dataVar, groupVar)
}
\arguments{
\item{dataVar}{A vector with the scores data}

\item{groupVar}{A vector with the group data}
}
\value{
Polychoric Correlation Coefficient value
}
\description{
Polychoric Correlation Coefficient
}
\details{
This simply uses the \emph{polychor()} function from the \emph{polycor} library
}
\examples{
scores = c(5, 12, 3, 4, 6, 1, 11, 13, NA)
groups = c("A","A","A","B","B","B","B", NA, "C")
r_polychoric(scores, groups)

}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
