% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rank_biserial_os.R
\name{r_rank_biserial_os}
\alias{r_rank_biserial_os}
\title{Rank biserial correlation coefficient (one-sample)}
\usage{
r_rank_biserial_os(data, levels = NULL, mu = NULL)
}
\arguments{
\item{data}{vector with the numeric scores}

\item{levels}{optional vector with levels in order}

\item{mu}{optional parameter to set the hypothesized median. If not used the midrange is used}
}
\value{
dataframe with the hypothesized median (mu) and the effect size measure
}
\description{
This function will calculate Rank biserial correlation coefficient (one-sample)
}
\details{
The formula used (Kerby, 2014, p. 5):
\deqn{r_{rb} = \frac{\left|R_{pos} - R_{neg}\right|}{R}}
This is actually the same as (King & Minium, 2008, p. 403):
\deqn{r_{rb} = \frac{4\times\left|R_{min} - \frac{R_{pos} + R_{min}}{2}\right|}{n\times\left(n + 1\right)}}

\emph{Symbols used:}
\itemize{
\item \eqn{R_{pos}} the sum of the ranks with a positive deviation from the hypothesized median
\item \eqn{R_{neg}} the sum of the ranks with a positive deviation from the hypothesized median
\item \eqn{R_{min}} the minimum of \eqn{R_{pos}},\eqn{R_{neg}}
\item \eqn{n} the number of ranks with a non-zero difference with the hypothesized median
\item \eqn{R} the sum of all ranks, i.e. \eqn{R_{pos} + R_{neg}}
}

If no hypothesized median is provided, the midrange is used, defined as:
\deqn{\frac{x_{max} - x_{min}}{2}}
Where \eqn{x_{max}} is the maximum value of the scores, and \eqn{x_{min}} the minimum

\strong{Alternative}

The \emph{effectsize} library has a similar function: \emph{rank_biserial()}
}
\examples{
 
data <- c(1, 2, 5, 1, 1, 5, 3, 1, 5, 1, 1, 5, 1, 1, 3, 3, 3, 4, 2, 4)
r_rank_biserial_os(data)
r_rank_biserial_os(data, mu=2)

}
\references{
Kerby, D. S. (2014). The simple difference formula: An approach to teaching nonparametric correlation. \emph{Comprehensive Psychology}, 3, 1–9. https://doi.org/10.2466/11.IT.3.1

King, B. M., & Minium, E. W. (2008). \emph{Statistical reasoning in the behavioral sciences} (5th ed.). John Wiley & Sons, Inc.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
