% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rosenthal.R
\name{r_rosenthal}
\alias{r_rosenthal}
\title{Rosenthal Correlation Coefficient}
\usage{
r_rosenthal(zVal, n)
}
\arguments{
\item{zVal}{z-value of test}

\item{n}{total sample size}
}
\value{
the effect size measure
}
\description{
Rosenthal Correlation Coefficient
}
\details{
The formula used (Rosenthal, 1991, p. 19):
\deqn{r = \frac{z}{\sqrt{n}}}

\emph{Symbols used:}
\itemize{
\item \eqn{n} the sample size
\item \eqn{z} the calculated z-statistic value
}

Rosenthal (1991) is the oldest reference I could find for this correlation coefficient.
However, Cohen (1988, p. 275) actually has a measure 'f' that has the same equation.

For a classification the same as for Pearson correlation use \emph{th_pearson_r()}
}
\examples{
 
z = 1.143943
n = 20
r_rosenthal(z, n)

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.

Rosenthal, R. (1991). \emph{Meta-analytic procedures for social research} (Rev. ed). Sage Publications.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
