% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tetrachoric.R
\name{r_tetrachoric}
\alias{r_tetrachoric}
\title{Tetrachoric Correlation Coefficient}
\usage{
r_tetrachoric(
  field1,
  field2,
  categories1 = NULL,
  categories2 = NULL,
  method = "search"
)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}

\item{method}{c("search", "kirk", "brown", "divgi") method to use (see details)}
}
\value{
Tetrachoric Correlation Coefficient
}
\description{
Tetrachoric Correlation Coefficient
}
\details{
The "search" method does a binary search for rt using the bivariate normal distribution
from the \emph{fMultivar} library.

"kirk" will use Kirk (1973) Fortran TET8 procedure, adapted by stikpet

"brown" will use Brown (1977) - Algorithm AS 116

"divgi" will use Divgi (1979) algorithm

Flow charts of these algorithms can be found at https://peterstatistics.com/CrashCourse/3-TwoVarUnpair/BinBin/BinBin-2b-EffectSize.html
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
r_tetrachoric(bin1, bin2)

}
\references{
Brown, M. B. (1977). Algorithm AS 116: The tetrachoric correlation and its asymptotic standard error. \emph{Applied Statistics, 26}(3), 343. https://doi.org/10.2307/2346985

Divgi, D. R. (1979). Calculation of the tetrachoric correlation coefficient. \emph{Psychometrika, 44}(2), 169–172. https://doi.org/10.1007/BF02293968

Kirk, D. B. (1973). On the numerical approximation of the bivariate normal (tetrachoric) correlation coefficient. \emph{Psychometrika, 38}(2), 259–268. https://doi.org/10.1007/BF02291118
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
