% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_wilcoxon.R
\name{srf}
\alias{srf}
\title{Wilcoxon Sum-of-Ranks Frequency Function}
\usage{
srf(x, y)
}
\arguments{
\item{T}{int with the sum of ranks}

\item{n}{int with the sample size}
}
\value{
A the requested count
}
\description{
This helper function will give the count for a sum of ranks of T, given a sample size of n, using the recursive formula.
}
\details{

}
\references{
McCornack, R. L. (1965). Extended tables of the Wilcoxon matched pair signed rank statistic. \emph{Journal of the American Statistical Association, 60}(311), 864–871. doi:10.2307/2283253
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
