% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cross.R
\name{tab_cross}
\alias{tab_cross}
\title{Cross Table / Contingency Table}
\usage{
tab_cross(
  field1,
  field2,
  order1 = NULL,
  order2 = NULL,
  percent = c(NULL, "all", "row", "column"),
  totals = "exclude"
)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{order1}{: optional list with order for categories of field1}

\item{order2}{: optional list with order for categories of field2}

\item{percent}{: optional which percentages to show. Either "none" (default), "all", "row", "column"}

\item{totals}{: optional to add margin totals. Either "exclude" (default), or "include"}
}
\value{
dataframe : the cross table
}
\description{
A contingency table can be defined as "tables arising when observations on a number of categorical variables are cross-classified" (Everitt, 2004, p.89).

There are quite a few variations on the name for this type of table. Perhaps the oldest name is actually contingency table, which was the name Pearson (1904, p. 34) gave to them. Another popular name is cross tabulation (Upton & Cook, 2002, p. 79), but also cross classification table (Zekeck, 2014, p. 71) and bivariate frequency table (Porkess, 1988, p. 48) are used. The one I used cross table which can for example be found in Newbold et al. (2013, p. 9) or Sa (2007, p. 52).
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
tab_cross(df1[['mar1']], df1[['sex']], percent="column", totals="include")

orderR = c("DIVORCED", "WIDOWED", "SEPARATED", "MARRIED", "NEVER MARRIED")
orderC = c("MALE", "FEMALE")
tab_cross(df1[['mar1']], df1[['sex']], order1=orderR, order2=orderC)

order = c("Not scientific at all", "Not too scientific", "Pretty scientific", "Very scientific")
tab_cross(df1[['mar1']], df1[['accntsci']], order2=order)

}
\references{
Everitt, B. (2004). \emph{The Cambridge dictionary of statistics} (2nd ed.). Cambridge University Press.
Newbold, P., Carlson, W. L., & Thorne, B. (2013). \emph{Statistics for business and economics} (8th ed). Pearson.

Pearson, K. (1904). \emph{Contributions to the Mathematical Theory of Evolution}. XIII. On the theory of contingency and its relation to association and normal correlation. Dulau and Co.

Porkess, R. (1988). \emph{Dictionary of statistics}. Collins.

Sa, J. P. M. de. (2007). \emph{Applied statistics: Using SPSS, Statistica, MATLAB, and R} (2nd ed.). Springer.

Upton, G., & Cook, I. (2002). \emph{Oxford: Dictionary of statistics}. Oxford University Press.

Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
