% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_frequency_bins.R
\name{tab_frequency_bins}
\alias{tab_frequency_bins}
\title{Binned Frequency Table}
\usage{
tab_frequency_bins(
  data,
  nbins = "sturges",
  bins = NULL,
  incl_lower = TRUE,
  adjust = 1
)
}
\arguments{
\item{data}{list or dataframe}

\item{nbins}{optional, either the number of bins to create, or a specific method from the \emph{tab_nbins()} function. Default is "sturges"}

\item{bins}{optional dataframe with lower and upper bounds}

\item{incl_lower}{optional boolean, to include the lower bound, otherwise the upper bound is included. Default is True}

\item{adjust}{optional value to add  or subtract to guarantee all scores will fit in a bin}
}
\value{
dataframe with:

\item{lower bound}{lower bound of class}
\item{upper bound}{upper bound of class}
\item{frequency}{count of scores in bin}
\item{frequency density}{count divided by bin range}
}
\description{
Bins data and creates a frequency table with frequency density.

This function is shown in this \href{https://youtu.be/-1kJrdDkImI}{YouTube video} and frequency tables are also described at \href{https://peterstatistics.com/Terms/Tables/FrequencyTable.html}{PeterStatistics.com}
}
\section{Before, After and Alternatives}{

Before this you might want to determine the number of bins you use:
\code{\link{tab_nbins}}, to determine the number of bins.

After this you might want to visualise the result:
\code{\link{vi_boxplot_single}}, for a Box (and Whisker) Plot.
\code{\link{vi_histogram}}, for a Histogram.
\code{\link{vi_stem_and_leaf}}, for a Stem-and-Leaf Display.
}

\examples{
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
df2 = read.csv(file2, sep=';', na.strings=c("", "NA"))
#Example 1: Numeric Dataframe
ex1a = df2['Gen_Age']
tab_frequency_bins(ex1a)

ex1b = df2['Gen_Age']
myBins = data.frame(c(0, 20, 25, 30), c(20, 25, 30, 120))
tab_frequency_bins(ex1b, bins=myBins)

#Example 2: Numeric list
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
tab_frequency_bins(ex2, adjust=0.1)

}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
