% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_nbins.R
\name{tab_nbins}
\alias{tab_nbins}
\title{Number of Bins}
\usage{
tab_nbins(data, method = "src", maxBins = 100, qmethod = "cdf")
}
\arguments{
\item{data}{vector or pandas series with numeric data}

\item{method}{optional to indicate the method to use. Either "src", "sturges", "qr", "rice", "ts", "exp", "velleman", "doane", "scott", "fd", "shinshim", "stone", or "knuth"}

\item{maxBins}{optional for in iterations with "shinshim", "stone" and "knuth"}

\item{qmethod}{optional quartile method calculation to use for IQR when "fd" is used. See me_quartiles for options}
}
\value{
k : integer with optimum number of bins according to chosen method
}
\description{
To decide on the appropriate number of bins, many different rules can be applied. This function will
determine the number of bins, based on the chosen method.
}
\details{

}
\references{
Doane, D. P. (1976). Aesthetic frequency classifications. \emph{The American Statistician, 30}(4), 181–183. https://doi.org/10.2307/2683757

Freedman, D., & Diaconis, P. (1981). On the histogram as a density estimator. \emph{Zeitschrift Für Wahrscheinlichkeitstheorie Und Verwandte Gebiete, 57}(4), 453–476. https://doi.org/10.1007/BF01025868

Iman, R. L., & Conover, W. J. (1989). \emph{Modern business statistics} (2nd ed.). Wiley.

Knuth, K. H. (2019). Optimal data-based binning for histograms and histogram-based probability density models. \emph{Digital Signal Processing, 95}, 1–30. https://doi.org/10.1016/j.dsp.2019.102581

Lohaka, H. O. (2007). Making a grouped-data frequency table: Development and examination of the iteration algorithm \link{Doctoral dissertation, Ohio University}. https://etd.ohiolink.edu

Scott, D. W. (1979). On optimal and data-based histograms. \emph{Biometrika, 66}(3), 605–610. https://doi.org/10.1093/biomet/66.3.605

Shimazaki, H., & Shinomoto, S. (2007). A method for selecting the bin size of a time histogram. \emph{Neural Computation, 19}(6), 1503–1527. https://doi.org/10.1162/neco.2007.19.6.1503

Stone, C. J. (1984). An asymptotically optimal window selection rule for kernel density estimates. \emph{The Annals of Statistics, 12}(4), 1285–1297.

Sturges, H. A. (1926). The choice of a class interval. \emph{Journal of the American Statistical Association, 21}(153), 65–66. https://doi.org/10.1080/01621459.1926.10502161

Terrell, G. R., & Scott, D. W. (1985). Oversmoothed nonparametric density estimates. \emph{Journal of the American Statistical Association, 80}(389), 209–214. https://doi.org/10.2307/2288074
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}
}
