% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cle.R
\name{th_cle}
\alias{th_cle}
\title{Rule-of-Thumb for Common Language Effect Size}
\usage{
th_cle(cle, qual = "vd", convert = "no")
}
\arguments{
\item{cle}{the Vargha-Delaney A value}

\item{qual}{{"vd", others via convert}, optional rules-of-thumb to use, currently only 'vd' for Vargha-Delaney, otherwise a converted measure}

\item{convert}{{"no", "rb", "cohen_d"}, optional list in case to use a rule-of-thumb from a converted measure. Either "no" for no conversion, "rb" for rank-biserial, or "cohen_d" for Cohen d.}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
This function will give a qualification (classification) for a Common Language Effect Size (/ Vargha-Delaney A / Probability of Superiority)
}
\details{
Vargha and Delaney (2000, p. 106):\tabular{ll}{
   |0.5 - A| \tab Interpretation \cr
   0.00 < 0.06 \tab negligible \cr
   0.11 < 0.14 \tab small \cr
   0.28 < 0.21 \tab medium \cr
   0.21 or more \tab large \cr
}


The CLE can be converted to a Rank Biserial Coefficient using:
\deqn{r_b = 2\times CLE - 1}

Rules of thumb from the \strong{th_rank_biserial()} function could then be used, by setting: \emph{convert="rb"}, and \emph{qual} is any of the options in th_rank_biserial()

This in turn can be converted to Cohen's d using (Marfo & Okyere, 2019, p.4):
\deqn{d = 2\times \phi^{-1}\left(-\frac{1}{r_b - 2}\right)}
Rules of thumb from the \strong{th_cohen_d()} function could then be used, by setting: \emph{convert="cohen_d"}, and \emph{qual} is any of the options in th_cohen_d()
}
\references{
Marfo, P., & Okyere, G. A. (2019). The accuracy of effect-size estimates under normals and contaminated normals in meta-analysis. \emph{Heliyon, 5}(6), e01838. doi:10.1016/j.heliyon.2019.e01838

Vargha, A., & Delaney, H. D. (2000). A critique and improvement of the CL common language effect size statistics of McGraw and Wong. \emph{Journal of Educational and Behavioral Statistics, 25}(2), 101–132. doi:10.3102/10769986025002101
}
\seealso{
\code{\link{es_common_language_is}}, to determine the CLE.

\code{\link{th_rank_biserial}}, for options for rules of thumb when converting to Rank Biserial.

\code{\link{th_cohen_d}}, for options for rules of thumb when converting to Cohen d.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
