% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cliff_delta.R
\name{th_cliff_delta}
\alias{th_cliff_delta}
\title{Rule-of-Thumb for Cliff Delta}
\usage{
th_cliff_delta(d, qual = "romano")
}
\arguments{
\item{d}{the Cliff Delta value}

\item{qual}{optional setting for which rule of thumb to use. Either "romano" (default), "metsamuuronen"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for Cliff Delta
}
\details{
The following rules-of-thumb can be used:

\emph{"romano"} => Uses Romano et al. (2006, p. 14):\tabular{ll}{
   \|d\| \tab Interpretation \cr
   0.00 < 0.15 \tab negligible \cr
   0.15 < 0.33 \tab small \cr
   0.33 < 0.47 \tab medium \cr
   0.47 or more \tab large \cr
}


\emph{"metsamuuronen"} => Metsämuuronen (2023, p. 17):\tabular{ll}{
   \|d\| \tab Interpretation \cr
   0.00 < 0.11 \tab negligible \cr
   0.11 < 0.28 \tab small \cr
   0.28 < 0.43 \tab medium \cr
   0.43 or more \tab large \cr
}
}
\section{Before, After and Alternatives}{

Cliff delta could be converted to Cohen d (Marfo & Okyere, 2019, p. 4) or Vargha-Delaney A (use \code{\link{es_convert}} to convert the effect size measure)
}

\references{
Marfo, P., & Okyere, G. A. (2019). The accuracy of effect-size estimates under normals and contaminated normals in meta-analysis. \emph{Heliyon, 5}(6), e01838. doi:10.1016/j.heliyon.2019.e01838

Metsämuuronen, J. (2023). Somers’ delta as a basis for nonparametric effect sizes: Grissom-Kim PS, Cliff’s d, and Vargha-Delaney A as specific cases of Somers delta. doi:10.13140/RG.2.2.36002.09925

Romano, J., Kromrey, J. D., Coraggio, J., Skowronek, J., & Devine, L. (2006). Exploring methods for evaluating group differences on the NSSE and other surveys: Are the t-test and Cohen’s d indices the most appropriate choices? 1–51.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
