% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cohen_d.R
\name{th_cohen_d}
\alias{th_cohen_d}
\title{Rules of Thumb for Cohen d}
\usage{
th_cohen_d(d, qual = "sawilowsky")
}
\arguments{
\item{d}{the Cohen d value}

\item{qual}{optional the rule of thumb to be used. Either "sawilowsky" (default), "brydges", cohen", "rosenthal", or "lovakov"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
This function will give a qualification (classification) for Cohen d
}
\details{
The following rules-of-thumb can be used:

\emph{"brydges"} => Brydges (2019, p. 5):\tabular{ll}{
   \|d\| \tab Interpretation \cr
   0.00 < 0.15 \tab negligible \cr
   0.15 < 0.40 \tab small \cr
   0.40 < 0.75 \tab medium \cr
   0.75 or more \tab large \cr
}


"cohen" => Cohen (1988, p. 40)\tabular{ll}{
   |d| \tab Interpretation \cr
   0.00 < 0.20 \tab negligible \cr
   0.20 < 0.50 \tab small \cr
   0.50 < 0.80 \tab medium \cr
   0.80 or more \tab large \cr
}


"sawilowsky" => Sawilowsky (2009, p. 599)\tabular{ll}{
   |d| \tab Interpretation \cr
   0.00 < 0.10 \tab negligible \cr
   0.10 < 0.20 \tab very small \cr
   0.20 < 0.50 \tab small \cr
   0.50 < 0.80 \tab medium \cr
   0.80 < 1.20 \tab large \cr
   1.20 < 2.00 \tab very large \cr
   2.00 or more \tab huge \cr
}


"lovakov" => Lovakov and Agadullina (2021, p. 501)\tabular{ll}{
   |d| \tab Interpretation \cr
   0.00 < 0.15 \tab negligible \cr
   0.15 < 0.36 \tab small \cr
   0.36 < 0.65 \tab medium \cr
   0.65 or more \tab large \cr
}


"rosenthal" => Rosenthal (1996, p. 45)\tabular{ll}{
   |d| \tab Interpretation \cr
   0.00 < 0.20 \tab negligible \cr
   0.20 < 0.50 \tab small \cr
   0.50 < 0.80 \tab medium \cr
   0.80 < 1.30 \tab large \cr
   0.80 or more \tab very large \cr
}
}
\section{Before, After and Alternatives}{

Cohen d for one-sample and Hedges g could be converted to Cohen d:
\code{\link{es_convert}}, for the conversions.
\code{\link{es_cohen_d}}, to determine Cohen d'
\code{\link{es_hedges_g_os}}, for Hedges g
}

\examples{
es = 0.6
th_cohen_d(es)
 
}
\references{
Brydges, C. R. (2019). Effect size guidelines, sample size calculations, and statistical power in gerontology. \emph{Innovation in Aging, 3}(4), 1-8. doi:10.1093/geroni/igz036

Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.

Lovakov, A., & Agadullina, E. R. (2021). Empirically derived guidelines for effect size interpretation in social psychology. \emph{European Journal of Social Psychology, 51}(3), 485-504. doi:10.1002/ejsp.2752

Rosenthal, J. A. (1996). Qualitative descriptors of strength of association and effect size. \emph{Journal of Social Service Research, 21}(4), 37-59. doi:10.1300/J079v21n04_02

Sawilowsky, S. (2009). New effect size rules of thumb. \emph{Journal of Modern Applied Statistical Methods, 8}(2). doi:10.22237/jmasm/1257035100
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
