% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cohen_f.R
\name{th_cohen_f}
\alias{th_cohen_f}
\title{Rule-of-Thumb for Cohen f}
\usage{
th_cohen_f(f, qual = "cohen")
}
\arguments{
\item{qual}{optional setting for which rule of thumb to use. Currently only "cohen"}

\item{g}{the Cohen f value}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Cohen f effect size.
}
\details{
Cohen's rule of thumb for Cohen f (1988, pp. 285-287):\tabular{ll}{
   |f| \tab Interpretation \cr
   0.00 < 0.10 \tab negligible \cr
   0.10 < 0.25 \tab small \cr
   0.25 < 0.40 \tab medium \cr
   0.40 or more \tab large \cr
}
}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\seealso{
\code{\link{es_cohen_f}}, to determine Cohen f

\code{\link{cor_rosenthal}}, to determine the Rosenthal correlation, which Cohen called also Cohen f
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
