% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cohen_g.R
\name{th_cohen_g}
\alias{th_cohen_g}
\title{Rule-of-Thumb for Cohen g}
\usage{
th_cohen_g(g, qual = "cohen")
}
\arguments{
\item{g}{the Cohen g value}

\item{qual}{optional setting for which rule of thumb to use. Currently only "cohen"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Cohen g effect size.

This function is shown in this \href{https://youtu.be/3DEmngmws2U}{YouTube video} and the effect size is also described at \href{https://peterstatistics.com/Terms/EffectSizes/CohenG.html}{PeterStatistics.com}
}
\details{
Cohen's rule of thumb for Cohen g (1988, pp. 147-149):\tabular{ll}{
   |g| \tab Interpretation \cr
   0.00 < 0.05 \tab negligible \cr
   0.05 < 0.15 \tab small \cr
   0.15 < 0.25 \tab medium \cr
   0.25 or more \tab large \cr
}
}
\section{Before, After and Alternatives}{

\code{\link{es_cohen_g}}, to determine Cohen g
}

\examples{
es = 0.6
th_cohen_g(es)
 
}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
