% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cohen_h.R
\name{th_cohen_h}
\alias{th_cohen_h}
\title{Rule-of-Thumb for Cohen h}
\usage{
th_cohen_h(h, qual = "cohen")
}
\arguments{
\item{h}{the Cohen h value}

\item{qual}{optional setting for which rule of thumb to use. Currently only 'cohen'}
}
\value{
dataframe with the qualification and source
}
\description{
Rule-of-Thumb for Cohen h
}
\details{
Cohen's rule of thumb for Cohen g (1988, p. 198):\tabular{ll}{
   |h| \tab Interpretation \cr
   0.00 < 0.20 \tab negligible \cr
   0.20 < 0.50 \tab small \cr
   0.50 < 0.80 \tab medium \cr
   0.80 or more \tab large \cr
}


Note that Cohen actually just lists small = 0.20, medium = 0.50, and large = 0.80.
}
\examples{
h <- -0.2391304
th_cohen_h(h)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
