% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cohen_w.R
\name{th_cohen_w}
\alias{th_cohen_w}
\title{Rule-of-Thumb for Cohen w}
\usage{
th_cohen_w(w, qual = "cohen")
}
\arguments{
\item{w}{the Cohen w value}

\item{qual}{optional setting for which rule of thumb to use. Currently only 'cohen'}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Cohen w effect size.
}
\details{
Cohen's rule of thumb for Cohen w (1988, p. 227):\tabular{ll}{
   |w| \tab Interpretation \cr
   0.00 < 0.10 \tab negligible \cr
   0.10 < 0.30 \tab small \cr
   0.30 < 0.50 \tab medium \cr
   0.50 or more \tab large \cr
}
}
\examples{
es = 0.6
th_cohen_w(es)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\seealso{
\code{\link{es_cohen_w}}, to determine Cohen w
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
