% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cramer_v.R
\name{th_cramer_v}
\alias{th_cramer_v}
\title{Rule-of-Thumb for Cramér V}
\usage{
th_cramer_v(v, qual = "rea-parker")
}
\arguments{
\item{v}{the Cramér V value}

\item{qual}{optional setting for which rule of thumb to use. Either "rea-parker" (default), "akoglu", "calamba-rustico"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Cramér V effect size. Note however that many will actually use the rule-of-thumb for Cohen w and convert Cramér V to Cohen w first.

The measure is also described at \href{https://peterstatistics.com/Terms/EffectSizes/CramerV.html}{PeterStatistics.com}
}
\details{
\emph{"rea-parker"} => Uses Rea and Parker (1992, p. 203):\tabular{ll}{
   \|v\| \tab Interpretation \cr
   0.00 < 0.10 \tab negligible \cr
   0.10 < 0.20 \tab weak \cr
   0.20 < 0.40 \tab moderate \cr
   0.40 < 0.60 \tab relatively strong \cr
   0.60 < 0.80 \tab strong \cr
   0.80 or more \tab very strong \cr
}


\emph{"akoglu"} => Uses Akoglu (2018, p. 92):\tabular{ll}{
   \|v\| \tab Interpretation \cr
   0.00 < 0.05 \tab very weak \cr
   0.05 < 0.10 \tab weak \cr
   0.10 < 0.15 \tab moderate \cr
   0.15 < 0.25 \tab strong \cr
   0.25 or more \tab very strong \cr
}


\emph{"calamba-rustico"} => Uses Calamba and Rustico (2019, p. 7):\tabular{ll}{
   \|v\| \tab Interpretation \cr
   0.00 < 0.15 \tab very weak \cr
   0.15 < 0.20 \tab weak \cr
   0.20 < 0.25 \tab moderate \cr
   0.25 < 0.30 \tab moderately strong \cr
   0.30 < 0.35 \tab strong \cr
   0.35 < 0.50 \tab worrisomely strong \cr
   0.50 or more \tab redundant \cr
}


Note that the original source has a gap from 0.40 < 0.50, I added this to the 'worrisomely strong' category.
}
\section{Before, After and Alternatives}{

Before using this function you need to obtain a Cramer v value:
\code{\link{es_cramer_v_gof}}, to determine Cramér V for a Goodness-of-Fit test.
\code{\link{es_cramer_v_ind}}, to determine Cramér V for a test of independence.
}

\examples{
es = 0.6
th_cramer_v(es)


}
\references{
Akoglu, H. (2018). User’s guide to correlation coefficients. \emph{Turkish Journal of Emergency Medicine, 18}(3), 91–93. doi:10.1016/j.tjem.2018.08.001

Calamba, S. S., & Rustico, E. M. P. (2019). Usefulness of code of ethics for professional accountants in resolving ethical conflicts in the Philippines.

Rea, L. M., & Parker, R. A. (1992). \emph{Designing and conducting survey research: A comprehensive guide}. Jossey-Bass Publishers.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
