% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_gk_gamma.R
\name{th_gk_gamma}
\alias{th_gk_gamma}
\title{Rule-of-Thumb for Goodman-Kruskal gamma}
\usage{
th_gk_gamma(g, qual = "blaikie")
}
\arguments{
\item{g}{the Goodman-Kruskal gamma value}

\item{qual}{optional setting for which rule of thumb to use. Either "blaikie" (default), "rea-parker", "metsamuuronen"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for Goodman-Kruskal gamma
}
\details{
The following rules-of-thumb can be used:

\emph{"blaikie"} => Uses Blaikie (2003, p. 100):\tabular{ll}{
   \|g\| \tab Interpretation \cr
   0.00 < 0.10 \tab negligible \cr
   0.10 < 0.30 \tab weak \cr
   0.30 < 0.60 \tab moderate \cr
   0.60 < 0.75 \tab strong \cr
   0.75 or more \tab very strong \cr
}


\emph{"rea-parker"} => Rea and Parker (2014, p. 229):\tabular{ll}{
   \|g\| \tab Interpretation \cr
   0.00 < 0.10 \tab negligible \cr
   0.10 < 0.30 \tab low \cr
   0.30 < 0.60 \tab moderate \cr
   0.60 < 0.75 \tab strong \cr
   0.75 or more \tab very strong \cr
}


\emph{"metsamuuronen"} => Metsämuuronen (2023, p. 17):\tabular{ll}{
   \|g\| \tab Interpretation \cr
   0.00 < 0.14 \tab negligible \cr
   0.14 < 0.31 \tab small \cr
   0.31 < 0.45 \tab medium \cr
   0.45 < 0.62 \tab large \cr
   0.62 < 0.84 \tab very large \cr
   0.84 or more \tab huge \cr
}
}
\references{
Blaikie, N. W. H. (2003). \emph{Analyzing quantitative data: From description to explanation}. Sage Publications Ltd.

Metsämuuronen, J. (2023). Somers’ delta as a basis for nonparametric effect sizes: Grissom-Kim PS, Cliff’s d, and Vargha-Delaney A as specific cases of Somers delta. doi:10.13140/RG.2.2.36002.09925

Rea, L. M., & Parker, R. A. (2014). \emph{Designing and conducting survey research: A comprehensive guide} (4th ed.). Jossey-Bass, a Wiley brand.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
