% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_kaiser_b.R
\name{th_kaiser_b}
\alias{th_kaiser_b}
\title{Rule-of-Thumb for Kaiser b}
\usage{
th_kaiser_b(b, qual = "kaiser")
}
\arguments{
\item{b}{the Cohen b value}

\item{qual}{optional setting for which rule of thumb to use. Currently only "kaiser"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Kaiser b variation measure.
}
\details{
Kaiser's rule of thumb for Kaiser b (1968, p. 212):\tabular{ll}{
   |b| \tab Interpretation \cr
   0.00 < 0.70 \tab terrible \cr
   0.70 < 0.80 \tab poor \cr
   0.80 < 0.90 \tab fair \cr
   0.90 < 0.95 \tab good \cr
   0.95 < 1.00 \tab excellent \cr
}
}
\references{
Kaiser, H. F. (1968). A measure of the population quality of legislative apportionment. \emph{American Political Science Review, 62}(1), 208–215. doi:10.2307/1953335
}
\seealso{
\code{\link{me_qv}}, to determine Kaiser b
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
