% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_odds_ratio.R
\name{th_odds_ratio}
\alias{th_odds_ratio}
\title{Rules of thumb for Odds Ratio}
\usage{
th_odds_ratio(or, qual = "chen")
}
\arguments{
\item{or}{the odds ratio}

\item{qual}{c("chen", "wuensch", "jones") optional rule of thumb to use (default is "chen")}
}
\value{
data frame with the qualification/classification  and source
}
\description{
Rules of thumb for Odds Ratio
}
\details{
If the OR is less than 1, the alternative is used, i.e. 1/OR.

Chen et al. (2010, p. 862):\tabular{ll}{
   OR \tab Interpretation \cr
   1.00 < 1.68 \tab negligible \cr
   1.68 < 3.47 \tab weak \cr
   3.47 < 6.71 \tab moderate \cr
   6.71 or more \tab strong \cr
}


Wuensch (2009, p. 2):\tabular{ll}{
   OR \tab Interpretation \cr
   1.00 < 1.49 \tab negligible \cr
   1.49 < 3.45 \tab small \cr
   3.45 < 9 \tab medium \cr
   9 or more \tab large \cr
}


Jones (2014):\tabular{ll}{
   OR \tab Interpretation \cr
   1.00 < 1.5 \tab negligible \cr
   1.5 < 2.5 \tab small \cr
   2.5 < 4.3 \tab medium \cr
   4.3 or more \tab large \cr
}
}
\examples{
th_odds_ratio(5.23)
th_odds_ratio(5.23, qual="wuensch")

}
\references{
Chen, H., Cohen, P., & Chen, S. (2010). How big is a big Odds Ratio? Interpreting the magnitudes of Odds Ratios in epidemiological studies. \emph{Communications in Statistics - Simulation and Computation, 39}(4), 860–864. https://doi.org/10.1080/03610911003650383

Jones, K. (2014, June 5). How do you interpret the odds ratio (OR)? ResearchGate. https://www.researchgate.net/post/How_do_you_interpret_the_odds_ratio_OR

Wuensch, K. (2009). Cohen’s conventions for small, medium, and large effects. https://imaging.mrc-cbu.cam.ac.uk/statswiki/FAQ/effectSize?action=AttachFile&do=get&target=esize.doc
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
