% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_point_biserial.R
\name{th_point_biserial}
\alias{th_point_biserial}
\title{Rule of thumb for Point Biserial Correlation}
\usage{
th_point_biserial(rp, qual = "cohen")
}
\arguments{
\item{rp}{the correlation coefficient}

\item{qual}{optional the rule of thumb to be used. Currently only "cohen"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Point Biserial Correlation.
}
\details{
Cohen's rule of thumb for rank biserial (1988, p. 82):\tabular{ll}{
   |r_p| \tab Interpretation \cr
   0.00 < 0.100 \tab negligible \cr
   0.100 < 0.243 \tab small \cr
   0.243 < 0.371 \tab medium \cr
   0.371 < 1 \tab large \cr
}
}
\examples{
es = 0.6
th_point_biserial(es)

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
