% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_post_hoc_gof.R
\name{th_post_hoc_gof}
\alias{th_post_hoc_gof}
\title{Post-Hoc Goodness-of-Fit Rules-of-Thumb}
\usage{
th_post_hoc_gof(eff_sizes, convert = FALSE, ph_results = NULL, ...)
}
\arguments{
\item{eff_sizes}{dataframe, the dataframe from es_post_hoc_gof()}

\item{convert}{boolean, optional. convert the effect size to use the rule-of-thumb from another, see details}

\item{ph_results}{dataframe, optional. the post-hoc analysis results, required for JBM-E and Fei.}

\item{...}{optional. additional arguments for the specific rule-of-thumb that are passed along. Most common 'qual=...' for a specific set of rules-of-thumb.}
}
\value{
df, dataframe with the same dataframe as the provided \emph{eff_sizes}, but added:

\item{qualification}{, the qualification using the rule-of-thumb.}
\item{reference}{, a reference to the source for the rule-of-thumb.}

If a conversion was done or needed:
\item{conversion description}{, the value of the converted measure}
}
\description{
This function will add a classification to the results of \strong{es_post_hoc_gof()} using a rules-of-thumb. This is frowned upon by some, and the rule-of-thumb can vary per discipline.
}
\details{
For Johnston-Berry-Mielke E and Fei, a conversion is always done to Cramér V, when setting \emph{convert=True} it will convert it again to Cohen w.

Other possible conversions are Cohen h' to Cohen h, and Cramér V to Cohen w.

See the separate documentation for each of the rules-of-thumb, or conversion.
}
\section{Before, After and Alternatives}{

Before using this the post-hoc effect sizes need to be made:
\code{\link{es_post_hoc_gof}}, to obtain post-hoc effect sizes.

Depending on the measure, the function will use the rules of thumb for:
\code{\link{th_cohen_g}}, for Cohen g.
\code{\link{th_cohen_h}}, for Cohen h.
\code{\link{th_cohen_w}}, for Cohen w.
\code{\link{th_cramer_v}}, for Cramer V.
\code{\link{th_pearson_r}}, for Pearson r.

It can also convert using:
\code{\link{es_convert}}, to convert various effect sizes.
}

\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
