% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_rank_biserial.R
\name{th_rank_biserial}
\alias{th_rank_biserial}
\title{Rule-of-Thumb for Rank Biserial Correlation}
\usage{
th_rank_biserial(rb, version = "glass", qual = NULL, convert = "no")
}
\arguments{
\item{rb}{the rank-biserial correlation value}

\item{version}{optional version of rank-biserial that was used. Either "glass" (default) or "cureton"}

\item{qual}{optional setting for which rule of thumb to use. Either "cohen" (default), "vd", "sawilowsky", "cohen-conv", "lovakov", "rosenthal", "brydges"}

\item{convert}{optional conversion to use (only for Glass version). Either "no", "cohen_d", "vda"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Rank Biserial Correlation.

This function is shown in this \href{https://youtu.be/31jr7jUCni4}{YouTube video} and the measure is also described at \href{https://peterstatistics.com/Terms/Correlations/RankBiserialCorrelation.html}{PeterStatistics.com}
}
\details{
If a Cureton version of rank-biserial was used, the result for independent samples is the same as Goodman-Kruskal gamma, so we can also use those rules-of-thumb.

If a Glass version is used, then we can either use Cliff Delta, Somers d, or a conversion to either Cohen d, or Vargha-Delaney A.

See the separate functions on each of these for various rules-of-thumbs.
}
\section{Before, After and Alternatives}{

Before this you might want to obtain the measure:
\code{\link{r_rank_biserial_is}}, to determine a the rank biserial for independent samples.
\code{\link{r_rank_biserial_os}}, to determine a the rank biserial for one-sample.

The function uses the convert function and corresponding rules of thumb:
\code{\link{es_convert}}, to convert this to Cohen d.
}

\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
