% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_somers_d.R
\name{th_somers_d}
\alias{th_somers_d}
\title{Rule-of-Thumb for Somers d}
\usage{
th_somers_d(d, qual = "metsamuuronen")
}
\arguments{
\item{d}{the Somers d value}

\item{qual}{optional setting for which rule of thumb to use. Currently only "vargha"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for Somers d
}
\details{
The following rules-of-thumb can be used:

\emph{"metsamuuronen"} => Uses Metsämuuronen (2023, p. 17):\tabular{ll}{
   \|d\| \tab Interpretation \cr
   0.00 < 0.13 \tab negligible \cr
   0.13 < 0.29 \tab small \cr
   0.29 < 0.43 \tab medium \cr
   0.43 < 0.59 \tab large \cr
   0.59 < 0.81 \tab very large \cr
   0.81 or more \tab huge \cr
}
}
\references{
Metsämuuronen, J. (2023). Somers’ delta as a basis for nonparametric effect sizes: Grissom-Kim PS, Cliff’s d, and Vargha-Delaney A as specific cases of Somers delta. doi:10.13140/RG.2.2.36002.09925
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
