% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_vda.R
\name{th_vda}
\alias{th_vda}
\title{Rule-of-Thumb for Vargha-Delaney A}
\usage{
th_vda(a, qual = "vd")
}
\arguments{
\item{a}{the Vargha-Delaney A value}

\item{qual}{optional setting for which rule of thumb to use. Either "vd" (default), "sawilowsky", "cohen", "lovakov", "rosenthal"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Vargha-Delaney A
}
\details{
Vargha and Delaney (2000, p. 106):\tabular{ll}{
   |0.5 - A| \tab Interpretation \cr
   0.00 < 0.06 \tab negligible \cr
   0.11 < 0.14 \tab small \cr
   0.28 < 0.21 \tab medium \cr
   0.21 or more \tab large \cr
}


"sawilowsky", "cohen", "lovakov", and "rosenthal" will use the rule-of-thumb from Cohen d, by converting VD-A first to a (Glass) Rank Biserial (Cliff delta) then to Cohen d, and use the rule of thumb from Cohen d.
}
\references{
Vargha, A., & Delaney, H. D. (2000). A critique and improvement of the CL common language effect size statistics of McGraw and Wong. \emph{Journal of Educational and Behavioral Statistics, 25}(2), 101–132. doi:10.3102/10769986025002101
}
\seealso{
\code{\link{es_vargha_delaney_a}}, to determine a Vargha-Delaney A

\code{\link{es_convert}}, to convert this to Cohen d

\code{\link{th_cohen_d}}, rules of thumb for Cohen d
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
