% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_yule_q.R
\name{th_yule_q}
\alias{th_yule_q}
\title{Rules of thumb for Yule Q}
\usage{
th_yule_q(q, qual = "glen")
}
\arguments{
\item{q}{the Yule Q value}

\item{qual}{optional for which rule-of-thumb to use. Currently only "glen"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for Yule Q effect size.
}
\details{
Glen rule of thumb for Yule Q (2017):\tabular{ll}{
   |Q| \tab Interpretation \cr
   0.00 < 0.30 \tab negligible \cr
   0.30 < 0.50 \tab moderate \cr
   0.50 < 0.70 \tab substantial \cr
   0.70 or more \tab very strong \cr
}
}
\examples{
q = 0.6
th_yule_q(q)

}
\references{
Glen, S. (2017, August 16). Gamma Coefficient (Goodman and Kruskal’s Gamma) & Yule’s Q. Statistics How To. https://www.statisticshowto.com/gamma-coefficient-goodman-kruskal/
}
\seealso{
\code{\link{es_bin_bin}}, to determine Yule Q
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
