% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_binomial_os.R
\name{ts_binomial_os}
\alias{ts_binomial_os}
\title{One-Sample Binomial Test}
\usage{
ts_binomial_os(
  data,
  codes = NULL,
  p0 = 0.5,
  twoSidedMethod = c("eqdist", "double", "smallp")
)
}
\arguments{
\item{data}{A vector with the data}

\item{codes}{Optional vector with the two codes to use}

\item{p0}{Optional hypothesized proportion for the first category (default is 0.5)}

\item{twoSidedMethod}{Optional method to be used for 2-sided significance (see details)}
}
\value{
Dataframe with:
\item{pValue}{two-sided p-value}
\item{testUsed}{a description of the test used}
}
\description{
Performs a one-sample (exact) binomial test.

This test can be useful with a single binary variable as input. The null hypothesis is usually that the proportions of the two categories in the population are equal (i.e. 0.5 for each). If the p-value of the test is below the pre-defined alpha level (usually 5\% = 0.05) the null hypothesis is rejected and the two categories differ in proportion significantly.

The input for the function doesn't have to be a binary variable. A nominal variable can also be used and the two categories to compare indicated.

A significance in general is the probability of a result as in the sample, or more extreme, if the null hypothesis is true. For a two-tailed binomial test the 'or more extreme' causes a bit of a complication. There are different methods to approach this problem. See the details for more information.

A \href{https://youtu.be/9OGCi1Q7tBQ}{YouTube} video on the binomial test.
}
\details{
A one sided p-value is calculated first:
\deqn{sig_{one-tail} = \text{Bin}\left(n, n_{min}, p_0^*\right)}
With:
\deqn{n_min = min\left\{n_s, n_f\right\}}
\deqn{p_0^* = \begin{cases}p_0 & \text{ if } n_{min}=n_s \\1 - p_0 & \text{ if } n_{min}= n_f\end{cases}}

\emph{Symbols used:}
\itemize{
\item \eqn{n} is the number of cases
\item \eqn{n_s} is the number of successes
\item \eqn{n_f} is the number of failures
\item \eqn{p_0} is the probability of a success according to the null hypothesis
\item \eqn{p_0^*} is the probability adjusted in case failures is used
\item \eqn{\text{Bin}\left(\dots\right)} the binomial cumulative distribution function
}

For the two sided significance three options can be used.

\emph{Option 1: Equal Distance Method (twoSidedMethod="eqdist")}
\deqn{sig_{two-tail} = B\left(n, n_{min}, p_0^* \right) + 1 - B\left(n, \left \lfloor 2 \times n_0 \right \rfloor - n_{min} - 1, p_0^*\right)}
With:
\deqn{n_0 = \left\lfloor n\times p_0\right\rfloor}

This method looks at the number of cases. In a sample of \eqn{n} people,
we’d then expect \eqn{n_0 = \left\lfloor n\times p_0\right\rfloor} successes (we round the result down to the nearest integer).
We only had \eqn{n_{min}}, so a difference of \eqn{n_0-n_{min}}.
The ‘equal distance method’ now means to look for the chance of having \eqn{k} or less,
and \eqn{n_0+n_0-n_{min}=2\times n_0-n{_min}} or more.
Each of these two probabilities can be found using a binomial distribution.
Adding these two together than gives the two-sided significance.

\emph{Option 2: Small p-method (twoSidedMethod="smallp")}
\deqn{sig_{two-tail} = B\left(n, n_{min}, p_0^*\right) + \sum_{i=n_{min}+1}^n \begin{cases} 0 & \text{ if } b\left(n, i, p_0^*\right)> b\left(n, n_{min}, p_0^*\right) \\ b\left(n, i, p_0^*\right)& \text{ if } x \leq  b\left(n, i, p_0^*\right)> b\left(n, n_{min}, p_0^*\right) \end{cases}}
With:
\eqn{b\left(\dots\right)} as the binomial probability mass function.

This method looks at the probabilities itself.
\eqn{b\left(n, n_{min},p_0^*\right)} is the probability of having exactly \eqn{n_{min}}
out of a group of n, with a chance \eqn{p_0^*} each time.
The method of small p-values now considers ‘or more extreme’ any number between 0 and n (the sample size)
that has a probability less or equal to this.
This means we need to go over each option, determine the probability and check if it is lower or equal.
So, the probability of 0 successes, the probability of 1 success, etc.
The sum for all of those will be the two-sided significance.
We can reduce the work a little since any value below \eqn{n_{min}} ,
will also have a lower probability, so we only need to sum over the ones above it and add the one-sided significance to the sum of those.

\emph{Option 3: Double single (twoSidedMethod="double")}
\deqn{sig_{two-tail} = 2\times sig_{one-tail}}

Fairly straight forward. Just double the one-sided significance.
}
\section{Alternatives}{


R \emph{stats} library: binom.test()
}

\examples{
#Example 1: Numeric list
ex1 = c(1, 1, 2, 1, 2, 1, 2, 1)
ts_binomial_os(ex1)
ts_binomial_os(ex1, p0=0.3, twoSidedMethod="eqdist")
ts_binomial_os(ex1, p0=0.3, twoSidedMethod="double")
ts_binomial_os(ex1, p0=0.3, twoSidedMethod="smallp")

#Example 2: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ts_binomial_os(df1['sex'])
ts_binomial_os(df1['mar1'], codes=c("DIVORCED", "NEVER MARRIED"))

}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
