% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_brunner_munzel.R
\name{ts_brunner_munzel_perm}
\alias{ts_brunner_munzel_perm}
\title{Brunner-Munzel Studentized Permutation Test}
\usage{
ts_brunner_munzel_perm(
  catField,
  ordField,
  categories = NULL,
  levels = NULL,
  n_iter = 1000
)
}
\arguments{
\item{catField}{A vector or dataframe with the group data}

\item{ordField}{A vector or dataframe with the scores data}

\item{categories}{optional list with the two categories to use from catField. If not set the first two found will be used}

\item{levels}{optional list with the scores in order}

\item{n_iter}{optional number of iterations. Default 1000}
}
\value{
A dataframe with:
\item{iters}{the number of iterations}
\item{observed}{the observed Brunner-Munzel statistic}
\item{n above}{the number of permutations that had a value above but not equal equal to the observed statistic}
\item{n below}{the number of permutations that had a value below but not equal equal to the observed statistic}
\item{p-appr.}{the significance (p-value), two-tailed}
}
\description{
This function performs a studentized permutation test of the Brunner-Munzel test. It is recommended to use this test above the regular version if sample sizes are too small in the data (e.g. less than 10 (or 15)).
}
\details{
The idea for a studentized permutation test was proposed by Neubert and Brunner (2007). The regular Brunner-Munzel test statistic is calculated, then the categories get shuffled and the statistic is re-calculated. This repeats.

The minimum of the number of permutations above vs. below the observed statistic is multiplied by two and then divided by the number of iterations, to obtain the p-value (Schüürhuis et al., 2025, p. 7)

See the ts_brunner_munzel for more details.
}
\section{Before, After and Alternatives}{

Before running the test you might first want to get an impression using a cross table using \code{\link{tab_cross}}, or a stacked bar chart using \code{\link{vi_bar_stacked_multiple}} for a visualisation.

After the test you might want an effect size measure, either \code{\link{es_common_language_is}} for the CLES, \code{\link{es_hodges_lehmann_is}} for Hodges-Lehmann, or \code{\link{r_rank_biserial_is}} for the (Glass) rank biserial (Cliff delta).
}

\examples{
dataFile = "https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv"
dfr <- read.csv(dataFile, sep=";", na.strings=c("", "NA"))
coding = c("Far too little", "too little", "Enough", "Too much", "Far too much")
ts_brunner_munzel_perm(dfr[['Gen_Gender']], dfr[['Mix_NrAct']], levels=coding)

}
\references{
Neubert, K., & Brunner, E. (2007). A studentized permutation test for the non-parametric Behrens–Fisher problem. \emph{Computational Statistics & Data Analysis, 51}(10), 5192–5204. https://doi.org/10.1016/j.csda.2006.05.024

Schüürhuis, S., Konietschke, F., & Brunner, E. (2025). A new approach to the Nonparametric Behrens-Fisher problem with compatible confidence intervals (arXiv:2504.01796). arXiv. https://doi.org/10.48550/arXiv.2504.01796
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
