% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cochran_q.R
\name{ts_cochran_q}
\alias{ts_cochran_q}
\title{Cochran Q Test}
\usage{
ts_cochran_q(data, success = NULL)
}
\arguments{
\item{data}{dataframe with the binary scores}

\item{success}{indicator for what is considered a success (default is first value found)}
}
\value{
A dataframe with:
\item{n}{the sample size}
\item{statistic}{the test statistic (chi-square value)}
\item{df}{the degrees of freedom}
\item{pValue}{the significance (p-value)}
}
\description{
A test for multiple binairy variables. The null hypothesis is that the proportion of successes is the same in all groups.

If the p-value (sig.) is below a certain threshold (usually .05) the assumption is rejected and at least one category has a significant different number of successes than at least one other group, in the population.

If the test is significant (below the threshold) a post-hoc Dunn test could be used, or pairwise McNemar-Bowker.
}
\details{
The formula used (Cochran, 1950, p. 259):
\deqn{Q = \frac{\left(k - 1\right)\times\sum_{j=1}^k\left(C_j - \bar{C}\right)^2}{k\times\sum_{i=1}^n R_i - \sum_{i=1}^n R_i^2}}
\deqn{sig. = 1 - \chi^2\left(Q, df\right)}
With:
\deqn{df = k - 1}

\emph{Symbols used:}
\itemize{
\item \eqn{C_j} the number of successes in category j
\item \eqn{k} the number of categories (factors)
\item \eqn{R_i} the number of succeses in case i
\item \eqn{n} the number of cases
}

\strong{Alternatives}

\emph{library(nonpar)}

matr = cbind(var1, var2, var3, var4)

cochrans.q(matr)

\emph{library(RVAideMemoire)}

myData.long<-reshape(dFr, varying=c("var1", "var2", "var3", "var4"), v.names="score", timevar="var", times=c("var1", "var2", "var3", "var4"),new.row.names = 1:1000, direction="long")

cochran.qtest(score~var |id, data=myData.long)
}
\references{
Cochran, W. G. (1950). The comparison of percentages in matched samples. \emph{Biometrika, 37}(3/4), 256–266. https://doi.org/10.2307/2332378
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
