% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cressie_read_ind.R
\name{ts_cressie_read_ind}
\alias{ts_cressie_read_ind}
\title{Cressie-Read Test of Independence / Power Divergence Test}
\usage{
ts_cressie_read_ind(nom1, nom2, cc = NULL, lambda = 2/3)
}
\arguments{
\item{nom1}{A vector with the data of the first variable}

\item{nom2}{A vector with the data of the second variable}

\item{cc}{c(NULL, "yates", "pearson", or "williams") Optional continuity correction (default is NULL)}

\item{lambda}{Optional value for lambda (default is 2/3)}
}
\value{
dataframe with test statistic, degrees of freedom, p-value, minimum expected count, proportion of expected counts below 5, and test used
}
\description{
Cressie-Read Test of Independence / Power Divergence Test
}
\details{
The formula used is (Cressie & Read, 1984, p. 442):
\deqn{\chi_{C}^{2} = \begin{cases} 2\times\sum_{i=1}^{r}\sum_{j=1}^c\left(F_{i,j}\times ln\left(\frac{F_{i,j}}{E_{i,j}}\right)\right) & \text{ if } \lambda=0 \\ 2\times\sum_{i=1}^{r}\sum_{j=1}^c\left(E_{i,j}\times ln\left(\frac{E_{i,j}}{F_{i,j}}\right)\right) & \text{ if } \lambda=-1 \\ \frac{2}{\lambda\times\left(\lambda + 1\right)} \times \sum_{i=1}^{r}\sum_{j=1}^{c} F_{i,j}\times\left(\left(\frac{F_{i,j}}{E_{i,j}}\right)^{\lambda} - 1\right) & \text{ else } \end{cases}}
\deqn{df = \left(r - 1\right)\times\left(c - 1\right)}
\deqn{sig. = 1 - \chi^2\left(\chi_{C}^{2},df\right)}

With:
\deqn{n = \sum_{i=1}^r \sum_{j=1}^c F_{i,j}}
\deqn{E_{i,j} = \frac{R_i\times C_j}{n}}
\deqn{R_i = \sum_{j=1}^c F_{i,j}}
\deqn{C_j = \sum_{i=1}^r F_{i,j}}

\emph{Symbols used:}
\itemize{
\item \eqn{r} the number of categories in the first variable (the number of rows)
\item \eqn{c} the number of categories in the second variable (the number of columns)
\item \eqn{F_{i,j}} the observed count in row i and column j
\item \eqn{E_{i,j}} the expected count in row i and column j
\item \eqn{R_i} the i-th row total
\item \eqn{C_j} the j-th column total
\item \eqn{n} the sum of all counts
\item \eqn{\chi^2\left(\dots\right)}	the chi-square cumulative density function
}

Cressie and Read (1984, p. 463) suggest to use \eqn{\lambda = \frac{2}{3}},  which
is therefor the default in this function.

The Yates correction (yates) is calculated using (Yates, 1934, p. 222):

Use instead of \eqn{F_{i,j}} the adjusted version defined by:
\deqn{F_{i,j}^\ast = \begin{cases} F_{i,j} - 0.5 & \text{ if } F_{i,j}>E_{i,j}  \\ F_{i,j} & \text{ if } F_{i,j}= E_{i,j}\\ F_{i,j} + 0.5 & \text{ if } F_{i,j}<E_{i,j} \end{cases}}

The Pearson correction (pearson) is calculated using (E.S. Pearson, 1947, p. 157):
\deqn{\chi_{PP}^2 = \chi_{P}^{2}\times\frac{n - 1}{n}}

The Williams correction (williams) is calculated using (Williams, 1976, p. 36):
\deqn{\chi_{PW}^2 = \frac{\chi_{P}^2}{q}}
With:
\deqn{q = 1 + \frac{\left(n\times\left(\sum_{i=1}^r \frac{1}{R_i}\right)-1\right) \times \left(n\times\left(\sum_{j=1}^c \frac{1}{C_j}\right)-1\right)}{6\times n\times df}}
}
\examples{
 
nom1 <- c("female", "female","female","female","female","female","female","female", 
"female","female","female", "male", "male", "male", "male", "male", "male", "male", 
"male", "male", "male", "male", "male", "male", "male", "male", "male","male", 
"male", "male", "male", "male", "male", "male", "male", "male", "male", "male", 
"male", "male", "male", "male")
nom2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", 
"other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl",
"nl","nl","nl","nl","nl","nl","nl","other", "other", "other", 
"other", "other", "other", "other", "other", "other", "other", 
"other", "other", "other", "other", "other")
ts_cressie_read_ind(nom1, nom2)
ts_cressie_read_ind(nom1, nom2, cc="yates")
ts_cressie_read_ind(nom1, nom2, cc="pearson")
ts_cressie_read_ind(nom1, nom2, cc="williams")
 
}
\references{
Cressie, N., & Read, T. R. C. (1984). Multinomial goodness-of-fit tests. \emph{Journal of the Royal Statistical Society: Series B (Methodological), 46}(3), 440-464. https://doi.org/10.1111/j.2517-6161.1984.tb01318.x

Pearson, E. S. (1947). The choice of statistical tests illustrated on the Interpretation of data classed in a 2 x 2 table. \emph{Biometrika, 34}(1/2), 139-167. https://doi.org/10.2307/2332518

Williams, D. A. (1976). Improved likelihood ratio tests for complete contingency tables. \emph{Biometrika, 63}(1), 33-37. https://doi.org/10.2307/2335081

Yates, F. (1934). Contingency tables involving small numbers and the chi square test. \emph{Supplement to the Journal of the Royal Statistical Society, 1}(2), 217-235. https://doi.org/10.2307/2983604
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
