% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fisher_freeman_halton.R
\name{ts_fisher_freeman_halton}
\alias{ts_fisher_freeman_halton}
\title{Fisher-Freeman-Halton Exact test}
\usage{
ts_fisher_freeman_halton(var1, var2)
}
\arguments{
\item{var1}{A vector with the data from the first variable}

\item{var2}{A vector with the data from the second variable}
}
\value{
the two-tailed p-value/sig.
}
\description{
Fisher-Freeman-Halton Exact test
}
\details{
This simply uses R's \emph{fisher.test()} function from the stats library.
}
\examples{
var1 <- c("female", "female","female","female","female","female","female","female",
          "female","female","female", "male", "male", "male", "male", "male", "male",
          "male", "male", "male", "male", "male", "male", "male", "male", "male", "male",
          "male", "male", "male", "male", "male", "male", "male", "male", "male", "male",
          "male", "male", "male", "male", "male")
var2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other",
          "nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl",
          "other", "other", "other", "other", "other", "other", "other", "other", "other",
          "other", "other", "other", "other", "other", "other")
ts_fisher(var1, var2)

}
\references{
Fisher, R. A. (1922). On the Interpretation of \eqn{\chi^2} from Contingency Tables, and the Calculation of P. \emph{Journal of the Royal Statistical Society, 85}(1), 87-94. https://doi.org/10.2307/2340521

Freeman, G. H., & Halton, J. H. (1951). Note on an exact treatment of contingency, goodness of fit and other problems of significance. \emph{Biometrika, 38}(1/2), 141-149. https://doi.org/10.2307/2332323
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
