% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_friedman.R
\name{ts_friedman}
\alias{ts_friedman}
\title{Friedman Test}
\usage{
ts_friedman(data, levels = NULL, ties = TRUE, dist = "chi")
}
\arguments{
\item{data}{dataframe. A column for each variable}

\item{levels}{vector, optional. Indication of what the levels are in order}

\item{ties}{boolean, optional. Apply a ties correction. Default is True}

\item{dist}{string, optional. Distribution to use. Either "chi" (default), "f", "normal"}
}
\value{
res : dataframe with the following columns
\item{n}{sample size}
\item{statistic}{test statistic used}
\item{df, df1, df2}{degrees of freedom (if applicable)}
\item{p-value}{the p-value (significance)}
}
\description{
A test to determine if any of the variables has a significant different average ranking than any of the others.

It is a paired-samples version of a Kruskal-Wallis test. If the p-value is below a pre-defined threshold (usually 0.05) it indicates at least one variable (column) is different than another.
}
\details{
The formula used in case of no ties (Friedman, 1937, p. 679):
\deqn{\chi_F^2 = \left(\frac{12}{n\times k\times\left(k+1\right)}\times\sum_{j=1}^k R_j^2\right)-3\times n\times\left(k+1\right)}
\deqn{df = k - 1}

With:
\deqn{R_j = \sum_{i=1}^n r_{i,j}}

In case a ties correction is used (Hollander & Wolfe, 1999, p. 274):
\deqn{\chi_{Fadj}^2 = \frac{12\times \sum_{j=1}^k R_j^2 - 3\times n^2\times\left(k +1\right)^2}{n\times\left(k+1\right) - \frac{\left(\sum t_{i,j}^3\right)-k}{k-1}}}

The ties correction used by IBM SPSS (2021, p. 811) will give the same result:
\deqn{\chi_{Fadj}^2 = \frac{\chi_F^2}{1 - \frac{\sum t_{i,j}^3 - t_{i,j}}{n\times\left(k^3-k\right)}}}

The function uses more of a one-way ANOVA approach in case of ties, but then on the ranks. It leads to the same result:
\deqn{\chi_{Fadj}^2 = \frac{n\times\sum_{j=1}^k\left(\bar{r}_j -\bar{r}\right)^2}{\left(\frac{\sum_{j=1}^k\sum_{i=1}^n \left(r_{i,j}-\bar{r}\right)^2}{n\times\left(k-1\right)}\right)}}

With:
\deqn{\bar{r}_j = \frac{R_j}{n}}
\deqn{\bar{r} = \frac{\sum_{j=1}^k R_j}{n\times k} = \frac{n\times\left(k+1\right)}{2}}

The significance is then determined using:
\deqn{sig. = 1 - \chi^2\left(\chi_F^2, df\right)}

A normal distribution approximation was proposed by Friedman (1937, p. 695; 1939, p. 109):
\deqn{z_F = \frac{\chi_F^2-\left(k-1\right)}{\sqrt{2\times\frac{n-1}{n}\times\left(k-1\right)}}}
\deqn{sig. = 2\times\left(1 - \Phi\left(\left|z_F\right|\right)\right)}

And an F distribution by Iman and Davenport (1980, p. 573):
\deqn{F_F = \frac{\left(n-1\right)\times\chi_F^2}{n\times\left(k-1\right)-\chi_F^2}}
\deqn{df_1 = k - 1}
\deqn{df_2 = \left(k-1\right)\times\left(n - 1\right)}
\deqn{sig. = 1 - F\left(F_{F}, df_1, df_2\right)}

Some might refer to Conover for this F-distribution, but in Conover (1980, p. 300) it seems Conover credits Iman and Davenport himself.

\emph{Symbols Used}

\itemize{
\item \eqn{n}, the number of cases
\item \eqn{k}, the number of variables
\item \eqn{r_{i,j}}, the rank of case i, in variable j. The ranks are determined for each case.
\item \eqn{t_{i,j}}, the frequency of unique rank j, in case i. For each row the frequencies of each rank is determined in the calculations.
}
}
\references{
Conover, W. J. (1980). \emph{Practical nonparametric statistics} (2nd ed.). Wiley.

Friedman, M. (1937). The use of ranks to avoid the assumption of normality implicit in the analysis of variance. \emph{Journal of the American Statistical Association, 32}(200), 675-701. doi:10.2307/2279372

Friedman, M. (1939). A correction. \emph{Journal of the American Statistical Association, 34}(205), 109-109. doi:10.1080/01621459.1939.10502372

Hollander, M., & Wolfe, D. A. (1999). \emph{Nonparametric statistical methods} (2nd ed.). Wiley.

IBM. (2021). IBM SPSS Statistics Algorithms. IBM.

Iman, R., & Davenport, J. (1980). Approximations of the critical region of the Friedman statistic. \emph{Communications in Statistics-Theory and Methods, 9}, 571-595.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
