% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_kruskal_wallis.R
\name{ts_kruskal_wallis}
\alias{ts_kruskal_wallis}
\title{Kruskal-Wallis H Test}
\usage{
ts_kruskal_wallis(
  scores,
  groups,
  tiescorr = c(TRUE, FALSE),
  approx = c("chi2", "kw-gamma", "kw-gamma-chi2", "kw-beta", "kw-beta-f",
    "wallace-I-beta", "wallace-II-beta", "wallace-III-beta", "wallace-I-f",
    "wallace-II-f", "wallace-III-f", "iman")
)
}
\arguments{
\item{scores}{the variable with the scores}

\item{groups}{the variable with the groups}

\item{tiescorr}{boolean to indicate the use of a ties correction}

\item{approx}{approximation to use (see details)}
}
\value{
Returns a dataframe with:
\item{n}{the sample size}
\item{H}{the H value}
\item{pValue}{the two-tailed significance, a.k.a. p-value}

Depending on the test used additional items might be added.
In case of a chi-square approximation
\item{statistic}{if not the same as H, the chi-square value used}
\item{df}{the degrees of freedom}

In case of a gamma, or beta approximation
\item{statistic}{the test statistic used}
\item{alpha}{the alpha value used}
\item{beta}{the beta value used}

In case of a F approximation
\item{statistic}{the test statistic used}
\item{df1}{the first degrees of freedom}
\item{df2}{the second degrees of freedom}
}
\description{
Kruskal-Wallis H Test
}
\details{
The formula used in case of no ties correction is (Kruskal & Wallis, 1952, p. 586):
\deqn{H = \frac{12}{n\times\left(n + 1\right)}\times\sum_{i = 1}^c \frac{R_i^2}{n_i} - 3\times\left(n + 1\right)}
With:
\deqn{R_i \sum_{j=1}^{n_i} r_{i,j}}

\emph{Symbols used}
\itemize{
\item \eqn{r_{i,j}} is the rank of the i-th score in category j
\item \eqn{k} is the number of categories
\item \eqn{n_i} is the number of ranks in the i-th category
\item \eqn{n} is the total number of scores
}

Applying a ties correction will adjust the formula to (Kruskal & Wallis, 1952, p. 586):
\deqn{H_{adj} = \frac{H}{1 - \frac{\sum T}{n^3 - n}}}
With:
\deqn{T_i = t_i^3 - t_i}

\emph{Symbols added}
\emph{Symbols used}
\itemize{
\item \eqn{t_i} is how often the i-th unique rank occurs
}

I have not been able to find an exact distribution for H in R. A good starting point
might be Choi et al. (2003) and let me know if you manage.

As for approximations, there are quite a few....

Kruskal and Wallis (1952) had a few of their own.
The most frequently used one (Kruskal & Wallis, 1952, p. 586):
\deqn{sig. \approx 1 - \chi\left(H, df\right)}
With:
\deqn{df = k - 1}
This is the one used when \emph{approx="chi2"}

Kruskal and Wallis (1952) also propose a incomplete gamma distribution approximation (p. 609):
\deqn{sig. \approx 1 - \Gamma\left(H, \alpha, \beta\right)}
With:
\deqn{\alpha = \frac{\mu^2}{\sigma^2}}
\deqn{\beta = \frac{\sigma^2}{\mu}}
\deqn{\mu = k - 1}
\deqn{\sigma^2 = 2\times\left(k - 1\right) - \frac{2\times\left(3\times k^2 - 6\times k + n\times\left(2\times k^2 - 6\times k + 1\right)\right)}{5\times n\left(n + 1\right)} - \frac{6}{5}\times \sum_{i=1}^k \frac{1}{n_i}}
This is the one used when \emph{approx="kw-gamma"}

And a chi-square approximation of this gamma approximation with:
\deqn{sig. \approx 1 - \chi\left(\chi_a^2, df\right)}
\deqn{\chi_a^2 = 2\times\frac{\mu}{\sigma^2}\times H}
\deqn{df = 2\times\frac{\mu^2}{\sigma^2}\times H}
This is the one used when \emph{approx="kw-gamma-chi2"}

They also provided an incomplete beta distribution approximation with (Kruskal & Wallis, 1952, pp. 609-610):
\deqn{sig. \approx 1 - B\left(\frac{H}{M}, \alpha, \beta \right)}
\deqn{M = \frac{n^3 - \sum_{i=1}^k n_i^3}{n\times\left(n + 1\right)}}
\deqn{\alpha = \frac{df_1}{2}}
\deqn{\beta = \frac{df_2}{2}}
\deqn{df_1 = \mu\times\frac{\mu\times\left(M - \mu\right) - \sigma^2}{\frac{1}{2}\times M\times\sigma^2}}
\deqn{df_2 = df_1\times\frac{M - \mu}{\mu}}
This is the one used when \emph{approx="kw-beta"}

And a F-distribution approximation of this using:
\deqn{sig. \approx 1 - F\left(F_H, df_1, df_2\right)}
\deqn{F_H = \frac{H\times\left(M - \mu\right)}{\mu\times\left(M - H\right)}}
This is the one used when \emph{approx="kw-beta-f"}

Wallace (1959) used a few different approximations using the F and Beta distributions.
First the test values are adjuste (Wallace, 1959, p. 226)
\deqn{B_2 = \frac{H}{n - 1}}
\deqn{F_2 = \frac{\left(n - k\right)\times H}{\left(k - 1\right)\times\left(n - 1 - H\right)}}
Wallace then has three different approximations.
First we define d for each (Wallace, 1959, pp. 226-227):
\deqn{d_{I} = \frac{\left(n - k\right)\times\left(k - 1\right) - \sigma^2}{\frac{1}{2}\times\left(n - 1\right)\times\sigma^2}}
\deqn{d_{II} = 1 - \frac{6\times\left(n + 1\right)}{5\times\left(n - 1\right)\times\left(n + 1.2\right)}}
\deqn{d_{III} = 1}

Then the degrees of freedom for the F-tests will be:
\deqn{df_1^i = \left(k - 1\right)\times d_i}
\deqn{df_2^i = \left(n - k\right)\times d_i}

And the test itself:
\deqn{sig. = 1 - F\left(F_2, df_1^i, df_2^i\right)}
This is used when \emph{approx="wallace-I-f"}, \emph{approx="wallace-II-f"}, or \emph{approx="wallace-III-f"}

For the beta approximations the same as before:
\deqn{\alpha_i = \frac{df_1^i}{2}}
\deqn{\beta_i = \frac{df_2^i}{2}}
\deqn{sig. \approx 1 - B\left(B_2, \alpha_i, \beta_i \right)}
This is used when \emph{approx="wallace-I-beta"}, \emph{approx="wallace-II-beta"}, or \emph{approx="wallace-III-beta"}

Finally, Iman and Davenport (1976) describe an approximation from Sattertwaite as:
\deqn{sig. = 1 - F\left(F_2, df_1, df_2\right)}
With:
\deqn{df_1 = k - 1}
\deqn{df_2 = \frac{\left(\sum_{i=1}^k \left(n_i -1\right)\times v_i\right)^2}{\sum_{i=1}^k \frac{\left(\left(n_i - 1\right)\times v_i\right)^2}{n_i - 1}}}
\deqn{v_i = \frac{\sum_{j=1}^{n_i} \left(r_{i,j} - \bar{r}_i\right)^2}{n_i -1}}
\deqn{\bar{r}_i = \frac{\sum_{j = 1}^{n_i} r_{i,j}}{n_i}}
This is the one used when \emph{approx="iman"}
}
\examples{
scores = c(1, 2, 5, 1, 1, 5, 3, 1, 5, 1, 1, 5, 1, 1, 3, 3, NA, 3, 4, 2, 4, 2, 1, 3, 2, 2, 4, 1, 1, 3, 1, 2, 4, 1, 5, 4, 2, 3, 4, 1, 2, 5, 1, 1, 3, 3, 3, 1, 4, 3, 1, 1, 2, 3, 1)
groups = c("Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Haarlem", "Diemen", "Haarlem", "Haarlem", "Haarlem", "Haarlem", "Haarlem")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="chi2")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="kw-gamma")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="kw-gamma-chi2")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="kw-beta")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="kw-beta-f")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="wallace-I-beta")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="wallace-II-beta")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="wallace-III-beta")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="wallace-I-f")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="wallace-II-f")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="wallace-III-f")
ts_kruskal_wallis(scores, groups, tiescorr=TRUE, approx="iman")

}
\references{
Choi, W., Lee, J. W., Huh, M.-H., & Kang, S.-H. (2003). An algorithm for computing the exact distribution of the Kruskal–Wallis test. \emph{Communications in Statistics - Simulation and Computation, 32}(4), 1029–1040. https://doi.org/10.1081/SAC-120023876

Iman, R. L., & Davenport, J. M. (1976). New approximations to the exact distribution of the kruskal-wallis test statistic. \emph{Communications in Statistics - Theory and Methods, 5}(14), 1335–1348. https://doi.org/10.1080/03610927608827446

Kruskal, W. H., & Wallis, W. A. (1952). Use of ranks in one-criterion variance analysis. \emph{Journal of the American Statistical Association, 47}(260), 583–621. https://doi.org/10.1080/01621459.1952.10483441

Wallace, D. L. (1959). Simplified beta-approximations to the Kruskal-Wallis H test. \emph{Journal of the American Statistical Association, 54}(285), 225. https://doi.org/10.2307/2282148
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
