% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mann_whitney.R
\name{ts_mann_whitney}
\alias{ts_mann_whitney}
\title{Mann-Whitney U Test}
\usage{
ts_mann_whitney(dataVar, groupVar, method = "exact", corr = TRUE)
}
\arguments{
\item{method}{c("exact", "appr") exact method or normal approximation}

\item{corr}{boolean to indicate the use of a continuity correction}

\item{var1}{A vector with the scores data}

\item{var2}{A vector with the group data}
}
\value{
dataframe with U, the test statistic, p-value, and the test used
}
\description{
Mann-Whitney U Test
}
\details{
The formula used is (Mann & Whitney, 1947, p. 51):
\deqn{U_i = R_i - \frac{n_i\times\left(n_i + 1\right)}{2}}
With:
\deqn{R_i = \sum_{j=1}^{n_i} r_{i,j}}

For an approximation the following is used:
\deqn{sig. = 2\times\left(1 - Z\left(z\right)\right)}
With:
\deqn{z = \frac{U_i - \frac{n_1\times n_2}{2}}{SE}}
\deqn{SE = \sqrt{\frac{n_1\times n_2}{n\times\left(n - 1\right)}\times\left(\frac{n^3 - n}{12} - \sum_i T_i\right)}}
\deqn{T_i = \frac{t_i^3 - t_i}{12}}
\deqn{n = n_1 + n_2}

If a continuity correction is used the z-value is calculated using:
\deqn{z_{cc} = z - \frac{0.5}{SE}}

\emph{Symbols used:}
\itemize{
\item \eqn{n_i} the sample size of category i
\item \eqn{n} the total sample size
\item \eqn{r_{i,j}} the j-th rank of category i
}

The ties correction (\eqn{T}) can be found in Lehmann and D'Abrera (1975, p. 20)
}
\examples{
scores = c(5, 12, 3, 4, 6, 1, 11, 13, NA)
groups = c("A","A","A","B","B","B","B", NA, "C")
ts_mann_whitney(scores, groups)

}
\references{
Lehmann, E. L., & D’Abrera, H. J. M. (1975). \emph{Nonparametrics: Statistical methods based on ranks}. Holden-Day.

Mann, H. B., & Whitney, D. R. (1947). On a Test of Whether one of Two Random Variables is Stochastically Larger than the Other. \emph{The Annals of Mathematical Statistics, 18}(1), 50–60. https://doi.org/10.1214/aoms/1177730491
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
