% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mcnemar_bowker.R
\name{ts_mcnemar_bowker}
\alias{ts_mcnemar_bowker}
\title{(McNemar-)Bowker Test}
\usage{
ts_mcnemar_bowker(field1, field2, categories = NULL, cc = FALSE)
}
\arguments{
\item{field1}{vector, the first categorical field}

\item{field2}{vector, the first categorical field}

\item{categories}{vector, optional, order and/or selection for categories of field1 and field2}

\item{cc}{boolean, optional, use of a continuity correction (default is False)}
}
\value{
Dataframe with:
\item{n}{the sample size}
\item{statistic}{the chi-squared value}
\item{df}{the degrees of freedom used in the test}
\item{p-value}{the significance (p-value)}
}
\description{
The Bowker test (Bowker, 1948) is an extension of the McNemar (1947) test, which was only for 2x2 tables.

It tests if there is a change in symmetric opinion changes. It assumes there is no change, and if the p-value is below a pre-set threshold (usually 0.05) this assumption is rejected.
}
\details{
The formula used is (Bowker, 1948, p. 573):
\deqn{\chi_{B}^2 = \sum_{i=1}^{r-1}\sum_{j=i+1}^c \frac{\left(F_{i,j} - F_{j,i}\right)^2}{F_{i,j} + F_{j,i}}}
\deqn{df = \frac{r \times \left(r - 1\right)}{2} = \frac{c \times \left(c - 1\right)}{2}}
\deqn{sig. = 1 - \chi^2\left(\chi_{B}\right)}

\emph{Symbols used}
\itemize{
\item \eqn{r} is the number of rows (categories in the first variable)
\item \eqn{c} is the number of columns (categories in the second variable)
\item \eqn{n} is the total number of scores
\item \eqn{F_{i,j}} is the frequency (count) of scores equal to the i-th category in the first variable, and the j-th category in the second.
}
}
\references{
Bowker, A. H. (1948). A test for symmetry in contingency tables. \emph{Journal of the American Statistical Association, 43}(244), 572-574. doi:10.2307/2280710

McNemar, Q. (1947). Note on the sampling error of the difference between correlated proportions or percentages. \emph{Psychometrika, 12}(2), 153-157. doi:10.1007/BF02295996
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
