% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mood_median.R
\name{ts_mood_median}
\alias{ts_mood_median}
\title{Mood Median Test}
\usage{
ts_mood_median(scores, groups, cc = c(NULL, "yates", "pearson", "williams"))
}
\arguments{
\item{scores}{the variable with the scores}

\item{groups}{the variable with the groups}

\item{cc}{optional continuity correction to use}
}
\value{
Returns a dataframe with:statistic, df, pValue, minExp, propBelow5
\item{statistic}{the test statistic}
\item{df}{the degrees of freedom}
\item{pValue}{the two-tailed significance, a.k.a. p-value}
\item{minExp}{the minimum expected count}
\item{propBelow5}{the proportion of cells with expected count below 5}
}
\description{
Mood Median Test
}
\details{
This test simply counts how many scores fall above the median in each category and
how many equal or below, and then perform a Pearson chi-square test on the table

The formula used is:
\deqn{\chi_M^2 = \sum_{i=1}^2 \sum_{j=1}^k \frac{\left(F_{i,j} - E_{i,j}\right)^2}{E_{i,j}}}
\deqn{df = k - 1}
\deqn{sig. = 1 - \chi^2\left(\chi_M^2, df\right)}
With:
\deqn{E_{i,j} = \frac{R_i\times C_j}{n}}
\deqn{R_i = \sum_{j=1}^k F_{i,j}}
\deqn{C_i = \sum_{i=1}^2 F_{i,j}}
\deqn{n = \sum_{i=1}^2 \sum_{j=1}^k F_{i,j}}

\emph{Symbols used}
\itemize{
\item \eqn{F_{1,j}} the number of scores is category j that are above the overall median
\item \eqn{F_{2,j}} the number of scores is category j that are equal or below the overall median
\item \eqn{E_{i,j}} the expected count for row i, column j
\item \eqn{R_i} the row total of row i
\item \eqn{C_i} the column total of column i
\item \eqn{n} the total sample size
\item \eqn{k} the number of categories
\item \eqn{\chi^2\left(\dots, \dots\right)} the cumulative distribution function of the chi-square distribution
}

The Yates correction (yates) is calculated using (Yates, 1934, p. 222):

Use instead of \eqn{F_{i,j}} the adjusted version defined by:
\deqn{F_{i,j}^\ast = \begin{cases} F_{i,j} - 0.5 & \text{ if } F_{i,j}>E_{i,j}  \\ F_{i,j} & \text{ if } F_{i,j}= E_{i,j}\\ F_{i,j} + 0.5 & \text{ if } F_{i,j}<E_{i,j} \end{cases}}

The Pearson correction (pearson) is calculated using (E.S. Pearson, 1947, p. 157):
\deqn{\chi_{PP}^2 = \chi_{P}^{2}\times\frac{n - 1}{n}}

The Williams correction (williams) is calculated using (Williams, 1976, p. 36):
\deqn{\chi_{PW}^2 = \frac{\chi_{P}^2}{q}}
With:
\deqn{q = 1 + \frac{\left(n\times\left(\sum_{i=1}^r \frac{1}{R_i}\right)-1\right) \times \left(n\times\left(\sum_{j=1}^c \frac{1}{C_j}\right)-1\right)}{6\times n\times df}}

Often Mood (1950) is used as a source, but I found the explanation clearer in Brown and Mood (1951)
}
\examples{
scores = c(1, 2, 5, 1, 1, 5, 3, 1, 5, 1, 1, 5, 1, 1, 3, 3, NA, 3, 4, 2, 4, 2, 1, 3, 2, 2, 4, 1, 1, 3, 1, 2, 4, 1, 5, 4, 2, 3, 4, 1, 2, 5, 1, 1, 3, 3, 3, 1, 4, 3, 1, 1, 2, 3, 1)
groups = c("Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Haarlem", "Diemen", "Haarlem", "Haarlem", "Haarlem", "Haarlem", "Haarlem")
ts_mood_median(scores, groups)
ts_mood_median(scores, groups, cc="yates")
ts_mood_median(scores, groups, cc="pearson")
ts_mood_median(scores, groups, cc="williams")

}
\references{
Brown, G. W., & Mood, A. M. (1951). On median tests for linear hypotheses. \emph{Proceedings of the Second Berkeley Symposium on Mathematical Statistics and Probability}, 2, 159–167.

Mood, A. M. (1950). \emph{Introduction to the theory of statistics}. McGraw-Hill.

Pearson, E. S. (1947). The choice of statistical tests illustrated on the Interpretation of data classed in a 2 × 2 table. \emph{Biometrika, 34}(1/2), 139–167. https://doi.org/10.2307/2332518

Williams, D. A. (1976). Improved likelihood ratio tests for complete contingency tables. \emph{Biometrika, 63}(1), 33–37. https://doi.org/10.2307/2335081

Yates, F. (1934). Contingency tables involving small numbers and the chi square test. \emph{Supplement to the Journal of the Royal Statistical Society, 1}(2), 217–235. https://doi.org/10.2307/2983604
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
