% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multinomial_gof.R
\name{ts_multinomial_gof}
\alias{ts_multinomial_gof}
\title{Exact Multinomial Test of Goodness-of-Fit}
\usage{
ts_multinomial_gof(data, expCount = NULL)
}
\arguments{
\item{data}{A vector with the data}

\item{expCount}{Optional dataframe with the categories and expected counts}
}
\value{
dataframe with probability of the observed data, number of combinations used, p-value and test used
}
\description{
Exact Multinomial Test of Goodness-of-Fit
}
\details{
The exact multinomial test of goodness of fit is done in four steps

Step 1: Determine the probability of the observed counts using
the probability mass function of the multinomial distribution

Step 2: Determine all possible permutations with repetition
that create a sum equal to the sample size over the k-categories.

Step 3: Determine the probability of each of these permutations using the probability mass function of the multinomial distribution.

Step 4: Sum all probabilities found in step 3 that are equal or less than the one found in step 1.

\strong{Alternatives}

The \emph{EMT} library has a similar function: \emph{multinomial.test()}

The \emph{XNomial} library has a similar function: \emph{xmulti()}
}
\examples{
data <- c("MARRIED", "DIVORCED", "MARRIED", "SEPARATED", "DIVORCED", "NEVER MARRIED", "DIVORCED", "DIVORCED", "NEVER MARRIED", "MARRIED", "MARRIED", "MARRIED", "SEPARATED", "DIVORCED", "NEVER MARRIED", "NEVER MARRIED", "DIVORCED", "DIVORCED", "MARRIED")
eCounts = data.frame(c("MARRIED", "DIVORCED", "NEVER MARRIED", "SEPARATED"), c(5,5,5,5))
ts_multinomial_gof(data)
ts_multinomial_gof(data, eCounts)

}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
