% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scott_smith_owa.R
\name{ts_scott_smith_owa}
\alias{ts_scott_smith_owa}
\title{Scott-Smith Test}
\usage{
ts_scott_smith_owa(nomField, scaleField, categories = NULL)
}
\arguments{
\item{nomField}{the groups variable}

\item{scaleField}{the numeric scores variable}

\item{categories}{vector, optional. the categories to use from catField}
}
\value{
A dataframe with:
\item{n}{the sample size}
\item{k}{the number of categories}
\item{statistic}{the test statistic (chi-square value)}
\item{df}{the degrees of freedom}
\item{pValue}{the significance (p-value)}
}
\description{
Tests if the means (averages) of each category could be the same in the population.

If the p-value is below a pre-defined threshold (usually 0.05), the null hypothesis is rejected, and there are then at least two categories who will have a different mean on the scaleField score in the population.

Yiğit and Gökpina (2010, p. 32) concluded that this test is inferior to some other alternatives when there is heteroscedasticity (variances in the groups not the same) are preferred (for example the Welch one-way ANOVA).

There are quite some alternatives for this, the stikpet library has Fisher, Welch, James, Box, Scott-Smith, Brown-Forsythe, Alexander-Govern, Mehrotra modified Brown-Forsythe, Hartung-Agac-Makabi, Özdemir-Kurt and Wilcox as options. See the notes from ts_fisher_owa() for some discussion on the differences.
}
\details{
The formula used (Scott & Smith, 1971, p. 277):
\deqn{\chi_{SS}^2 = \sum_{j=1}^k d_j^2}
\deqn{df = k}
\deqn{sig. = 1 - \chi^2\left(\chi_{SS}^2, df\right)}
With:
\deqn{d_j = t_j\times\sqrt{\frac{n_j - 3}{n_j - 1}}}
\deqn{t_j = \frac{\left(\bar{x}_j - \bar{x}\right)\times\sqrt{n_j}}{s_j}}
\deqn{s_j^2 = \frac{\sum_{i=1}^{n_j} \left(x_{i,j} - \bar{x}_j\right)^2}{n_j - 1}}
\deqn{\bar{x}_j = \frac{\sum_{i=1}^{n_j} x_{i,j}}{n_j}}
\deqn{\bar{x} = \frac{\sum_{j=1}^k n_j\times\bar{x}_j}{n}}
\deqn{n = \sum_{j=1}^k n_j}

\emph{Symbols:}
\itemize{
\item \eqn{x_{i,j}} the i-th score in category j
\item \eqn{k} the number of categories
\item \eqn{n_j} the sample size of category j
\item \eqn{x_j} the sample mean of category j
\item \eqn{s_j^2} the sample variance of the scores in category j
\item \eqn{df} the degrees of freedom
\item \eqn{\chi^2\left(\dots,\dots\right)} the cumulative distribution function of the chi-square distribution.
}

I couldn't find the chi-square test itself in the original article, but the calculation for an independent samples test repeated usually indeed leads to a chi-square distribution. The formula and chi-square distribution can also be found in Adepoju et al. (2016, p. 64), Cavus and Yazici (2020, p. 7) and Yiğit and Gökpina (2010, p. 17)
}
\references{
Adepoju, K. A., Shittu, O. I., & Chukwu, A. U. (2016). On the development of an exponentiated F test for one-way ANOVA in the presence of outlier(s). \emph{Mathematics and Statistics, 4}(2), 62–69. https://doi.org/10.13189/ms.2016.040203

Cavus, M., & Yazıcı, B. (2020). Testing the equality of normal distributed and independent groups’ means under unequal variances by doex package. \emph{The R Journal, 12}(2), 134. https://doi.org/10.32614/RJ-2021-008

Scott, A. J., & Smith, T. M. F. (1971). Interval estimates for linear combinations of means. \emph{Applied Statistics, 20}(3), 276–285. https://doi.org/10.2307/2346757

Yiğit, E., & Gökpinar, F. (2010). A simulation study on tests for one-way ANOVA under the unequal variance assumption. \emph{Communications, Faculty Of Science, University of Ankara}, 15–34. https://doi.org/10.1501/Commua1_0000000660
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
