% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sign_ps.R
\name{ts_sign_ps}
\alias{ts_sign_ps}
\title{Paired Samples Sign Test}
\usage{
ts_sign_ps(field1, field2, levels = NULL, dmu = 0, method = "exact")
}
\arguments{
\item{field1}{the numeric scores of the first variable}

\item{field2}{the numeric scores of the second variable}

\item{levels}{vector, optional. the levels from field1 and field2}

\item{dmu}{float, optional. The difference according to the null hypothesis (default is 0)}

\item{method}{string, optional. Test to be used. Either "exact" (default), "appr".}
}
\value{
A dataframe with:
\item{n pos}{the number of scores with a positive difference}
\item{n neg}{the number of scores with a negative difference}
\item{statistic}{the test statistic (only applicable if method="appr")}
\item{p-Value}{the significance (p-value)}
}
\description{
This test compares the number of pairs that have a difference above the hypothesized difference, with those below the difference. It can be considered an alternative for the paired samples t-test.
}
\details{
If method="exact" the binomial distribution will be used. The formula used is (Dixon & Mood, 1946):
\deqn{sig. = 2\times \text{Bin}\left(n, \min\left(n_{pos}, n_{neg}\right), \frac{1}{2}\right)}

When using the approximation, the standard normal distribution is used (SPSS, 2006, p. 483):
\deqn{z = \frac{\max\left(n_{pos}, n_{neg}\right)-0.5\times\left(n_{pos} + n_{neg}\right)-0.5}{0.5\times\sqrt{n_{pos}+n_{neg}}}}
\deqn{sig. = 2\times\left(1 - \Phi\left(\left|z\right|\right)\right)}

With:
\deqn{n_{pos} = \sum_{i=1}^n \begin{cases} 1 & \text{ if } d_i>d_{H0} \ 0 & \text{ if } d_i \le d_{H0} \end{cases}}
\deqn{n_{neg} = \sum_{i=1}^n \begin{cases} 0 & \text{ if } d_i \ge d_{H0} \ 1 & \text{ if } d_i < d_{H0} \end{cases}}
\deqn{d_i = x_i - y_i}

\emph{Symbols used:}

\itemize{
\item \eqn{n} is the number of pairs with a difference unequal to zero
\item \eqn{n_{pos}} the number of pairs with a positive difference
\item \eqn{n_{neg}} the number of pairs with a negative difference
\item \eqn{d_{H0}} the difference according to the null hypothesis, usually 0
\item \eqn{x_i} the i-th score from the first variable
\item \eqn{y_i} the i-th score from the second variable
\item \eqn{\text{Bin}\left(\dots,\dots\right)} the cumulative probability mass function of the binomial distribution
}

The test was described by Arbuthnott (1710) but with more modern notation see Dixon and Mood (1946).

\strong{Alternatives}

\emph{library(DescTools)}

SignTest(ord1, ord2)

\emph{library(EnvStats)}

signTest(ord1, ord2, paired=TRUE)

\emph{library(BSDA)}

SIGN.test(ord1, ord2)
}
\references{
Arbuthnott, J. (1710). An argument for divine providence, taken from the constant regularity observed in the births of both sexes. \emph{Philosophical Transactions of the Royal Society of London, 27}(328), 186–190. doi:10.1098/rstl.1710.0011

Dixon, W. J., & Mood, A. M. (1946). The statistical sign test. \emph{Journal of the American Statistical Association, 41}(236), 557–566. doi:10.1080/01621459.1946.10501898

SPSS. (2006). SPSS 15.0 algorithms.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
