% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_student_t_os.R
\name{ts_student_t_os}
\alias{ts_student_t_os}
\title{One-Sample Student t-Test}
\usage{
ts_student_t_os(data, mu = NULL)
}
\arguments{
\item{data}{A vector with the data as numbers}

\item{mu}{optional hypothesized mean, otherwise the midrange will be used}
}
\value{
dataframe with the hypothesized mean, sample mean, test statistic, degrees of freedom, p-value (sig.) and name of test used
}
\description{
One-Sample Student t-Test
}
\details{
The formula used is:
\deqn{t = \frac{\bar{x} - \mu_{H_0}}{SE}}
\deqn{sig = 2\times\left(1 - T\left(\left|t\right|, df\right)\right)}

With:
\deqn{df = n - 1}
\deqn{SE = \frac{s}{\sqrt{n}}}
\deqn{s = \sqrt{\frac{\sum_{i=1}^n\left(x_i - \bar{x}\right)^2}{n - 1}}}
\deqn{\bar{x} = \frac{\sum_{i=1}^nx_i}{n}}

\emph{Symbols used:}
\itemize{
\item \eqn{T\left(\dots, \dots\right)} the cumulative distribution function of the t-distribution
\item \eqn{\bar{x}} the sample mean
\item \eqn{μ_{H_0}} the hypothesized mean in the population
\item \eqn{SE} the standard error (i.e. the standard deviation of the sampling distribution)
\item \eqn{df} the degrees of freedom
\item \eqn{n} the sample size (i.e. the number of scores)
\item \eqn{s} the unbiased sample standard deviation
\item \eqn{x_i} the i-th score
}

The Student t test (Student, 1908) was described by Gosset under the pseudo name Student.

\strong{Alternative}
R stats library has a similar function: \emph{t.test()}
}
\examples{
 
data <- c(1, 2, 5, 1, 1, 5, 3, 1, 5, 1, 1, 5, 1, 1, 3, 3, 3, 4, 2, 4)
ts_student_t_os(data)

}
\references{
Student. (1908). The probable error of a mean. \emph{Biometrika, 6}(1), 1–25. https://doi.org/10.1093/biomet/6.1.1
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
