% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_trimmed_mean_os.R
\name{ts_trimmed_mean_os}
\alias{ts_trimmed_mean_os}
\title{One-Sample (Yuen or Yuen-Welch) Trimmed Mean Test}
\usage{
ts_trimmed_mean_os(data, mu = NULL, trimProp = 0.1, se = "yuen")
}
\arguments{
\item{data}{A vector or dataframe}

\item{mu}{optional hypothesized trimmed mean, otherwise the midrange will be used}

\item{trimProp}{optional proportion to trim in total (half will be trimmed from each side)}

\item{se}{optional method to use to determine standard error. Either "yuen" (default) or "wilcox"}
}
\value{
A dataframe with:
\item{trim. mean}{the sample trimmed mean}
\item{mu}{hypothesized trimmed mean}
\item{SE}{the standard error}
\item{statistic}{test statistic}
\item{df}{degrees of freedom}
\item{p-value}{p-value (sig.)}
\item{test used}{test used}
}
\description{
A variation on a one-sample Student t-test where the data is first trimmed, and the Winsorized variance is used.

The assumption about the population for this test is that the mean in the population is equal to the provide mu value. The test will show the probability of the found test statistic, or more extreme, if this assumption would be true. If this is below a specific threshold (usually 0.05) the assumption is rejected.
}
\details{
The formula used is:
\deqn{\frac{\bar{x}_t - \mu_{H_0}}{SE}}
\deqn{sig = 2\times\left(1 - T\left(\left|t\right|, df\right)\right)}

With:
\deqn{\bar{x}_t = \frac{\sum_{i=g+1}^{n - g}y_i}{}}
\deqn{g = \lfloor n\times p_t\rfloor}
\deqn{m = n - 2\times g}
\deqn{SE = \sqrt{\frac{SSD_w}{m}\times\left(m - 1\right)}}
or:
\deqn{SE = \frac{\sqrt{SSD_w}}{\left(1 - 2\times p_t\right)\times\sqrt{n}}}
\deqn{SSD_w = g\times\left(y_{g+1} - \bar{x}_w\right)^2 + g\times\left(y_{n-g} - \bar{x}_w\right)^2 + \sum_{i=g+1}^{n - g} \left(y_i - \bar{x}_w\right)^2}
\deqn{\bar{x}_w = \frac{\bar{x}_t\times m + g\times\left(y_{g+1} + y_{n-g}\right)}{n}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_t} the trimmed mean of the scores
\item \eqn{x_w} The Winsorized mean
\item \eqn{SSD_w} the sum of squared deviations from the Winsorized mean
\item \eqn{m} the number of scores in the trimmed data set from category i
\item \eqn{y_i} the i-th score after the scores are sorted from low to high
\item \eqn{p} the proportion of trimming on each side, we can define
}

The test is often also referred to as a Yuen test, or Yuen-Welch test.

The standard error can either be calculated using the first SE, which for example can be found in
Tukey and McLaughlin (1963, p. 342), and seems similar to the independent samples version of this
test as proposed by Yuen (1974, p. 167)

The second version is used in the other libraries, and can be found in Wilcox (2012, p. 157),
or Peró-Cebollero and Guàrdia-Olmos (2013, p. 409).
}
\section{Alternatives}{

R stats library has a similar function: \emph{t.test()}

The \emph{DescTools} library has a similar function: \emph{YuenTTest()}

The \emph{PairedData} library has a similar functions: \emph{yuen.t.test()} and \emph{yuen1.test}

The \emph{WRS2} library has a similar function: \emph{yuen()}
}

\examples{
Example 1: Numeric dataframe
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
df2 = read.csv(file2, sep=';', na.strings=c("", "NA"))
#Example 1: Numeric dataframe
ex1 = df2['Gen_Age']
ts_trimmed_mean_os(ex1)
ts_trimmed_mean_os(ex1, mu=23, trimProp=0.15, se="wilcox")

#Example 2: Numeric list
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
ts_trimmed_mean_os(ex2, trimProp=0.05)

}
\references{
Peró-Cebollero, M., & Guàrdia-Olmos, J. (2013). The adequacy of different robust statistical tests in comparing two independent groups. \emph{Psicológica}, 34, 407–424.

Tukey, J. W., & McLaughlin, D. H. (1963). Less vulnerable confidence and significance procedures for location based on a single sample: Trimming/Winsorization 1. \emph{Sankhyā: The Indian Journal of Statistics, 25}(3), 331–352.

Wilcox, R. R. (2012). \emph{Introduction to robust estimation and hypothesis testing} (3rd ed.). Academic Press.

Yuen, K. K. (1974). The two-sample trimmed t for unequal population variances. \emph{Biometrika, 61}(1), 165–170. https://doi.org/10.1093/biomet/61.1.165
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
