% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_trimmed_mean_t_is.R
\name{ts_trimmed_mean_t_is}
\alias{ts_trimmed_mean_t_is}
\title{Independent Samples Trimmed/Yuen Mean Test}
\usage{
ts_trimmed_mean_t_is(
  catField,
  scaleField,
  categories = NULL,
  dmu = 0,
  trimProp = 0.1,
  se = c("yuen", "wilcox")
)
}
\arguments{
\item{catField}{A vector with the categorical data}

\item{scaleField}{A vector with the scores}

\item{categories}{Optional to indicate which two categories of catField to use, otherwise first two found will be used.}

\item{dmu}{Optional difference according to null hypothesis (default is 0)}

\item{trimProp}{Optional proportion to trim in total for each category. If for example set to 0.1 then 0.05 from each side for each category will be trimmed. Default is 0.1.}

\item{se}{Optional to indicate which standard error to use. Either "yuen" (default) or "yuen-dixon".}
}
\value{
A dataframe with:
\item{n cat. 1}{the sample size of the first category}
\item{n cat. 2}{the sample size of the second category}
\item{trim mean cat. 1}{the sample trimmed mean of the first category}
\item{trim mean cat. 2}{the sample trimmed mean of the second category}
\item{diff.}{difference between the two sample means}
\item{hyp. diff.}{hypothesized difference between the two population means}
\item{statistic}{the test statistic (t-value)}
\item{df}{degrees of freedom}
\item{pValue}{the significance (p-value)}
\item{test}{name of test used}
}
\description{
A test to compare two means. The null hypothesis would be that the means of each category are equal in the population.

There are four similar tests, with different assumptions.\tabular{lll}{
   test \tab equal variance \tab normality \cr
   Student \tab yes \tab yes \cr
   Welch \tab no \tab yes \cr
   Trimmed \tab yes \tab no \cr
   Yuen-Welch \tab no \tab no \cr
}


The Student and Welch are available as separate functions. The Trimmed Means and Yuen-Welch test are available in this one.
}
\details{
\strong{YUEN}

The default \code{se="yuen"} will perform a Yuen-Welch test.

The formula used is (Yuen, 1974, p. 167):
\deqn{t = \frac{\bar{x}_{t,1} - \bar{x}_{t,2}}{SE}}
\deqn{sig = 2\times\left(1 - T\left(\left|t\right|, df\right)\right)}

With:
\deqn{SE = \sqrt{\frac{s_{w,1}^2}{m_1} + \frac{s_{w,2}^2}{m_2}}}
\deqn{s_{w,i}^2 = \frac{SSD_{w,i}}{m_i - 1}}
\deqn{df = \frac{1}{\frac{c^2}{m_1 - 1} + \frac{\left(1 - c\right)^2}{m_2 -1}}}
\deqn{c = \frac{\frac{s_{w,1}^2}{m_1}}{\frac{s_{w,1}^2}{m_1} + \frac{s_{w,2}^2}{m_2}}}
\deqn{\bar{x}_{t,i} = \frac{\sum_{j=g_i+1}^{n_i - g_i}y_{i,j}}{}}
\deqn{g_i = \lfloor n_i\times p_t\rfloor}
\deqn{m_i = n_ - 2\times g_i}
\deqn{SSD_{w,i} = g_i\times\left(y_{i,g_i+1} - \bar{x}_{wi}\right)^2 + g_i\times\left(y_{i,n_i-g_i} - \bar{x}_{w,i}\right)^2 + \sum_{j=g+1}^{n_i - g_i} \left(y_{i,j} - \bar{x}_{w,i}\right)^2}
\deqn{\bar{x}_{w,i} = \frac{\bar{x}_{t,i}\times m_i + g_i\times\left(y_{i, g_i+1} + y_{i, n_i-g_i}\right)}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{t,i}} the trimmed mean of the scores in category i
\item \eqn{x_{w,i}} The Winsorized mean of the scores in category i
\item \eqn{SSD_{w,i}} the sum of squared deviations from the Winsorized mean of category i
\item \eqn{m_i} the number of scores in the trimmed data set from category i
\item \eqn{y_{i,j}} the j-th score after the scores in category i, after they are sorted from low to high
\item \eqn{p_t} the proportion of trimming on each side, we can define
}

\strong{YUEN-DIXON}

If \verb{se="yuen-dixon} a trimmed means test will be performed.

The formula used is (Yuen & Dixon, 1973, p. 394):
\deqn{t = \frac{\bar{x}_{t,1} - \bar{x}_{t,2}}{SE}}
\deqn{sig = 2\times\left(1 - T\left(\left|t\right|, df\right)\right)}

With:
\deqn{SE = \sqrt{\frac{SSD_{w,1} + SSD_{w,2}}{m_1 + m_2 - 2}\times\left(\frac{1}{m_1} + \frac{1}{m_2}\right)}}
\deqn{df = m_1 + m_2 - 2}
\deqn{\bar{x}_{t,i} = \frac{\sum_{j=g_i+1}^{n_i - g_i}y_{i,j}}{}}
\deqn{g_i = \lfloor n_i\times p_t\rfloor}
\deqn{m_i = n_ - 2\times g_i}
\deqn{SSD_{w,i} = g_i\times\left(y_{i,g_i+1} - \bar{x}_{w,i}\right)^2 + g_i\times\left(y_{i,n_i-g_i} - \bar{x}_{w,i}\right)^2 + \sum_{j=g+1}^{n_i - g_i} \left(y_{i,j} - \bar{x}_{w,i}\right)^2}
\deqn{\bar{x}_{w,i} = \frac{\bar{x}_{t,i}\times m_i + g_i\times\left(y_{i, g_i+1} + y_{i, n_i-g_i}\right)}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{ti}} the trimmed mean of the scores in category i
\item \eqn{x_{wi}} The Winsorized mean of the scores in category i
\item \eqn{SSD_{wi}} the sum of squared deviations from the Winsorized mean of category i
\item \eqn{m_i} the number of scores in the trimmed data set from category i
\item \eqn{y_{i,j}} the j-th score after the scores in category i, after they are sorted from low to high
\item \eqn{p_t} the proportion of trimming on each side, we can define
}
}
\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = df1['age']
ex1 = replace(ex1, ex1=="89 OR OLDER", "90")
ts_trimmed_mean_t_is(df1['sex'], ex1)
ts_trimmed_mean_t_is(df1['sex'], ex1, se="yuen-dixon")

#Example 2: vectors
scores = c(20,50,80,15,40,85,30,45,70,60, NA, 90,25,40,70,65, NA, 70,98,40)
groups = c("nat.","int.","int.","nat.","int.", "int.","nat.","nat.","int.","int.","int.","int.","int.","int.","nat.", "int." ,NA,"nat.","int.","int.")
ts_trimmed_mean_t_is(groups, scores)
ts_trimmed_mean_t_is(groups, scores, se="yuen-dixon")

}
\references{
Yuen, K. K. (1974). The two-sample trimmed t for unequal population variances. \emph{Biometrika, 61}(1), 165–170. https://doi.org/10.1093/biomet/61.1.165

Yuen, K. K., & Dixon, W. J. (1973). The approximate behaviour and performance of the two-sample trimmed t. \emph{Biometrika, 60}(2), 369–374. https://doi.org/10.2307/2334550
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
