% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_trimmed_means_t_is.R
\name{ts_trimmed_means_t_is}
\alias{ts_trimmed_means_t_is}
\title{Independent Samples Trimmed/Yuen Mean Test}
\usage{
ts_trimmed_means_t_is(scores, groups, trimProp = 0.1)
}
\arguments{
\item{scores}{A vector with the scores data}

\item{groups}{A vector with the group data}

\item{trimProp}{optional proportion to trim from each side (so in total twice this will be trimmed)}
}
\value{
dataframe test statistic, degrees of freedom, p-value (sig.) and name of test used
}
\description{
Independent Samples Trimmed/Yuen Mean Test
}
\details{
The formula used is (Yuen & Dixon, 1973, p. 394):
\deqn{t = \frac{\bar{x}_{t,1} - \bar{x}_{t,2}}{SE}}
\deqn{sig = 2\times\left(1 - T\left(\left|t\right|, df\right)\right)}

With:
\deqn{SE = \sqrt{\frac{SSD_{w,1} + SSD_{w,2}}{m_1 + m_2 - 2}\times\left(\frac{1}{m_1} + \frac{1}{m_2}\right)}}
\deqn{df = m_1 + m_2 - 2}
\deqn{\bar{x}_{t,i} = \frac{\sum_{j=g_i+1}^{n_i - g_i}y_{i,j}}{}}
\deqn{g_i = \lfloor n_i\times p_t\rfloor}
\deqn{m_i = n_ - 2\times g_i}
\deqn{SSD_{w,i} = g_i\times\left(y_{i,g_i+1} - \bar{x}_{w,i}\right)^2 + g_i\times\left(y_{i,n_i-g_i} - \bar{x}_{w,i}\right)^2 + \sum_{j=g+1}^{n_i - g_i} \left(y_{i,j} - \bar{x}_{w,i}\right)^2}
\deqn{\bar{x}_{w,i} = \frac{\bar{x}_{t,i}\times m_i + g_i\times\left(y_{i, g_i+1} + y_{i, n_i-g_i}\right)}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{ti}} the trimmed mean of the scores in category i
\item \eqn{x_{wi}} The Winsorized mean of the scores in category i
\item \eqn{SSD_{wi}} the sum of squared deviations from the Winsorized mean of category i
\item \eqn{m_i} the number of scores in the trimmed data set from category i
\item \eqn{y_{i,j}} the j-th score after the scores in category i, after they are sorted from low to high
\item \eqn{p_t} the proportion of trimming on each side, we can define
}
}
\examples{
scores = c(20,50,80,15,40,85,30,45,70,60, NA, 90,25,40,70,65, NA, 70,98,40)
groups = c("national","international","international","national","international", "international","national","national","international","international","international","international","international","international","national", "international" ,NA,"national","international","international")
ts_trimmed_means_t_is(scores, groups)

}
\references{
Yuen, K. K., & Dixon, W. J. (1973). The approximate behaviour and performance of the two-sample trimmed t. \emph{Biometrika, 60}(2), 369–374. https://doi.org/10.2307/2334550
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
