% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_welch_t_is.R
\name{ts_welch_t_is}
\alias{ts_welch_t_is}
\title{Welch t Test (Independent Samples)}
\usage{
ts_welch_t_is(catField, scaleField, categories = NULL, dmu = 0, df_ver = "ws")
}
\arguments{
\item{catField}{A vector with the categorical data}

\item{scaleField}{A vector with the scores}

\item{categories}{Optional to indicate which two categories of catField to use, otherwise first two found will be used.}

\item{dmu}{Optional difference according to null hypothesis (default is 0)}

\item{df_ver}{c('ws', 'w', welch-satterthwaite, 'welch') optional version of degrees of freedom to use}
}
\value{
A dataframe with:
\item{n cat. 1}{the sample size of the first category}
\item{n cat. 2}{the sample size of the second category}
\item{mean cat. 1}{the sample mean of the first category}
\item{mean cat. 2}{the sample mean of the second category}
\item{diff.}{difference between the two sample means}
\item{hyp. diff.}{hypothesized difference between the two population means}
\item{statistic}{the test statistic (t-value)}
\item{df}{degrees of freedom}
\item{pValue}{the significance (p-value)}
\item{test}{name of test used}
}
\description{
A test to compare two means. The null hypothesis would be that the means of each category are equal in the population.

Unlike the Student t-test, the Welch test does not assume the variances of the two categories to be equal in the population. Ruxten (2006) even argues that the Welch t-test should always be prefered over the Student t-test.

There are four similar tests, with different assumptions.\tabular{lll}{
   test \tab equal variance \tab normality \cr
   Student \tab yes \tab yes \cr
   Welch \tab no \tab yes \cr
   Trimmed \tab yes \tab no \cr
   Yuen-Welch \tab no \tab no \cr
}


The Trimmed and Yuen-Welch can be found in the \strong{ts_trimmed_mean_is()}, and the Student t-test with the \strong{ts_student_t_is().}

This function is shown in this \href{https://youtu.be/yKCkSfNz6-Q}{YouTube video} and the test is also described at \href{https://peterstatistics.com/Terms/Tests/WelchIS.html}{PeterStatistics.com}
}
\details{
The formula used is:
\deqn{t = \frac{\bar{x}_1 - \bar{x}_2}{SE}}
\deqn{df = \frac{SE^4}{\frac{\left(s_1^2\right)^2}{n_1^2\times\left(n_1 - 1\right)} + \frac{\left(s_2^2\right)^2}{n_2^2\times\left(n_2 - 1\right)}}}
\deqn{sig. = 2\times\left(1 - T\left(\left|t\right|, df\right)\right)}

With:
\deqn{SE = \sqrt{\frac{s_1^2}{n_1} + \frac{s_2^2}{n_2}}}
\deqn{s_i^2 = \frac{\sum_{j=1}^{n_i} \left(x_{i,j} - \bar{x}_i\right)^2}{n_i - 1}}
\deqn{\bar{x}_i = \frac{\sum_{j=1}^{n_i} x_{i,j}}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{i,j}} the j-th score in category i
\item \eqn{n_i} the number of scores in category i
}

The degrees of freedom can be found in Welch (1938, p. 353, eq. 9; 1947, p. 32, eq. 28) and Satterthwaite (1946, p. 114, eq. 17). Welch (1947, p. 32, eq. 29) also suggests another formula for the degrees of freedom, which can be written as:
\deqn{df_w = \frac{\left( \frac{s_1^2}{n_1}+\frac{s_2^2}{n_2}\right)^2}{\frac{s_1^2}{n_1^2 \times \left(n_1 + 1\right)} + \frac{s_2^2}{n_2^2 \times \left(n_2 + 1\right)}} - 2}
Later, Aspin and Welch (1949, p. 295) indicate this alternative version has little to know advantage over the other version.
}
\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = df1['age']
ex1 = replace(ex1, ex1=="89 OR OLDER", "90")
ts_welch_t_is(df1['sex'], ex1)

#Example 2: vectors
scores = c(20,50,80,15,40,85,30,45,70,60, NA, 90,25,40,70,65, NA, 70,98,40)
groups = c("nat.","int.","int.","nat.","int.", "int.","nat.","nat.","int.",
"int.","int.","int.","int.","int.","nat.", "int." ,NA,"nat.","int.","int.")
ts_welch_t_is(groups, scores)

}
\references{
Aspin, A. A., & Welch, B. L. (1949). Tables for use in comparisons whose accuracy involves two variances, separately estimated. \emph{Biometrika, 36}(3/4), 290. https://doi.org/10.2307/2332668

Ruxton, G. D. (2006). The unequal variance t-test is an underused alternative to Student's t-test and the Mann-Whitney U test. \emph{Behavioral Ecology, 17}(4), 688-690. https://doi.org/10.1093/beheco/ark016

Satterthwaite, F. E. (1946). An approximate distribution of estimates of variance components. \emph{Biometrics Bulletin, 2}(6), 110. https://doi.org/10.2307/3002019

Welch, B. L. (1938). The significance of the difference between two means when the population variances are unequal. \emph{Biometrika, 29}(3-4), 350-362. https://doi.org/10.1093/biomet/29.3-4.350

Welch, B. L. (1947). The generalization of `Student's' problem when several different population variances are involved. \emph{Biometrika, 34}(1/2), 28-35. https://doi.org/10.2307/2332510
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
