% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wilcox.R
\name{ts_wilcox}
\alias{ts_wilcox}
\title{Wilcox Test}
\usage{
ts_wilcox(scores, groups)
}
\arguments{
\item{scores}{the numeric scores variable}

\item{groups}{the groups variable}
}
\value{
A dataframe with:
\item{statistic}{the chi-square-statistic from the test}
\item{df}{the degrees of freedom}
\item{pValue}{the significance (p-value)}
}
\description{
Wilcox Test
}
\details{
The formula used (Wilcox, 1988, pp. 110-111)
\deqn{H = \frac{\sum_{j=1}^k \left(W_j - \bar{W}\right)^2}{\hat{\theta}}}
\deqn{df = k - 1}
\deqn{sig. = 1 - \chi^2\left(H, df\right)}
With:
\deqn{W_j = b_j\times x_{n_j,j} + \frac{1 - b_j}{n_j}\times\sum_{i=1}^{n_j-1} x_{i,j}}
\deqn{\bar{W} = \frac{\sum_{j=1}^k W_j}{k}}
\deqn{b_j = \frac{1 + \sqrt{\frac{\left(n_j - 1\right)\times\left(n_j\times\hat{\theta}\right)}{s_j^2}}}{n_j}}
\deqn{\hat{\theta} = \text{max}\left\{\frac{s_1^2}{n_1}, \frac{s_2^2}{n_2}, \dots, \frac{s_k^2}{n_k}\right\}}
\deqn{s_j^2 = \frac{\sum_{i=1}^{n_j} \left(x_{i,j} - \bar{x}_j\right)^2}{n_j - 1}}
\deqn{\bar{x}_j = \frac{\sum_{j=1}^{n_j} x_{i,j}}{n_j}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{i,j}} the i-th score in category j
\item \eqn{k} the number of categories
\item \eqn{n_j} the sample size of category j
\item \eqn{\bar{x}_j} the sample mean of category j
\item \eqn{s_j^2} the sample variance of the scores in category j
\item \eqn{df} the degrees of freedom
\item \eqn{\chi^2\left(\dots\right)} the cumulative density function of the chi-square distribution
}

The original article has an error in the formula for \eqn{b_j}. There are missing brackets.
Using the population version in the article of \eqn{c_j} the formula used here was adapted.
}
\examples{
scores = c(20, 50, 80, 15, 40, 85, 30, 45, 70, 60, NA, 90, 25, 40, 70, 65, NA, 70, 98, 40, 65, 60, 35, NA, 50, 40, 75, NA, 65, 70, NA, 20, 80, 35, NA, 68, 70, 60, 70, NA, 80, 98, 10, 40, 63, 75, 80, 40, 90, 100, 33, 36, 65, 78, 50)
groups = c("Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Rotterdam", "Haarlem", "Diemen", "Haarlem", "Diemen", "Haarlem", "Haarlem", "Haarlem", "Haarlem", "Haarlem")
ts_wilcox(scores, groups)

}
\references{
Wilcox, R. R. (1988). A new alternative to the ANOVA F and new results on James’s second-order method. \emph{British Journal of Mathematical and Statistical Psychology, 41}(1), 109–117. https://doi.org/10.1111/j.2044-8317.1988.tb00890.x
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
