% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wilcox_owa.R
\name{ts_wilcox_owa}
\alias{ts_wilcox_owa}
\title{Wilcox Test}
\usage{
ts_wilcox_owa(nomField, scaleField, categories = NULL)
}
\arguments{
\item{nomField}{the groups variable}

\item{scaleField}{the numeric scores variable}

\item{categories}{vector, optional. the categories to use from catField}
}
\value{
A dataframe with:
\item{n}{the sample size}
\item{statistic}{the test statistic (chi-square value value)}
\item{df}{the degrees of freedom}
\item{p-value}{the significance (p-value)}
}
\description{
Tests if the means (averages) of each category could be the same in the population.

If the p-value is below a pre-defined threshold (usually 0.05), the null hypothesis is rejected, and there are then at least two categories who will have a different mean on the scaleField score in the population.

There are quite some alternatives for this, the stikpet library has Fisher, Welch, James, Box, Scott-Smith, Brown-Forsythe, Alexander-Govern, Mehrotra modified Brown-Forsythe, Hartung-Agac-Makabi, Özdemir-Kurt and Wilcox as options. See the notes from ts_fisher_owa() for some discussion on the differences.
}
\details{
The formula used (Wilcox, 1988, pp. 110-111)
\deqn{H = \frac{\sum_{j=1}^k \left(W_j - \bar{W}\right)^2}{\hat{\theta}}}
\deqn{df = k - 1}
\deqn{sig. = 1 - \chi^2\left(H, df\right)}
With:
\deqn{W_j = b_j\times x_{n_j,j} + \frac{1 - b_j}{n_j}\times\sum_{i=1}^{n_j-1} x_{i,j}}
\deqn{\bar{W} = \frac{\sum_{j=1}^k W_j}{k}}
\deqn{b_j = \frac{1 + \sqrt{\frac{\left(n_j - 1\right)\times\left(n_j\times\hat{\theta}\right)}{s_j^2}}}{n_j}}
\deqn{\hat{\theta} = \text{max}\left\{\frac{s_1^2}{n_1}, \frac{s_2^2}{n_2}, \dots, \frac{s_k^2}{n_k}\right\}}
\deqn{s_j^2 = \frac{\sum_{i=1}^{n_j} \left(x_{i,j} - \bar{x}_j\right)^2}{n_j - 1}}
\deqn{\bar{x}_j = \frac{\sum_{j=1}^{n_j} x_{i,j}}{n_j}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{i,j}} the i-th score in category j
\item \eqn{k} the number of categories
\item \eqn{n_j} the sample size of category j
\item \eqn{\bar{x}_j} the sample mean of category j
\item \eqn{s_j^2} the sample variance of the scores in category j
\item \eqn{df} the degrees of freedom
\item \eqn{\chi^2\left(\dots\right)} the cumulative density function of the chi-square distribution
}

The original article has an error in the formula for \eqn{b_j}. There are missing brackets.
Using the population version in the article of \eqn{c_j} the formula used here was adapted.
}
\references{
Wilcox, R. R. (1988). A new alternative to the ANOVA F and new results on James’s second-order method. \emph{British Journal of Mathematical and Statistical Psychology, 41}(1), 109–117. https://doi.org/10.1111/j.2044-8317.1988.tb00890.x
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
