% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wilcoxon_os.R
\name{ts_wilcoxon_os}
\alias{ts_wilcoxon_os}
\title{One-Sample Wilcoxon Signed Rank Test}
\usage{
ts_wilcoxon_os(
  data,
  levels = NULL,
  mu = NULL,
  ties = TRUE,
  appr = c("wilcoxon", "exact", "imanz", "imant"),
  eqMed = c("wilcoxon", "zsplit", "pratt"),
  cc = FALSE
)
}
\arguments{
\item{data}{dataframe with scores as numbers, or if text also provide levels}

\item{levels}{optional vector with levels in order}

\item{mu}{optional hypothesized median, otherwise the midrange will be used}

\item{ties}{optional boolean to use a tie correction (default is True)}

\item{appr}{optional which method to use for approximation (default is "wilcoxon")}

\item{eqMed}{optional method to deal with scores equal to hypMed (default is "wilcoxon")}

\item{cc}{optional boolean to use a continuity correction (default is FALSE)}
}
\value{
A dataframe with:
\item{mu}{the hypothesized median according to the null}
\item{W}{the Wilcoxon W value}
\item{statistic}{test statistic}
\item{df}{degrees of freedom (only applicable for Iman t approximation)}
\item{pValue}{significance (p-value)}
\item{testUsed}{description of the test used}
}
\description{
The one-sample Wilcoxon signed rank test is often considered the non-parametric version of a one-sample t-test.
It can be used to determine if the median is significantly different from an hypothesized value. It actually
doesn't always tests this specifically, but more if the mean rank is significantly different.

If the p-value is the probability of a result as in the sample, or more extreme, if the assumption about the
population would be true. If this is below a certain threshold (usually 0.05) the assumption about the
population is rejected. For this test the assumed median for the population is then incorrect.

Results in software packages for this test can vary, since there are a few different approaches. Especially if there are so-called ties. See the details for more information.

This function is shown in this \href{https://youtu.be/1r2lpPkJ1Vo}{YouTube video} and the test is also described at \href{https://peterstatistics.com/Terms/Tests/WilcoxonSignedRankOneSample.html}{PeterStatistics.com}.
}
\details{
The unadjusted test statistic is given by:
\deqn{W=\sum_{i=1}^{n_{r}^{+}}r_{i}^{+}}
With:
\deqn{r=\textup{rank}(|d|)}
\deqn{d_{i}=y_{i}-\theta}

\emph{Symbols used:}
\itemize{
\item \eqn{n_{r}^{+}} is the number of ranks with a positive deviation from the hypothesized median
\item \eqn{r_{i}^{+}} the i-th rank of the ranks with a positive deviation from the hypothesized median
\item \eqn{\theta} is the median tested (the hypothesized median).
\item \eqn{y_i} is the i-th score of the variable after removing scores that were equal to \eqn{\theta}
}

If there are no ties, an exact method can be used, using the Sign Rank Distribution.
R has this available with \emph{psignrank()}.
The exact test can be found in Zaiontz (n.d.)

\strong{Approximations}

If the sample size is large enough, we can use a normal approximation.
What is large enough varies quite per author. A few examples: n > 8 (slideplayer, 2015),
n > 15 (SigMaxl, n.d.), n > 20 (Wikipedia, n.d.), n > 25 (Harris & Hardin, 2013),
n > 30 (Winthrop, n.d.) .

The z-statistic is given by (appr="wilcoxon", ties=FALSE, cc=FALSE):
\deqn{Z = \frac{W - \mu_w}{\sigma_w}}
or with a ties correction (appr="wilcoxon", ties=TRUE, cc=FALSE):
\deqn{Z_{adj} = \frac{W - \mu_w}{\sigma_w^*}}
With:
\deqn{\mu_w = \frac{n_r\times\left(n_r + 1\right)}{4}}
\deqn{\sigma_w^2 = \frac{n_r\times\left(n_r + 1\right)\times\left(2\times n_r + 1\right)}{24}}
\deqn{\sigma_w^{*2} = \sigma_w^2 - A}
\deqn{A = \frac{\sum_{i=1}^k \left(t_i^3 - t_i\right)}{48}}

\emph{Additional symbols used}
\itemize{
\item \eqn{n_{r}} is the number of ranks used
\item \eqn{k} the number of unique ranks
\item \eqn{t_i} the frequency of the i-th unique rank
}

A Yates continuity correction can simply be applied:
In case of no ties (appr="wilcoxon", ties=FALSE, cc=TRUE):
\deqn{Z = \frac{\left|W - \mu_w\right| - 0.5}{\sigma_w}}
In case of ties (appr="wilcoxon", ties=TRUE, cc=TRUE):
\deqn{Z_{adj} = \frac{\left|W - \mu_w\right| - 0.5}{\sigma_w^*}}

An alternative approximation using the Student t distribution is given by Iman (1974, p. 799).
The formula is (appr="imant", ties=FALSE, cc=FALSE):
\deqn{t = \frac{W - \mu_w}{\sqrt{\frac{\sigma_w^2\times n_r - \left(W - \mu_w\right)^2}{n_r - 1}}}}
or with the ties correction (appr="imant", ties=TRUE, cc=FALSE):
\deqn{t = \frac{W - \mu_w}{\sqrt{\frac{\sigma_w^{*2}\times n_r - \left(W - \mu_w\right)^2}{n_r - 1}}}}

The two versions for with a continuity correction are:
No ties correction, but continuity (appr="imant", ties=FALSE, cc=TRUE):
\deqn{t = \frac{\left|W - \mu_w\right| - 0.5}{\sqrt{\frac{\sigma_w^2\times n_r - \left(\left|W - \mu_w\right| - 0.5\right)^2}{n_r - 1}}}}
Both corrections (appr="imant", ties=TRUE, cc=TRUE):
\deqn{t = \frac{\left|W - \mu_w\right| - 0.5}{\sqrt{\frac{\sigma_w^{*2}\times n_r - \left(\left|W - \mu_w\right| - 0.5\right)^2}{n_r - 1}}}}

Iman (1974, p. 803) also provides a combination of the t-approximation and the regular z-approximation
The equation is given by (appr="imanz"):
\deqn{Z_{I} = \frac{Z}{2}\times\left(1 + \sqrt{\frac{n_r - 1}{n_r - Z^2}}\right)}
The \eqn{Z} is any of the previous methods.

\strong{Ties with mu}

The default (eqMed="wilcoxon") removes first any scores that are equal to the hypothesized median.
There are two alternative methods for this.Both re-define \eqn{d_i} to:
\deqn{d_i = x_i - \theta}
Where \eqn{x_i} is simply the i-th score.

For the z-split method we only need to re-define:
\deqn{W = \frac{\sum_{i=1}^{n_{d_0}}r_{i,0}}{2} + \sum_{i=1}^{n_{r}^{+}}r_{i}^{+}}
Where \eqn{n_{d_0}} is the number of scores that equal the hypothesized median, and
\eqn{r_{i,0}} is the rank of the i-th score that equals the hypothesized median.

In essence we added half the sum of the ranks that were equal to the hypothesized median.

For the z-split method all other calculations than go the same.

For the Pratt (1959) method we also re-define:
\deqn{\mu_w = \frac{n_r\times\left(n_r + 1\right) - n_{d_0}\times\left(n_{d_0} + 1\right)}{4}}
\deqn{\sigma_w^2 = \frac{n_r\times\left(n_r + 1\right)\times\left(2\times n_r + 1\right) - n_{d_0}\times\left(n_{d_0} + 1\right)\times\left(2\times n_{d_0} + 1\right)}{24}}

For the Pratt method, the ties correction still excludes the ties
for the scores that equal the hypothesized median,
but for the z-split method it will include them.

For both methods now \eqn{n_r=n}, where n is the number of scores.

The Pratt (1959) method and z-split method were found in Python’s documentation
for scipy’s Wilcoxon function (scipy, n.d.). They also refer to Cureton (1967) for the Pratt method.
}
\section{Before, After and Alternatives}{

Before this measure you might want an impression using a frequency table or a visualisation:
\code{\link{tab_frequency}}, for a frequency table
\code{\link{vi_bar_stacked_single}}, or Single Stacked Bar-Chart.
\code{\link{vi_bar_dual_axis}}, for Dual-Axis Bar Chart.

After this you might want to determine an effect size measure:
\code{\link{es_common_language_os}}, for the Common Language Effect Size.
\code{\link{es_dominance}}, for the Dominance score.
\code{\link{r_rank_biserial_os}}, for the Rank-Biserial Correlation

Alternative tests:
\code{\link{ts_sign_os}}, for One-Sample Sign Test.
\code{\link{ts_trinomial_os}}, for One-Sample Trinomial Test.
}

\examples{
#Example 1: Text dataframe
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
df2 = read.csv(file2, sep=';', na.strings=c("", "NA"))
ex1 = df2[['Teach_Motivate']]
order = c("Fully Disagree", "Disagree", "Neither disagree nor agree", "Agree", "Fully agree")
ts_wilcoxon_os(ex1, levels=order)

#Example 2: Numeric data
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
ts_wilcoxon_os(ex2)

}
\references{
Cureton, E. E. (1967). The normal approximation to the signed-rank sampling distribution when zero differences are present. \emph{Journal of the American Statistical Association, 62}(319), 1068–1069. https://doi.org/10.1080/01621459.1967.10500917

Harris, T., & Hardin, J. W. (2013). Exact Wilcoxon Signed-Rank and Wilcoxon Mann–Whitney Ranksum Tests. \emph{The Stata Journal, 13}(2), 337–343. https://doi.org/10.1177/1536867X1301300208

Iman, R. L. (1974). Use of a t-statistic as an approximation to the exact distribution of the wildcoxon signed ranks test statistic. \emph{Communications in Statistics, 3}(8), 795–806. https://doi.org/10.1080/03610927408827178

Pratt, J. W. (1959). Remarks on zeros and ties in the Wilcoxon signed rank procedures. \emph{Journal of the American Statistical Association, 54}(287), 655–667. https://doi.org/10.1080/01621459.1959.10501526

scipy. (n.d.). Scipy.stats.wilcoxon. Scipy. https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.wilcoxon.html

SigMaxl. (n.d.). One Sample Wilcoxon Sign Test Exact. Retrieved August 30, 2020, from https://www.sigmaxl.com/OneSampleSignWilcoxonExact.shtml

slideplayer. (2015, June 13). Using statistics to make inferences 6.

Wikipedia. (n.d.). Wilcoxon signed-rank test. In Wikipedia. Retrieved August 30, 2020, from https://en.wikipedia.org/w/index.php?title=Wilcoxon_signed-rank_test&oldid=974561084

Wilcoxon, F. (1945). Individual comparisons by ranking methods. \emph{Biometrics Bulletin, 1}(6), 80. https://doi.org/10.2307/3001968

Winthrop. (n.d.). The Wilcoxon signed rank test for one sample. Winthrop Univerisy Hospital. https://nyuwinthrop.org/wp-content/uploads/2019/08/wilcoxon-sign-rank-test-one-sample.pdf

Zaiontz, C. (n.d.). Wilcoxon signed ranks exact test. Real Statistics Using Excel. Retrieved January 25, 2023, from https://real-statistics.com/non-parametric-tests/wilcoxon-signed-ranks-test/wilcoxon-signed-ranks-exact-test/
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
