% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wilcoxon_ps.R
\name{ts_wilcoxon_ps}
\alias{ts_wilcoxon_ps}
\title{Paired Samples Wilcoxon Signed Rank Test}
\usage{
ts_wilcoxon_ps(
  field1,
  field2,
  levels = NULL,
  dmu = 0,
  appr = c("wilcoxon", "exact", "imant", "imanz"),
  noDiff = c("wilcoxon", "zsplit", "pratt"),
  ties = TRUE,
  cc = FALSE
)
}
\arguments{
\item{field1}{the numeric scores of the first variable}

\item{field2}{the numeric scores of the second variable}

\item{levels}{vector, optional. the levels from field1 and field2}

\item{dmu}{float, optional. The difference according to the null hypothesis (default is 0)}

\item{appr}{c("wilcoxon", "exact", "imanz", "imant") optional which method to use for approximation (default is "wilcoxon")}

\item{noDiff}{c("wilcoxon", "pratt", "zsplit") optional method to deal with scores equal on both variables (default is "wilcoxon")}

\item{ties}{optional boolean to use a tie correction (default is True)}

\item{cc}{optional boolean to use a continuity correction (default is False)}
}
\value{
A dataframe with:
\item{nr}{the number of ranks used in calculation}
\item{mu}{the median according to the null hypothesis}
\item{W}{the Wilcoxon W}
\item{statistic}{the test statistic}
\item{df}{degrees of freedom (only applicable for Iman t approximation)}
\item{p-value}{the significance (p-value)}
\item{test}{description of the test used}
}
\description{
The paired-sample Wilcoxon signed rank test is often considered the non-parametric version of a paired-samples t-test. It can be used to determine if the median is significantly different between the two variables. It actually doesn't always tests this specifically, but more if the mean rank is significantly different.

If the p-value is the probability of a result as in the sample, or more extreme, if the assumption about the population would be true. If this is below a certain threshold (usually 0.05) the assumption about the population is rejected.

Results in software packages for this test can vary, since there are a few different approaches. Especially if there are so-called ties.

This function simply determines the differences between the two provided variables, and then passes these differences along to the one-sample version. See ts_wilcoxon_os() for details on this.
}
\details{
The unadjusted test statistic is given by:
\deqn{W=\min\left(W_{neg}, W_{pos}\right)}
With:
\deqn{W_{pos} = \sum_{i=1}^n \begin{cases} r_i & \text{ if } d_i > 0 \\ 0 & \text{ if } d_i \le 0 \end{cases}}
\deqn{W_{neg} = \sum_{i=1}^n \begin{cases} r_i & \text{ if } d_i < 0 \\ 0 & \text{ if } d_i \ge 0 \end{cases}}
\deqn{d_i = x_i - y_i}

\emph{Symbols used:}
\itemize{
\item \eqn{n} the number of scores (equal for each variable)
\item \eqn{x_i} the i-th score on the first variable
\item \eqn{y_i} the i-th score on the second variable
\item \eqn{r_i} the i-th rank of the absolute differences (\eqn{d_i})
\item \eqn{W_{pos}} is the number of ranks with a positive difference
\item \eqn{W_{neg}} is the number of ranks with a negative difference
}

The distribution and test for \eqn{W} can now be performed the same way as
for the one-sample case. See \emph{ts_wilcoxon_os()} for details on the calculations.
The \eqn{d_i} scores are now the one-sample, and the hypothesized median would be 0.

\strong{Alternatives}

*R's stats library *

wilcox.test(ord1, ord2, paired=TRUE, exact=FALSE, correct=TRUE)

wilcox.test(ord1, ord2, paired=TRUE, exact=FALSE, correct=FALSE)

\emph{library(coin)}

wilcoxsign_test(ord1 ~ ord2, zero.method = "Wilcoxon")

wilcoxsign_test(ord1 ~ ord2, zero.method = "Pratt")
}
\references{
Wilcoxon, F. (1945). Individual comparisons by ranking methods. \emph{Biometrics Bulletin, 1}(6), 80. doi:10.2307/3001968
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
