% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_z_is.R
\name{ts_z_is}
\alias{ts_z_is}
\title{Independent Samples Z Test}
\usage{
ts_z_is(scores, groups, dmu = 0, sigma1 = NULL, sigma2 = NULL)
}
\arguments{
\item{scores}{A vector with the scores data}

\item{groups}{A vector with the group data}

\item{dmu}{difference according to null hypothesis (default is 0)}

\item{sigma1}{population standard deviation of the first group, if NULL sample results will be used}

\item{sigma2}{population standard deviation of the second group, if NULL sample results will be used}
}
\value{
A dataframe with:
\item{statistic}{the test statistic}
\item{pValue}{the significance (p-value)}
}
\description{
Independent Samples Z Test
}
\details{
The formula used is:
\deqn{z = \frac{\bar{x}_1 - \bar{x}_2}{SE}}
\deqn{sig. = 2\times\left(1 - \Phi\left(\left|z\right|\right)\right)}

With:
\deqn{SE = \sqrt{\frac{\sigma_1^2}{n_1} + \frac{\sigma_2^2}{n_2}}}
\deqn{\sigma_i^2 \approx s_i^2 = \frac{\sum_{j=1}^{n_i} \left(x_{i,j} - \bar{x}_i\right)^2}{n_i - 1}}
\deqn{\bar{x}_i = \frac{\sum_{j=1}^{n_i} x_{i,j}}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{i,j}} the j-th score in category i
\item \eqn{n_i} the number of scores in category i
}
}
\examples{
scores = c(20,50,80,15,40,85,30,45,70,60, NA, 90,25,40,70,65, NA, 70,98,40)
groups = c("national","international","international","national","international", "international","national","national","international","international","international","international","international","international","national", "international" ,NA,"national","international","international")
ts_z_is(scores, groups)

}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
