% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_z_os.R
\name{ts_z_os}
\alias{ts_z_os}
\title{One-Sample Z Test}
\usage{
ts_z_os(data, mu = NULL, sigma = NULL)
}
\arguments{
\item{data}{A vector with the data as numbers}

\item{mu}{optional hypothesized mean, otherwise the midrange will be used}

\item{sigma}{population standard deviation, if NULL sample results will be used}
}
\value{
A dataframe with:
\item{mu}{the hypothesized mean}
\item{sample.mean}{the sample mean}
\item{statistic}{the test statistic}
\item{pValue}{the significance (p-value)}
\item{testUsed}{name of test used}
}
\description{
This test is often used if there is a large sample size. For smaller sample sizes, a Student t-test
is usually used.
}
\details{
The assumption about the population (null hypothesis) for this test is a pre-defined mean, i.e. the (arithmetic) mean
that is expected in the population. If the p-value (significance) is then below a pre-defined threhold
(usually 0.05), the assumption is rejected.

The formula used is:
\deqn{z = \frac{\bar{x} - \mu_{H_0}}{SE}}
\deqn{sig = 2\times\left(1 - \Phi\left(\left|z\right|\right)\right)}

With:
\deqn{SE = \frac{\sigma}{\sqrt{n}}}
\deqn{\sigma \approx s = \sqrt{\frac{\sum_{i=1}^n\left(x_i - \bar{x}\right)^2}{n - 1}}}
\deqn{\bar{x} = \frac{\sum_{i=1}^nx_i}{n}}

\emph{Symbols used:}
\itemize{
\item \eqn{\Phi\left(\dots\right)} the cumulative distribution function of the standard normal distribution
\item \eqn{\bar{x}} the sample mean
\item \eqn{\mu_{H_0}} the hypothesized mean in the population
\item \eqn{SE} the standard error (i.e. the standard deviation of the sampling distribution)
\item \eqn{n} the sample size (i.e. the number of scores)
\item \eqn{s} the unbiased sample standard deviation
\item \eqn{x_i} the i-th score
}
}
\examples{
data <- c(1, 2, 5, 1, 1, 5, 3, 1, 5, 1, 1, 5, 1, 1, 3, 3, 3, 4, 2, 4)
ts_z_os(data)

}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
