% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_bar_clustered.R
\name{vi_bar_clustered}
\alias{vi_bar_clustered}
\title{Clustered / Multiple Bar Chart}
\usage{
vi_bar_clustered(
  field1,
  field2,
  order1 = NULL,
  order2 = NULL,
  percent = c(NULL, "all", "row", "column")
)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{order1}{: optional list with order for categories of field1}

\item{order2}{: optional list with order for categories of field2}

\item{percent}{: optional which percentages to show. Either "none" (default), "all", "row", "column"}
}
\value{
clustered bar chart
}
\description{
A bar-chart is defined as “a graph in which bars of varying height with spaces between them are used to display data for variables defined by qualities or categories” (Zedeck, 2014, p. 20).

The bars can be split into multiple bars based on another variable. This is then known as a multiple bar-chart (Kemp, 2004, p. 150) or clustered bar-chart (Brase, 2009, p. 50; Griffith, 2007, p. 168).

It can be defined as “a bar chart for comparing the frequencies of a categorical variable in two or more situations” (Upton & Cook, 2014, p. 283).
}
\examples{
#Example 1: Clustered Bar Chart in percentages
vi_bar_clustered(df1[['mar1']], df1[['sex']], percent="column")

#Example 2: Specified order
orderR = c("DIVORCED", "WIDOWED", "SEPARATED", "MARRIED", "NEVER MARRIED")
orderC = c("MALE", "FEMALE")
vi_bar_clustered(df1[['mar1']], df1[['sex']], order1=orderR, order2=orderC)

}
\references{
Brase, C. (2009). \emph{Understandable statistics} (9th ed.). Houghton MIfflin.

Griffith, A. (2007). \emph{SPSS for dummies}. Wiley.

Kemp, S. M., & Kemp, S. (2004). \emph{Business statistics demystified}. McGraw-Hill.

Upton, G., & Cook, I. (2014). \emph{Oxford: Dictionary of statistics} (3rd ed.). Oxford University Press.

Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
