% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_bar_stacked_single.R
\name{vi_bar_stacked_single}
\alias{vi_bar_stacked_single}
\title{Single Stacked Bar-Chart}
\usage{
vi_bar_stacked_single(data, varname = NULL, height = c("percent", "count"))
}
\arguments{
\item{data}{the data from which to create the bar-chart}

\item{varname}{name of the variable, if not provided name of data is used.}

\item{height}{which values to show on axis, either percent or count}
}
\value{
The chart.
}
\description{
A regular bar-chart but with the bars on top of each other, instead of next to each other.
This is called a compound bar chart, stacked bar chart (Wilkinson, 2005, p. 157) or
component bar chart (Zedeck, 2014, p. 54).

It can be defined as: “a bar chart showing multiple bars stacked at each x-axis category,
each representing a value of the stacking variable” (Upton & Cook, 2014, p. 88).
}
\details{
This function basically uses barplot(...,beside = FALSE) from R's \emph{graphics} library
}
\examples{
ordData <- c(1, 2, 5, 1, 1, 5, 3, 1, 5, 1, 1, 5, 1, 1, 3, 3, 3, 4, 2, 4)
vi_bar_stacked_single(ordData)
vi_bar_stacked_single(ordData, height="count")
vi_bar_stacked_single(ordData, varname="scores", height="percent")

}
\references{
Upton, G. J. G., & Cook, I. (2014). \emph{Dictionary of statistics} (3rd ed.). Oxford University Press.

Wilkinson, L. (2005). \emph{The grammar of graphics} (2nd ed). Springer.

Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
\seealso{
An alternative chart for a single ordinal variable could be a dual axis bar chart, see \code{\link{vi_bar_dual_axis}}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}
}
