% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_boxplot_split.R
\name{vi_boxplot_split}
\alias{vi_boxplot_split}
\title{Split Box Plot}
\usage{
vi_boxplot_split(catField, scaleField, categories = NULL, ...)
}
\arguments{
\item{catField}{list or dataframe with the categories}

\item{scaleField}{list or dataframe with the scores}

\item{categories}{optional list with categories to use}

\item{...}{other parameters for use in boxplot function}
}
\value{
The split boxplot
}
\description{
Based on a categorical field the scores for each category are plotted in a separate boxplot and each of them is placed underneath each other.

See \strong{vi_boxplot_single()} for more details on boxplots.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
