% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_butterfly_chart.R
\name{vi_butterfly_chart}
\alias{vi_butterfly_chart}
\title{Butterfly Chart / Tornado Chart / Pyramid Chart}
\usage{
vi_butterfly_chart(
  field1,
  field2,
  categories1 = NULL,
  categories2 = NULL,
  variation = "butterfly",
  roundHigh = 5
)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection of categories of field1}

\item{categories2}{: optional list with selection of categories of field2}

\item{variation}{: optional order of the bars. Either "butterfly" (default), "tornado", or "pyramid"}

\item{roundHigh}{: optional to adjust number of tickmarks on horizontal axis}
}
\value{
plot
}
\description{
A special case of diverging bar charts when only comparing two categories.

Depending on the ordering of the results different names exist. I've chosen to use 'butterfly' if no ordering is done, 'pyramid' if they are ordered from small to large, and 'tornado' when going from large to small.

This function is shown in this \href{https://youtu.be/QeyqW5Vk69o}{YouTube video} and the diagram is also discussed at \link{PeterStatistics.com}(https://peterstatistics.com/Terms/Visualisations/PyramidChart.html
}
\details{
The term \emph{butterfly chart} can for example be found in Hwang and Yoon (2021, p. 25).

The term \emph{tornado diagrom} can be found in the guide from the Project Management Institute (2013, p. 338). The term \emph{funnel chart} is also sometimes used (for example Jamsa (2020, p. 135)), but this is also a term sometimes used for a more analytical scatterplot used for some specific analysis.

The term \emph{pyramid chart} can for example be found in Schwabish (2021, p. 185). It is very often used for comparing age distributions.
}
\examples{
#Example
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
vi_butterfly_chart(df1[['mar1']], df1[['sex']], roundHigh=100)
vi_butterfly_chart(df1[['mar1']], df1[['sex']], variation="tornado", roundHigh=100)
vi_butterfly_chart(df1[['mar1']], df1[['sex']], variation="pyramid", roundHigh=100)

}
\references{
Hwang, J., & Yoon, Y. (2021). Data analytics and visualization in quality analysis using Tableau. CRC Press.

Jamsa, K. (2020). Introduction to data mining and analytics: With machine learning in R and Python. Jones & Bartlett Learning.

Project Management Institute (Ed.). (2013). A guide to the project management body of knowledge (5th ed.). Project Management Institute, Inc.

Schwabish, J. (2021). Better data visualizations: A guide for scholars, researchers, and wonks. Columbia University Press.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
