% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dot_plot.R
\name{vi_dot_plot}
\alias{vi_dot_plot}
\title{Dot Plot}
\usage{
vi_dot_plot(data, stackRatio = 1, dotSize = 1)
}
\arguments{
\item{data}{the data from which to create the dot plot}

\item{stackRatio}{ratio on how close the dots are to each other}

\item{dotSize}{indicator for how big the dots need to be}
}
\value{
chart the dot plot
}
\description{
The Oxford Dictionary of Statistics defines a dot plot as "an alternative to a bar chart
or line graph when there are very few data values. Each value is recorded as a dot,
so that the frequencies for each value can easily be counted" (Upton & Cook, 2014, p. 129).

This function uses ggplot2 geom_dotplot() to create a simple dot plot.

A \href{https://youtu.be/qs1nh0CMiIY}{YouTube} video on pie charts.
}
\details{
In the definition a \emph{bar chart} is mentioned. A bar chart can be defined as
“a graph in which bars of varying height with spaces between them are used to
display data for variables defined by qualities or categories” (Zedeck, 2014, p. 20).
Together this indicates that a dot plot is used for categorical data.

However, Zedeck sees the dot plot as an alternative name for a scatterplot,
which is for continuous data. A third version comes from the
Cambridge Dictionary of Statistics: "A more effective display than a number of
other methods, for example, pie charts and bar charts, for displaying quantitative
data which are labelled" (Everitt, 2004, p. 123). They also show an example where
we see a categorical variable on one axis, and a continuous variable on another.

This function was only for the original definition for categorical data.
}
\examples{
data <- c("MARRIED", "DIVORCED", "MARRIED", "SEPARATED", "DIVORCED", "NEVER MARRIED", "DIVORCED", "DIVORCED", "NEVER MARRIED", "MARRIED", "MARRIED", "MARRIED", "SEPARATED", "DIVORCED", "NEVER MARRIED", "NEVER MARRIED", "DIVORCED", "DIVORCED", "MARRIED")
vi_dot_plot(data, stackRatio=1.5, dotSize=2)

}
\references{
Everitt, B. (2004). \emph{The Cambridge dictionary of statistics} (2nd ed). Cambridge University Press.

Upton, G. J. G., & Cook, I. (2014). \emph{Dictionary of statistics} (3rd ed.). Oxford University Press.

Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}
}
