% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_histogram_overlay.R
\name{vi_histogram_overlay}
\alias{vi_histogram_overlay}
\title{Overlaid Histogram}
\usage{
vi_histogram_overlay(
  catField,
  scaleField,
  categories = NULL,
  bins = NULL,
  show = "count",
  density = "auto",
  title = NULL,
  xlbl = NULL,
  colors = c(rgb(0, 0, 1, 1/4), rgb(1, 0, 0, 1/4)),
  ...
)
}
\arguments{
\item{catField}{list or dataframe with the categories}

\item{scaleField}{list or dataframe with the scores}

\item{categories}{optional list with the two categories to use from bin_field. If not set the first two found will be used}

\item{bins}{optional list with upper bounds of bins to be used, or any of the pre-set options from hist()}

\item{show}{c('count', 'relative'), show either counts or relative count on vertical scale}

\item{density}{c('auto', FALSE, TRUE), show (relative) frequency or use (relative) frequency density.}

\item{title}{string, title on top of chart, default is 'histogram'}

\item{xlbl}{string, title for the horizontal axis. If set to NULL the name of the catField variable is used.}

\item{colors}{vector, two colors, one for each category, default is c(rgb(0, 0, 1, 1/4), rgb(1, 0, 0, 1/4))}

\item{...}{optional additional parameters to pass to plot()}
}
\value{
The overlaid histograms
}
\description{
This function creates a simple overlaid (overlapping) histogram of two categories. The bins will be including the lower bound and excluding the upper, except the last bin.

See \strong{vi_histogram()} for more details on histograms.

The function is shown in this \href{https://youtu.be/IWIPIFRwaWw}{YouTube video} and the visualisation is also described at \href{https://peterstatistics.com/Terms/Visualisations/histogram.html}{PeterStatistics.com}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
