% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_histogram_split.R
\name{vi_histogram_split}
\alias{vi_histogram_split}
\title{Split Histogram}
\usage{
vi_histogram_split(
  catField,
  scaleField,
  categories = NULL,
  bins = NULL,
  show = "count",
  density = "auto",
  shareY = TRUE,
  xlbl = NULL,
  ...
)
}
\arguments{
\item{catField}{list or dataframe with the categories}

\item{scaleField}{list or dataframe with the scores}

\item{categories}{optional list with the two categories to use from bin_field. If not set the first two found will be used}

\item{bins}{optional list with upper bounds of bins to be used, or any of the pre-set options from hist()}

\item{show}{c('count', 'relative'), show either counts or relative count on vertical scale}

\item{density}{c('auto', FALSE, TRUE), show (relative) frequency or use (relative) frequency density.}

\item{shareY}{boolean, use same vertical axis range for each category}

\item{xlbl}{string, title for the horizontal axis. If set to NULL the name of the catField variable is used.}

\item{...}{optional additional parameters to pass to plot()}
}
\value{
The split histogram
}
\description{
Based on a categorical field the scores for each category are plotted in a separate histogram and each of the histograms is placed underneath each other.

See \strong{vi_histogram()} for more details on histograms.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
