% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pie.R
\name{vi_pie}
\alias{vi_pie}
\title{Pie Chart}
\usage{
vi_pie(data, labels = c("count", "percent", "both", "none"))
}
\arguments{
\item{data}{the data for which to create a pie-chart from}

\item{labels}{what to show besides the labels Either "count" (default), "percent", "none", or "both"}
}
\value{
chart the pie chart
}
\description{
A pie-chart is a “graphic display in which a circle is cut into wedges with the area of
each wedge being proportional to the percentage of cases in the category represented by
that wedge” (Zedeck, 2014, p. 260).

A video on pie charts is available \href{https://youtu.be/e6JtJsh-6iw}{here}.
}
\details{
It is possible to either show only the labels (label="none"), the counts (label="counts"),
the percentages (label="percent"), or both count and percent (label="both").

The function uses the basic R's graphics library \emph{pie} function, rotated and counter clockwise.

The pie-chart is quite popular and often used, but actually has a few disadvantages.
It can only show relative frequencies. To show other frequencies the numbers themselves
have to be added. A circle has 360 degrees, equal to 100\%. So by multiplying the relative
frequencies with 360, the degrees for each category can be found. This means that visually
the pie-chart can only show the relative frequencies.

Another disadvantage is when the relative frequencies are close to each other,
the differences are not easily seen in a circle diagram.

As a third disadvantage, when there are many categories the circle diagram will
look very busy and not easily to read.

People also have more difficulty with comparing areas and angles
(what you do when looking at a pie-chart) than comparing heights (what is done with a bar-chart).

Also often a 3D effect is added, but this actually makes comparisons of the slices even more difficult.

he earliest found circle diagram is found on the inlay of a book by William Playfair (1801).
The name 'pie chart' might come from a misspelling of the word Pi.
Pi is often associated with a circle. It might also simply come from the resemblances
with a pie (as in apple-pie).
However Srivastava and Rego (2011) put forward another belief that it is named after a
royal French cook Pie, who served dishes in a pie-chart shape.
}
\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = df1['mar1']
vi_pie(ex1);
vi_pie(ex1, labels="percent");
vi_pie(ex1, labels="none");
vi_pie(ex1, labels="both");

#Example 2: a list
ex2 = c("MARRIED", "DIVORCED", "MARRIED", "SEPARATED", "DIVORCED", "NEVER MARRIED", 
"DIVORCED", "DIVORCED", "NEVER MARRIED", "MARRIED", "MARRIED", "MARRIED", "SEPARATED", 
"DIVORCED", "NEVER MARRIED", "NEVER MARRIED", "DIVORCED", "DIVORCED", "MARRIED")
vi_pie(ex2);

}
\references{
Playfair, W. (1801). \emph{The statistical breviary: Shewing the resources of every state and kingdom}. T. Bensley. http://archive.org/details/statisticalbrev00playgoog

Srivastava, T. N., & Rego, S. (2011). \emph{Business research methodology}. Tata McGraw-Hill.

Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
