% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_stem_and_leaf.R
\name{vi_stem_and_leaf}
\alias{vi_stem_and_leaf}
\title{Stem-and-Leaf Display}
\usage{
vi_stem_and_leaf(data, key = NULL)
}
\arguments{
\item{data}{list with the numeric data}

\item{key}{optional factor to use for the stems}
}
\value{
prints the display in console and returns a dataframe with the stems and leafs
}
\description{
A stem-and-leaf display is defined as: "a method of displaying data in which each observation is split into two parts labelled the ‘stem’ and the ‘leaf’" (Everitt, 2004, p. 362). A diagram that could be used to visualize scale variables, created by Tukey (1972, p. 296).

In some variations of this, the cumulative frequencies are also shown, but currently this function does not provide for that.
}
\references{
Everitt, B. (2004). \emph{The Cambridge dictionary of statistics} (2nd ed.). Cambridge University Press.

Tukey, J. W. (1972). Some graphic and semigraphic displays. In T. A. Bancroft & S. A. Brown (Eds.), \emph{Statistical Papers in Honor of George W. Snedecor} (pp. 293–316). Iowa State University Press.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
