% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_alroy_f.R
\name{es_alroy_f}
\alias{es_alroy_f}
\title{Alroy's Forbes Adjustment}
\usage{
es_alroy_f(var1, var2)
}
\arguments{
\item{var1}{A vector with the data from the first variable}

\item{var2}{A vector with the data from the second variable}
}
\value{
Alroy F
}
\description{
Alroy's Forbes Adjustment
}
\details{
The formula used (Alroy, 2015):
\deqn{F' = \frac{a\times\left(n' + \sqrt{n'}\right)}{a\times\left(n' + \sqrt{n'}\right) + \frac{3}{2}\times b\times c}}
With:
\deqn{n' = a + b + c}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
}

Note Alroy refers to the Forbes coefficient as a measure of similarity.
Alroy then sets out to improve the measure by disregarding the lower left value (d).
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", 
"female","female","female", "male", "male", "male", "male", "male", "male", "male", 
"male", "male", "male", "male", "male", "male", "male", "male", "male", "male", 
"male", "male", "male", "male", "male", "male", "male", "male", "male", "male", 
"male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other",
"nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", 
"other", "other", "other", "other", "other", "other", "other", "other", "other", 
"other", "other", "other", "other", "other", "other")
es_alroy_f(bin1, bin2)

}
\references{
Alroy, J. (2015). A new twist on a very old binary similarity coefficient. \emph{Ecology, 96}(2), 575–586. https://doi.org/10.1890/14-0471.1
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
