% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_bonett_price_y.R
\name{es_bonett_price_y}
\alias{es_bonett_price_y}
\title{Bonett and Price Y*}
\usage{
es_bonett_price_y(var1, var2)
}
\arguments{
\item{var1}{A vector with the binary data from the first variable}

\item{var2}{A vector with the binary data from the second variable}
}
\value{
Bonett and Price Y*
}
\description{
Bonett and Price Y*
}
\details{
The formula used (Bonett and Price, 2007, pp. 433-434):
\deqn{Y* = \frac{\hat{\omega}^x-1}{\hat{\omega}^x+1}}
With:
\deqn{x = \frac{1}{2}-\left(\frac{1}{2}-p_{min}\right)^2}
\deqn{p_{min} = \frac{\text{MIN}\left(R_1, R_2, C_1, C_2\right)}{n}}
\deqn{\hat{\omega} = \frac{\left(a+0.1\right)\times\left(d+0.1\right)}{\left(b+0.1\right)\times\left(c+0.1\right)}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell
\item \eqn{b} the count in the top-right cell
\item \eqn{c} the count in the bottom-left cell
\item \eqn{d} the count in the bottom-right cell
\item \eqn{R_i} the sum of counts in the i-th row
\item \eqn{C_i} the sum of counts in the i-th column
}

Note that \eqn{\hat{\omega}} is a biased corrected version of the Odds Ratio
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female",
"female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", 
"male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", 
"male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl",
"nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", 
"other", "other", "other", "other", "other", "other", "other", "other", "other", "other", 
"other", "other")
es_bonett_price_y(bin1, bin2)

}
\references{
Bonett, D. G., & Price, R. M. (2007). Statistical inference for generalized yule coefficients in 2 × 2 contingency tables. \emph{Sociological Methods & Research, 35}(3), 429–446. https://doi.org/10.1177/0049124106292358
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
