% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_d_os.R
\name{es_cohen_d_os}
\alias{es_cohen_d_os}
\title{Cohen d' (for one-sample)}
\usage{
es_cohen_d_os(data, mu = NULL)
}
\arguments{
\item{data}{pandas series with the numeric scores}

\item{mu}{optional parameter to set the hypothesized mean. If not used the midrange is used}
}
\value{
Cohen d'. mu is also printed if not provided.
}
\description{
Cohen d' (for one-sample)
}
\details{
The formula used (Cohen, 1988, p. 46):
\deqn{d'=\frac{\bar{x}-\mu_{H_{0}}}{s}}
With:
\deqn{s = \sqrt{\frac{\sum_{i=1}^n \left(x_i - \bar{x}\right)^2}{n - 1}}}
\deqn{\bar{x} = \frac{\sum_{i=1}^n x_i}{n}}

\emph{Symbols used:}
\itemize{
\item \eqn{\bar{x}} the sample mean
\item \eqn{μ_{H_0}} the hypothesized mean in the population
\item \eqn{n} the sample size (i.e. the number of scores)
\item \eqn{s} the unbiased sample standard deviation
\item \eqn{x_i} the i-th score
}

Note to use a rule-of-thumb from Cohen d, first convert this to a regular Cohen d using
\emph{es_convert(d', from="cohendos", to="cohend")}

Then use \emph{th_cohen_d(d)}

Or convert it further to an Odds Ratio using:
\emph{es_convert(d, from="cohend", to="or", ex1="chinn)} or \emph{es_convert(d, from="cohend", to="or", ex1="borenstein)}

Then use \emph{th_odds_ratio(or)}

\strong{Alternative}

The \emph{lsr} library has a similar function: \emph{cohensD()}
}
\examples{
data <- c(20, 50, 80, 15, 40, 85, 30, 45, 70, 60, 90, 25, 40, 70, 65, 70, 98, 40, 65, 60)
es_cohen_d_os(data)
es_cohen_d_os(data, mu=70)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
