% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_d.R
\name{es_cohen_d_ow}
\alias{es_cohen_d_ow}
\title{Cohen d (for one-way ANOVA)}
\usage{
es_cohen_d_ow(nomField, scaleField, categories = NULL)
}
\arguments{
\item{nomField}{the groups variable}

\item{scaleField}{the numeric scores variable}

\item{categories}{vector, optional. the categories to use from catField}
}
\value{
the Cohen d value
}
\description{
An effect size measure for a one-way ANOVA. It simply compares the largest possible difference between two categories means and divides this over the total variance.

Note that most often Cohen d is reported with pairwise tests, but that is actually Cohen d_z. That version is available using es_cohen_d_ps().
}
\details{
The formula used is (Cohen, 1988, p. 276):
\deqn{d = \frac{\bar{x}_{max} - \bar{x}_{min}}{\sigma}}
With:
\deqn{\sigma = \sqrt{\frac{SS_w}{n}}}
\deqn{SS_w = \sum_{j=1}^k \sum_{i=1}^{n_j} \left(x_{i,j}-\bar{x}_j\right)^2}
\deqn{\bar{x}_j = \frac{\sum_{i=1}^{n_j} x_{i,j}}{n_j}}

\emph{Symbols}
\itemize{
\item \eqn{x_{i,j}} the i-th score in category j
\item \eqn{n_j} the number of scores in category j
\item \eqn{k} the number of categories
\item \eqn{\bar{x}_j} the mean of the scores in category j
\item \eqn{SS_w} the within sum of squares (sum of squared deviation of the mean)
}
}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
