% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_f.R
\name{es_cohen_f}
\alias{es_cohen_f}
\title{Cohen f}
\usage{
es_cohen_f(nomField, scaleField, categories = NULL, useRanks = FALSE)
}
\arguments{
\item{nomField}{the groups variable}

\item{scaleField}{the numeric scores variable}

\item{categories}{vector, optional. the categories to use from catField}

\item{useRanks}{boolean, optional. Use of ranks or original scores. Default is FALSE}
}
\value{
the Cohen f value
}
\description{
An effect size measure for regression analysis or an ANOVA test. It gives roughly the proportion of variance explained by the categorical variable.

The Cohen f is often used with ANOVA, while Cohen f-squared with regression.
}
\details{
The formula used (Cohen, 1988, p. 284):
\deqn{f = \sqrt{\frac{\eta^2}{1-\eta^2}}}

Where \eqn{\eta^2} is the value of eta-squared.

It can also be calculated using (Cohen, 1988, p. 371):
\deqn{f = \frac{\sigma_{\mu}}{\sigma}}

With:
\deqn{\sigma_{\mu} = \sqrt{\frac{SS_b}{n}}}
\deqn{\sigma = \sqrt{\frac{SS_w}{n}}}

Where \eqn{SS_i} is the sum of squared differences, see the Fisher one-way ANOVA for details on how to calculate these.

The \eqn{f^2} can be found in Cohen (1988, p. 410).

\strong{Conversions}

Cohen f can be converted to eta-squared using: \emph{es_convert(f, from="cohenf", to="etasq")}

\strong{Alternatives}

\emph{library(effectsize)}

anova_stats(aov(scores~groups))

cohens_f(aov(scores~groups))
}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
