% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_g.R
\name{es_cohen_g}
\alias{es_cohen_g}
\title{Cohen's g}
\usage{
es_cohen_g(data, codes = NULL)
}
\arguments{
\item{data}{vector with the data}

\item{codes}{optional vector with the two codes to use}
}
\value{
Cohen's g as a single value
}
\description{
Cohen’s g (Cohen, 1988) is an effect size measure that could be accompanying a one-sample binomial (see Rosnow & Rosenthal, 2003),
score or Wald test. It is simply the difference of the sample proportion with 0.5.

A video explanation of Cohen g can be found \href{https://youtu.be/tPZMvB8QrM0}{here on YouTube}
}
\details{
The formula used is (Cohen, 1988, p. 147):
\deqn{g=p-0.5}

\emph{Symbols used}:
\itemize{
\item \eqn{p} is the sample proportion
}
}
\examples{
#Example 1: Numeric list
ex1 = c(1, 1, 2, 1, 2, 1, 2, 1)
es_cohen_g(ex1)

#Example 2: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_cohen_g(df1['sex'])
#Using two specific categories:
es_cohen_g(df1['mar1'], codes=c("DIVORCED", "NEVER MARRIED"))

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\seealso{
\code{\link{th_cohen_g}}, rules-of-thumb for Cohen g
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
