% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_g.R
\name{es_cohen_g}
\alias{es_cohen_g}
\title{Cohen's g}
\usage{
es_cohen_g(data, p0Cat = NULL, codes = NULL)
}
\arguments{
\item{data}{vector with the data}

\item{p0Cat}{Optional the category for which p0=0.5 was used}

\item{codes}{Optional vector with the two codes to use}
}
\value{
Dataframe with:
\item{g for cat 1}{Cohen g for category 1}
\item{g for cat 2}{Cohen g for category 2}
}
\description{
Cohen's g (Cohen, 1988) is an effect size measure that could be accompanying a one-sample binomial (see Rosnow & Rosenthal, 2003),
score or Wald test. It is simply the difference of the sample proportion with 0.5.

A video explanation of Cohen g can be found \href{https://youtu.be/tPZMvB8QrM0}{here on YouTube}

This function is shown in this \href{https://youtu.be/3DEmngmws2U}{YouTube video} and the effect size is also described at \href{https://peterstatistics.com/Terms/EffectSizes/CohenG.html}{PeterStatistics.com}
}
\details{
To decide on which category is associated with p0 the following is used:
\itemize{
\item If codes are provided, the first code is assumed to be the category for the p0.
\item If p0Cat is specified that will be used for p0 and all other categories will be considered as category 2, this means if there are more than two categories the remaining two or more (besides p0Cat) will be merged as one large category.
\item If neither codes or p0Cat is specified and more than two categories are in the data a warning is printed and no results.
\item If neither codes or p0Cat is specified and there are two categories, p0 is assumed to be for the category closest matching the p0 value (i.e. if p0 is above 0.5 the category with the highest count is assumed to be used for p0)
}

The formula used is (Cohen, 1988, p. 147):
\deqn{g=p-0.5}

\emph{Symbols used}:
\itemize{
\item \eqn{p} is the sample proportion
}
}
\section{Before, After and Alternatives}{

Before this effect size you might first want to perform a test:
\code{\link{ts_binomial_os}}, for One-Sample Binomial Test
\code{\link{ts_score_os}}, for One-Sample Score Test
\code{\link{ts_wald_os}}, for One-Sample Wald Test

After this, you might want a rule-of-thumb:
\code{\link{th_cohen_g}}, for rules-of-thumb for Cohen g

Alternatives for this effect size could be:
\code{\link{es_cohen_h_os}}, for Cohen h'
\code{\link{es_alt_ratio}}, for Alternative Ratio
\code{\link{r_rosenthal}}, for Rosenthal Correlation if a z-value is available
}

\examples{
#Example 1: Numeric list
ex1 = c(1, 1, 2, 1, 2, 1, 2, 1)
es_cohen_g(ex1)

#Example 2: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_cohen_g(df1['sex'])
#Using two specific categories:
es_cohen_g(df1['mar1'], codes=c("DIVORCED", "NEVER MARRIED"))

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
