% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_g.R
\name{es_cohen_g}
\alias{es_cohen_g}
\title{Cohen's g}
\usage{
es_cohen_g(data, codes = NULL)
}
\arguments{
\item{data}{A vector with the data}

\item{codes}{Optional vector with the two codes to use}
}
\value{
Cohen's g as a single value
}
\description{
Cohen’s g (Cohen, 1988) is an effect size measure that could be accompanying a one-sample binomial (see Rosnow & Rosenthal, 2003),
score or Wald test. It is simply the difference of the sample proportion with 0.5.

A video explanation of Cohen g can be found \href{https://youtu.be/tPZMvB8QrM0}{here on YouTube}
}
\details{
The formula used is (Cohen, 1988, p. 147):
\deqn{g=p-0.5}

\emph{Symbols used}:
\itemize{
\item \eqn{p} is the sample proportion
}
}
\section{Alternatives}{


I'm not aware of any alternative library that has this function.
}

\examples{
data <- c("Female", "Male", "Male", "Female", "Male", "Male")
es_cohen_g(data)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\seealso{
This effect size could be used with a one-sample binomial test (\code{\link{ts_binomial_os}}),
score test (\code{\link{ts_score_os}}), or Wald test (\code{\link{ts_wald_os}}).

Alternatives for Cohen g are the Alternative Ratio (\code{\link{es_alt_ratio}}) or Cohen's h' (\code{\link{es_cohen_h_os}})

For a rule of thumb interpretation use \code{\link{th_cohen_g}}

\href{https://peterstatistics.com/CrashCourse/2-SingleVar/Binary/Binary-2b-EffectSize.html}{Companion Website} on Cohen g.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}
}
