% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_h_os.R
\name{es_cohen_h_os}
\alias{es_cohen_h_os}
\title{Cohen's h'}
\usage{
es_cohen_h_os(data, codes = NULL, p0 = 0.5)
}
\arguments{
\item{data}{a vector with the data}

\item{codes}{optional vector with the two codes to use}

\item{p0}{optional the hypothesized proportion for the first category (default is 0.5)}
}
\value{
Cohen's h'
}
\description{
An adaptation of Cohen h (\code{\link{es_cohen_h}}) for a one-sample case. It is an effect size measure that could
be accompanying a one-sample binomial, score or Wald test.

A \href{https://youtu.be/ddWe94VKX_8}{YouTube} video on Cohen h'.
}
\details{
Formula used (Cohen, 1988, p. 202):
\deqn{h'=\phi_{1}-\phi_{h_0}}
With:
\deqn{\phi_{i}=2\times\textup{arcsin}\sqrt{p_{i}}}
\deqn{p_i = \frac{F_i}{n}}
\deqn{n = \sum_{i=1}^k F_i}

\emph{Symbols used}:
\itemize{
\item \eqn{F_i} is the (absolute) frequency (count) of category i
\item \eqn{n} is the sample size, i.e. the sum of all frequencies
\item \eqn{p_i} the proportion of cases in category i
\item \eqn{p_{h_0}} the expected proportion (i.e. the proportion according to the null hypothesis)
}
}
\section{Alternatives}{


The \emph{'pwr'} library has a similar function: \emph{ES.h()}
}

\examples{
data <- c("Female", "Male", "Male", "Female", "Male", "Male")
es_cohen_h_os(data)

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\seealso{
\code{\link{es_convert}}, to convert Cohen h' to Cohen h, use fr="cohenhos" and to=cohenh

\code{\link{th_cohen_h}}, rules-of-thumb for Cohen g
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}
}
