% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_h_os.R
\name{es_cohen_h_os}
\alias{es_cohen_h_os}
\title{Cohen's h'}
\usage{
es_cohen_h_os(data, p0 = 0.5, p0Cat = NULL, codes = NULL)
}
\arguments{
\item{data}{a vector with the data}

\item{p0}{Optional hypothesized proportion for the first category (default is 0.5)}

\item{p0Cat}{Optional the category for which p0 was used}

\item{codes}{Optional vector with the two codes to use}
}
\value{
Dataframe with:
\item{Cohen h'}{the Cohen h' value}
\item{comment}{the category for which p0 was}
}
\description{
An adaptation of Cohen h (\code{\link{es_cohen_h}}) for a one-sample case. It is an effect size measure that could
be accompanying a one-sample binomial, score or Wald test.

A \href{https://youtu.be/ddWe94VKX_8}{YouTube} video on Cohen h'.
}
\details{
To decide on which category is associated with p0 the following is used:
\itemize{
\item If codes are provided, the first code is assumed to be the category for the p0.
\item If p0Cat is specified that will be used for p0 and all other categories will be considered as category 2, this means if there are more than two categories the remaining two or more (besides p0Cat) will be merged as one large category.
\item If neither codes or p0Cat is specified and more than two categories are in the data a warning is printed and no results.
\item If neither codes or p0Cat is specified and there are two categories, p0 is assumed to be for the category closest matching the p0 value (i.e. if p0 is above 0.5 the category with the highest count is assumed to be used for p0)
}

Formula used (Cohen, 1988, p. 202):
\deqn{h'=\phi_{1}-\phi_{h_0}}
With:
\deqn{\phi_{i}=2\times\textup{arcsin}\sqrt{p_{i}}}
\deqn{p_i = \frac{F_i}{n}}
\deqn{n = \sum_{i=1}^k F_i}

\emph{Symbols used}:
\itemize{
\item \eqn{F_i} is the (absolute) frequency (count) of category i
\item \eqn{n} is the sample size, i.e. the sum of all frequencies
\item \eqn{p_i} the proportion of cases in category i
\item \eqn{p_{h_0}} the expected proportion (i.e. the proportion according to the null hypothesis)
}
}
\section{Before, After and Alternatives}{

Before this effect size you might first want to perform a test:
\code{\link{ts_binomial_os}}, for One-Sample Binomial Test
\code{\link{ts_score_os}}, for One-Sample Score Test
\code{\link{ts_wald_os}}, for One-Sample Wald Test

After this, you might want a rule-of-thumb or first convert this to a 'regular' Cohen h:
\code{\link{es_convert}}, to convert Cohen h' to Cohen h, use fr="cohenhos" and to=cohenh
\code{\link{th_cohen_h}}, for rules-of-thumb for Cohen h

Alternatives for this effect size could be:
\code{\link{es_cohen_g}}, for Cohen g
\code{\link{es_alt_ratio}}, for Alternative Ratio
\code{\link{r_rosenthal}}, for Rosenthal Correlation if a z-value is available
}

\examples{
#Example 1: Numeric list
ex1 = c(1, 1, 2, 1, 2, 1, 2, 1)
es_cohen_h_os(ex1)
es_cohen_h_os(ex1, p0=0.3)

#Example 2: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_cohen_h_os(df1['sex'])

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
