% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_w.R
\name{es_cohen_w}
\alias{es_cohen_w}
\title{Cohen's w}
\usage{
es_cohen_w(chi2, n)
}
\arguments{
\item{chi2}{the chi-square test statistic}

\item{n}{the sample size}
}
\value{
value of Cohen's w
}
\description{
Cohen's w
}
\details{
The formula used is (Cohen, 1988, p. 216):
\deqn{w = \sqrt\frac{\chi_{GoF}^{2}}{n}}

\emph{Symbols used}:
\itemize{
\item \eqn{\chi_{GoF}^{2}} the Pearson chi-square goodness-of-fit value
\item \eqn{n} the sample size, i.e. the sum of all frequencies
}

The chi-square value could for example be obtained using the \emph{ts_pearson_gof()} function.

For a classification using some rule-of-thumb run \emph{th_cohen_w(w)}

\strong{Alternative}

The \emph{'rcompanion'} library also has a function for this: \emph{cohenW()}
}
\examples{
chi2Value <- 3.105263
n <- 19
es_cohen_w(chi2Value, n)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
