% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cohen_w.R
\name{es_cohen_w}
\alias{es_cohen_w}
\title{Cohen's w}
\usage{
es_cohen_w(chi2, n)
}
\arguments{
\item{chi2}{the chi-square test statistic}

\item{n}{the sample size}
}
\value{
value of Cohen's w
}
\description{
An effect size measure that could be used with a chi-square test. It has no upper limit, but can be compared to
Cohen's rules-of-thumb.
}
\details{
The formula used is (Cohen, 1988, p. 216):
\deqn{w = \sqrt\frac{\chi_{GoF}^{2}}{n}}

\emph{Symbols used}:
\itemize{
\item \eqn{\chi_{GoF}^{2}} the Pearson chi-square goodness-of-fit value
\item \eqn{n} the sample size, i.e. the sum of all frequencies
}
}
\section{Alternative}{


The \emph{'rcompanion'} library also has a function for this: \emph{cohenW()}
}

\examples{
chi2Value <- 3.105263
n <- 19
es_cohen_w(chi2Value, n)

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.
}
\seealso{
\code{\link{ts_pearson_gof}}, to obtain a chi-square value

\code{\link{th_cohen_w}}, rules-of-thumb for Cohen w
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
