% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cole_c5.R
\name{es_cole_c5}
\alias{es_cole_c5}
\title{Cole C5}
\usage{
es_cole_c5(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Cole C5
}
\description{
A measure of association between two binary variables.

This is actually an adjustment of the contingency coefficient, by dividing it over \eqn{\sqrt{\frac{1}{2}}}, which is the maximum value in 2x2 tables.
}
\details{
The formula used (Cole, 1949, p. 416):
\deqn{C_5 = \frac{\sqrt{2}\times\left(a\times d - b\times c\right)}{\sqrt{\left(a\times d - b\times c\right)^2 + R_1\times R_2\times C_1\times C_2}}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
\item \eqn{R_i} the sum of counts in the i-th row
\item \eqn{C_i} the sum of counts in the i-th column
}

Note that Cole C2 is the phi coefficient, Cole C3 is McEwen-Michael coefficient, Cole C4 is Yule Q, and Cole C6 is Yule r.
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_cole_c5(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Cole, L. C. (1949). The measurement of interspecific associaton. \emph{Ecology, 30}(4), 411–424. https://doi.org/10.2307/1932444
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
