% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_cole_c7.R
\name{es_cole_c7}
\alias{es_cole_c7}
\title{Cole C7 / Coefficient of Interspecific Association}
\usage{
es_cole_c7(var1, var2)
}
\arguments{
\item{var1}{A vector with the binary data from the first variable}

\item{var2}{A vector with the binary data from the second variable}
}
\value{
Cole C7
}
\description{
Cole C7 / Coefficient of Interspecific Association
}
\details{
The formula used is (Cole, 1949, pp. 420-421).
\deqn{C_7 = \begin{cases} \frac{a\times d - b\times c}{R_1\times C_2} & \text{ if } a\times d\geq b\times c \\ \frac{a\times d - b\times c}{R_1 \times C_1} & \text{ if } a\times d < b\times c \text{ and } a\leq d  \\ \frac{a\times d - b\times c}{R_2\times C_2} & \text{ if } a\times d < b\times c \text{ and } a >  d \end{cases}}

#' \emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
\item \eqn{R_i} the sum of counts in the i-th row
\item \eqn{C_i} the sum of counts in the i-th column
}

Cole refers to this as the 'coefficient of interspecific association'

Note that Cole C2 is the phi coefficient, Cole C3 is McEwen-Michael coefficient, Cole C4 is Yule Q, and Cole C6 is Yule r.
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
es_cole_c7(bin1, bin2)

}
\references{
Cole, L. C. (1949). The measurement of interspecific associaton. \emph{Ecology, 30}(4), 411–424. https://doi.org/10.2307/1932444
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
