% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_digby_h.R
\name{es_digby_h}
\alias{es_digby_h}
\title{Digby H}
\usage{
es_digby_h(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Digby H
}
\description{
A measure of association between two binary variables.

Yule Q and Yule Y can each be written in the format of:
\deqn{\frac{OR^x - 1}{OR^x + 1}}
With OR being the Odds Ratio. For Yule Q the \eqn{x=1} and for Yule Y \eqn{x=0.5}. Digby (1983, p. 754) showed that Yule’s Q consistently overestimates the association, while Yule’s Y underestimates it It seems that a better approximation might be somewhere between 0.5 and 1 as the power to use on the Odds Ratio. Digby proposed to use 0.75
}
\details{
The formula used (Digby, 1983, pp. 753-754):
\deqn{H = \frac{\left(a\times d\right)^{3/4} - \left(b\times c\right)^{3/4}}{\left(a\times d\right)^{3/4} + \left(b\times c\right)^{3/4}}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell
\item \eqn{b} the count in the top-right cell
\item \eqn{c} the count in the bottom-left cell
\item \eqn{d} the count in the bottom-right cell
}
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_digby_h(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Digby, P. G. N. (1983). Approximating the tetrachoric correlation coefficient. \emph{Biometrics, 39}(3), 753–757. https://doi.org/10.2307/2531104
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
