% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_digby_h.R
\name{es_digby_h}
\alias{es_digby_h}
\title{Digby H}
\usage{
es_digby_h(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Digby H
}
\description{
Digby H
}
\details{
The formula used (Digby, 1983, pp. 753-754):
\deqn{H = \frac{\left(a\times d\right)^{3/4} - \left(b\times c\right)^{3/4}}{\left(a\times d\right)^{3/4} + \left(b\times c\right)^{3/4}}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell
\item \eqn{b} the count in the top-right cell
\item \eqn{c} the count in the bottom-left cell
\item \eqn{d} the count in the bottom-right cell
}
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
es_digby_h(bin1, bin2)

}
\references{
Digby, P. G. N. (1983). Approximating the tetrachoric correlation coefficient. \emph{Biometrics, 39}(3), 753–757. https://doi.org/10.2307/2531104
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
