% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_edward_q.R
\name{es_edward_q}
\alias{es_edward_q}
\title{Edward Q}
\usage{
es_edward_q(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Edward Q
}
\description{
Edward Q
}
\details{
The formula used (Edwards 1957, as cited in Becker & Clogg, 1988, p. 408):
\deqn{Q_E = \frac{OR^{\frac{\pi}{4}} - 1}{OR^{\frac{\pi}{4}} + 1}}
With:
\deqn{OR = \frac{a\times d}{b \times c}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell
\item \eqn{b} the count in the top-right cell
\item \eqn{c} the count in the bottom-left cell
\item \eqn{d} the count in the bottom-right cell
\item \eqn{OR} the Odds Ratio
}
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
es_edward_q(bin1, bin2)

}
\references{
Becker, M. P., & Clogg, C. C. (1988). A note on approximating correlations from Odds Ratios. \emph{Sociological Methods & Research, 16}(3), 407–424. https://doi.org/10.1177/0049124188016003003

Edwards, J. H. (1957). A note on the practical interpretation of 2 x 2 tables. \emph{Journal of Epidemiology & Community Health, 11}(2), 73–78. https://doi.org/10.1136/jech.11.2.73
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
