% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_fei.R
\name{es_fei}
\alias{es_fei}
\title{Fei}
\usage{
es_fei(chi2, n, minExp)
}
\arguments{
\item{chi2}{the chi-square test statistic}

\item{n}{the sample size}

\item{minExp}{the minimum expected count}
}
\value{
the value of Fei
}
\description{
An effect size measure that could be used with a chi-square test or g-test.

Ben-Shachar et al. (2023) developed this effect size measure for goodness-of-fit tests. They divide the found chi-square test statistic by the maximum possible chi-square value, and then take the square root from this.

The measure is therefor simply the square root from Johnston-Berry-Mielke E. Although Ben-Shachar et al. reference the article from Johnston et al (2006), in personal communication with them (pers. comm. 2024) they indicated that it was a case of independent discovery and only later found that similar attempts were made as theirs that were very close.

Taking the square root makes sense, since as the name implies the chi-square is a squared value, and also other effect sizes use the square root, making Fei more aligned with those.

This function is shown in this \href{https://youtu.be/FFGlvMd13LM}{YouTube video} and the measure is also described at \href{https://peterstatistics.com/Terms/EffectSizes/Fei.html}{PeterStatistics.com}
}
\details{
The formula used (Ben-Shachar et al., 2023, p. 6):
\deqn{Fei = \\sqrt{\\frac{\\chi_{GoF}^2}{n\\times\\left(\\frac{1}{\\min\\left(p_E\\right)}-1\\right)}}}

\emph{Symbols used}
\itemize{
\item \eqn{\\chi_{GoF}^2}, the chi-square value of the goodness-of-fit chi-square test
\item \eqn{n}, the sample size
\item \eqn{p_E}, the expected proportions
}

\emph{Classification}
A qualification rule-of-thumb could be obtained by converting this to Cohen's w (use \strong{es_convert(Fei, fr="fei", to="cohenw", ex1=minExp/n)})
}
\section{Before, After and Alternatives}{

Before this you will need a chi-square value. From either:
\code{\link{ts_freeman_tukey_gof}}, for Freeman-Tukey Test of Goodness-of-Fit.
\code{\link{ts_freeman_tukey_read}}, for Freeman-Tukey-Read Test of Goodness-of-Fit.
\code{\link{ts_g_gof}}, for G (Likelihood Ratio) Goodness-of-Fit Test.
\code{\link{ts_mod_log_likelihood_gof}}, for Mod-Log Likelihood Test of Goodness-of-Fit.
\code{\link{ts_neyman_gof}}, for Neyman Test of Goodness-of-Fit.
\code{\link{ts_pearson_gof}}, for Pearson Test of Goodness-of-Fit.
\code{\link{ts_powerdivergence_gof}}, for Power Divergence GoF Test.
\code{\link{ph_pairwise_gof}} for Pairwise Goodness-of-Fit Tests.
\code{\link{ph_residual_gof_gof}} for Residuals Using Goodness-of-Fit Tests

After this you might want to use some rule-of-thumb for the interpretation by converting it to Cohen w:
\code{\link{es_convert}} to convert Fei to Cohen w (using fr="fei", to="cohenw", ex1=minExp/n).
\code{\link{th_cohen_w}} for various rules-of-thumb for Cohen w.

Alternative effect sizes that use a chi-square value:
\code{\link{es_cohen_w}}, for Cohen w.
\code{\link{es_cramer_v_gof}},  for Cramer's V for Goodness-of-Fit.
\code{\link{es_jbm_e}}, for Johnston-Berry-Mielke E.
}

\examples{
chi2 = 23.5
n = 53
minExp = 14
es_fei(chi2=chi2, n=n, minExp=minExp)

}
\references{
Ben-Shachar, M. S., Patil, I., Theriault, R., Wiernik, B. M., & Ludecke, D. (2023). Phi, fei, fo, fum: Effect sizes for categorical data that use the chi-squared statistic. \emph{Mathematics, 11}(1982), 1-10. doi:10.3390/math11091982

Johnston, J. E., Berry, K. J., & Mielke, P. W. (2006). Measures of effect size for chi-squared and likelihood-ratio goodness-of-fit tests. \emph{Perceptual and Motor Skills, 103}(2), 412-414. doi:10.2466/pms.103.2.412-414
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
