% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_forbes.R
\name{es_forbes}
\alias{es_forbes}
\title{Forbes Coefficient}
\usage{
es_forbes(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Forbes Coefficient
}
\description{
An effect size measure for two binary variables.

A measure for the association between two binary variables. If all values are equal in the cross table, there is no association. Forbes uses that if all values are the same, then:
\deqn{1 = \frac{\left(a+b\right)\times\left(a+c\right)}{n}}

While the above equation would result in 0 or 2 if there is a perfect association (i.e. b or c is 0).
}
\details{
The formula used (Forbes, 1907, p. 279):
\deqn{F = \frac{n\times\min\left(a, d\right)}{C_1\times R_1}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
\item \eqn{R_1} the sum of counts in the 1st row
\item \eqn{C_1} the sum of counts in the 1st column
}
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_forbes(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Forbes, S. A. (1907). On the local distribution of certain Illinois fishes: An essay in statistical ecology. \emph{Illinois Natural History Survey Bulletin, 7}(8), 273–303.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
