% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_glass_delta.R
\name{es_glass_delta}
\alias{es_glass_delta}
\title{Glass Delta}
\usage{
es_glass_delta(
  catField,
  scaleField,
  categories = NULL,
  dmu = 0,
  control = NULL
)
}
\arguments{
\item{catField}{A vector with the categorical data}

\item{scaleField}{A vector with the scores}

\item{categories}{Optional to indicate which two categories of catField to use, otherwise first two found will be used.}

\item{dmu}{Optional difference according to null hypothesis (default is 0)}

\item{control}{Optional to indicate which category to use as control group. Default is first category found.}
}
\value{
Glass Delata value
}
\description{
An effect size measure when comparing two means, with a specified control group.
}
\details{
The formula used is (Glass, 1976, p. 7):
\deqn{\delta = \frac{\bar{x}_1 - \bar{x}_2}{s_2}}

With:
\deqn{s_2 = \sqrt{\frac{\sum_{i=1}^{n_2} \left(x_{2,i} - \bar{x}_2\right)^2}{n_2 - 1}}}
\deqn{\bar{x}_i = \frac{\sum_{j=1}^{n_i} x_{i,j}}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{i,j}} the j-th score in category i
\item \eqn{n_i} the number of scores in category i
}

Glass actually uses a ‘control group’ and \eqn{s_2} is then the standard deviation of the control group.
}
\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = df1['age']
ex1 = replace(ex1, ex1=="89 OR OLDER", "90")
es_glass_delta(df1['sex'], ex1)

#Example 2: vectors
scores = c(20,50,80,15,40,85,30,45,70,60, NA, 90,25,40,70,65, NA, 70,98,40)
groups = c("nat","int","int","nat","int","int","nat","nat","int","int",
"int","int","int","int","nat","int",NA,"nat","int","int")
es_glass_delta(groups, scores)

}
\references{
Glass, G. V. (1976). Primary, secondary, and meta-analysis of research. \emph{Educational Researcher, 5}(10), 3–8. https://doi.org/10.3102/0013189X005010003
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
