% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_mcewen_michael.R
\name{es_mcewen_michael}
\alias{es_mcewen_michael}
\title{McEwen-Michael Coefficient / Cole C3}
\usage{
es_mcewen_michael(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
McEwen-Michael Coefficient
}
\description{
A measure of association between two binary variables.

A problem with measures that use the Forbes coefficient or Odds Ratio is that if only one cell is very large compared to the others, or 0, the association will be quite large (close to -1 or 1). Michael and McEwen attempted to overcome this by adding a correction.

Note that Cole (1949, p. 415) refers to this as C3. Cole (1949, p. 417) critized this approach and proposed some alternatives himself.
}
\details{
The formula used (Michael, 1920, p. 57)):
\deqn{\frac{a\times d - b\times c}{\left(\frac{a + d}{2}\right)^2 + \left(\frac{b + c}{2}\right)^2}}
\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
}
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_mcewen_michael(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Cole, L. C. (1949). The measurement of interspecific associaton. \emph{Ecology, 30}(4), 411–424. https://doi.org/10.2307/1932444

Michael, E. L. (1920). Marine Ecology and the coefficient of association: A plea in behalf of quantitative biology. \emph{Journal of Ecology, 8}(1), 54–59. https://doi.org/10.2307/2255213
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
