% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_pearson_q1.R
\name{es_pearson_q1}
\alias{es_pearson_q1}
\title{Pearson Q1}
\usage{
es_pearson_q1(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Pearson Q1
}
\description{
An effect size measure for two binary variables.

Note that Pearson (1900) stated: "Q1 was found of little service" (p. 16).
}
\details{
The formula used (Pearson, 1900, p. 15):
\deqn{Q_1 = \sin\left(\frac{\pi}{2} \times \frac{a\times d - b\times c}{\left(a+b\right)\times\left(b+d\right)}\right)}

With:
The cross table arranged such that \eqn{a\times d > b\times c}, \eqn{a>d}, and \eqn{c>b}.

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
}

Note that Pearson Q2 is the same as Yule Q and Pearson Q3 is the same as Yule r.
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_pearson_q1(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Pearson, K. (1900). Mathematical contributions to the theory of evolution. VII. On the correlation of characters not quantitatively measurable. \emph{Philosophical Transactions of the Royal Society of London, 195}, 1–405. https://doi.org/10.1098/rsta.1900.0022
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
