% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_pearson_q1.R
\name{es_pearson_q1}
\alias{es_pearson_q1}
\title{Pearson Q1}
\usage{
es_pearson_q1(var1, var2)
}
\arguments{
\item{var1}{A vector with the binary data from the first variable}

\item{var2}{A vector with the binary data from the second variable}
}
\value{
Pearson Q1
}
\description{
Pearson Q1
}
\details{
The formula used (Pearson, 1900, p. 15):
\deqn{Q_1 = \sin\left(\frac{\pi}{2} \times \frac{a\times d - b\times c}{\left(a+b\right)\times\left(b+d\right)}\right)}

With:
The cross table arranged such that \eqn{a\times d > b\times c}, \eqn{a>d}, and \eqn{c>b}.

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
}

Note that Pearson (1900) stated: "Q1 was found of little service" (p. 16).

Note that Pearson Q2 is the same as Yule Q and Pearson Q3 is the same as Yule r.
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
es_pearson_q1(bin1, bin2)

}
\references{
Pearson, K. (1900). Mathematical contributions to the theory of evolution. VII. On the correlation of characters not quantitatively measurable. \emph{Philosophical Transactions of the Royal Society of London, 195}, 1–405. https://doi.org/10.1098/rsta.1900.0022
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
