% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_vargha_delaney_a.R
\name{es_vargha_delaney_a}
\alias{es_vargha_delaney_a}
\title{Vargha and Delaney A}
\usage{
es_vargha_delaney_a(dataVar, groupVar)
}
\arguments{
\item{var1}{A vector with the scores data}

\item{var2}{A vector with the group data}
}
\value{
Vargha and Delaney A value
}
\description{
Vargha and Delaney A
}
\details{
The formula used is (Vargha & Delaney, 2000, p. 107):
\deqn{A = \frac{1}{n_j}\times\left(\frac{R_i}{n_i} - \frac{n_i + 1}{2}\right)}

\emph{Symbols used:}
\itemize{
\item \eqn{n_i} the number of scores in category i
\item \eqn{R_i} the sum of the ranks in category i
}

This effect size is an adaptation of the Common Language effect size,
adapted for ordinal data.

It could also be calculated from the Mann-Whitney U value:
\deqn{A = \frac{U}{n_1\times n_2}}
}
\examples{
scores = c(5, 12, 3, 4, 6, 1, 11, 13, NA)
groups = c("A","A","A","B","B","B","B", NA, "C")
es_vargha_delaney_a(scores, groups)

}
\references{
Vargha, A., & Delaney, H. D. (2000). A critique and improvement of the CL common language effect size statistics of McGraw and Wong. \emph{Journal of Educational and Behavioral Statistics, 25}(2), 101–132. https://doi.org/10.3102/10769986025002101
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
