% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_yule_y.R
\name{es_yule_y}
\alias{es_yule_y}
\title{Yule's Y / Coefficient of Colligation}
\usage{
es_yule_y(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Yule's Y
}
\description{
A measure of association between two binary variables. This method is based on the Odds Ratio. It converts the sample data cross table to a similar one but having the diagonal values the same, while keeping the same odds ratio as the original.
}
\details{
The formula used (Yule, 1912, p. 592):
\deqn{Y = \frac{\sqrt{a\times d} - \sqrt{b\times c}}{\sqrt{a\times d} + \sqrt{b\times c}}}

\emph{Symbols used:}

\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
}

Yule Y can be converted to Yule Q for which rules-of-thumb (\emph{th_yule_q(q)}) can be used, or converted to an Odds Ratio using also with rules of thumb available via \emph{th_odds_ratio(or)}
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
es_yule_y(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Yule, G. U. (1912). On the methods of measuring association between two attributes. \emph{Journal of the Royal Statistical Society, 75}(6), 579–652. https://doi.org/10.2307/2340126
}
\seealso{
\code{\link{es_convert}}, to convert Yule Y to Yule Q or an Odds Ratio.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
