% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_yule_y.R
\name{es_yule_y}
\alias{es_yule_y}
\title{Yule's Y / Coefficient of Colligation}
\usage{
es_yule_y(field1, field2, categories1 = NULL, categories2 = NULL)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}
}
\value{
Yule's Y
}
\description{
Yule's Y / Coefficient of Colligation
}
\details{
The formula used (Yule, 1912, p. 592):
\deqn{Y = \frac{\sqrt{a\times d} - \sqrt{b\times c}}{\sqrt{a\times d} + \sqrt{b\times c}}}

\emph{Symbols used:}
\itemize{
\item \eqn{a} the count in the top-left cell of the cross table
\item \eqn{b} the count in the top-right cell of the cross table
\item \eqn{c} the count in the bottom-left cell of the cross table
\item \eqn{d} the count in the bottom-right cell of the cross table
}

Yule Y can be converted to Yule Q using: \emph{es_convert(q, "yuley", "yuleq")}

Rules-of-thumb for Yule Q can then be used for classification/qualification: \emph{th_yule_q(q)}

Yule Y can be converted to Odds Ratio using: \emph{es_convert(q, "yuley", "or")}

Rules-of-thumb for Odds Ratio can then be used for classification/qualification: \emph{th_odds_ratio(or)}
}
\examples{
bin1 <- c("female", "female","female","female","female","female","female","female", "female","female","female", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male", "male")
bin2 <- c("nl", "nl","nl","nl","nl","nl","nl","nl", "other", "other", "other","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl","nl", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other", "other")
es_yule_y(bin1, bin2)

}
\references{
Yule, G. U. (1912). On the methods of measuring association between two attributes. \emph{Journal of the Royal Statistical Society, 75}(6), 579–652. https://doi.org/10.2307/2340126
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
