% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_consensus.R
\name{me_consensus}
\alias{me_consensus}
\title{Consensus}
\usage{
me_consensus(data, levels = NULL)
}
\arguments{
\item{data}{a vector with the data}

\item{levels}{optional to indicate the categories in order if data is non-numeric}
}
\value{
cns the consensus score
}
\description{
The Consensus is a measure of agreement or dispersion for ordinal data. If there is no agreement the value is 0, and with full agreement 1.

This function is shown in this \href{https://youtu.be/81iJtQtZbGI}{YouTube video} and the measure is also described at \href{https://peterstatistics.com/Terms/Measures/Consensus.html}{PeterStatistics.com}
}
\details{
The formula used (Tastle et al., 2005, p. 98):
\deqn{\text{Cns}\left(X\right) = 1 + \sum_{i=1}^k p_i \log_2\left(1 - \frac{\left|i - \mu_X\right|}{d_X}\right)}
With:
\deqn{\mu_X = \frac{\sum_{i=1}^k i\times F_i}{n}}
\deqn{d_X = k - 1}
\deqn{p_i = \frac{F_i}{n}}

\emph{Symbols used:}
\itemize{
\item \eqn{F_i} the frequency (count) of the i-th category (after they have been sorted)
\item \eqn{n} the sample size
\item \eqn{k} the number of categories.
}
}
\section{Before, After and Alternatives}{

Before this measure you might want an impression using a frequency table or a visualisation:
\code{\link{tab_frequency}}, for a frequency table
\code{\link{vi_bar_stacked_single}}, or Single Stacked Bar-Chart.
\code{\link{vi_bar_dual_axis}}, for Dual-Axis Bar Chart.

After this you might want some other descriptive measures:
\code{\link{me_hodges_lehmann_os}}, for the Hodges-Lehmann Estimate (One-Sample).
\code{\link{me_median}}, for the Median.
\code{\link{me_quantiles}}, for Quantiles.
\code{\link{me_quartiles}}, for Quartiles / Hinges.
\code{\link{me_quartile_range}}, for Interquartile Range, Semi-Interquartile Range and Mid-Quartile Range.

or perform a test:
\code{\link{ts_sign_os}}, for One-Sample Sign Test.
\code{\link{ts_trinomial_os}}, for One-Sample Trinomial Test.
\code{\link{ts_wilcoxon_os}}, for One-Sample Wilcoxon Signed Rank Test.
}

\examples{
# Example 1: Dataframe
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
studentDf = read.csv(file2, sep=';', na.strings=c("", "NA"))
ex1 = studentDf[['Teach_Motivate']]
order = c("Fully Disagree", "Disagree", "Neither disagree nor agree", "Agree", "Fully agree")
me_consensus(ex1, levels=order)

#Example 2: Numeric data
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
me_consensus(ex2)

#Example 3: Text data with
ex3 = c("a", "b", "f", "d", "e", "c")
order = c("a", "b", "c", "d", "e", "f")
me_consensus(ex3, levels=order)

}
\references{
Tastle, W. J., & Wierman, M. J. (2007). Consensus and dissention: A measure of ordinal dispersion. \emph{International Journal of Approximate Reasoning, 45}(3), 531-545. https://doi.org/10.1016/j.ijar.2006.06.024
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
