% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_quartile_range.R
\name{me_quartile_range}
\alias{me_quartile_range}
\title{Interquartile Range, Semi-Interquartile Range and Mid-Quartile Range}
\usage{
me_quartile_range(
  data,
  levels = NULL,
  measure = c("iqr", "siqr", "qd", "mqr"),
  method = "cdf"
)
}
\arguments{
\item{data}{vector or dataframe with scores as numbers, or if text also provide levels}

\item{levels}{optional vector with levels in order}

\item{measure}{the specific measure to determine. Either \code{"iqr"} (default), \code{"siqr"}, \code{"qd"}, or \code{"mqr"}}

\item{method}{the method to use to determine the quartiles}
}
\value{
A dataframe with:
\item{Q1}{the first (lower) quartile}
\item{Q3}{the third (upper/higher) quartile}
\item{range}{the range determined}
}
\description{
Three different ranges that can be used with quartiles. The Interquartile Range (IQR) is simply the third minus the first quartile.

The Semi-Interquartile range (a.k.a. Quartile Deviation) divides the IQR by 2.

The Mid-Quartile Range adds the two quartiles and then divides by 2.

A special case is the H-spread, which is if Hinges are used the IQR. This can be obtained by
setting \emph{method="tukey"}.

The function uses the \emph{me_quartiles()} function and any of the methods from that function can be used.
}
\details{
The formula used for the Interquartile Range is:
\deqn{IQR = Q_3 - Q_1}
This can be obtained by setting \emph{range="iqr"}.

The IQR is mentioned in Galton (1881, p. 245) and the H-spread in Tukey (1977, p. 44).

The H-spread can be obtained by setting \emph{range="iqr"} and \emph{method="tukey"}.

The formula used for the Semi-Interquartile Range (Quartile Deviation) is (Yule, 1911, p. 147):
\deqn{SIQR = \frac{Q_3 - Q_1}{2}}
This can be obtained by setting \emph{range="siqr"} or \emph{range="qd"}.

The formula for the mid-quartile range used is:
\deqn{MQR = \frac{Q_3 + Q_1}{2}}
This can be obtained by setting \emph{range="mqr"}.

This formula can be found in Parzen (1980, p. 19), but there are probably older references.
}
\examples{
data = runif(n=10, min=1, max=50)
me_quartile_range(data)
me_quartile_range(data, method="tukey")
me_quartile_range(data, range="siqr")
me_quartile_range(data, range="mqr")

}
\references{
Galton, F. (1881). Report of the anthropometric committee. Report of the British Association for the Advancement of Science, 51, 225–272.

Parzen, E. (1980). \emph{Data modeling using quantile and density-quantile functions}. Institute of Statistics, Texas A&M University.

Tukey, J. W. (1977). \emph{Exploratory data analysis}. Addison-Wesley Pub. Co.

Yule, G. U. (1911). \emph{An introduction to the theory of statistics}. Charles Griffin.
}
\seealso{
\code{\link{me_quartiles}}, calculating quartiles
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}
}
