% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_binomial.R
\name{ph_binomial}
\alias{ph_binomial}
\title{Pairwise Binomial Test for Post-Hoc Analysis}
\usage{
ph_binomial(
  data,
  expCount = NULL,
  twoSidedMethod = "eqdist",
  posthoc = "bonferroni"
)
}
\arguments{
\item{data}{dataframe with scores}

\item{expCount}{optional pandas dataframe with categories and expected counts}

\item{twoSidedMethod}{method to use for determining two-sided p-value of binomial test}

\item{posthoc}{the correction to use, currently only "bonferroni" available}
}
\value{
a dataframe with:
\item{category 1}{the label of the first category}
\item{category 2}{the label of the second category}
\item{n1}{the sample size of the first category}
\item{n2}{the sample size of the second category}
\item{obs. prop. cat. 1}{the proportion in the sample of the first category}
\item{exp. prop. cat. 1}{the expected proportion for the first category}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
}
\description{
This function will perform a one-sample binomial test for each possible pair in the
data. It makes use of the ts_binomial_os() function.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}
}
