% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_pairwise_bin.R
\name{ph_pairwise_bin}
\alias{ph_pairwise_bin}
\title{Pairwise Binary Test for Post-Hoc Analysis}
\usage{
ph_pairwise_bin(
  data,
  test = "binomial",
  expCount = NULL,
  mtc = "bonferroni",
  ...
)
}
\arguments{
\item{data}{dataframe with scores}

\item{test}{{"binomial", "score", "wald"}, optional test to use for each pair}

\item{expCount}{optional dataframe with categories and expected counts}

\item{mtc}{optional string. Any of the methods available in p_adjust() to correct for multiple tests}

\item{...}{optional additional arguments for the specific test that are passed along.}
}
\value{
a dataframe with:
\item{category 1}{the label of the first category}
\item{category 2}{the label of the second category}
\item{n1}{the sample size of the first category}
\item{n2}{the sample size of the second category}
\item{n pair}{the sample size of of the pair}
\item{obs. prop. 1}{the proportion in the sample of the first category}
\item{exp. prop. 1}{the expected proportion for the first category}
\item{statistic}{the test statistic}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
\item{test}{description of the test used}
}
\description{
This function will perform a one-sample binary test for each possible pair in the data. This could either be a binomial, Wald or score test.

The unadjusted p-values and Bonferroni adjusted p-values are both determined.

This function is shown in this \href{https://youtu.be/Li_I2D52Kjk}{YouTube video} and the test is also described at \href{https://peterstatistics.com/Terms/Tests/PostHocAfterGoF.html}{PeterStatistics.com}
}
\section{Before, After and Alternatives}{

Before this an omnibus test might be helpful:
\code{\link{ts_freeman_tukey_gof}}, for Freeman-Tukey Test of Goodness-of-Fit.
\code{\link{ts_freeman_tukey_read}}, for Freeman-Tukey-Read Test of Goodness-of-Fit.
\code{\link{ts_g_gof}}, for G (Likelihood Ratio) Goodness-of-Fit Test.
\code{\link{ts_mod_log_likelihood_gof}}, for Mod-Log Likelihood Test of Goodness-of-Fit.
\code{\link{ts_neyman_gof}}, for Neyman Test of Goodness-of-Fit.
\code{\link{ts_pearson_gof}}, for Pearson Test of Goodness-of-Fit.
\code{\link{ts_powerdivergence_gof}}, for Power Divergence GoF Test.

After this you might want to add an effect size measure:
\code{\link{es_post_hoc_gof}} for various effect sizes.

Alternative post-hoc tests:
\code{\link{ph_pairwise_gof}} for Pairwise Goodness-of-Fit Tests.
\code{\link{ph_residual_gof_bin}} for Residuals Tests using Binary Tests
\code{\link{ph_residual_gof_gof}} for Residuals Using Goodness-of-Fit Tests

The binary test that is performed on each pair:
\code{\link{ts_binomial_os}} for One-Sample Binomial Test.
\code{\link{ts_score_os}} for One-Sample Score Test.
\code{\link{ts_wald_os}} for One-Sample Wald Test.

More info on the adjustment for multiple testing:
\code{\link{p_adjust}}, various adjustment methods.
}

\examples{
# Examples: get data
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
gssDf <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = gssDf['mar1']

#Example 1 using default settings (one-sample binomial tests with equal-distance method)
ph_pairwise_bin(ex1)

#Example 2 using a score test with Yates correction:
ph_pairwise_bin(ex1, test="score", mtc='holm', cc='yates')

}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
