% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_pairwise_gof.R
\name{ph_pairwise_gof}
\alias{ph_pairwise_gof}
\title{Pairwise Goodness-of-Fit Tests for Post-Hoc Analysis}
\usage{
ph_pairwise_gof(
  data,
  test = "pearson",
  expCount = NULL,
  mtc = "bonferroni",
  ...
)
}
\arguments{
\item{data}{dataframe with scores}

\item{test}{{"pearson", "freeman-tukey", "freeman-tukey-read", "g", "mod-log-g", "neyman", "powerdivergence", "multinomial"}, optional test to use for each pair}

\item{expCount}{optional dataframe with categories and expected counts}

\item{mtc}{optional string. Any of the methods available in p_adjust() to correct for multiple tests}

\item{...}{optional additional arguments for the specific test that are passed along.}
}
\value{
a dataframe with:
\item{category 1}{the label of the first category}
\item{category 2}{the label of the second category}
\item{n1}{the sample size of the first category}
\item{n2}{the sample size of the second category}
\item{obs. prop. 1}{the observed proportion of category 1}
\item{exp. prop. 1}{the expected proportion of category 1}
\item{statistic}{the chi-square test statistic}
\item{df}{the degrees of freedom}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
\item{minExp}{the minimum expected count}
\item{propBelow5}{the proportion of cells with an expected count below 5}
\item{test}{description of the test used}

In case of a multinomial test, the same columns except there are no \emph{minExp} and \emph{propBelow5} columns and:
\item{p obs}{instead of \emph{statistic}, showing the probability of the observed sample table}
\item{n combs.}{instead of \emph{df}, showing the number of possible tables}
}
\description{
This function will perform a goodness-of-fit test for each possible pair in the data. This could be any of the goodness-of-fit tests, e.g. a Pearson chi-square.

The unadjusted p-values and Bonferroni adjusted p-values are both determined.
}
\seealso{
\code{\link{ts_multinomial_gof}}
\code{\link{ts_powerdivergence_gof}}
\code{\link{ts_neyman_gof}}
\code{\link{ts_mod_log_likelihood_gof}}
\code{\link{ts_g_gof}}
\code{\link{ts_freeman_tukey_read}}
\code{\link{ts_freeman_tukey_gof}}
\code{\link{ts_pearson_gof}}
\code{\link{p_adjust}}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
