% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_residual_gof_bin.R
\name{ph_residual_gof_bin}
\alias{ph_residual_gof_bin}
\title{Post-Hoc Residuals Using Binary Tests for GoF}
\usage{
ph_residual_gof_bin(data, test = "adj-residual", expCount = NULL, ...)
}
\arguments{
\item{data}{dataframe with scores}

\item{test}{{"adj-residual", "binomial", "wald", "score"} optional test to use}

\item{expCount}{optional dataframe with categories and expected counts}

\item{...}{optional additional parameters to be passed to the test}
}
\value{
a dataframe with:
\item{category}{the label of the first category}
\item{obs. count}{the observed count}
\item{exp. count}{the expected count}
\item{statistic}{the test statistic}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
\item{test}{description of the test used}
}
\description{
This function will perform a goodness-of-fit test using each category and collapsing the other categories. It also includes z-test on the standardized residuals.

The unadjusted p-values and Bonferroni adjusted p-values are both determined.
}
\details{
The formula used for the residual test:
\deqn{z = \frac{F_i - E_i}{\sqrt{E_i}}}
\deqn{sig = 2\times\left(1 - \Phi\left(\left|z\right|\right)\right)}

With:
\itemize{
\item \eqn{F_i}, the observed count for category $i$
\item \eqn{E_i}, the expected count for category $i$
\item \eqn{\Phi\left(\dots\right)}, the cumulative distribution function of the standard normal distribution
}

If no expected counts are provide it is assumed they are all equal for each category, i.e. \eqn{E_i = \frac{n}{k}}

The Bonferroni adjustment is calculated using:
\deqn{p_{adj} = \min \left(p \times k, 1\right)}

The other tests use the formula from the one-sample test variant, using the expected count/n as the expected proportion.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
