% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_residual_gof_bin.R
\name{ph_residual_gof_bin}
\alias{ph_residual_gof_bin}
\title{Post-Hoc Residuals Using Binary Tests for GoF}
\usage{
ph_residual_gof_bin(
  data,
  test = "std-residual",
  expCount = NULL,
  mtc = "bonferroni",
  ...
)
}
\arguments{
\item{data}{dataframe with scores}

\item{test}{c("adj-residual", "std-residual", "binomial", "wald", "score") optional test to use}

\item{expCount}{optional dataframe with categories and expected counts}

\item{mtc}{optional string. Any of the methods available in p_adjust() to correct for multiple tests}

\item{...}{optional additional parameters to be passed to the test}
}
\value{
a dataframe with:
\item{category}{the label of the first category}
\item{obs. count}{the observed count}
\item{exp. count}{the expected count}
\item{statistic}{the test statistic}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
\item{test}{description of the test used}
}
\description{
This function will perform a residuals post-hoc test for each of the categories in a nominal field. This could either be a z-test using the standardized residuals, the adjusted residuals, or any of the one-sample binary tests.

The unadjusted p-values and Bonferroni adjusted p-values are both determined.

This function is shown in this \href{https://youtu.be/mioMqScuhZ8}{YouTube video} and test is also described at \href{https://peterstatistics.com/Terms/Tests/PostHocAfterGoF.html}{PeterStatistics.com}
}
\details{
The formula used is for the adjusted residual test:
\deqn{z = \frac{F_i - E_i}{\sqrt{E_i\times\left(1 - \frac{E_i}{n}\right)}}}
\deqn{sig = 2\times\left(1 - \Phi\left(\left|z\right|\right)\right)}

The formula used for the standardized residual test:
\deqn{z = \frac{F_i - E_i}{\sqrt{E_i}}}
\deqn{sig = 2\times\left(1 - \Phi\left(\left|z\right|\right)\right)}

With:
\itemize{
\item \eqn{F_i}, the observed count for category $i$
\item \eqn{E_i}, the expected count for category $i$
\item \eqn{\Phi\left(\dots\right)}, the cumulative distribution function of the standard normal distribution
}

If no expected counts are provide it is assumed they are all equal for each category, i.e. \eqn{E_i = \frac{n}{k}}

The Bonferroni adjustment is calculated using:
\deqn{p_{adj} = \min \left(p \times k, 1\right)}

The other tests use the formula from the one-sample test variant, using the expected count/n as the expected proportion.

The adjusted residuals will gave the same result as using a one-sample score test. Some sources will also call these adjusted residuals as standardized residuals (Agresti, 2007, p. 38), and the standardized residuals used in this function as Pearson residuals (R, n.d.). Haberman (1973, p. 205) and Sharpe (2015, p. 3) are sources for the terminology used in this function.
}
\section{Before, After and Alternatives}{

Before this an omnibus test might be helpful:
\code{\link{ts_freeman_tukey_gof}}, for Freeman-Tukey Test of Goodness-of-Fit.
\code{\link{ts_freeman_tukey_read}}, for Freeman-Tukey-Read Test of Goodness-of-Fit.
\code{\link{ts_g_gof}}, for G (Likelihood Ratio) Goodness-of-Fit Test.
\code{\link{ts_mod_log_likelihood_gof}}, for Mod-Log Likelihood Test of Goodness-of-Fit.
\code{\link{ts_neyman_gof}}, for Neyman Test of Goodness-of-Fit.
\code{\link{ts_pearson_gof}}, for Pearson Test of Goodness-of-Fit.
\code{\link{ts_powerdivergence_gof}}, for Power Divergence GoF Test.

After this you might want to add an effect size measure:
\code{\link{es_post_hoc_gof}} for various effect sizes.

Alternative post-hoc tests:
\code{\link{ph_pairwise_bin}} for Pairwise Binary Tests.
\code{\link{ph_pairwise_gof}} for Pairwise Goodness-of-Fit Tests.
\code{\link{ph_residual_gof_gof}} for Residuals Using Goodness-of-Fit Tests

The binary test that is performed on each category:
\code{\link{ts_binomial_os}} for One-Sample Binomial Test.
\code{\link{ts_score_os}} for One-Sample Score Test.
\code{\link{ts_wald_os}} for One-Sample Wald Test.

More info on the adjustment for multiple testing:
\code{\link{p_adjust}}, various adjustment methods.
}

\examples{
# Examples: get data
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
gssDf <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = gssDf['mar1']

#Example 1 using default settings
ph_residual_gof_bin(ex1)

#Example 2 using a binomial test and Holm correction:
ph_residual_gof_bin(ex1, test="binomial", mtc='holm')

}
\references{
Agresti, A. (2007). \emph{An introduction to categorical data analysis} (2nd ed.). Wiley-Interscience.
Haberman, S. J. (1973). The analysis of residuals in cross-classified tables. \emph{Biometrics, 29}(1), 205-220. doi:10.2307/2529686
R. (n.d.). Chisq.test [Computer software]. https://stat.ethz.ch/R-manual/R-devel/library/stats/html/chisq.test.html
Sharpe, D. (2015). Your chi-square test is statistically significant: Now what? Practical Assessment, \emph{Research & Evaluation, 20}(8), 1-10.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
