% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poho_residual_gof_gof.R
\name{ph_residual_gof_gof}
\alias{ph_residual_gof_gof}
\title{Post-Hoc Residuals Using GoF for GoF}
\usage{
ph_residual_gof_gof(
  data,
  test = "pearson",
  expCount = NULL,
  mtc = "bonferroni",
  ...
)
}
\arguments{
\item{data}{dataframe with scores}

\item{test}{{"pearson", "freeman-tukey", "freeman-tukey-read", "g", "mod-log-g", "neyman", "powerdivergence", "multinomial"} optional test to use}

\item{expCount}{optional dataframe with categories and expected counts}

\item{mtc}{optional string. Any of the methods available in p_adjust() to correct for multiple tests}

\item{...}{optional additional parameters to be passed to the test}
}
\value{
a dataframe with:
\item{category}{the label of the first category}
\item{obs. count}{the observed count of the category}
\item{exp. count}{the expected count of the category}
\item{statistic}{the chi-square test statistic}
\item{df}{the degrees of freedom}
\item{p-value}{the unadjusted significance}
\item{adj. p-value}{the adjusted significance}
\item{minExp}{the minimum expected count}
\item{propBelow5}{the proportion of cells with an expected count below 5}
\item{test}{description of the test used}

In case of a multinomial test, the same columns except there are no \emph{minExp} and \emph{propBelow5} columns and:
\item{p obs}{instead of \emph{statistic}, showing the probability of the observed sample table}
\item{n combs.}{instead of \emph{df}, showing the number of possible tables}
}
\description{
This function will perform a goodness-of-fit test using each category and collapsing the other categories.

The unadjusted p-values and Bonferroni adjusted p-values are both determined.
}
\seealso{
\code{\link{ts_multinomial_gof}}
\code{\link{ts_powerdivergence_gof}}
\code{\link{ts_neyman_gof}}
\code{\link{ts_mod_log_likelihood_gof}}
\code{\link{ts_g_gof}}
\code{\link{ts_freeman_tukey_read}}
\code{\link{ts_freeman_tukey_gof}}
\code{\link{ts_pearson_gof}}
\code{\link{ps_adjust}}
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
