% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_goodman_kruskal_gamma.R
\name{r_goodman_kruskal_gamma}
\alias{r_goodman_kruskal_gamma}
\title{Goodman-Kruskal Gamma}
\usage{
r_goodman_kruskal_gamma(
  ordField1,
  ordField2,
  levels1 = NULL,
  levels2 = NULL,
  ase = "appr",
  useRanks = FALSE
)
}
\arguments{
\item{ordField1}{the numeric scores of the first variable}

\item{ordField2}{the numeric scores of the second variable}

\item{levels1}{vector, optional. the categories to use from ordField1}

\item{levels2}{vector, optional. the categories to use from ordField2}

\item{ase}{optional. Which asymptotic standard error to use. Either "appr" (default), 0, 1}

\item{useRanks}{boolean, optional. rank the data first or not. Default is False}
}
\value{
A dataframe with:
\item{g}{the Goodman-Kruskal Gamma value}
\item{statistic}{the z-value used for the test}
\item{pValue}{the significance (p-value)}
}
\description{
A rank correlation coefficient. It ranges from -1 (perfect negative association) to 1 (perfect positive association). A zero would indicate no correlation at all.

A positive correlation indicates that if someone scored high on the first field, they also likely score high on the second, while a negative correlation would indicate a high score on the first would give a low score on the second.

Alternatives for Gamma are Kendall Tau, Stuart-Kendall Tau and Somers D, but also Spearman rho could be considered.

Gamma looks at so-called discordant and concordant pairs, and ignores tied pairs. Kendall Tau b does the same, but applies a correction for ties. Stuart-Kendall Tau c also, but also takes the size of the table into consideration. Somers d only makes a correction for tied pairs in one of the two directions. Spearman rho is more of a variation on Pearson correlation, but applied to ranks. See Göktaş and İşçi. (2011) for more information on the comparisons.
}
\details{
The formula used (Goodman & Kruskal, 1954, p. 749):
\deqn{\gamma = \frac{P - Q}{P + Q}}
With:
\deqn{P = \sum_{i=1}^r \sum_{j=1}^c P_{i,j}}
\deqn{Q = \sum_{i=1}^r \sum_{j=1}^c Q_{i,j}}
\deqn{P_{i,j} = F_{i,j}\times C_{i,j}}
\deqn{Q_{i,j} = F_{i,j}\times D_{i,j}}
\deqn{C_{i,j} = \sum_{h<i}\sum_{k<j} F_{h,k} + \sum_{h>i}\sum_{k>j} F_{h,k}}
\deqn{D_{i,j} = \sum_{h<i}\sum_{k>j} F_{h,k} + \sum_{h>i}\sum_{k<j} F_{h,k}}

The test can be done with a generic approximation:
\deqn{z_{\gamma} = \gamma\times\sqrt{\frac{P+Q}{n\times\left(1-\gamma^2\right)}}}

If we assume the alternative hypothesis we can obtain (Goodman & Kruskal, 1963, p. 324; Goodman & Kruskal, 1972, p. 416; Brown & Benedetti, 1977, p. 310):
\deqn{z_{\gamma} = \frac{\gamma}{ASE_1}}
\deqn{ASE_1 = \frac{4}{\left(P+Q\right)^2}\times\sqrt{\sum_{i=1}^r\sum_{j=1}^c F_{i,j}\times\left(Q\times C_{i,j}-P\times D_{i,j}\right)^2}}

While if we assume the null hypothesis we can obtain (Brown & Benedetti, 1977, p. 311):
\deqn{z_{\gamma} = \frac{\gamma}{ASE_0}}
\deqn{ASE_0 = \frac{2}{P+Q}\times\sqrt{\sum_{i=1}^r\sum_{j=1}^c F_{i,j}\times\left(C_{i,j}- D_{i,j}\right)^2 - \frac{\left(P-Q\right)^2}{n}}}

The significance (p-value) in each case is then determined using:
\deqn{sig. = 2\times\left(1 - \Phi\left(\left|z_{gamma}\right|\right)\right)}

\emph{Symbols}

\itemize{
\item \eqn{F_{i,j}} the count of scores equal to i in the first variable and j in the second
\item \eqn{n} the total sample size
\item \eqn{r} the number of unique categories in the first variable (number of rows)
\item \eqn{c} the number of unique categories in the second variable (number of columns)
\item \eqn{P} double the number of concordant pairs
\item \eqn{Q} double the number of discordant pairs
}

Note that Kendall \eqn{\tau_a} is the same as Goodman-Kruskall gamma.

\strong{Alternatives}

\emph{library(DescTools)}

GoodmanKruskalGamma(table(ord1, ord2), conf.level=0.95)

\emph{library(MESS)}

gkgamma(table(ord1, ord2))

\emph{library(ryouready)}

ord.gamma(table(ord1, ord2))
}
\references{
Brown, M. B., & Benedetti, J. K. (1977). Sampling behavior of test for correlation in two-way contingency tables. \emph{Journal of the American Statistical Association, 72}(358), 309–315. doi:10.2307/2286793

Göktaş, A., & İşçi, Ö. (2011). A comparison of the most commonly used measures of association for doubly ordered square contingency tables via simulation. \emph{Advances in Methodology and Statistics, 8}(1). doi:10.51936/milh5641

Goodman, L. A., & Kruskal, W. H. (1954). Measures of association for cross classifications. J\emph{ournal of the American Statistical Association, 49}(268), 732–764. doi:10.2307/2281536

Goodman, L. A., & Kruskal, W. H. (1963). Measures of Association for Cross Classifications III: Approximate Sampling Theory. \emph{Journal of the American Statistical Association, 58}(302), 310–364. doi:10.1080/01621459.1963.10500850

Goodman, L. A., & Kruskal, W. H. (1972). Measures of Association for Cross Classifications IV: Simplification of Asymptotic Variances. \emph{Journal of the American Statistical Association, 67}(338), 415–421. doi:10.1080/01621459.1972.10482401
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
