% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rank_biserial_is.R
\name{r_rank_biserial_is}
\alias{r_rank_biserial_is}
\title{(Glass) Rank Biserial Correlation / Cliff Delta}
\usage{
r_rank_biserial_is(catField, ordField, categories = NULL, levels = NULL)
}
\arguments{
\item{catField}{A vector with the scores data}

\item{ordField}{A vector with the group data}

\item{categories}{optional vector with categories to use and order for the categorical field. Otherwise the first two found will be used.}

\item{levels}{optional vector with the labels of the ordinal field in order.}
}
\value{
(Glass) Rank Biserial Correlation / Cliff Delta value
}
\description{
This function will calculate Rank biserial correlation coefficient (independent-samples)
}
\details{
The formula used is (Glass, 1966, p. 626):
\deqn{r_b = \frac{2\times\left(\bar{R}_1 - \bar{R}_2\right)}{n}}
With:
\deqn{\bar{R}_i=\frac{R_i}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{\bar{R}_i} the average of ranks in category i
\item \eqn{R_i} the sum of ranks in category i
\item \eqn{n} the total sample size
\item \eqn{n_i} the number of scores in category i
}

Glass (1966) showed that the formula was the same as that of the rank biserial from Cureton (1956). Cliff's delta (Cliff, 1993, p. 495) is actually also the same.

The rank biserial can be converted to a Cohen d (using the \strong{es_convert()} function), and then the rules-of-thumb for Cohen d could be used (\strong{th_cohen_d()})
}
\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
myLevels = c('Not scientific at all', 'Not too scientific', 'Pretty scientific', 'Very scientific')
r_rank_biserial_is(df1['sex'], df1['accntsci'], levels=myLevels)

#Example 2: vectors
binary = c("apple", "apple", "apple", "peer", "peer", "peer", "peer")
ordinal = c(4, 3, 1, 6, 5, 7, 2)
r_rank_biserial_is(binary, ordinal, categories=c("peer", "apple"))

}
\references{
Cliff, N. (1993). Dominance statistics: Ordinal analyses to answer ordinal questions. \emph{Psychological Bulletin, 114}(3), 494-509. https://doi.org/10.1037/0033-2909.114.3.494

Cureton, E. E. (1956). Rank-biserial correlation. \emph{Psychometrika, 21}(3), 287-290. https://doi.org/10.1007/BF02289138

Glass, G. V. (1966). Note on rank biserial correlation. \emph{Educational and Psychological Measurement, 26}(3), 623-631. https://doi.org/10.1177/001316446602600307
}
\seealso{
\code{\link{es_convert}}, to convert to Cohen d, use \verb{fr="rb", to="cohend"}.

\code{\link{th_cohen_d}}, rules of thumb for Cohen d
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
