% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rosenthal.R
\name{r_rosenthal}
\alias{r_rosenthal}
\title{Rosenthal Correlation Coefficient}
\usage{
r_rosenthal(zVal, n)
}
\arguments{
\item{zVal}{z-value of test}

\item{n}{total sample size}
}
\value{
r the effect size measure
}
\description{
This function will calculate Rosenthal Correlation Coefficient. A simple correlation coefficient that divides a z-score by the square root of the sample size.
}
\details{
The formula used (Rosenthal, 1991, p. 19):
\deqn{r = \frac{z}{\sqrt{n}}}

\emph{Symbols used:}
\itemize{
\item \eqn{n} the sample size
\item \eqn{z} the calculated z-statistic value
}

Rosenthal (1991) is the oldest reference I could find for this correlation coefficient. However, Cohen (1988, p. 275) actually has a measure 'f' that has the same equation.

For a classification the same as for Pearson correlation use \emph{th_pearson_r()}
}
\examples{
z = 1.143943
n = 20
r_rosenthal(z, n)
 
}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.

Rosenthal, R. (1991). \emph{Meta-analytic procedures for social research} (Rev. ed). Sage Publications.
}
\seealso{
\code{\link{th_pearson_r}}, rules of thumb for a Pearson correlation coefficient.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
