% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_spearman_rho.R
\name{r_spearman_rho}
\alias{r_spearman_rho}
\title{Spearman Rho / Rank Correlation Coefficient}
\usage{
r_spearman_rho(
  ordField1,
  ordField2,
  levels1 = NULL,
  levels2 = NULL,
  test = c("t", "z-fieller", "z-olds", "iman-conover", "as89", "exact"),
  cc = FALSE
)
}
\arguments{
\item{ordField1}{the numeric scores of the first variable}

\item{ordField2}{the numeric scores of the second variable}

\item{levels1}{vector, optional. the categories to use from ordField1}

\item{levels2}{vector, optional. the categories to use from ordField2}

\item{test}{the test to be used. Either "t" (default), "as89", "exact", "iman-conover", "z-fieller", "z-olds", "none"}

\item{cc}{boolean to indicate the use of a continuity correction}
}
\value{
A dataframe with:
\item{rs}{the correlation coefficient}
\item{pValue}{the significance (p-value)}
\item{statistic}{the statistic from the test (only if applicable)}
\item{df}{the degrees of freedom (only if applicable)}
}
\description{
The Spearman Rank Correlation Coefficient is the Pearson Correlation Coefficient,
after the scores first have been converted to ranks.

This function makes use of \emph{di_spearman()} for the test of this correlation,
which requires the \emph{pspearman} library for exact computations.
}
\details{
The formula used is (Spearman, 1904, p. 77):
\deqn{r_s = \frac{SS_{r_x, r_y}}{SS_{r_x}\times SS_{r_y}}}
With:
\deqn{SS_{r_x} = \sum_{i=1}^n \left(r_{x_i} - \bar{r}_x\right)^2}
\deqn{SS_{r_y} = \sum_{i=1}^n \left(r_{y_i} - \bar{r}_y\right)^2}
\deqn{SS_{r_x, r_y} = \sum_{i=1}^n \left(r_{x_i} - \bar{r}_x\right) \times \left(r_{y_i} - \bar{r}_y\right)}

\emph{Symbols}

\itemize{
\item \eqn{r_{x_i}} the i-th rank of the scores of the first variable
\item \eqn{r_{y_i}} the i-th rank of the scores of the second variable
\item \eqn{n} the total sample size (number of ranks)
}

If all the ranks are distinct (i.e. no ties) the formula can also be written as:

\deqn{r_s = 1 - \frac{6}{n\times\left(n^2 - 1\right)}\times S}
With:
\deqn{S = \sum_{i=1}^n d_i^2}
\deqn{d_i^2 = \left(r_{x_i} - r_{y_i}\right)^2}

The test can be performed in different ways. Options to choose from are:
\itemize{
\item "t" uses a Student t distribution approximation
\item "z-fieller" uses a standard normal approximation from Fieller
\item "z-old" uses standard normal approximation from Old
\item "iman-conover" a combination of z and t distribution from Iman and Conover
\item "AS89" uses the AS 89 algorithm
\item "exact" uses an exact distribution
}

See for the details of each the \emph{di_spearman()} function

A continuity correction can be applied (Zar, 1972, p. 579):
\deqn{r_s^{cc} = \left|r_s\right| - \frac{6}{n^3 - n}}

\strong{Alternatives}

\emph{R's stats}

Using the t-approximation:

cor.test(ord1, ord2, method="spearman")

Using AS89

cor.test(ord1, ord2, method="spearman", exact=TRUE)

\emph{library(pspearman)}

spearman.test(ord1, ord2, approximation="t-distribution")

spearman.test(ord1, ord2, approximation="AS89")

spearman.test(ord1, ord2, approximation="exact")
}
\references{
Göktaş, A., & İşçi, Ö. (2011). A comparison of the most commonly used measures of association for doubly ordered square contingency tables via simulation. \emph{Advances in Methodology and Statistics, 8}(1). doi:10.51936/milh5641

Spearman, C. (1904). The proof and measurement of association between two things. \emph{The American Journal of Psychology, 15}(1), 72–101.

Zar, J. H. (1972). Significance testing of the Spearman rank correlation coefficient. \emph{Journal of the American Statistical Association, 67}(339), 578–580. doi:10.1080/01621459.1972.10481251
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
