% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tetrachoric.R
\name{r_tetrachoric}
\alias{r_tetrachoric}
\title{Tetrachoric Correlation Coefficient}
\usage{
r_tetrachoric(
  field1,
  field2,
  categories1 = NULL,
  categories2 = NULL,
  method = "search"
)
}
\arguments{
\item{field1}{: dataframe field with categories for the rows}

\item{field2}{: dataframe field with categories for the columns}

\item{categories1}{: optional list with selection and/or order for categories of field1}

\item{categories2}{: optional list with selection and/or order for categories of field2}

\item{method}{c("search", "kirk", "brown", "divgi") method to use (see details)}
}
\value{
Tetrachoric Correlation Coefficient
}
\description{
In essence this attempts to mimic a correlation coefficient between two scale variables. It can be defined as "An estimate of the correlation between two random variables having a bivariate normal distribution, obtained from the information from a double dichotomy of their bivariate distribution" (Everitt, 2004, p. 372).

This assumes the two binary variables have ‘hidden’ underlying normal distribution. If so, the combination of the two forms a bivariate normal distribution with a specific correlation between them. The quest is then to find the correlation, such that the cumulative density function of the z-values of the two marginal totals of the top-left cell (a) match that value.

This is quite tricky to do, so a few have proposed an approximation for this. These include Yule r (\code{\link{es_yule_r}}), Pearson Q4 (\code{\link{es_pearson_q4}}) and Q5 (\code{\link{es_pearson_q5}}), Camp (\code{\link{es_camp_r}}), Becker and Clogg (\code{\link{es_becker_clogg_r}}), and Bonett and Price (\code{\link{es_bonett_price_r}}).

Besides closed form approximation formula's, various algorithms have been designed as well. The three most often mentioned are Brown (1977), Kirk (1973), and Divgi (1979), available in this function.
}
\details{
The "search" method does a binary search for rt using the bivariate normal distribution
from the \emph{fMultivar} library.

"kirk" will use Kirk (1973) Fortran TET8 procedure, adapted by stikpet

"brown" will use Brown (1977) - Algorithm AS 116

"divgi" will use Divgi (1979) algorithm

Flow charts of these algorithms can be found at https://peterstatistics.com/CrashCourse/3-TwoVarUnpair/BinBin/BinBin-2b-EffectSize.html
}
\examples{
#Example: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
r_tetrachoric(df1[['mar1']], df1[['sex']], categories1=c("WIDOWED", "DIVORCED"))

}
\references{
Brown, M. B. (1977). Algorithm AS 116: The tetrachoric correlation and its asymptotic standard error. \emph{Applied Statistics, 26}(3), 343. https://doi.org/10.2307/2346985

Divgi, D. R. (1979). Calculation of the tetrachoric correlation coefficient. \emph{Psychometrika, 44}(2), 169–172. https://doi.org/10.1007/BF02293968

Kirk, D. B. (1973). On the numerical approximation of the bivariate normal (tetrachoric) correlation coefficient. \emph{Psychometrika, 38}(2), 259–268. https://doi.org/10.1007/BF02291118
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
