% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_frequency_bins.R
\name{tab_frequency_bins}
\alias{tab_frequency_bins}
\title{Binned Frequency Table}
\usage{
tab_frequency_bins(
  data,
  nbins = "sturges",
  bins = NULL,
  incl_lower = TRUE,
  adjust = 1
)
}
\arguments{
\item{data}{list or dataframe}

\item{nbins}{optional, either the number of bins to create, or a specific method from the \emph{tab_nbins()} function. Default is "sturges"}

\item{bins}{optional list of tuples, optional}

\item{incl_lower}{optional boolean, to include the lower bound, otherwise the upper bound is included. Default is True}

\item{adjust}{optional value to add  or subtract to guarantee all scores will fit in a bin}
}
\value{
dataframe with:

\item{lower bound}
\item{upper bound}
\item{frequency}
\item{frequency density}
}
\description{
Bins data and creates a frequency table with frequency density.
}
\seealso{
\code{\link{tab_nbins}}, to determine the number of bins
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}
}
