% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cohen_d.R
\name{th_cohen_d}
\alias{th_cohen_d}
\title{Rules of Thumb for Cohen d}
\usage{
th_cohen_d(d, qual = "sawilowsky")
}
\arguments{
\item{d}{the Cohen d value}

\item{qual}{c("cohen", "sawilowsky", "lovakov", "rosenthal") optional the rule of thumb to be used}
}
\value{
dataframe with the qualification and source
}
\description{
Rules of Thumb for Cohen d
}
\details{
The following rules-of-thumb can be used:

"cohen" => Cohen (1988, p. 40)\tabular{ll}{
   |d| \tab Interpretation \cr
   0.00 < 0.20 \tab negligible \cr
   0.20 < 0.50 \tab small \cr
   0.50 < 0.80 \tab medium \cr
   0.80 or more \tab large \cr
}


"sawilowsky" => Sawilowsky (2009, p. 599)\tabular{ll}{
   |d| \tab Interpretation \cr
   0.00 < 0.10 \tab negligible \cr
   0.10 < 0.20 \tab very small \cr
   0.20 < 0.50 \tab small \cr
   0.50 < 0.80 \tab medium \cr
   0.80 < 1.20 \tab large \cr
   1.20 < 2.00 \tab very large \cr
   2.00 or more \tab huge \cr
}


"lovakov" => Lovakov and Agadullina (2021, p. 501)\tabular{ll}{
   |d| \tab Interpretation \cr
   0.00 < 0.15 \tab negligible \cr
   0.15 < 0.36 \tab small \cr
   0.36 < 0.65 \tab medium \cr
   0.65 or more \tab large \cr
}


"rosenthal" => Rosenthal (1996, p. 45)\tabular{ll}{
   |d| \tab Interpretation \cr
   0.00 < 0.20 \tab negligible \cr
   0.20 < 0.50 \tab small \cr
   0.50 < 0.80 \tab medium \cr
   0.80 < 1.30 \tab large \cr
   0.80 or more \tab very large \cr
}
}
\examples{
 
th_cohen_d(0.34)

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed.). L. Erlbaum Associates.

Lovakov, A., & Agadullina, E. R. (2021). Empirically derived guidelines for effect size interpretation in social psychology. \emph{European Journal of Social Psychology, 51}(3), 485–504. https://doi.org/10.1002/ejsp.2752

Rosenthal, J. A. (1996). Qualitative descriptors of strength of association and effect size. \emph{Journal of Social Service Research, 21}(4), 37–59. https://doi.org/10.1300/J079v21n04_02

Sawilowsky, S. (2009). New effect size rules of thumb. \emph{Journal of Modern Applied Statistical Methods, 8}(2). https://doi.org/10.22237/jmasm/1257035100
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
