% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cohen_g.R
\name{th_cohen_g}
\alias{th_cohen_g}
\title{Rule-of-Thumb for Cohen g}
\usage{
th_cohen_g(g, qual = "cohen")
}
\arguments{
\item{g}{the Cohen g value}

\item{qual}{optional setting for which rule of thumb to use. Currently only "cohen"}
}
\value{
dataframe with the qualification and source
}
\description{
Rule-of-Thumb for Cohen g
}
\details{
Cohen's rule of thumb for Cohen g (1988, pp. 147–149):\tabular{ll}{
   |g| \tab Interpretation \cr
   0.00 < 0.05 \tab negligible \cr
   0.05 < 0.15 \tab small \cr
   0.15 < 0.25 \tab medium \cr
   0.25 or more \tab large \cr
}
}
\examples{
g <- -0.2391304
th_cohen_g(g)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
