% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_cohen_w.R
\name{th_cohen_w}
\alias{th_cohen_w}
\title{Rule-of-Thumb for Cohen w}
\usage{
th_cohen_w(w, qual = "cohen")
}
\arguments{
\item{w}{the Cohen w value}

\item{qual}{optional setting for which rule of thumb to use. Currently only 'cohen'}
}
\value{
dataframe with the qualification and source
}
\description{
Rule-of-Thumb for Cohen w
}
\details{
Cohen's rule of thumb for Cohen w (1988, p. 227):\tabular{ll}{
   |w| \tab Interpretation \cr
   0.00 < 0.10 \tab negligible \cr
   0.10 < 0.30 \tab small \cr
   0.30 < 0.50 \tab medium \cr
   0.50 or more \tab large \cr
}
}
\examples{
w <- -0.2391304
th_cohen_w(w)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
