% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_rank_biserial.R
\name{th_rank_biserial}
\alias{th_rank_biserial}
\title{Rule-of-Thumb for Rank Biserial Correlation}
\usage{
th_rank_biserial(rb, qual = "cohen")
}
\arguments{
\item{rb}{the rank-biserial correlation value}

\item{qual}{optional setting for which rule of thumb to use. Currently only 'cohen'}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for the Rank Biserial Correlation.
}
\details{
Cohen's rule of thumb for rank-biserial correlation (1988, p. 82):\tabular{ll}{
   |r_b| \tab Interpretation \cr
   0.000 < 0.125 \tab negligible \cr
   0.125 < 0.304 \tab small \cr
   0.304 < 0.465 \tab medium \cr
   0.465 or more \tab large \cr
}
}
\examples{
es = 0.6
th_rank_biserial(es)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). L. Erlbaum Associates.
}
\seealso{
\code{\link{es_cohen_w}}, to determine Cohen w
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
