% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumb_vda.R
\name{th_vda}
\alias{th_vda}
\title{Rule-of-Thumb for Vargha-Delaney A}
\usage{
th_vda(a, qual = "vargha")
}
\arguments{
\item{a}{the Vargha-Delaney A value}

\item{qual}{optional setting for which rule of thumb to use. Currently only "vargha"}
}
\value{
A dataframe with:
\item{classification}{the qualification of the effect size}
\item{reference}{a reference for the rule of thumb used}
}
\description{
Simple function to use a rule-of-thumb for Vargha-Delaney A
}
\details{
The following rules-of-thumb can be used:

\emph{"vargha"} => Uses Vargha and Delaney (2000, p. 106):\tabular{ll}{
   \|0.5 - A\| \tab Interpretation \cr
   0.00 < 0.06 \tab negligible \cr
   0.06 < 0.14 \tab small \cr
   0.14 < 0.21 \tab medium \cr
   0.21 or more \tab large \cr
}
}
\references{
Vargha, A., & Delaney, H. D. (2000). A critique and improvement of the CL common language effect size statistics of McGraw and Wong. \emph{Journal of Educational and Behavioral Statistics, 25}(2), 101–132. doi:10.3102/10769986025002101
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
