% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fligner_policello.R
\name{ts_fligner_policello}
\alias{ts_fligner_policello}
\title{Fligner-Policello Test}
\usage{
ts_fligner_policello(
  catField,
  ordField,
  categories = NULL,
  levels = NULL,
  ties = TRUE,
  cc = FALSE
)
}
\arguments{
\item{catField}{A vector with the scores data}

\item{ordField}{A vector with the group data}

\item{categories}{optional vector with categories to use and order for the categorical field. Otherwise the first two found will be used.}

\item{levels}{optional vector with the labels of the ordinal field in order.}

\item{ties}{boolean to indicate the use of a ties correction. Default is TRUE}

\item{cc}{boolean to indicate the use of a continuity correction. Default is FALSE}
}
\value{
A dataframe with:
\item{n}{the sample size}
\item{statistic}{test statistic}
\item{p-value}{significance (p-value)}
\item{test}{description of the test used}
}
\description{
An alternative for the more famous Mann-Whitney U test. The MWU test has as an assumption that the scores in the two categories have the same shape and have unequal variances (Fong & Huang, 2019). The Fligner-Policello test does not, although the distribution around their medians should be symmetric in the population Zaiontz (n.d.).

Roughly put the assumption for this test is that the two categories have the same median in the population.
}
\details{
The formula used is:
\deqn{z = \frac{N_Y - N_X}{2\times\sqrt{SS_X + SS_Y - M_X\times M_Y}}}
With:
\deqn{SS_X = \sum_{x\in X} \left(N_X - M_X\right)^2, SS_Y = \sum_{y\in Y} \left(N_Y - M_Y\right)^2}
\deqn{M_X = \frac{N_X}{n_x}, M_Y = \frac{N_Y}{n_y}}
\deqn{N_X = \sum_{x \in X} N\left(x\right), N_Y = \sum_{y \in Y} N\left(y\right)}
\deqn{N\left(y\right) = \sum_{x\in X} f\left(y, x\right)}
\deqn{N\left(x\right) = \sum_{y\in Y} f\left(x, y\right)}
\deqn{f\left(a, b\right) = \begin{cases} 1 & \text{ if } a> b \\ 0 & \text{ if } a\leq b \end{cases}}

In case of a tie correction (Hollander et al., 2014, p. 146):
\deqn{z = \frac{\left|N_Y - N_X\right| - 0.5}{2\times\sqrt{SS_X + SS_Y - M_X\times M_Y}}}
\deqn{f\left(a, b\right) = \begin{cases} 1 & \text{ if } a> b \\ 0.5 & \text{ if } a = b \\ 0 & \text{ if } a\leq b \end{cases}}

\emph{Symbols used:}
\itemize{
\item \eqn{X} the scores in the first category
\item \eqn{Y} the scores in the second category
\item \eqn{n_i} the number of scores in the i category
}

The test is described by Fligner and Policello (1981), and can also be found in Kloke and McKean (2015, p. 68)
}
\section{Before, After and Alternatives}{

Before running the test you might first want to get an impression using a cross table using \code{\link{tab_cross}}, or a stacked bar chart using \code{\link{vi_bar_stacked_multiple}} for a visualisation.

After the test you might want an effect size measure, either \code{\link{es_common_language_is}} for the CLES, \code{\link{es_hodges_lehmann_is}} for Hodges-Lehmann, or \code{\link{r_rank_biserial_is}} for the (Glass) rank biserial (Cliff delta).

Independent samples tests for a binary vs ordinal variable include Brunner-Munzel (\code{\link{ts_brunner_munzel}}), C-square (\code{\link{ts_c_square}}), Cliff-Delta (\code{\link{ts_cliff_delta_is}}), Fligner-Policello (\code{\link{ts_fligner_policello}}), Mann-Whitney U (\code{\link{ts_mann_whitney}}), Mood-Median (\code{\link{ts_mood_median}})
}

\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
myLevels = c('Not scientific at all', 'Not too scientific', 'Pretty scientific', 'Very scientific')
ts_fligner_policello(df1[['sex']], df1[['accntsci']], levels = myLevels)
ts_fligner_policello(df1[['sex']], df1[['accntsci']], levels = myLevels, ties= FALSE, cc=TRUE)
ts_fligner_policello(df1[['sex']], df1[['accntsci']], levels = myLevels, ties= TRUE, cc=FALSE)

#Example 2: vectors
binary = c("apple", "apple", "apple", "peer", "peer", "peer", "peer")
ordinal = c(4, 3, 1, 6, 5, 7, 2)
ts_fligner_policello(binary, ordinal, categories=c("peer", "apple"))

}
\references{
Fligner, M. A., & Policello, G. E. (1981). Robust rank procedures for the Behrens-Fisher problem. \emph{Journal of the American Statistical Association, 76}(373), 162–168. https://doi.org/10.1080/01621459.1981.10477623

Hollander, M., Wolfe, D. A., & Chicken, E. (2014). \emph{Nonparametric statistical methods} (3rd ed.). John Wiley & Sons, Inc.

Kloke, J., & McKean, J. W. (2015). \emph{Nonparametric statistical methods using R}. CRC Press, Taylor & Francis.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
