% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mann_whitney.R
\name{ts_mann_whitney}
\alias{ts_mann_whitney}
\title{Mann-Whitney U Test}
\usage{
ts_mann_whitney(
  catField,
  ordField,
  categories = NULL,
  levels = NULL,
  method = "exact",
  cc = TRUE
)
}
\arguments{
\item{catField}{A vector or dataframe with the group data}

\item{ordField}{A vector or dataframe with the scores data}

\item{categories}{: optional list with the two categories to use from catField. If not set the first two found will be used}

\item{levels}{optional list with the scores in order}

\item{method}{c("exact", "appr") exact method or normal approximation}

\item{cc}{boolean to indicate the use of a continuity correction}
}
\value{
A dataframe with:
\item{n}{the sample size}
\item{U1}{the Mann-Whitney U score of the first category}
\item{U2}{the Mann-Whitney U score of the second category}
\item{statistic}{test statistic}
\item{pValue}{significance (p-value)}
\item{test}{description of the test used}
}
\description{
The Mann-Whitney U and Wilcoxon Rank Sum test are the same. Mann and Whitney simply expanded on the ideas from Wilcoxon.

The test will compare the distribution of ranks between two categories. The assumption is that the two categories have the same mean rank (which often is stated simplified as having the same median in the population).

The function is shown in this \href{https://youtu.be/VwnR7f1sU8w}{YouTube video} and the test is also described at \href{https://peterstatistics.com/Terms/Tests/MannWhitneyUtest.html}{PeterStatistics.com}
}
\details{
The formula used is (Mann & Whitney, 1947, p. 51):
\deqn{U_i = R_i - \frac{n_i\times\left(n_i + 1\right)}{2}}
With:
\deqn{R_i = \sum_{j=1}^{n_i} r_{i,j}}

For an approximation the following is used:
\deqn{sig. = 2\times\left(1 - Z\left(z\right)\right)}
With:
\deqn{z = \frac{U_i - \frac{n_1\times n_2}{2}}{SE}}
\deqn{SE = \sqrt{\frac{n_1\times n_2}{n\times\left(n - 1\right)}\times\left(\frac{n^3 - n}{12} - \sum_i T_i\right)}}
\deqn{T_i = \frac{t_i^3 - t_i}{12}}
\deqn{n = n_1 + n_2}

If a continuity correction is used the z-value is calculated using:
\deqn{z_{cc} = z - \frac{0.5}{SE}}

\emph{Symbols used:}
\itemize{
\item \eqn{n_i} the sample size of category i
\item \eqn{n} the total sample size
\item \eqn{r_{i,j}} the j-th rank of category i
}

The ties correction (\eqn{T}) can be found in Lehmann and D'Abrera (1975, p. 20)

For the exact distribution the Mann-Whitney-Wilcoxon distribution is used, from the \strong{pwilcox()} function from R.

Wilcoxon (1945) had developed this test earlier for the case when both categories have the same sample size, and Mann and Whitney expanded on this.
}
\section{Before, After and Alternatives}{

Before running the test you might first want to get an impression using a cross table using \code{\link{tab_cross}}, or a stacked bar chart using \code{\link{vi_bar_stacked_multiple}} for a visualisation.

After the test you might want an effect size measure, either \code{\link{es_common_language_is}} for the CLES, \code{\link{me_hodges_lehmann_is}} for Hodges-Lehmann, or \code{\link{r_rank_biserial_is}} for the (Glass) rank biserial (Cliff delta).

Independent samples tests for a binary vs ordinal variable include Brunner-Munzel (\code{\link{ts_brunner_munzel}}), C-square (\code{\link{ts_c_square}}), Cliff-Delta (\code{\link{ts_cliff_delta_is}}), Fligner-Policello (\code{\link{ts_fligner_policello}}), Mann-Whitney U (\code{\link{ts_mann_whitney}}), Mood-Median (\code{\link{ts_mood_median}})
}

\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
myLevels = c('Not scientific at all', 'Not too scientific', 'Pretty scientific', 'Very scientific')
ts_mann_whitney(df1['sex'], df1['accntsci'], levels=myLevels)

#Example 2: vectors
binary = c("apple", "apple", "apple", "peer", "peer", "peer", "peer")
ordinal = c(4, 3, 1, 6, 5, 7, 2)
ts_mann_whitney(binary, ordinal, categories=c("peer", "apple"))

}
\references{
Lehmann, E. L., & D'Abrera, H. J. M. (1975). \emph{Nonparametrics: Statistical methods based on ranks}. Holden-Day.

Mann, H. B., & Whitney, D. R. (1947). On a Test of Whether one of Two Random Variables is Stochastically Larger than the Other. \emph{The Annals of Mathematical Statistics, 18}(1), 50-60. https://doi.org/10.1214/aoms/1177730491

Wilcoxon, F. (1945). Individual comparisons by ranking methods. \emph{Biometrics Bulletin, 1}(6), 80. https://doi.org/10.2307/3001968
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
