% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mcnemar_bowker.R
\name{ts_mcnemar_bowker}
\alias{ts_mcnemar_bowker}
\title{McNemar-Bowker Test}
\usage{
ts_mcnemar_bowker(nom1, nom2)
}
\arguments{
\item{nom1}{the scores on the first variable}

\item{nom2}{the scores on the second variable}
}
\value{
dataframe with the test statistic, degrees of freedom, and p-value (sig.)
}
\description{
McNemar-Bowker Test
}
\details{
The formula used is (Bowker, 1948, p. 573):
\deqn{\chi_{B}^2 = \sum_{i=1}^{r-1}\sum_{j=i+1}^c \frac{\left(F_{i,j} - F_{j,i}\right)^2}{F_{i,j} + F_{j,i}}}
\deqn{df = \frac{r \times \left(r - 1\right)}{2} = \frac{c \times \left(c - 1\right)}{2}}
\deqn{sig. = 1 - \chi^2\left(\chi_{B}\right)}

\emph{Symbols used}
\itemize{
\item \eqn{r} is the number of rows (categories in the first variable)
\item \eqn{c} is the number of columns (categories in the second variable)
\item \eqn{n} is the total number of scores
\item \eqn{F_{i,j}} is the frequency (count) of scores equal to the i-th category in the first variable, and the j-th category in the second.
}

This test is an extension of the McNemar (1947) test, which is only for 2x2 tables
}
\examples{
nom1 = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3)
nom2 = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3)
ts_mcnemar_bowker(nom1, nom2)

}
\references{
Bowker, A. H. (1948). A test for symmetry in contingency tables. \emph{Journal of the American Statistical Association, 43}(244), 572–574. https://doi.org/10.2307/2280710

McNemar, Q. (1947). Note on the sampling error of the difference between correlated proportions or percentages. \emph{Psychometrika, 12}(2), 153–157. https://doi.org/10.1007/BF02295996
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
