% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sign_os.R
\name{ts_sign_os}
\alias{ts_sign_os}
\title{one-sample sign test}
\usage{
ts_sign_os(data, levels = NULL, mu = NULL)
}
\arguments{
\item{data}{A vector with the data as numbers}

\item{mu}{optional hypothesized median, otherwise the midrange will be used}
}
\value{
dataframe with mu, the significance (p-value) and test used
}
\description{
one-sample sign test
}
\details{
The test statistic is calculated using (Stewart, 1941, p. 236):
\deqn{p = 2\times B\left(n, \text{min}\left(n_+, n_-\right), \frac{1}{2}\right)}

\emph{Symbols used:}
\itemize{
\item \eqn{B\left(\dots\right)} is the binomial cumulative distribution function
\item \eqn{n} is the number of cases
\item \eqn{n_+} is the number of cases above the hypothesized median
\item \eqn{n_-} is the number of cases below the hypothesized median
\item \eqn{min} is the minimum value of the two values
}

The test is described in Stewart (1941), although there are earlier uses.
The paired version for example was already described by Arbuthnott (1710)

\strong{Alternatives}

\emph{library(BSDA)}

SIGN.test()

\emph{library(DescTools)}

SignTest()

\emph{library(nonpar)}

signtest(data, m=2)
}
\examples{
data <- c(1, 2, 5, 1, 1, 5, 3, 1, 5, 1, 1, 5, 1, 1, 3, 3, 3, 4, 2, 4)
ts_sign_os(data)
ts_sign_os(data, mu = 2)

}
\references{
Arbuthnott, J. (1710). An argument for divine providence, taken from the constant regularity observ’d in the births of both sexes. \emph{Philosophical Transactions of the Royal Society of London, 27}(328), 186–190. https://doi.org/10.1098/rstl.1710.0011

Stewart, W. M. (1941). A note on the power of the sign test. \emph{The Annals of Mathematical Statistics, 12}(2), 236–239. https://doi.org/10.1214/aoms/1177731755
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
