% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_student_t_is.R
\name{ts_student_t_is}
\alias{ts_student_t_is}
\title{Student t Test (Independent Samples)}
\usage{
ts_student_t_is(scores, groups)
}
\arguments{
\item{scores}{A vector with the scores data}

\item{groups}{A vector with the group data}
}
\value{
dataframe with the test statistic, degrees of freedom and p-value
}
\description{
Student t Test (Independent Samples)
}
\details{
The formula used is:
\deqn{t = \frac{\bar{x}_1 - \bar{x}_2}{SE}}
\deqn{df = n_1 + n_2 - 2}
\deqn{sig. = 2\times\left(1 - T\left(\left|t\right|, df\right)\right)}

With:
\deqn{SE = s_p\times\sqrt{\frac{1}{n_1} + \frac{1}{n_2}}}
\deqn{s_p = \sqrt{\frac{\left(n_1 - 1\right)\times s_1^2 + \left(n_2 - 1\right)\times s_2^2}{df}}}
\deqn{s_i^2 = \frac{\sum_{j=1}^{n_i} \left(x_{i,j} - \bar{x}_i\right)^2}{n_i - 1}}
\deqn{\bar{x}_i = \frac{\sum_{j=1}^{n_i} x_{i,j}}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{i,j}} the j-th score in category i
\item \eqn{n_i} the number of scores in category i
}
}
\examples{
scores = c(20,50,80,15,40,85,30,45,70,60, NA, 90,25,40,70,65, NA, 70,98,40)
groups = c("national","international","international","national","international", "international","national","national","international","international","international","international","international","international","national", "international" ,NA,"national","international","international")
ts_student_t_is(scores, groups)

}
\references{
Student. (1908). The probable error of a mean. \emph{Biometrika, 6}(1), 1–25. https://doi.org/10.1093/biomet/6.1.1
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
