% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_welch_t_is.R
\name{ts_welch_t_is}
\alias{ts_welch_t_is}
\title{Welch t Test (Independent Samples)}
\usage{
ts_welch_t_is(scores, groups)
}
\arguments{
\item{scores}{A vector with the scores data}

\item{groups}{A vector with the group data}
}
\value{
dataframe with the test statistic, degrees of freedom and p-value
}
\description{
Welch t Test (Independent Samples)
}
\details{
The formula used is:
\deqn{t = \frac{\bar{x}_1 - \bar{x}_2}{SE}}
\deqn{df = \frac{SE^4}{\frac{\left(s_1^2\right)^2}{n_1^2\times\left(n_1 - 1\right)} + \frac{\left(s_2^2\right)^2}{n_2^2\times\left(n_2 - 1\right)}}}
\deqn{sig. = 2\times\left(1 - T\left(\left|t\right|, df\right)\right)}

With:
\deqn{SE = \sqrt{\frac{s_1^2}{n_1} + \frac{s_2^2}{n_2}}}
\deqn{s_i^2 = \frac{\sum_{j=1}^{n_i} \left(x_{i,j} - \bar{x}_i\right)^2}{n_i - 1}}
\deqn{\bar{x}_i = \frac{\sum_{j=1}^{n_i} x_{i,j}}{n_i}}

\emph{Symbols used:}
\itemize{
\item \eqn{x_{i,j}} the j-th score in category i
\item \eqn{n_i} the number of scores in category i
}
}
\examples{
scores = c(20,50,80,15,40,85,30,45,70,60, NA, 90,25,40,70,65, NA, 70,98,40)
groups = c("national","international","international","national","international", "international","national","national","international","international","international","international","international","international","national", "international" ,NA,"national","international","international")
ts_welch_t_is(scores, groups)

}
\references{
Welch, B. L. (1947). The generalization of `Student’s’ problem when several different population variances are involved. \emph{Biometrika, 34}(1/2), 28–35. https://doi.org/10.2307/2332510
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
