% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wilcoxon_ps.R
\name{ts_wilcoxon_ps}
\alias{ts_wilcoxon_ps}
\title{Paired Samples Wilcoxon Signed Rank Test}
\usage{
ts_wilcoxon_ps(
  ord1,
  ord2,
  appr = c("wilcoxon", "exact", "imant", "imanz"),
  noDiff = c("wilcoxon", "zsplit", "pratt"),
  ties = TRUE,
  cc = FALSE
)
}
\arguments{
\item{ord1}{the scores on the first variable}

\item{ord2}{the scores on the second variable}

\item{appr}{c("wilcoxon", "exact", "imanz", "imant") optional which method to use for approximation (default is "wilcoxon")}

\item{noDiff}{c("wilcoxon", "pratt", "zsplit") optional method to deal with scores equal on both variables (default is "wilcoxon")}

\item{ties}{optional boolean to use a tie correction (default is True)}

\item{cc}{optional boolean to use a continuity correction (default is False)}
}
\value{
A dataframe with:
\item{W}{the Wilcoxon W}
\item{statistic}{the test statistic}
\item{df}{degrees of freedom (only applicable for Iman t approximation)}
\item{pValue}{the significance (p-value)}
\item{testUsed}{the test that was used}
}
\description{
Paired Samples Wilcoxon Signed Rank Test
}
\details{
The unadjusted test statistic is given by:
\deqn{W=\min\left(W_{neg}, W_{pos}\right)}
With:
\deqn{W_{pos} = \sum_{i=1}^n \begin{cases} r_i & \text{ if } d_i > 0 \\ 0 & \text{ if } d_i \le 0 \end{cases}}
\deqn{W_{neg} = \sum_{i=1}^n \begin{cases} r_i & \text{ if } d_i < 0 \\ 0 & \text{ if } d_i \ge 0 \end{cases}}
\deqn{d_i = x_i - y_i}

\emph{Symbols used:}
\itemize{
\item \eqn{n} the number of scores (equal for each variable)
\item \eqn{x_i} the i-th score on the first variable
\item \eqn{y_i} the i-th score on the second variable
\item \eqn{r_i} the i-th rank of the absolute differences (\eqn{d_i})
\item \eqn{W_{pos}} is the number of ranks with a positive difference
\item \eqn{W_{neg}} is the number of ranks with a negative difference
}

The distribution and test for \eqn{W} can now be performed the same way as
for the one-sample case. See \emph{ts_wilcoxon_os()} for details on the calculations.
The \eqn{d_i} scores are now the one-sample, and the hypothesized median would be 0.

\strong{Alternatives}

*R's stats library *

wilcox.test(ord1, ord2, paired=TRUE, exact=FALSE, correct=TRUE)

wilcox.test(ord1, ord2, paired=TRUE, exact=FALSE, correct=FALSE)

\emph{library(coin)}

wilcoxsign_test(ord1 ~ ord2, zero.method = "Wilcoxon")

wilcoxsign_test(ord1 ~ ord2, zero.method = "Pratt")
}
\examples{
ord1 = c(50,45,33,22,99,79,4,36,62,51,27,15,26,83,86)
ord2 = c(47,45,31,24,78,76,13,46,45,44,23,14,34,79,81)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="wilcoxon", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="wilcoxon", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="wilcoxon", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="wilcoxon", ties = FALSE, cc = TRUE)

ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="pratt", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="pratt", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="pratt", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="pratt", ties = FALSE, cc = TRUE)

ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="zsplit", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="zsplit", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="zsplit", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="wilcoxon", noDiff="zsplit", ties = FALSE, cc = TRUE)

ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="wilcoxon", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="wilcoxon", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="wilcoxon", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="wilcoxon", ties = FALSE, cc = TRUE)

ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="pratt", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="pratt", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="pratt", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="pratt", ties = FALSE, cc = TRUE)

ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="zsplit", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="zsplit", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="zsplit", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imanz", noDiff="zsplit", ties = FALSE, cc = TRUE)

ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="wilcoxon", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="wilcoxon", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="wilcoxon", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="wilcoxon", ties = FALSE, cc = TRUE)

ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="pratt", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="pratt", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="pratt", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="pratt", ties = FALSE, cc = TRUE)

ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="zsplit", ties = TRUE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="zsplit", ties = TRUE, cc = TRUE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="zsplit", ties = FALSE, cc = FALSE)
ts_wilcoxon_ps(ord1, ord2, appr="imant", noDiff="zsplit", ties = FALSE, cc = TRUE)

ord1 = c(5, 8, 6, 3, 9, 9, 18, 13, 16)
ord2 = c(4, 11, 8, 7, 3, 16, 5, 18, 7)
ts_wilcoxon_mp(ord1, ord2, appr="exact")

}
\references{
Wilcoxon, F. (1945). Individual comparisons by ranking methods. \emph{Biometrics Bulletin, 1}(6), 80. https://doi.org/10.2307/3001968
}
\author{
P. Stikker

Please visit: https://PeterStatistics.com

YouTube channel: https://www.youtube.com/stikpet
}
