% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_bar_dual_axis.R
\name{vi_bar_dual_axis}
\alias{vi_bar_dual_axis}
\title{Dual-Axis Bar Chart}
\usage{
vi_bar_dual_axis(data, varname = NULL)
}
\arguments{
\item{data}{the data from which to create a Pareto chart}

\item{varname}{a name for the data, if not provided the name of the data variable is used}
}
\value{
a chart in the plot window
}
\description{
A dual axis bar-chart is a bar-chart with two vertical axis. In this function it will
show both the count and cumulative proportion.

This chart could be used with a single ordinal variable.
}
\section{Alternatives}{

It is also possible to do this with the library \emph{ggplot2}.
A video on how to use ggplot2 for this can be found \href{https://youtu.be/EhK9K9jwsco}{here}
}

\examples{
#Example 1: dataframe
dataFile = "https://peterstatistics.com/Packages/ExampleData/GSS2012a.csv"
df1 <- read.csv(dataFile, sep=",", na.strings=c("", "NA"))
ex1 = df1['mar1']
vi_bar_dual_axis(ex1);
vi_bar_dual_axis(ex1, varname="marital status");

}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
