% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_boxplot_single.R
\name{vi_boxplot_single}
\alias{vi_boxplot_single}
\title{Box (and Whisker) Plot}
\usage{
vi_boxplot_single(data, varname = NULL)
}
\arguments{
\item{data}{list or dataframe}

\item{varname}{optional name to display on vertical axis}
}
\value{
boxplot
}
\description{
A box plot is a little more complex visualisation than a histogram. It shows the five quartiles (e.g. minimum, 1st quartile, median, 3rd quartile, and maximum). It can also be adjusted to show so-called outliers.
}
\details{
This was actually a 'range chart' (Spear, 1952, p. 166) but somehow it is these days referred to as a box-and-whisker plot as named by Tukey (1977, p. 39)

The function uses the \strong{boxplot()} function from the \emph{graphics} library. If you want to modify more things you might want to use that function.
}
\examples{
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
df2 = read.csv(file2, sep=';', na.strings=c("", "NA"))
#Example 1: dataframe
ex1 = df2['Gen_Age']
vi_boxplot_single(ex1);

#Example 2: Numeric list
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
vi_boxplot_single(ex2);

}
\references{
Spear, M. E. (1952). \emph{Charting statistics}. McGraw-Hill.

Tukey, J. W. (1977). \emph{Exploratory data analysis}. Addison-Wesley Pub. Co.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
