% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_histogram.R
\name{vi_histogram}
\alias{vi_histogram}
\title{Histogram}
\usage{
vi_histogram(data, show = "count", density = "auto", xlbl = NULL, ...)
}
\arguments{
\item{data}{list or dataframe}

\item{show}{c('count', 'relative'), show either counts or relative count on vertical scale}

\item{density}{c('auto', FALSE, TRUE), show (relative) frequency or use (relative) frequency density.}

\item{xlbl}{string, optional label for horizontal axis (default is variable name of data)}

\item{...}{other parameters for use in plot function}
}
\value{
The histogram
}
\description{
A histogram is a bit like a bar chart for a scale variable. You would create some bins, and then plot these as bars.

This function is shown in this \href{https://youtu.be/v13UlQvbOvs}{YouTube video} and the visualisation is described at \href{https://peterstatistics.com/Terms/Visualisations/histogram.html}{PeterStatistics.com}
}
\details{
This function is just using some defaults for the \strong{hist()} function from R's \emph{graphics} library.

To set the bins, the \emph{breaks} argument can be used. This could be a pre-set number based on a calculation, a specific rule (e.g. bins="sturges"), or a list with the cut-off points.

If your bins are of equal width, a true histogram than actually should show frequency densities (Pearson, 1895, p. 399). These are the frequencies divided by the bin-width. This can be done using \emph{density=TRUE} parameter.
}
\section{Before, After and Alternatives}{

Before this you might want to create a binned frequency table
\code{\link{tab_frequency_bins}}, to create a binned frequency table.

After this you might want some descriptive measures:
\code{\link{me_mode_bin}}, for Mode for Binned Data.
\code{\link{me_mean}}, for different types of mean.
\code{\link{me_variation}}, for different Measures of Quantitative Variation.

Or a perform a test:
\code{\link{ts_student_t_os}}, for One-Sample Student t-Test.
\code{\link{ts_trimmed_mean_os}}, for One-Sample Trimmed (Yuen or Yuen-Welch) Mean Test.
\code{\link{ts_z_os}}, for One-Sample Z Test.

Alternative Visualisations:
\code{\link{vi_boxplot_single}}, for a Box (and Whisker) Plot.
\code{\link{vi_stem_and_leaf}}, for a Stem-and-Leaf Display.
}

\references{
Pearson, K. (1895). Contributions to the mathematical theory of evolution. II. Skew variation in homogeneous material. \emph{Philosophical Transactions of the Royal Society of London. (A.)}, 186, 343-414. doi:10.1098/rsta.1895.0010
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
