% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_histogram_b2b.R
\name{vi_histogram_b2b}
\alias{vi_histogram_b2b}
\title{Back-to-Back Histogram}
\usage{
vi_histogram_b2b(
  catField,
  scaleField,
  categories = NULL,
  bins = NULL,
  show = "count",
  density = "auto",
  xlbl = NULL,
  rotate = TRUE,
  colors = c("orange", "blue"),
  ...
)
}
\arguments{
\item{catField}{list or dataframe with the categories}

\item{scaleField}{list or dataframe with the scores}

\item{categories}{optional list with the two categories to use from bin_field. If not set the first two found will be used}

\item{bins}{optional either a list with parameters to pass to tab_frequency_bins, a dataframe with upper and lower bounds, or simply NULL (default)}

\item{show}{c('count', 'relative'), show either counts or relative count on vertical scale}

\item{density}{c('auto', FALSE, TRUE), show (relative) frequency or use (relative) frequency density.}

\item{xlbl}{: string, optional label for the field axis, if not set the name of the field is used.}

\item{rotate}{: bool, optional rotate the bars so they appear horizontal. Default is True}

\item{colors}{vector, two colors, one for each category, default is c(rgb(0, 0, 1, 1/4), rgb(1, 0, 0, 1/4))}

\item{...}{optional additional parameters to pass to geom_histogram}
}
\value{
The back-to-back histogram
}
\description{
This function creates a simple back-to-back histogram. This is sometimes also referred to as a Pyramid chart or a Dual-Sided histogram (Jelen, 2005).

The back-to-back histogram together with back-to-back stem-and-leaf and split box-plots are described as effective ways to compare two distributions by Lane and Sándor (2009).

The function is shown in this \href{https://youtu.be/nJuel_XSCdo}{YouTube video} and the visualisation is also described at \href{https://peterstatistics.com/Terms/Visualisations/histogram.html}{PeterStatistics.com}

See \strong{vi_histogram()} for more details on histograms.
}
\references{
Jelen, B. (2005, December 24). Dual Sided Histogram in Excel. MrExcel. https://www.mrexcel.com/tech-tv/dual-sided-histogram-in-excel/

Lane, D. M., & Sándor, A. (2009). Designing better graphs by including distributional information and integrating words, numbers, and images. \emph{Psychological Methods, 14}(3), 239–257. doi:10.1037/a0016620
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
