% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_histogram_b2b.R
\name{vi_histogram_b2b}
\alias{vi_histogram_b2b}
\title{Back-to-Back Histogram}
\usage{
vi_histogram_b2b(
  catField,
  scaleField,
  categories = NULL,
  bins = NULL,
  equal_bins = TRUE,
  density = FALSE,
  ...
)
}
\arguments{
\item{catField}{list or dataframe with the categories}

\item{scaleField}{list or dataframe with the scores}

\item{categories}{optional list with the two categories to use from bin_field. If not set the first two found will be used}

\item{bins}{optional list with upper bounds of bins to be used, or any of the pre-set options from hist()}

\item{equal_bins}{optional boolean. use the same bins for each sample}

\item{density}{optional boolean. use frequency density instead of counts}

\item{...}{other parameters for use in geom_histogram function}
}
\value{
The back-to-back histogram
}
\description{
This function creates a simple back-to-back histogram. This is sometimes also referred to as a Pyramid chart or a Dual-Sided histogram (Jelen, 2005).

The back-to-back histogram together with back-to-back stem-and-leaf and split box-plots are described as effective ways to compare two distributions by Lane and Sándor (2009).

The visualisation is also described at \href{https://peterstatistics.com/Terms/Visualisations/histogram.html}{PeterStatistics.com}

See \strong{vi_histogram()} for more details on histograms.
}
\references{
Jelen, B. (2005, December 24). Dual Sided Histogram in Excel. MrExcel. https://www.mrexcel.com/tech-tv/dual-sided-histogram-in-excel/

Lane, D. M., & Sándor, A. (2009). Designing better graphs by including distributional information and integrating words, numbers, and images. \emph{Psychological Methods, 14}(3), 239–257. doi:10.1037/a0016620
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
