% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_stem_and_leaf.R
\name{vi_stem_and_leaf}
\alias{vi_stem_and_leaf}
\title{Stem-and-Leaf Display}
\usage{
vi_stem_and_leaf(data, key = NULL)
}
\arguments{
\item{data}{list with the numeric data}

\item{key}{optional factor to use for the stems}
}
\value{
prints the display in console and returns a dataframe with the stems and leafs
}
\description{
A stem-and-leaf display is defined as: "a method of displaying data in which each observation is split into two parts labelled the ‘stem’ and the ‘leaf’" (Everitt, 2004, p. 362). A diagram that could be used to visualize scale variables, created by Tukey (1972, p. 296).

In some variations of this, the cumulative frequencies are also shown, but currently this function does not provide for that.

This function is shown in this \href{https://youtu.be/u4XOYJz-aNE}{YouTube video} and the visualisation is described at \href{https://peterstatistics.com/Terms/Visualisations/stemAndLeafDisplay.html}{PeterStatistics.com}
}
\section{Before, After and Alternatives}{

Before this you might want to create a binned frequency table
\code{\link{tab_frequency_bins}}, to create a binned frequency table.

After this you might want some descriptive measures:
\code{\link{me_mode_bin}}, for Mode for Binned Data.
\code{\link{me_mean}}, for different types of mean.
\code{\link{me_variation}}, for different Measures of Quantitative Variation.

Or a perform a test:
\code{\link{ts_student_t_os}}, for One-Sample Student t-Test.
\code{\link{ts_trimmed_mean_os}}, for One-Sample Trimmed (Yuen or Yuen-Welch) Mean Test.
\code{\link{ts_z_os}}, for One-Sample Z Test.

Alternative Visualisations:
\code{\link{vi_boxplot_single}}, for a Box (and Whisker) Plot.
\code{\link{vi_histogram}}, for a Histogram.
}

\examples{
file2 = 'https://peterstatistics.com/Packages/ExampleData/StudentStatistics.csv'
studentDf = read.csv(file2, sep=';', na.strings=c("", "NA"))
# Example 1: dataframe
ex1 = studentDf[['Gen_Age']]
vi_stem_and_leaf(ex1);

# Example 2: Numeric list
ex2 = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5)
vi_stem_and_leaf(ex2);

}
\references{
Everitt, B. (2004). \emph{The Cambridge dictionary of statistics} (2nd ed.). Cambridge University Press.

Tukey, J. W. (1972). Some graphic and semigraphic displays. In T. A. Bancroft & S. A. Brown (Eds.), \emph{Statistical Papers in Honor of George W. Snedecor} (pp. 293–316). Iowa State University Press.
}
\author{
P. Stikker. \href{https://PeterStatistics.com}{Companion Website}, \href{https://www.youtube.com/stikpet}{YouTube Channel}, \href{https://www.patreon.com/bePatron?u=19398076}{Patreon donations}
}
